Four Fitting-Functions that can be used as an input of fitfun.SP argument to obtain the solution paths for the SPSP algorithm. The users can also customize a function to generate the solution paths. As long as the customized function take arguments x, y, family, standardize, and intercept, and return an object of class glmnet, lars (or SCAD, MCP in the future).

Four Fitting-Functions that can be used as an input of fitfun.SP argument to obtain the solution paths for the SPSP algorithm. The users can also customize a function to generate the solution paths. As long as the customized function take arguments x, y, family, standardize, and intercept, and return an object of class glmnet, lars (or SCAD, MCP in the future).

lasso.glmnet uses lasso selection from glmnet.

adalasso.glmnet the function to conduct the adaptive lasso selection using the lambda.1se from cross-validation lasso method to obtain initial coefficients. It uses package glmnet.

adalassoCV.glmnet adaptive lasso selection using the lambda.1se from cross-validation adaptive lasso method to obtain initial coefficients. It uses package glmnet.

ridge.glmnet uses ridge regression to obtain the solution path.

lasso.lars uses lasso selection in lars to obtain the solution path.

Usage

lasso.glmnet(x, y, family, standardize, intercept, ...)

adalasso.glmnet(x, y, family, standardize, intercept, ...)

adalassoCV.glmnet(x, y, family, standardize, intercept, ...)

ridge.glmnet(x, y, family, standardize, intercept, ...)

lasso.lars(x, y, family, standardize, intercept, ...)

Arguments

x

a matrix of the independent variables. The dimensions are (nobs) and (nvars); each row is an observation vector.

y

Response variable. Quantitative for family="gaussian" or family="poisson" (non-negative counts). For family="binomial" should be either a factor with two levels.

family

Response type. Either a character string representing one of the built-in families, or else a glm() family object.

standardize

logical argument. Should conduct standardization before the estimation? Default is TRUE.

intercept

logical. If x is a data.frame, this argument determines if the resulting model matrix should contain a separate intercept or not. Default is TRUE.

...

Additional optional arguments.

Value

An object of class "glmnet" is returned to provide solution paths for the SPSP algorithm. An object of class "glmnet" is returned to provide solution paths for the SPSP algorithm. An object of class "glmnet" is returned to provide solution paths for the SPSP algorithm. An object of class "glmnet" is returned to provide solution paths for the SPSP algorithm. An object of class "lars" is returned to provide solution paths for the SPSP algorithm.