A high dimensional dataset with n equals to 200 and p equals to 500.

A dataset with 200 observations and 500 dimensions is generated from the following process: linear regression model with only first three non-zero coefficients equal to 3, 2, and 1.5 respectively. The covariates are correlated with AR structure (rho=0.3). The error term is normally distributed with zero mean and sd equals to 0.5.

data(HighDim)

Examples

# HighDim dataset is generated from the following process:
n <- 200; p <- 500; sigma <- 0.5
beta <- rep(0, p); nonzero <- c(1, 2, 3); zero <- setdiff(1:p, nonzero)
beta[nonzero] <- c(3, 2, 1.5)
Sigma <- 0.3^(abs(outer(1:p,1:p,"-")))
library(MASS)
X <- mvrnorm(n, rep(0,p), Sigma)
error <- rnorm(n, 0, sigma)

X <- apply(X, 2, scale) * sqrt(n)/sqrt(n-1)
error <- error - mean(error)

Y <- X %*% beta + error
HighDim <- data.frame(Y, X)
head(HighDim)
#>           Y         X1         X2         X3           X4         X5         X6
#> 1  1.231731  2.2242640 -0.9717483 -1.9172081  0.886508947  0.3149891  1.5775887
#> 2  6.205267  1.9726453  0.4125698 -0.1086922 -0.197173483 -1.1493503  0.6965587
#> 3 -1.891698 -1.5467757  0.8551281  0.7653365  2.295717739 -0.9950162 -0.4310555
#> 4  1.403851 -0.4669693  0.8516669  0.5733587 -1.161523069 -1.2059007  0.2618885
#> 5 -4.196336 -0.4822487 -1.1673326 -0.2174506  0.447216964 -0.3511484 -0.3796366
#> 6  8.511040  1.5970207  1.1364553  0.6705758  0.009940064 -0.1251793 -1.2137649
#>           X7          X8         X9        X10        X11        X12
#> 1 -0.4249186  0.36610471  0.5681776 -1.3438444 -1.5439637  0.5415930
#> 2  0.9430580  0.79091890  1.1791096  0.7055706 -1.0582287  0.8627606
#> 3 -0.6906046 -0.07504468 -1.0236781 -0.5929731 -0.7362851 -1.7369246
#> 4 -0.3102682 -1.00481741  0.0725473  0.6223741  0.4685574  0.8877164
#> 5  0.7581888  0.50386505  2.1270388 -1.2760473 -0.4769631 -0.0529032
#> 6 -0.9093267  0.38837884  0.4347974  1.7577909  0.5887000  0.9574163
#>           X13         X14        X15          X16         X17          X18
#> 1 -0.94182726 -2.21863522 -2.3198930 -0.007619012 -0.38573176 -0.002135633
#> 2 -0.60199390 -0.12543711 -2.3293448  0.184445857 -0.14571933  0.410943068
#> 3  0.08520095 -0.16012738 -0.5559131 -2.035180416 -0.43050226  0.282582634
#> 4  0.91811695 -0.61984475  0.5453336  1.321242292 -0.27510500 -0.106084236
#> 5 -0.30803898  1.24956127  0.2599081 -0.142546938  0.08823354 -1.537246869
#> 6  0.51004581  0.07723188  0.9522927 -0.574601617 -1.55353211 -1.016328397
#>          X19         X20         X21        X22         X23        X24
#> 1 -0.8973722  0.09196979 -1.12520140 -0.3232023 -0.32062251 -2.0905281
#> 2 -1.1722638  0.73819707  0.06254948  0.2088387 -1.11844590  1.3797362
#> 3  0.1289385 -0.17071497 -0.02430450 -1.3726372 -1.09532379 -1.8303950
#> 4  0.7114437  0.66141170  0.02144238  0.7362930 -0.03900712  0.6415109
#> 5 -1.4981049 -1.23353698  0.96795522 -0.8171765  0.61805516 -0.3394902
#> 6  0.8571707 -1.01623485  0.46420796  1.3034536  0.56983598 -1.1579757
#>          X25         X26         X27        X28         X29        X30
#> 1 -0.6781616 -0.06376075 -0.48683052 -0.3489853  1.05189824 -0.7559336
#> 2  0.2666887  1.75887231 -0.04100925  0.8255365  0.74640088 -0.6044855
#> 3 -3.3793722 -1.28305621 -2.01393487 -0.4927416 -0.33341740 -0.1863699
#> 4  1.3031717  1.04339056  2.62906321  0.3064477  0.28676341  1.2554209
#> 5  0.7288556 -0.19225956 -0.55555252  0.3073821  0.08205796 -2.2309271
#> 6 -0.1715092 -0.50787257 -0.15450918 -0.3192300 -0.02876546  0.4409376
#>          X31        X32         X33        X34        X35        X36        X37
#> 1 -0.4516093 -0.5264192  0.79050648 -0.9067472 -0.8512358 -1.1085759 -0.6531913
#> 2 -0.1588938 -0.2670745 -0.39055954 -0.4047924 -0.7716019 -0.7568968 -1.7088148
#> 3  0.3971079 -1.0393757 -0.83366490  1.1885053 -0.5064691 -1.1217765 -1.4144748
#> 4  0.1163412 -0.6193406 -1.29379004 -1.0230109 -0.8943140  1.6862907  2.1677381
#> 5 -0.5488772  0.2659501  0.04336562 -1.1112368  0.5641700  1.3210881  0.3570081
#> 6  0.7258721 -1.1199499 -0.43778750  1.6618952 -0.5424865  0.2078410  0.8602175
#>          X38        X39        X40        X41        X42         X43        X44
#> 1  0.4161551  2.1434654  1.1441011 -1.0353828  1.2424519 -1.32917464 -2.1981934
#> 2  0.7096222 -1.5191134 -0.5994505 -0.5319437  0.2651967 -0.54907280  1.7959288
#> 3 -1.0086870 -1.0977642  0.1758804  1.9259746  0.8393086  0.12710386 -1.2518719
#> 4  0.2707099 -1.1252872 -1.1583144 -0.5695171  0.6599533  1.64999132  0.6375881
#> 5 -1.8235875  0.5241888 -0.3406572 -0.2052045 -0.8682224  0.13131192  0.0247909
#> 6 -0.3852214 -0.1407929  0.4353806  0.1856386  0.7799259 -0.02602255  2.0130448
#>           X45        X46         X47        X48         X49         X50
#> 1 -0.07124362  0.5660141  1.42666760  0.8361223  0.08378314  1.25105703
#> 2 -0.92154433  0.6708122 -0.67782914 -0.5353898 -0.79669102 -1.68688950
#> 3  1.25356543 -0.8612511  1.31850312 -1.4739905 -0.20486370  0.30812637
#> 4  0.71314463  0.7332826 -1.08870215  1.2868228 -1.01061627 -0.02881078
#> 5  1.69202131 -1.8047396 -0.05604502  0.2029707  1.51811642  1.44155145
#> 6  0.12845656 -0.5380072 -1.01224278 -1.3171443 -0.05071174 -1.50403304
#>           X51         X52         X53        X54         X55        X56
#> 1 -0.80440571 -0.77594928  0.56269485  0.6703816 -0.03389403  0.3264154
#> 2  0.42887389  0.44505813  0.03488502 -1.7302921  0.53155595 -0.1011064
#> 3 -0.72973789 -1.56138368 -0.81653897 -1.6285452 -1.16017010 -0.8995793
#> 4  0.05705414  2.22900286  0.68075579 -0.1094033 -0.55385706 -0.8140026
#> 5 -0.95916487 -1.38589001 -0.79865104  0.4265803 -1.24142728  0.1780090
#> 6 -1.14074967 -0.04064272  0.13563138  1.0643053  1.22161622  0.4233738
#>           X57          X58         X59         X60         X61        X62
#> 1  0.24126023  0.741435671  1.04193729 -0.54765519  0.86460418 -1.3115782
#> 2 -0.02307907 -0.156091363  0.55362635  0.04946237 -1.34358234  0.2276912
#> 3 -0.91372011 -0.297621630 -1.00506426 -0.49396946  1.00598506 -0.3308822
#> 4 -0.54587944  0.009367841  0.25776364  0.43451579  0.02647806 -0.3882433
#> 5 -0.04204621  0.145283496  0.05338012 -0.53583897  0.65189819  1.3202504
#> 6 -1.31775003 -0.029771689 -0.03147656  0.87651813 -1.34224311 -0.5773862
#>          X63        X64        X65        X66        X67         X68        X69
#> 1  0.1073052 -0.5590633  0.7608648  0.6667398 -0.3677362 -0.85717712  0.4288088
#> 2 -0.6914165  1.2902853 -0.9627273 -0.5550902  0.3186271 -0.43005275  2.1429596
#> 3 -0.6000434 -0.3108479 -0.8322178  0.1415949  0.4814085  0.36625081  1.2445008
#> 4 -1.8647995 -1.5635418 -0.8007465 -1.4568426 -0.7629376 -0.64640168 -2.1329841
#> 5  1.1734191  0.2779450  0.2812951  0.1654208 -2.0959402 -0.49500676 -0.6914869
#> 6 -1.7634135 -2.5055385 -1.7691857 -0.5468504 -1.1326422  0.02746729 -0.0488516
#>           X70        X71        X72        X73        X74        X75        X76
#> 1  0.09198306  1.6120917  1.5357365  1.7242758  0.7310965 -0.1394002 -0.5726909
#> 2  0.22002701 -0.6742727 -1.8283578 -1.2580752 -1.0242637 -1.4388449 -0.3359192
#> 3  2.19799260  0.8774381 -0.5428685 -0.1931627  1.4388019  0.1809847 -0.9724335
#> 4 -1.22125249 -0.7477778  1.5779998  0.8786932  0.9954962  0.2824587  0.8177665
#> 5  0.45797081  1.0118777 -0.7349732 -0.5022708  0.3229667 -0.7265777  0.3711609
#> 6  0.25588762 -0.7042605  1.3994144  1.4452469  0.4763606  0.2721659  0.1878445
#>           X77         X78         X79         X80         X81        X82
#> 1  0.25976556 -0.54855129  0.66539469 -0.87627588 -0.03848065  0.4921307
#> 2 -0.08779222 -0.03603916 -1.95886161 -0.08477027  0.51819597  1.0547069
#> 3  1.18919912 -1.03512148 -0.02712222 -1.33485317  0.84737829  0.8818746
#> 4  0.05723590  0.19546717  1.16181337 -0.69377878 -0.70086885  0.0993519
#> 5  0.56680899 -0.28588166 -1.09764091 -0.44847513 -1.62889826 -1.0863192
#> 6  0.04016741 -0.06925041 -0.19625898  0.59635164  0.80802355  0.6150150
#>           X83          X84         X85        X86        X87        X88
#> 1  0.39481248 -0.935261616 -0.28530659  0.8358765 -1.3689332 -0.3051490
#> 2  0.85926933 -1.365740406  0.32145319 -1.1294575 -0.4264498  1.1807854
#> 3  0.05185716  0.305641312  2.01942262 -0.2986404  1.1437987 -1.2792634
#> 4 -0.54655880  0.008509939 -1.20483732 -0.2205067 -0.5055729  1.0704216
#> 5 -0.61406418 -2.518080407 -1.81474529 -0.5345623  0.6729512  1.3935184
#> 6 -0.95591176  0.314599271 -0.02077581  0.6186535  0.4417060  0.5733147
#>          X89        X90          X91        X92         X93        X94
#> 1  0.2958469 -1.2732061  0.715179961 -0.1525909 -1.01148124  1.6848627
#> 2 -0.1088503  0.9491989  1.380175389  0.8637619  0.24041499  0.5255384
#> 3 -0.9339842  0.3558937  0.006740514  1.8243484  1.11470292 -0.8011211
#> 4  0.7245976  1.1129771 -0.553204221 -0.5767044 -2.10912037 -2.4061090
#> 5  2.0182390  0.2993799 -0.561792712  1.0824436 -0.75029417  1.2340042
#> 6  1.2367745  2.0734630 -0.294977564 -0.2691502  0.02962269 -0.5074858
#>           X95        X96         X97         X98        X99       X100
#> 1  0.09313243 0.01685716  0.82205767 -1.09721999 -0.5878475 -1.5432156
#> 2  1.36461474 1.11570337  0.71969687  0.50645556  0.9691558  0.3949414
#> 3 -0.62795362 0.37126102  0.06982468 -0.61002633 -0.9725509  1.9198221
#> 4 -0.39948763 0.13698111 -1.90911970 -1.63709252  0.8932136  2.0590025
#> 5  0.31618785 0.32549738 -1.06698086  0.32060138  1.1727437  1.0350823
#> 6 -1.15004913 0.17075552  0.27401081  0.01363978  0.1603119  1.7545650
#>         X101       X102        X103        X104        X105       X106
#> 1 -0.6589583 -0.3291567  0.24708145  0.05644638  1.09699830 -0.3044732
#> 2 -0.6975058 -0.1808723 -0.51615260 -1.44404570 -1.21596821  0.1889406
#> 3  1.9598901  0.8548672  0.02321504  0.71768676  0.01856243 -1.1016096
#> 4  1.2143239 -1.1757909 -1.58935393  0.74575947  0.67835454 -0.6517423
#> 5 -0.6453078 -1.6876304  0.06546302 -0.08220588 -0.30429871 -1.0767225
#> 6  1.4670890  1.1848970 -0.47431221 -0.70252835  0.85860562  1.4526684
#>         X107       X108       X109        X110       X111       X112
#> 1  0.5077762  0.4965733 -1.1112884  1.14334272  0.8335019  1.0075011
#> 2 -0.2251987 -0.5899875  0.8746572 -0.80813604 -1.6017426 -2.8246538
#> 3  0.5509432  0.7405574  1.1239506  0.58477778 -0.6331548 -0.4580814
#> 4 -0.6607408 -0.7168775  0.2579126 -0.04198743 -0.6613706  0.1935634
#> 5  0.2416682  1.0412735  1.6366639 -0.79182984 -0.5895886  0.2324182
#> 6  1.8126423  0.6320376 -0.5912453 -0.61617919  0.9856388  1.4913117
#>          X113         X114        X115        X116       X117        X118
#> 1 -0.01505369 -0.871215249 -0.28652664 -0.24045321 -0.5802121 -1.54013245
#> 2 -2.47411611 -0.448394820 -0.93467117  0.68571863 -0.6063078 -2.47193136
#> 3  0.63908490  1.116149846  0.08475766 -0.68721015 -0.7078886 -0.02696291
#> 4  0.12909649 -1.054123832  0.04026084 -0.08199274 -0.4632966  0.15373700
#> 5 -0.39962153 -0.820981711 -0.95965057  1.13977923  0.5480702  0.56739115
#> 6  0.63873702 -0.002250663 -0.29442978  1.47303270  0.4913816 -0.14423627
#>         X119        X120       X121       X122       X123        X124
#> 1  0.1612545  0.73348029  3.1648429  0.9320113  0.5822983 -0.48610604
#> 2  0.3131302 -0.68389898 -0.5728650 -0.8832644 -1.2436099 -0.36853318
#> 3 -0.3922286 -1.83373489 -2.5563128 -2.1111293 -2.7515490  0.19154595
#> 4 -0.2106024  0.05200849 -0.1824971  0.7689517  0.5231774 -0.10173370
#> 5  0.6455878  2.50893002  1.0427956  0.6429293  0.3468256 -0.03073328
#> 6 -0.1277472 -1.46851123  0.2759155  0.9010464  2.1985169  0.44394703
#>          X125       X126      X127        X128        X129       X130
#> 1 -0.40543754  1.1474259 1.3948981  0.90692260  1.38473494 -0.2662050
#> 2 -0.07484683  0.8375683 0.8405047  0.03597797 -0.05969104 -0.7824867
#> 3 -0.86690573 -1.2000160 0.6790595  0.59372822  0.85414732 -0.8169641
#> 4 -1.36399499  0.8472874 0.3775873 -0.18492705 -0.16671600  1.1506866
#> 5  0.16545364 -0.2378764 0.1951309 -0.85237695 -0.05259284  0.7393936
#> 6  0.11339447  1.1226435 0.9389657  0.07863046  0.13574451  0.9259086
#>          X131        X132       X133         X134       X135       X136
#> 1 -0.09848674 -1.00914175  1.0404573 -1.192100754 -0.9073251  0.4473414
#> 2  0.75611318 -0.03251461  0.4842493  1.722707455  0.8465866 -2.2788383
#> 3 -0.11141945 -0.60633193  0.7899084 -0.581583292  0.4136994  1.2029278
#> 4 -0.64909755  0.87380529 -0.5902933  1.400422764  1.3217864  1.4553390
#> 5  0.10377872  1.20112335 -0.6753456  0.006060217  0.8970759 -0.4747520
#> 6  0.49943491  2.02724948  1.4476187 -0.402908311  1.2071099 -0.8350891
#>         X137        X138       X139        X140       X141       X142
#> 1 -0.2488772  0.85305884 -1.5308590  0.08202657  0.9451145 -0.4680033
#> 2  0.7945097  0.02024568  0.1435399 -0.13201304  1.4864219  0.6098219
#> 3  0.2353851 -0.48859323  0.5402949 -0.50377371 -0.5960718  0.1285989
#> 4 -0.2283690  0.34810293 -1.3091957  0.32183182 -0.1236591  2.5576582
#> 5  0.3252999 -0.62423671 -0.1265981 -0.25646277 -1.3741845 -2.5957222
#> 6 -0.3964920  0.20516824 -0.7161208  0.06556292 -0.2376977 -0.2669738
#>          X143       X144       X145       X146         X147       X148
#> 1 -0.84221702 -0.7313886 -1.2498933 -0.4915673  0.212408538 -0.2796454
#> 2  0.33310356  1.3159974  1.8012852 -0.4181970 -0.385686750  0.2815069
#> 3  0.94682336  0.1249687  0.3583630 -0.9644627 -0.309056982  1.0157137
#> 4 -0.06563247  0.6388370  0.9111558  1.8034505  1.543500407 -1.1515276
#> 5 -1.02173011 -1.1166144 -0.1599759  0.6799330 -0.279038556 -2.0155593
#> 6  1.58432415  0.7890778  0.2709612 -0.3437125 -0.005027825  0.3867639
#>           X149       X150       X151       X152       X153        X154
#> 1  0.320726163  1.5043722  1.6408172 -0.3198641 -1.4033150 -0.90642950
#> 2  1.742769358  1.5355945  1.2704590  2.1060446 -0.7470363  0.04657719
#> 3  0.028078091 -0.5498591 -0.2841361 -0.1892005 -1.7683374 -0.48995944
#> 4  0.008403876 -0.3775684 -0.1083686  0.1202809  1.1025782 -0.23675755
#> 5  0.732535146  0.6892154  1.0129580  2.0848948  0.2039583  0.01397331
#> 6 -0.553860785  0.8695791 -0.4362437  1.3780368  0.5223016  0.17707002
#>           X155       X156       X157        X158       X159       X160
#> 1 -0.200207163 -0.7620355  0.8014141 -0.75065873 -0.9664188 -0.2341509
#> 2  0.501903062 -0.4964064  0.4429347 -0.87662129  0.3091394 -1.0522848
#> 3  0.008298437 -0.8420321  0.8732508  1.56979543 -0.1920161 -2.4439149
#> 4 -1.098952954 -0.7435608 -1.5014116 -1.35963937 -1.2703114 -0.4592027
#> 5 -1.234038172 -0.3528733 -0.1483356  0.08013746 -1.0089100 -0.7221527
#> 6 -0.690280775  1.3489488  0.7357441 -0.12016382  0.4062344 -1.4603484
#>         X161        X162       X163       X164         X165       X166
#> 1 -0.8250809 -1.06030656  1.2590552 -0.5553811 -0.813652005 -1.4183821
#> 2  1.0188075  0.90412877  0.9283886  1.1422148 -1.787814282  0.3019494
#> 3 -1.1265558 -0.08781804 -1.7703920 -0.5276798 -1.533909312 -0.1056704
#> 4  0.2683019  0.60315535 -0.3314949 -0.1438118  0.425386649  0.3735244
#> 5  0.4083276  2.26745696  2.7997228  0.9662279  0.004108954 -1.2240163
#> 6 -2.2890567 -0.90531660 -0.5405862 -1.4806332  0.091187620 -1.4503061
#>          X167       X168        X169        X170       X171        X172
#> 1  1.86575092 -0.8285222  0.05163581 -0.86854733  0.9927492  0.31741732
#> 2 -0.05837239  0.9340636  0.41404568  0.19026031 -0.8441754 -1.27496039
#> 3 -0.80750428 -1.4106172  0.42456612 -0.06124626 -0.9271349 -0.57434365
#> 4 -0.26697847  1.8957028  1.58201524 -0.09003717 -0.7358531  0.60114178
#> 5 -0.56626326 -1.7179406 -0.99221503 -0.40569689  0.1918519 -0.08012571
#> 6 -0.70109409 -0.5118165 -0.13347040  1.80043570  0.7610009 -0.85003288
#>         X173        X174         X175       X176       X177       X178
#> 1  0.4036876  1.31526996  1.590521085 -0.2651257 -1.1414241 -1.7452554
#> 2 -0.7367167 -0.09763451  0.413871708 -0.4931618 -0.7501336 -0.4850525
#> 3 -0.5281798 -1.34032005 -0.595154081  0.3542458  0.5899678  0.3093023
#> 4  1.0825684  0.87010186  0.165174862  1.9604351  0.5916852  0.7534998
#> 5  0.7596467  0.02444454 -0.005569657 -0.8934482 -0.4468906  1.1297033
#> 6  1.3028269  0.85044900  0.555236926  1.3454330  0.1273505 -0.8137215
#>         X179       X180        X181        X182       X183        X184
#> 1 -1.6524917 -2.0489503 -1.81571681  0.50025886  1.4387943 -0.69119846
#> 2  1.3936130 -1.1924188 -0.83376450 -0.24109785  0.2975812  0.57843502
#> 3 -0.4097093 -0.2431783  0.02464259 -1.45686093  0.6038120  0.72020895
#> 4 -0.3994448 -0.8457392 -0.08835897  0.07926127  0.1794236 -1.09569863
#> 5 -0.2201659  1.0339532 -0.53592267 -1.24857802 -0.5252966 -1.68286327
#> 6  0.8124653 -0.3777985 -0.48542692 -1.71521952 -0.2245200  0.06425548
#>          X185       X186        X187        X188        X189       X190
#> 1  0.62658636 -0.6657029 -0.04924837  1.03769835  0.09778741  1.2930786
#> 2  0.47736920 -0.5609529  0.58872287 -0.43240062 -0.77135610  0.2930779
#> 3 -0.03182042  0.6138850 -0.10227013  0.87183066  0.99228245  0.7075231
#> 4 -0.73639131  0.6463371  2.38656268  0.21529976  1.63138915 -0.6302936
#> 5  0.13075539  0.4515670 -0.04877495 -0.01489581  0.88401790  0.6068667
#> 6 -0.97844186 -0.2994434 -0.55224971 -1.44840041 -1.31904852 -1.0387026
#>         X191       X192        X193       X194        X195       X196
#> 1 -1.3804181 -0.5441990  1.68120474  1.3494377 -1.18512314 -2.7105921
#> 2 -0.3788819  1.5723125 -0.63781956  0.0834478  1.84131165 -0.3927456
#> 3 -0.4287359  1.3377399 -0.05298802 -1.0648507 -1.15500929 -0.4158353
#> 4  0.1601673 -0.7612037  0.54477561 -0.8918672  0.65838787 -0.1038153
#> 5 -0.5236287  0.3957964  0.20906769  1.6476769 -0.02053072 -0.9326889
#> 6  1.2072524  1.0407503  1.90052553  0.3504464  0.10569748  0.2199274
#>         X197       X198       X199        X200        X201       X202
#> 1 -2.2466017 -1.3142572 -1.4916704  0.18941240  1.32271927 -1.6515635
#> 2 -0.3073722  1.4255467  0.7597426 -1.51628627 -1.46019476 -1.5155475
#> 3 -0.2840605 -0.4563254 -1.2849345 -0.07356816 -0.01147627  0.9550411
#> 4 -0.1143649  0.7550358 -0.1328533 -0.20151602  0.75421217  0.2171555
#> 5 -0.1591568 -0.5727965 -1.3239760  0.12595105 -1.82956944 -0.7599737
#> 6  1.1370096  1.4286829 -0.1070395  0.10762740  1.10520882  0.1498379
#>         X203        X204       X205       X206       X207       X208
#> 1  0.3686107  0.05199155  0.2759943 -1.5873571 -0.5660908 -1.0219007
#> 2 -1.0859890 -0.70780549  1.7660389  1.4377699  2.0470148 -1.1717577
#> 3 -0.1480125 -0.95908152  0.7435426 -0.3304697  0.4744828 -1.2493432
#> 4 -0.4622009  0.60199826  0.8992688  0.3691631 -0.5950806  0.5826154
#> 5  0.8429307 -0.02644788 -0.8018910  1.4702579  0.5466921 -1.7003060
#> 6  0.2132765 -0.61812058  0.7038619  0.5157051  1.6058027  1.5050068
#>           X209       X210        X211       X212        X213      X214
#> 1  0.009740324 -0.1583615  0.58896266  1.2046458  0.32214740 -1.028133
#> 2 -0.069838115 -0.8053010  0.03114488  0.4877765  0.25123331 -1.030138
#> 3 -0.256278855  0.6669141 -0.05580956 -0.6591600 -0.76738782 -1.053784
#> 4 -0.778794637 -0.2561989 -0.13714086  0.7560375 -0.06290661 -1.339881
#> 5 -0.394616704 -1.8748860 -1.93711528 -0.5006561  0.79917262  0.581731
#> 6 -0.309369905  0.1767364  0.55819426  0.6345520  1.02228204  1.155836
#>         X215       X216       X217        X218        X219       X220
#> 1  0.9804069  1.5106305  0.3541241 -0.43115137  0.26213407 -0.4575319
#> 2  0.3470812  0.9005857  0.2818658 -0.12731332 -0.06995360 -1.7136995
#> 3 -0.0945928  0.4163076  0.4814429  1.26819528  0.03562515 -0.2695178
#> 4 -0.4792530 -0.8523257  0.3433027  0.47890969 -0.07123055 -0.2380835
#> 5 -0.7238729 -1.0925272 -0.8144631 -0.02373767  1.38049892 -0.1222187
#> 6  1.1620944  0.2084242  1.2816643  1.67182510  1.80244799  0.5153799
#>         X221        X222       X223       X224        X225       X226
#> 1 -0.8737185 -1.41655389 -1.0700494  0.4428056  0.60766571  1.2595862
#> 2 -1.0284652  1.46422303  0.3171192 -1.2461488 -1.20003943 -0.5443153
#> 3 -0.1496834 -0.04057334 -0.7777674  0.1220830 -0.58162326  1.0326965
#> 4  1.0559996  1.88599249 -2.3882343 -1.9464066  0.67042117  1.0330945
#> 5  1.5029140 -0.00516436  1.5239016 -0.5365965 -0.08899157 -0.6606082
#> 6 -0.4166218  0.92503895 -0.9065375  0.1034042  0.11760833 -1.3067983
#>         X227       X228        X229       X230       X231        X232
#> 1  1.6979925  1.3838515  0.68088089 -1.4502164  1.0882566  0.38822260
#> 2  1.2044287 -0.6920360  0.32872231  1.5702008 -0.6618968 -1.34382643
#> 3 -0.8500277 -0.8169123 -0.85941699 -0.9535682 -0.5900585 -0.07825340
#> 4  0.7506631  0.2869815 -0.04918026 -0.1572404 -0.1718585  0.34087095
#> 5 -2.0577167 -1.9247629  0.13037623  0.9160114 -0.3789089  0.02486891
#> 6 -1.4313135 -0.9290514  0.57318132 -0.5330030  0.2900347 -1.30723803
#>         X233       X234       X235        X236       X237       X238       X239
#> 1 -0.9825257 -0.5325432  0.3908327 -0.66891792 -0.5982088 -2.9119994 -2.9366480
#> 2 -0.3799223  0.2280749 -0.4806353  0.24931793 -0.6968500  0.1317336  1.2689246
#> 3 -0.5198807  2.4927624  0.8251724  0.45257620 -0.6764584  0.5304473 -0.5479467
#> 4 -0.7947212 -1.3292761  0.3126550 -0.23429515 -1.1279504 -1.4726406 -0.5212690
#> 5  0.7227157  1.4228650  1.0408743  0.01515234  0.3887321  0.7494994  1.4465322
#> 6 -1.4174778 -0.9880085  0.7486920 -1.17068270  0.5697434  0.6404232 -0.3187893
#>         X240        X241       X242       X243       X244        X245
#> 1  0.5613479 -1.17771837  1.6173379  0.7715806 -1.1247399 -0.08118441
#> 2 -0.4185202  0.07998532 -0.6233910  0.5945910 -0.3658858 -0.31857690
#> 3 -1.4691110 -1.45941086 -1.3738852 -0.4056386 -0.4414636 -1.57236172
#> 4 -0.7119451  1.08275410  0.5462053  0.6624797 -0.2014752 -0.56054711
#> 5  0.7484274  1.78173683 -0.2980009  1.1971387 -0.4107532 -1.40222967
#> 6  0.5396164  0.70139622  0.7571806  0.3298992 -1.6372462 -0.89055758
#>         X246        X247       X248        X249        X250       X251
#> 1 -1.5762190 -1.19145664  0.5962736 -0.38260596 -2.30874937 -0.6921582
#> 2 -1.2072677 -0.50637946 -1.1054845 -1.13433841 -0.17874768 -1.1754054
#> 3 -0.3058662  0.02907274  0.9838201 -0.05631997  1.21405549  1.5326112
#> 4  0.2564963  1.99745440 -1.0390855  0.54097882 -0.03264308 -0.6947586
#> 5  0.7392645 -0.71075056  0.7965674  0.61410177  0.64357419  0.4401214
#> 6  0.5832545  1.06743315  1.8016605  0.35812430  0.26414012  0.1935580
#>          X252       X253       X254        X255       X256        X257
#> 1 -0.98890916  0.4910441  0.9260515  0.18782904  0.1473049  0.04999948
#> 2  0.03003175 -0.7361750 -2.3809914 -1.11526073 -1.5485759  1.72184425
#> 3  0.35338831  0.2107511  0.3045041 -0.41345124 -0.2369453 -0.16835680
#> 4 -0.21288252  1.3261868 -0.3354246 -1.74095813 -0.1087515 -1.36947455
#> 5 -0.99771819  0.1207097 -1.3837666 -0.24399717  0.4538367 -1.04820365
#> 6  0.05652064 -0.2273719 -0.5014211 -0.01175445 -0.4685406  0.66220070
#>         X258       X259       X260        X261       X262       X263
#> 1 -0.9403396  1.0732977 -0.1962702  0.30762472  0.9695659 -0.4525661
#> 2  1.0890747 -0.6152549 -0.5684589  1.28128974 -1.6335017 -2.1453240
#> 3 -1.9260336 -0.6275568  1.7165308  0.08008778  0.9959072  1.3180254
#> 4 -0.3910011 -0.3446456  1.5370324 -0.47700689  0.6451554  1.1828200
#> 5 -1.9165150 -0.4691098  0.2675970  1.51685712  1.3782822  1.5631324
#> 6  0.3767804 -1.1669104 -0.3834572  0.19002046  0.1795984 -1.3294421
#>          X264        X265       X266        X267       X268        X269
#> 1 -0.60595419  0.15871519 -0.5410606  1.43091768  0.9046670  0.23707708
#> 2  0.59290894 -0.28071117  1.7282244  0.08566197 -0.7686998  0.40597493
#> 3  1.20680427  0.53882493 -0.0212002 -0.62309873 -0.3328886  1.22306673
#> 4 -0.31483864 -0.24598985  0.0580984  0.09109744  1.1631968  0.01742295
#> 5  0.25095254 -0.06788799 -0.5071356 -0.67911250 -0.2973733 -2.20524824
#> 6 -0.08560288 -0.56188402 -0.8369467 -0.06406936  2.0450470  1.02698078
#>         X270         X271       X272       X273       X274          X275
#> 1 -0.2950760  0.921717367  0.9806351  1.6203492  1.0900025  0.0740853531
#> 2 -0.1076297 -0.664613733 -2.1531358 -2.2508184 -0.8540582  0.0001129259
#> 3  0.7302670 -0.098321187 -0.7513112  0.2682068  1.0150068 -0.9431174017
#> 4  0.7256971 -0.850982545 -1.8932807 -0.7589215 -1.8962347 -0.8571833967
#> 5 -1.0967658 -0.296590876 -0.3970358  0.8406677  1.8334646 -0.9711721618
#> 6 -0.9611426 -0.008229088  0.1183128  0.8626846  0.9968315  0.6389441744
#>         X276       X277       X278       X279       X280       X281        X282
#> 1 -1.4879461 -0.5205016 -0.3321467 -1.1502136 -1.0723320  1.0644327  0.47621076
#> 2 -0.5530185 -0.9614307 -2.1001437 -2.1543783 -0.4681826  0.0331278  0.46204035
#> 3 -2.3139952 -1.1327202  0.2762150 -1.1149690 -1.4877288 -1.2399874  0.07312562
#> 4  0.4521203 -0.2366549  0.1588002 -1.6663003 -2.2152403  1.2538183  1.09464949
#> 5 -0.3757871  0.9103318 -1.3027541  0.9300761  0.2456875 -1.0553113 -0.43556330
#> 6  0.5181162  0.1771232  1.1763566  0.1426907  0.5134893 -0.3754427 -1.38793588
#>         X283       X284       X285       X286       X287       X288       X289
#> 1 -0.4114005 -0.6361792  0.9412756  0.8371748 -1.1576215  0.8090767  1.5769734
#> 2 -0.2786797 -0.2218032 -0.1451769 -0.7804999  0.3857698  0.6293152 -0.1478698
#> 3  0.5412628 -0.1947026 -0.6185939 -0.2313810  0.6253733 -1.0331848 -0.3349219
#> 4 -0.1217139 -1.3937964 -0.3350272  0.2572855  1.7185292  0.9799359  0.8125442
#> 5  1.4874764  0.3493011 -0.7520083 -1.5358549 -1.4200221 -1.3693955  0.2654054
#> 6 -0.6282443  1.4156091 -0.3620061  0.6157258  0.4189200  0.7056710  0.8548724
#>         X290        X291         X292        X293       X294         X295
#> 1  2.1178840  0.87050942  0.093407335 -0.08003218 -0.2279245 -1.117338078
#> 2  0.5278643  0.59093508  0.611014038  1.14713480  0.5227297  0.006391417
#> 3  0.4647945  0.07098946 -0.009371939  1.22217421  0.5564521 -0.160753505
#> 4 -1.5402375 -0.14463356 -0.742056351 -0.43142425 -0.8092002  0.855869471
#> 5  1.5612594  0.82334378 -0.400723464 -0.65604856  2.2617453  1.060540463
#> 6  0.5400587  1.45350504 -0.214436557 -0.31536594  0.9895675  0.339590471
#>          X296       X297        X298         X299       X300       X301
#> 1 -2.20192075 -1.1701202  0.13606482 -0.387851648 -1.8309601 -1.1333241
#> 2 -1.22379720  0.7958707 -0.60141410 -0.289670430 -1.0761916 -0.3489316
#> 3  0.36621123  0.1568247 -0.28027714 -1.428184367  0.2580904 -0.7345910
#> 4  0.04578732  0.2434682  0.29278725  1.178975130  0.1216062  1.3200815
#> 5  0.07460306  0.5023944 -0.09392733  0.007929475 -0.8493926 -0.8322507
#> 6  0.91226868  0.9865272 -1.02151272 -0.568165501 -0.8680835  0.1086494
#>         X302        X303       X304       X305       X306        X307
#> 1 -0.1295555 -0.67486249  1.7071091 -0.4699473 -1.2791561 -0.76953637
#> 2  0.5046963 -1.17054805  0.9505737  0.9352037  1.0765185 -0.06987788
#> 3 -1.0789876  2.10969813 -0.2072914 -1.1414583 -0.1389639 -0.68185593
#> 4  0.3211746 -1.00013822 -1.6158089  0.6569712 -0.8754429  0.94778922
#> 5  0.6390780 -0.26831813 -0.2368617 -2.1613189 -1.0035390 -0.69221528
#> 6  0.1154450  0.09501391 -0.6750834  0.4211466 -1.2743623 -1.56308674
#>         X308         X309       X310       X311        X312       X313
#> 1  0.7554649 -0.613513649 -0.1089068 -0.2301253  0.52138232  2.5799296
#> 2 -0.5432268  1.044358894 -0.2246119  0.2848355 -0.02995538  0.2372344
#> 3 -1.1862121  0.593379619 -0.5847259  0.5748389 -0.63185395  0.3414766
#> 4  1.6129498  0.001883357  2.1348600  1.3742939 -0.67921851 -0.1036589
#> 5 -0.3784765 -0.298189207 -1.4658924 -0.2623992  0.31528818 -0.5529301
#> 6 -0.5328179 -0.660791342  2.1168518  1.2257049  0.87962481 -0.6843275
#>         X314        X315        X316        X317       X318       X319
#> 1  1.8396241 -0.06625529 -0.53079500 -0.55578504 -0.3485270 -2.4149134
#> 2 -2.0169987 -0.41788376  0.01120773 -0.04914114 -0.6685490  1.0039386
#> 3  1.0458516  0.53877984  0.20907642 -1.62610026 -2.0009538 -0.2909028
#> 4 -0.9780435  0.21104979 -0.24625563  0.16335660 -2.3115474 -0.8031062
#> 5  0.6344741  1.14237098  0.03474184 -0.10409057  0.4612139  0.4472314
#> 6  0.6645587 -0.07186736 -0.68995893  2.51727442  0.8295431  1.3139326
#>         X320        X321       X322        X323       X324       X325
#> 1 -2.2146442 -1.76081833 -1.4411519 -0.10370867 -0.3376094 -0.7379486
#> 2  2.4671396  0.74654866  0.3017768 -0.03808528  0.5868532  0.7957194
#> 3 -0.1583614 -0.06892949  0.6809256  2.11681267  1.3957223  0.3937518
#> 4 -1.4583313 -0.89065390 -1.6749429 -1.34194450 -0.9552307 -0.5477856
#> 5  1.3941949  0.04433161 -0.6627509  1.03291867  0.7295116 -0.4989383
#> 6  0.8443646 -1.60978149  0.1241128  0.62116115 -0.2337668 -0.3476382
#>         X326       X327       X328       X329       X330       X331       X332
#> 1 -0.2857327 -0.7515615 -2.6448047 -2.1498556 -0.4549837  1.3260989  0.4390848
#> 2 -0.8228970 -0.9434112 -0.7227815 -0.9309733 -2.0714009 -0.4455404 -0.3642051
#> 3  0.1018520  0.4206586  1.1753261 -1.0504429 -0.9458544  1.2504051  3.4674987
#> 4 -0.8696307 -0.6812939 -1.1518454 -0.5900021 -0.3597455 -0.3091880 -1.6486602
#> 5  1.6669731  1.2383481  0.1082777 -0.1547623  1.2048648  1.9101940  0.6176707
#> 6 -0.8486874  0.3348629 -0.3044659 -0.2960767  0.5369355 -0.4066315  0.1728794
#>         X333       X334       X335        X336       X337        X338
#> 1 -0.1517198 -0.8718593  0.1447423  0.03859174  0.9684291  0.49948204
#> 2  0.3261807 -0.3352177 -0.9038801 -0.52028756  0.3377483 -0.03651622
#> 3  1.2961758 -0.9937618  0.5659594 -0.35858441 -1.9233440 -0.66410747
#> 4 -0.4432119 -1.6040194 -0.6804276  0.22677326  0.1599112 -0.27344775
#> 5 -0.3118507 -0.1568536  0.2347507 -0.31934934 -1.4979632 -0.23232534
#> 6 -0.3314521  1.5591317 -0.1681369  1.43236353  1.2361981  0.34645585
#>          X339       X340        X341        X342       X343       X344
#> 1  0.57864661 -0.5470919 -0.71406193 -0.59543886  1.2296609 -1.3628365
#> 2  0.48947347 -0.5928042  1.36579866  0.34684380  2.0535515  0.2664432
#> 3 -0.24683096 -0.5113775 -2.20895354  0.08505545  1.1410060  1.0801293
#> 4 -0.67236594  1.5802953 -0.69576222 -0.46564796 -1.7536785 -0.6731452
#> 5  0.05233252  0.2657490 -0.05504135  1.74441412  0.4866159  0.6523321
#> 6 -0.24469651 -1.2125589 -0.09356967 -0.43900143  0.8702234  0.4836215
#>         X345        X346       X347       X348       X349        X350
#> 1  0.4246818 -0.30052948 -1.0181244 -0.5801765 -0.5325446 -1.63907253
#> 2 -0.4756938 -2.19988640 -2.0025578 -0.9847654 -1.5112875  0.07709142
#> 3 -1.0910085  0.06864763  0.8478838  0.7521769 -0.9841055 -1.09628144
#> 4  0.9831841  1.88854006 -0.2300141 -0.9491307  0.3068017  0.19624221
#> 5  0.9134145 -0.97857857  0.6476162 -1.0254470 -0.6798179 -1.47476977
#> 6  0.7328751 -0.14298280 -1.1928421 -0.1149980 -0.6409754 -0.43009445
#>         X351       X352       X353        X354         X355       X356
#> 1 -0.9859769 -0.9041005 -0.5655536  0.75055543 -0.165016741 -0.8125843
#> 2 -0.2684969  0.3221462 -0.2678694  1.39983474 -0.480513822 -2.7514221
#> 3  0.6164262 -1.0677844 -0.6846504 -0.19676771  0.954269364  0.0413788
#> 4  1.3988124 -2.4809421 -0.3555018  0.15135034  0.006436724 -0.7975242
#> 5  0.2681790 -0.5327119  0.7475874  0.07407135 -0.243076660 -1.3645101
#> 6  1.7570466  2.9568356  1.4117133  0.62425631 -0.933134779 -0.4000298
#>          X357       X358       X359       X360       X361       X362       X363
#> 1 -1.41438419 -1.1094041 -1.3996694 -1.4630476  0.7706643 -0.6665517  0.7317518
#> 2  0.90629976 -0.3103903  0.3005255  0.9697660 -1.0483664  0.2086444 -0.1708566
#> 3 -0.37175604  0.1967420 -0.5140715 -1.0334055 -0.2275283 -0.1465913  0.4770984
#> 4  0.01991421 -0.8532921  0.7302298  1.6661312  0.9542551 -0.1892339 -0.7395921
#> 5  0.74084843  0.3328452 -0.7715736  0.2827398 -0.6513145 -0.4888270 -1.4760146
#> 6 -0.04635777  0.6829554  2.7153040 -0.6444137 -1.3013746 -1.3471833 -0.3604940
#>          X364        X365        X366        X367       X368        X369
#> 1 -0.07491949  1.08003548 -0.23774255  0.83256813 -0.7520300 -0.65459233
#> 2 -1.14419810  0.16236486  0.07771483  0.34160242 -0.2695568  1.34251399
#> 3 -0.09581861 -0.37001962 -0.68396683 -1.38039860 -0.7034797  0.02236174
#> 4  0.31136437 -0.08194743 -0.64231112  0.73903039 -0.4610919  1.02291026
#> 5 -0.97206059 -1.83551108 -0.31074552  0.05261196  1.0829329 -0.02605987
#> 6 -1.05261295 -0.45474574  0.15884039 -1.81212084 -0.7936022 -1.05674887
#>          X370         X371       X372       X373       X374       X375
#> 1  0.31664039  1.030076122  1.6219062 -0.1559956 -0.3917831  1.1030107
#> 2  0.27026703 -0.115388810 -0.8553267 -0.3710189  0.3529083 -0.5876000
#> 3  0.83595146  0.095504109  1.3369162  1.3211873  0.8792672 -0.8398817
#> 4  0.99192152  2.652454568  1.5592794 -0.8031174 -2.4531691 -0.4227760
#> 5  0.01790027 -0.001000357  0.7500279  0.7398043 -0.7540808 -2.4252322
#> 6 -0.31434925 -0.427587932 -0.5775529 -1.5218727  0.5437818  0.1609044
#>         X376       X377       X378       X379         X380       X381
#> 1  0.5486860 -2.1476017 -0.9381026 -0.9661741 -0.698163234 -0.6723805
#> 2 -1.1122011 -0.5552436 -1.5125125 -1.2063323 -0.278024692  1.1241675
#> 3 -0.5303737  1.1397676 -1.2280586 -0.2777012  0.006302475 -0.2594059
#> 4 -1.0031138  0.3331094  1.9701288  1.8893015 -0.356502273 -1.2313906
#> 5 -0.8286717 -0.4236886 -0.9176849  0.6583191 -0.679237314  0.3783803
#> 6  0.1191457  1.1134207 -0.1259902 -1.0969157 -1.055416120 -1.9493749
#>         X382       X383       X384       X385        X386       X387       X388
#> 1  0.6715532  1.1176762  0.3501384 -1.1255622 -0.19488344 -1.7244768  1.0748322
#> 2  0.6095046  1.4713010  0.5980030 -0.8354251 -0.52394360 -1.0712617 -1.3420548
#> 3  0.2513550 -0.3541116 -0.6369149 -0.7861432 -1.20092914 -0.4083104 -0.3390904
#> 4 -0.3160742 -0.5396520  0.6171274 -1.5525768  1.16004509  1.6729443  2.6213996
#> 5  0.4971466  0.4182315 -0.1318076  0.4633773 -0.02784577  0.3146765 -0.9153178
#> 6 -2.3356903 -2.5573265 -0.4220237 -0.6195463  0.23057665  0.4259582  0.4473803
#>         X389      X390       X391       X392       X393        X394        X395
#> 1 -0.8323601 0.3113357 -1.0849150 -0.6080929  1.1737467  0.88155224  1.96753859
#> 2  0.1629225 1.1709557  1.3141217 -0.9503594 -0.6282559 -0.56569828 -1.14971453
#> 3  0.7826443 0.2964051  0.9350345  0.4879103  1.0181502  1.39211510 -0.02846824
#> 4  1.0559376 1.2335712 -0.3901510 -0.8157539 -1.2886368 -0.52809382 -0.06735146
#> 5 -0.7531465 0.7990813  0.2474213  2.4447503  1.3218069  0.02515632 -0.11002181
#> 6  1.2348844 1.8062879  2.1358684  1.1170281  0.1544100  1.40888525  1.15470295
#>         X396        X397        X398        X399       X400       X401
#> 1  0.6364878  1.09487105  0.44672150  0.47579575 -0.7125989 -2.3296849
#> 2 -0.1050233 -0.02005871  0.06420737 -0.82542206  0.7370463 -0.6954457
#> 3 -1.6160153 -0.27681394 -1.62426048 -0.26984839  1.2160586 -1.3594109
#> 4  0.5750184  0.73134896  0.27812903 -1.19930002  1.3285981  0.5832207
#> 5  0.1170448 -1.93933846 -0.82959130  0.09157703 -0.2020881 -1.3536779
#> 6  0.6887655 -0.79967992 -2.03602520 -1.23355224 -0.5714286 -0.8399855
#>         X402       X403        X404       X405        X406        X407
#> 1 -0.2238537  0.2747720  0.93281704  0.1688604  0.16176536  0.02526382
#> 2 -0.2964936  0.9720433 -1.69982235  0.4621121 -0.25101498 -0.82003440
#> 3 -0.9461777 -1.9538630 -0.71196260 -1.1727493 -0.08308036  1.65486805
#> 4  0.9825259 -0.1603068 -0.97301971  0.9021843 -0.37111157  1.02157135
#> 5  0.7015322  0.4366484 -1.85360372 -0.8644414  0.11723477  0.78200067
#> 6 -0.6481517  1.0382453 -0.07686252 -0.4262007  0.90020810  1.50998995
#>         X408       X409       X410       X411        X412       X413       X414
#> 1  0.5458858  1.2123828  0.2825824 -0.7396926 -0.01279808  1.4671700  1.3615282
#> 2 -1.1567408  0.1267785 -0.4933228  0.1035576 -0.92172320 -1.1503168 -0.7641438
#> 3  1.1984060  0.7053950 -0.7587118 -1.3043675 -0.03691944 -2.6057106 -1.7016818
#> 4 -0.5897232 -1.6424774  0.6518950 -0.3155463 -0.46514103 -0.3490049  0.5253788
#> 5  1.4326685  0.2091840 -2.0951664  0.3483817  0.27306155 -0.4354102 -1.7967971
#> 6 -0.7025066  0.1017188 -1.0401471 -0.2271072 -1.89437494 -1.6858986 -0.2147227
#>         X415        X416       X417       X418       X419       X420
#> 1  0.3029563  1.79461840 -0.7019230 -0.2912567  0.9560404  0.1893585
#> 2  0.8940555  0.65278531  0.3754426  1.7673293  0.4500004  1.4634983
#> 3 -1.1975201  0.84141578 -0.9590328 -0.9442143  0.2195497  0.7221952
#> 4  1.0099750  0.62181859 -0.5357804 -1.8680425  2.0771994  0.1062712
#> 5  0.1844817  0.87015379 -0.3575131 -0.6807627  0.5903926 -1.1197612
#> 6 -0.2410333 -0.02644824  0.9319008 -0.7733402 -1.0239245  0.4274387
#>          X421       X422       X423       X424       X425        X426
#> 1  0.50112468 -0.5199798  0.4191951  0.0831275 -0.8359135  0.98313437
#> 2  0.54831182  1.3450427  0.5242917  1.8261443 -1.0544252 -1.93569406
#> 3 -0.13056855 -0.8647410 -1.3802123  0.4625527 -0.2485553  0.74575288
#> 4  0.80408804  0.1070491 -0.5786786  0.2403973 -0.5912598 -0.04342799
#> 5 -1.29632797 -0.9000919 -0.8516478 -2.0153805 -0.5414077 -0.77583789
#> 6 -0.05939908 -0.9633581 -1.7654583  1.6674514 -0.1371943  0.16284778
#>         X427       X428       X429       X430       X431        X432
#> 1  0.1790127 -0.3749497  1.0095588 -0.1848103 -1.0385936 -1.36872851
#> 2 -0.3883658  0.4929537 -1.4188782  0.9276990 -0.2282638 -1.42455773
#> 3 -0.4569813 -0.2890160  1.3270788  1.9493395 -0.2234410  0.07285737
#> 4 -0.6446805 -1.4906176  1.4418622 -0.6245140  1.0624099  0.56803984
#> 5  0.5575863 -0.3258826  0.4605247  0.6708330  1.1674169 -1.14028793
#> 6  0.6760125 -0.6848342 -0.8889140  0.2305343 -0.7743193  0.68861695
#>          X433         X434       X435        X436       X437       X438
#> 1 -0.06040154  1.609839537  0.9095172 -0.11517792  1.8205220 -0.7856497
#> 2 -0.82991618  0.669530577  0.5385454 -0.04491663  1.4084613  0.9557873
#> 3  0.06455259 -0.495270166  0.5347195 -0.22824321 -1.1660435 -0.7347371
#> 4 -0.28180839 -0.001547957 -0.2033898  0.70104153  1.1883503  0.2118240
#> 5  1.11179942 -0.693931379  0.7825567  0.89047362  0.7729545  1.0921395
#> 6  0.34407924 -0.756241920 -0.7104176 -1.50804407  0.7779602  1.3202890
#>         X439       X440       X441       X442       X443        X444       X445
#> 1  0.5403801 -0.4495166  0.3022358  0.2306449  1.7146510 -0.66313125 0.40473223
#> 2  2.0982946  0.4090651 -0.1653653  0.5354821  0.3789694  1.20131296 1.40654268
#> 3  0.3823880  0.3483427  0.6023216 -0.8010192 -0.8011633  0.06190688 1.64537715
#> 4 -0.2341529  0.7070729  0.2027683  0.2981728  0.8587581 -0.51867612 1.89115641
#> 5  1.0016320 -0.7015509 -0.4144033  0.8720972 -0.7808348 -0.20426915 0.42647826
#> 6  1.5464311  0.2123509  1.1744810  0.4654500  0.4491099 -0.14720066 0.04407636
#>           X446       X447       X448       X449         X450        X451
#> 1 -0.143990036 -0.3907548 -0.3803689 -1.1181957 -0.626348226 -1.77976266
#> 2  0.379080728  0.4236626  1.6535624  0.2829181 -0.005069406  1.45806887
#> 3  1.336578247  0.3940201 -2.7134172 -0.9376679 -1.114948071 -1.30062453
#> 4  0.778113742  1.4043868 -0.8812750  0.3977761  0.483750159 -0.35353632
#> 5 -0.003575775  0.2084897  0.7476252 -0.4319577  0.287384725  0.06233871
#> 6 -0.298369463  1.1864448  0.2348409 -1.3673825 -1.244357082  0.58362602
#>         X452        X453        X454       X455       X456        X457
#> 1 -0.2393898 -1.29247801 -0.88717842 -0.4121703  0.1454338 -1.44257200
#> 2  1.1094205  0.40251475 -0.02802213  0.6649789  0.5344498 -0.16193579
#> 3 -1.0503637  1.38963901 -0.09191220  0.5730023 -0.0381789 -0.83433608
#> 4 -0.1595006 -0.06959075  0.03403521 -1.6078795 -0.2296580  0.01669433
#> 5  0.9550102  0.35597570 -0.39164067  0.9767209 -0.2675973 -0.23783625
#> 6  1.4456938 -0.92562992 -0.35659351 -0.7849868  0.8899357  0.25847582
#>         X458       X459        X460       X461       X462         X463
#> 1 -1.7704599 -0.4713945 -0.18731991  0.4168217  0.3457166 -0.008892267
#> 2 -1.0886195 -0.3990821 -0.34794728  0.7994823  1.3592727  1.337027433
#> 3 -0.5293958 -1.3141528 -1.16752215  0.3236681  1.3628847  0.730642468
#> 4  0.3646950 -0.8952392 -1.64014094  1.0491968  2.2784543 -0.699434310
#> 5 -0.5494211 -0.5672391 -0.70970429 -1.3790683 -0.8612513 -0.434970649
#> 6  1.3013937  0.7659885 -0.06640324  0.4355583  1.9567437  1.559828034
#>          X464        X465        X466       X467       X468       X469
#> 1 -1.09399267  0.19083110  0.73929424  0.2088560 -0.7832494  0.3445355
#> 2  0.55458728  0.77454803  0.81997106  0.1509531  0.8083908  0.3460523
#> 3  0.85673980  1.03308584  1.65581772  0.2372048  0.2721333  0.6887398
#> 4 -1.16989623 -1.09938534 -0.51508971  1.0541827  1.4306470 -1.3852409
#> 5  0.05423217 -0.74228841 -1.69427628 -0.3454695 -2.5047143 -1.8696222
#> 6  0.55293863  0.04403015 -0.05354919  0.1893378  1.1402612 -0.4026354
#>         X470        X471       X472       X473        X474       X475
#> 1 -0.5339870 -1.86934904 -0.1549899 -0.6788673  0.02478029 -1.7916752
#> 2 -0.6859531  0.69261596  0.8737082  0.7571569  1.41755267  0.7991328
#> 3 -0.3750811 -0.03723616  0.7124473 -0.7897932  1.32899352  0.7151591
#> 4 -0.2349242  0.74501047  2.3084115  1.0374221  2.72999344  1.0959275
#> 5 -1.0575289 -0.30626274  0.9318456 -2.0056499 -0.26906640  1.2553372
#> 6 -0.3093295  0.49171409 -1.5241126 -1.8366835 -1.74980199  1.9346987
#>         X476       X477       X478       X479       X480        X481
#> 1 -1.0310632 -1.6035931 -0.9495442  0.3960185  0.9682897  0.31314496
#> 2 -0.4805303  0.4399199 -0.1727029 -1.1237428  0.8525049 -0.03023303
#> 3  0.9468539  0.1351162  2.2290449 -0.2891256 -0.3607669  0.32752764
#> 4 -0.2232211 -0.5203520  0.3763548 -0.1900212  0.2903727 -0.88568947
#> 5  1.2180206  1.6824752  0.4135011 -0.7863797 -0.4400936 -0.14499615
#> 6 -1.6755324 -1.4229710 -1.8146352  0.9857013  0.7732035  0.50436873
#>          X482       X483       X484       X485       X486       X487       X488
#> 1 -0.88587272 -0.4521818 -1.3831202 -0.7479050 -1.4606057  0.5512568  1.9803336
#> 2  0.08676647  0.4661569  0.6067631  0.2198268 -0.6950273 -0.4686857 -0.4571359
#> 3 -0.51618371  0.3454940  2.2344695  0.5902286 -0.9874858 -1.0724852 -0.6476944
#> 4  1.80199270  0.3242946  0.7845263 -1.3506598 -0.5971810 -0.5931219 -1.6723738
#> 5 -0.67953922  0.1705498  1.2133069  0.4201785  0.3192117 -2.5870586 -1.4893139
#> 6 -0.88608890  1.2341399 -0.3207206 -0.8597113 -0.5535381 -2.0180770 -0.6645738
#>         X489       X490       X491       X492       X493       X494       X495
#> 1 -0.3551877 -0.1544249  1.9363195  1.2973811 -0.6209040  1.6263711  1.4502053
#> 2  0.1978957 -0.8922163  0.7334411 -0.3692212  1.8750037  1.4465752 -0.8977225
#> 3  0.1168683  0.7829177 -1.3046724  0.2580584  0.9923054 -0.8774984 -1.2999775
#> 4 -0.9796202 -0.6806968  1.4154297  1.9136576  1.9597549  0.2971967  0.6253316
#> 5 -1.0476493 -1.4399225 -0.8295193  1.1733533 -1.1461937 -1.2470641 -0.9039293
#> 6  0.5305254 -0.6595557 -1.3434185 -0.8553414 -1.7906306 -0.6142223  1.0277695
#>         X496        X497       X498       X499       X500
#> 1  0.4339441  1.33623015  0.4411173 -0.9623490 -0.1176083
#> 2  1.3540772  1.61231696 -1.5407585  0.7553471  0.7770937
#> 3  0.8494918 -0.58025098  0.7233693  0.2402952  0.4582340
#> 4 -0.4796536  0.26886228  0.8139327  1.1461832  0.4295677
#> 5 -0.2503990 -0.01114487 -0.2921242  0.6313194  1.7032363
#> 6  0.1312858  0.53734878  1.4582260  0.6163070 -0.7381629