A high dimensional dataset with n equals to 200 and p equals to 500.

A dataset with 200 observations and 500 dimensions is generated from the following process: linear regression model with only first three non-zero coefficients equal to 3, 2, and 1.5 respectively. The covariates are correlated with AR structure (rho=0.3). The error term is normally distributed with zero mean and sd equals to 0.5.

Usage

data(HighDim)

Examples

# HighDim dataset is generated from the following process:
n <- 200; p <- 500; sigma <- 0.5
beta <- rep(0, p); nonzero <- c(1, 2, 3); zero <- setdiff(1:p, nonzero)
beta[nonzero] <- c(3, 2, 1.5)
Sigma <- 0.3^(abs(outer(1:p,1:p,"-")))
library(MASS)
X <- mvrnorm(n, rep(0,p), Sigma)
error <- rnorm(n, 0, sigma)

X <- apply(X, 2, scale) * sqrt(n)/sqrt(n-1)
error <- error - mean(error)

Y <- X %*% beta + error
HighDim <- data.frame(Y, X)
head(HighDim)
#>            Y         X1          X2          X3         X4          X5
#> 1 -5.2144685 -0.6992240 -1.10693187 -0.42897892  0.8240160 -0.04278301
#> 2  0.3379996 -0.4642745  0.93847232  0.51124151  1.3934988  1.24806265
#> 3 -5.7092227 -0.9446990 -1.63792917  0.08770152  1.2704197 -1.30481456
#> 4 -4.1883117 -0.2448360 -1.91218272  0.27969057 -0.4981043  1.31984882
#> 5  7.8953391  2.4906716  0.05002897  0.18634994 -1.0504947 -2.65869931
#> 6  1.3362787  0.2100373  0.13763145 -0.04469690 -1.3714997 -0.81826957
#>            X6         X7         X8         X9        X10          X11
#> 1  1.21340336 -0.2767008  0.3098810  1.3911223  1.5082470  0.190644890
#> 2 -0.03146763  0.4401973  0.6109438 -1.6410780  0.7129020 -0.183582668
#> 3 -0.57874096 -1.2516198  0.4013279 -0.2038560  0.3037437 -0.227033510
#> 4  1.33296792 -1.3529998  0.1855562  2.1451039  1.0474491  0.965322542
#> 5 -0.91944677 -1.0257919 -0.8420894 -0.2425171 -2.3228093 -0.572319268
#> 6  0.24026219 -0.8082704 -0.4427554 -0.9168057 -0.3450439 -0.006817472
#>           X12        X13         X14        X15        X16         X17
#> 1 -0.70369403 -0.6537557  0.75108663  0.6080218 -0.5531744 -0.36549284
#> 2 -0.01303313 -0.7199550  1.30965323  2.4305587  2.4570159 -0.57136646
#> 3 -2.56074146 -0.1460951 -0.10189081  0.9792404 -0.2424095  0.10834697
#> 4  0.66946923  0.9323505  1.45944678  0.7788307  1.7986455  0.45341763
#> 5  1.79871406  0.7628009  0.01296338 -1.0918406  0.5317240 -0.05505679
#> 6  1.08403367 -0.7503472 -1.13324611 -1.7808183 -0.7566007 -1.11144354
#>          X18         X19        X20        X21        X22        X23
#> 1 -1.3164983 -1.55100903 -0.8048435  0.6077642 -1.3323196  1.0360005
#> 2 -0.9747489 -1.68346563 -1.1225604 -0.3902885 -0.1550368  0.7003986
#> 3  2.1619722  0.04788973  0.7050863  1.1767787  0.5214734 -0.5865307
#> 4 -0.2641882 -0.16884987 -0.9133071 -0.1470683  0.5668455 -2.6704150
#> 5 -0.2550384 -0.94377012 -0.6289882 -0.8440221  0.2483461  1.4388218
#> 6 -1.7391553 -0.37573436  0.8294732  1.8127413  0.2582896 -0.6417039
#>           X24        X25          X26        X27        X28        X29
#> 1  0.60167926  0.9818896  1.711902198  0.5693331 -0.4966702  0.5204416
#> 2 -0.11388636 -3.1996457  0.753298011  0.5233514 -1.6037643 -1.7112433
#> 3 -0.42942075 -0.3288347 -0.684418715  0.0996813 -0.8151328 -0.5133788
#> 4 -0.06724853  0.3142786 -1.144341219 -0.8728443 -0.6032357  0.8751416
#> 5 -0.34026522 -1.9096233  0.274897012 -0.7066244  0.1913990  0.5860746
#> 6  0.46268300  1.2773458  0.001807013 -0.6246586  0.2044777 -0.6174130
#>           X30         X31        X32        X33        X34         X35
#> 1 -0.38953364 -0.07069399 -0.3776508 -2.0908388  0.6840642 -0.03605611
#> 2  0.04614196  0.05248485  1.4553941  0.8833500 -0.3158047  1.38940572
#> 3  0.76055753 -0.79408167  1.4775704  0.7950508 -0.6721185  1.38634925
#> 4  1.63679075  1.18078746  0.9375755 -0.2294180  0.5405558 -0.69170292
#> 5 -0.13538491 -0.37886357 -0.3641337  0.5361837 -0.5147421  0.89419636
#> 6 -0.31081550  1.26865520  1.3352330  1.0394937  0.5800391  0.57804619
#>          X36         X37        X38         X39        X40         X41
#> 1 -0.3639577  0.27390430 -0.4002055 -0.92521665  0.1606853  0.09611333
#> 2  0.6250252  0.02967633 -0.6496811  0.84280708 -0.5262265 -0.61150256
#> 3  1.5972251  2.06375556  1.1198621  0.91165291 -1.5669872 -1.57380397
#> 4 -0.5275017  1.08803330 -0.1402331  0.08654391  0.4947830  1.98977136
#> 5  1.6408441  0.83338224  0.3829829  0.25147343 -1.6803108  1.96943910
#> 6 -0.4080827 -0.22469220 -0.8166515  1.07611399  1.4196984  0.25626426
#>          X42         X43         X44        X45         X46        X47
#> 1  0.4523584  0.40482621  1.40873883  3.1991404 0.723325173  0.3624496
#> 2  0.2536805  1.17101472  0.29158669 -0.3031197 0.002610309 -0.5714298
#> 3 -3.9212966 -2.35383864 -0.73634787 -0.1527040 0.367581345 -0.4083439
#> 4  1.1555723  1.47067031 -0.05960978  1.0047504 0.038486876  0.6819968
#> 5  2.2131562  0.02822312 -0.61003786  1.1157440 0.002099080 -1.2979921
#> 6  0.5298033 -1.12421908 -0.90267521 -0.5043917 0.638719970 -0.1806951
#>          X48        X49         X50         X51        X52        X53
#> 1 -1.9946125 -1.1498880  0.08461021 -0.06919475 -1.0002978 -0.4409936
#> 2 -0.6198597 -1.3167446 -2.43562024 -1.04718000 -1.9139334  0.4860126
#> 3  0.5753457  2.1187977 -1.25304681  0.95229021  1.3140763  0.7250420
#> 4 -0.8586296  0.9229014  0.49218302  0.05676256  0.4310616 -0.9758102
#> 5 -1.2771624 -0.7751230  0.44495439  0.93283132  0.4001345  0.2105761
#> 6 -0.3051128  0.3711021 -0.08088526 -0.43831425  1.4198598 -0.9344872
#>          X54        X55        X56        X57        X58        X59         X60
#> 1 -1.6891480 0.08588834 -0.8779062 -0.7328128  0.7791154  0.7277954 -0.07557065
#> 2  0.6158409 2.37667709  0.7997882  0.7342839 -0.2420239 -0.5592648  0.29182043
#> 3  1.6345725 0.18063412 -1.5459915 -0.3388858  0.3881167 -0.2674385 -1.77950148
#> 4 -0.1830154 0.42017620  0.6020311 -0.2621936 -0.3577581 -0.5658565  1.32208500
#> 5 -0.1955470 0.98170579 -0.3988371 -0.2938541 -0.1519048  0.1206346  1.25883444
#> 6  0.2112603 1.04866323 -1.1558989 -0.3835870  1.3391773  0.7174895  0.79911954
#>          X61        X62         X63         X64         X65         X66
#> 1 -0.8125274  0.9079056 -0.41950421 -0.69425009  0.59999999 -0.67504463
#> 2 -0.3251114 -0.8760096 -0.12451386  0.15556781  0.45662011  0.24937409
#> 3  0.4240294 -0.4227326 -1.99842495 -0.88004780 -2.23689502 -1.32459654
#> 4  1.3262161 -0.3146113 -0.03816725 -0.14281762 -0.21479558 -0.07008852
#> 5 -0.1348990  0.9078244 -0.48462201 -0.18468838 -0.08507428  0.32152086
#> 6  0.1317216  0.3360522  0.80586757  0.05295362 -0.12726741 -1.22316054
#>           X67        X68        X69        X70         X71          X72
#> 1  0.32025933 -0.2354227 -0.4826757 -1.1922669 -0.38860508 -0.264443838
#> 2 -0.18586698 -1.3578688 -2.6249379 -0.6869406 -0.05767076 -0.544293122
#> 3 -0.06477554  0.1480608 -0.2404217  1.1445040  0.75158468 -1.437729208
#> 4  1.71634413 -0.3446815 -0.9725571 -0.1653351 -1.72351909  0.750731449
#> 5 -0.97918334 -1.1403211  0.8081413 -0.2606741  0.57227403 -0.008670859
#> 6 -1.55676840 -2.7715112  0.4143226  0.3836340  1.17981839  0.512782893
#>            X73        X74         X75         X76         X77         X78
#> 1 -0.770378000  1.7171757  0.92344229  0.93718015 -0.03683777  0.01575321
#> 2 -0.666895226 -0.1742834 -1.35098058 -0.19530151 -0.23554664  0.56065512
#> 3 -1.565332853 -1.2761026  0.03470984 -0.88182603  0.90041500  1.65619075
#> 4  0.323358605  0.5510529  0.67991649  0.06127543  1.51029756 -0.75199346
#> 5 -0.389838795 -1.2256675 -1.22828021 -0.70181781  0.04469833 -0.21505064
#> 6 -0.002359446  0.5108424  0.62566208  1.34645441  1.67470454 -0.39592499
#>           X79        X80        X81         X82         X83        X84
#> 1 -1.80272359 -0.9700235 -0.9904611  0.70557503 -0.50532724  0.9680781
#> 2 -0.40441282 -0.2788889  1.0186078 -0.54352448  0.26173586 -1.0923439
#> 3 -0.23282842 -1.6319385  0.1282562 -0.07083815  1.23134511  2.0351206
#> 4  0.03724141  0.9603665  0.6531745 -0.62385848 -0.15168751  0.6330165
#> 5 -0.62464944 -1.2500327 -1.4684570 -1.08307858 -0.41240821  0.4370668
#> 6 -0.39476447 -0.3765418 -0.7677947 -0.17890193  0.03541072 -1.5531772
#>          X85        X86         X87         X88        X89         X90
#> 1  1.6846591  2.0278957  0.00487375  0.23901160 -0.7782002 -0.45972327
#> 2  0.1684525  0.7486722 -1.23477399 -0.50269785 -0.9690629 -1.05820528
#> 3 -0.8202437  0.1361100 -0.66416960 -1.65353378 -1.0907012 -0.13935581
#> 4 -0.3474989 -0.6420140  0.50696701 -0.99043920 -1.2237826 -0.87632985
#> 5  1.4721601  0.2662238 -0.94963579  0.05752729  1.7686167 -0.06169592
#> 6 -0.5555203 -0.2825923  0.54870831 -1.38742688 -0.5228766 -0.45805698
#>           X91        X92         X93        X94         X95        X96
#> 1  0.61790810  0.4111583  1.28382036  1.2472547 -0.03424144  0.9647083
#> 2  0.97767472  0.7062744 -0.11243398 -0.2177337 -0.92959346  1.3663603
#> 3  0.44469637  1.1223437  0.23677618  2.1938844 -1.67958661  0.3469892
#> 4  0.09343849 -0.1679974  1.21356065  0.2073148  0.28067185 -0.1396682
#> 5 -2.23735821 -2.0472249 -2.01252916 -1.2866714 -0.50856813  0.8825031
#> 6 -0.25980849  0.6872912  0.03735988  1.6530925 -0.50800480  1.1274088
#>           X97        X98        X99         X100        X101       X102
#> 1  0.68293490  0.6800507  0.2107024 -0.714782617  1.47015147 -0.5950559
#> 2 -0.94469929  0.0317876 -1.1462692 -0.383064590 -0.93829736 -1.2740775
#> 3  0.32857773  0.1063998  0.2705857 -0.532791569 -0.45705283  0.7568560
#> 4  0.07765325  2.0487188 -0.4575440  0.008913269  1.27835966  1.4192339
#> 5  0.79961754 -0.3102093  0.1153658 -0.854841252 -0.03494727 -1.0216104
#> 6  0.48626935 -0.1492458 -1.6979407 -0.384032130  0.90240219  1.3573754
#>         X103         X104       X105       X106       X107       X108
#> 1  1.5411487  0.007583927  0.2266697 -0.9256921 -0.8524000 -0.5341951
#> 2  1.0855890  1.421246426  1.2244841 -1.5729647 -0.5517459  1.4328146
#> 3  0.2103744  1.429696435  0.3576719 -0.4230437 -0.4995483  0.7517487
#> 4  1.9683707  1.836872954  0.1080235 -1.9546339 -2.0831118 -0.8177438
#> 5 -1.2825157 -0.808514111 -0.2750920  1.4891250  1.2527753  0.8586055
#> 6  0.5551032 -0.895490969  0.5434529  1.2431924  0.5385138 -0.8659265
#>         X109       X110       X111       X112         X113       X114
#> 1 -0.8728007 -0.1547384 -0.8314220 -1.4710165  0.498100353  0.1778940
#> 2  1.3463351  0.5588950 -2.2046914 -1.4050280 -0.959441112 -0.2110394
#> 3  0.3427951  0.5293920 -1.6986258 -1.0798728 -1.046925475 -1.7289421
#> 4  0.3889006  0.8276989 -0.7898691 -0.7338968 -0.199024795  0.7499279
#> 5  1.2070404  0.1723291  0.6673907 -0.1163819 -1.131171183 -0.4047850
#> 6  0.9388199  0.5051034  0.4517472  1.3929159 -0.007347736 -0.7688771
#>           X115      X116       X117        X118        X119       X120
#> 1  1.545944537 2.1314161  0.5273982  1.89248328  0.49741858 -1.0586993
#> 2 -0.506590259 0.7734198  0.6347421  1.08242174  1.30453079  0.5249109
#> 3 -0.001744141 0.7808743  2.2450192  1.03957527 -0.23668189  0.3449096
#> 4  0.065552718 0.1103362  0.8293884 -0.02856608 -0.04309536  0.4786373
#> 5 -0.024928446 2.3428401  0.5386826  0.03303639  0.74353917 -0.8000983
#> 6 -0.186101983 0.1462739 -0.8385394 -0.50332042  0.62549786  1.1403346
#>          X121       X122       X123       X124       X125       X126
#> 1  1.10396530  0.1169832  1.4313218 -0.4472191 -0.5947037  1.3205692
#> 2  0.62755043  0.2920989  0.3743567  0.7796902  0.6597788 -0.3430154
#> 3  1.89068949 -0.4937248 -1.0824477  0.6664446 -0.3444321  0.4063098
#> 4  0.10388036 -0.7318512 -0.4432145 -0.8806781 -0.1304396 -0.7191296
#> 5 -1.07763311 -0.4581638  0.7506521  0.9723029  0.7576018  0.6737663
#> 6 -0.09979851 -0.6013423  0.4030517  0.9855298  0.8660207 -0.3791549
#>           X127       X128       X129       X130        X131        X132
#> 1 -0.853027857  1.3312555  0.1322630  0.1615196 -1.19451260 -0.01973251
#> 2  0.001187442 -1.0974838  0.3133281  1.7017674  0.88512862  0.35619820
#> 3 -1.870916448 -0.2491318  0.6325227 -0.5412661  0.09740491 -1.28187749
#> 4  1.720200294 -0.5120549 -0.4218964 -1.3749159  0.80420348  1.42275533
#> 5 -0.461040915 -1.6171646  0.7156152 -0.2451430 -0.77349758 -0.58576899
#> 6  0.701840110  0.5354080 -0.3279059  1.0868633  0.38321827 -0.35903509
#>          X133       X134       X135       X136       X137       X138
#> 1 -1.23632331  0.6408091  0.9314610  0.4187235 -0.2759131  0.7373551
#> 2  0.77307032  0.8819398 -0.5624340 -0.5356571 -0.1650452 -0.2824833
#> 3 -1.59833194  0.1539318  0.3474567 -0.5451513 -1.6024744 -0.6443640
#> 4  0.45924030  0.9454336  0.6346384 -0.3508439 -1.7376772 -0.9221761
#> 5  0.09241779 -1.5534913 -0.5524640 -1.3470141 -1.1335802 -2.0050215
#> 6 -1.21400887 -0.1234684  0.5272452  0.6215405  0.2966576 -0.2155612
#>          X139        X140       X141        X142        X143        X144
#> 1  0.14532018 -0.01483541 -1.4295506  0.01280594 -1.27581888 -0.37543450
#> 2 -0.09349959 -0.29046678  0.0242747 -0.07668793 -0.94907662 -0.07212669
#> 3 -0.33756913  0.50186493 -0.8542393  0.16818479  0.44285134 -1.24824997
#> 4 -0.12643912 -1.87508376  0.1579369  0.43947201 -0.05811752 -1.49640232
#> 5  0.11731319 -0.45990744  1.2272990  0.81374763  0.51618862 -0.22546692
#> 6  0.77580958 -0.36059275  0.4720569  0.33634172 -0.35055823  1.76030164
#>         X145       X146       X147       X148       X149        X150       X151
#> 1 -0.3583220 -1.6975853  0.2954308 -0.9437296  0.1677236  1.88633382  1.2585858
#> 2 -0.4900857 -1.9778616 -2.2436615 -0.6706149 -0.1284756 -0.97064755  0.6656097
#> 3 -1.5648600  1.0619458  0.2990823  0.8166122  1.3034139  0.04235726 -0.3589363
#> 4 -1.5865807 -0.2492368  0.9081364  0.6574422  0.4263302 -0.09237054 -0.3756140
#> 5 -1.0071637 -1.9197759 -0.8346601 -0.6849633  0.7728035  0.83401669  1.5616595
#> 6  1.4336710  2.0853696 -0.8227465 -0.9635854 -1.3265385 -0.45067969  0.3848731
#>         X152        X153       X154       X155       X156       X157       X158
#> 1  2.6769570  0.51042350  0.6630487 -0.8024139  1.2177842  1.2754594 -0.2057395
#> 2  0.3224271  0.04922508 -1.5084300  0.1829333 -0.6639552 -0.6679540 -0.2511848
#> 3 -1.0558634  1.35456260 -1.0939943 -0.2789334  0.1661766  1.1004892  1.0410116
#> 4  0.4763789  0.14090318  0.9135114  2.1868816 -0.4763325 -0.2528243  0.5164727
#> 5 -0.9181938 -0.59794550 -0.2039135 -0.7796214 -1.0242587 -0.5070987 -0.6996917
#> 6 -0.1068451 -0.78655989 -2.0508952 -0.4064759 -1.4166566 -0.1777516  0.2674557
#>         X159        X160      X161       X162       X163      X164       X165
#> 1  1.2267323  0.10895636  1.549437  1.1250769  0.8819252  1.277566 -0.3168034
#> 2  0.7676509  0.04984201 -1.198103 -0.7320447  0.2550502 -1.821839 -0.5370869
#> 3 -0.3983242  1.19259384  1.437546  0.1019962 -1.0595206 -0.935010 -1.6032387
#> 4  1.5109048  1.44571587  1.963247 -0.9154255 -0.5192576  1.359689  0.7552408
#> 5 -0.1607748 -0.85584415 -1.113636 -1.4764585  1.5193909  1.513101  0.5079188
#> 6 -0.4112910 -0.35931927 -0.567605 -2.1739444  0.5896049  0.677605  0.6174605
#>         X166       X167        X168       X169        X170        X171
#> 1 -2.6041527  0.8159444  0.26160400  1.8640058  0.50982863 -0.68469944
#> 2  0.1231178 -0.8597462  0.54727794  0.8718854 -0.38677353 -1.21436920
#> 3 -0.4734024 -1.0302547  1.60453852  1.8610982 -1.19674246  0.45294252
#> 4  0.2551392  1.2067660 -0.03018905 -0.6983153  0.14462557  1.09948210
#> 5 -0.1117539  1.0521601  0.24284483  0.8923145 -0.24466955  0.07776146
#> 6  0.5585672  1.3743074  1.00387017  0.5627516 -0.04236477 -0.13014786
#>         X172       X173       X174       X175        X176       X177       X178
#> 1 -1.0422351 -1.4077711 -1.2179777  0.8362709  0.69407252 -1.6431567 -0.3320427
#> 2 -1.5531437  3.5327067  1.5050901  0.3751060  1.78952397  2.2729334  0.7681219
#> 3 -0.1378037  0.3967926  1.3187582 -0.6378704  0.57468629  0.8989987  0.8005000
#> 4  1.3494024  2.3423060 -0.0323106 -0.3856434 -1.03677964 -0.9366263 -0.5758114
#> 5 -0.4326799 -0.1888962  0.7121986  0.3788221 -0.04829265  0.8148799  0.3983599
#> 6  0.1911324  0.1632841 -0.6866281  0.3619373 -0.02417657 -0.9319037 -0.3227338
#>          X179       X180       X181       X182       X183       X184
#> 1 -0.31741543  2.0534713  0.7749787 0.58043113  1.2196041  0.3208152
#> 2  0.03071649 -0.3558585 -0.7013893 0.49460101  1.0495156  0.7825131
#> 3  0.04164012 -1.0843185 -0.5925747 0.13465814  0.8170147  0.0331933
#> 4 -0.41744985  0.3828181 -1.4866440 0.69469312 -1.9615887 -0.5983734
#> 5  1.28177369  2.2692999  1.3240128 0.04538412  0.3322240  1.3044230
#> 6 -0.15111224 -1.8997395 -0.4805932 1.09210125 -0.3945602 -1.2271856
#>          X185       X186       X187       X188       X189        X190
#> 1  0.04024469 -0.2317381  1.4538044  0.9765137  0.5952898  0.49078301
#> 2  0.05005687 -1.3578204 -0.2438818 -0.8998788  0.2141096  0.09479316
#> 3  1.36626138  1.4388540  0.5232731 -0.1226652  0.8719242  1.42866642
#> 4  1.12764758 -0.7926292  0.6741236  0.7707907 -0.3881811 -0.08761134
#> 5  0.86578051 -1.2789830 -0.6730280  1.2968415  1.1331816  0.58372106
#> 6 -0.19027698 -0.2661550 -0.2452833  1.4702333  0.8318511 -0.24687602
#>         X191       X192       X193       X194       X195        X196       X197
#> 1  0.9415272 -1.2592072 -0.7044794 -0.5444843 0.74632708  0.76871139 -0.1745726
#> 2  1.4939214  0.7726923  0.6438138  2.9808504 0.68898162  1.40800846 -0.0102366
#> 3  1.9006388  0.1584535  1.1174402  1.0073449 0.69138732  0.59783889  1.2564435
#> 4 -0.2092994  0.4228469 -0.5814529 -0.6461698 0.01957806 -0.09728842 -0.4256993
#> 5  0.3594670 -0.5205762  0.9061061  0.5422425 2.12037275 -1.01118032 -1.3346181
#> 6  0.8508030  1.8373574 -1.1289246  0.3974914 0.25623849  1.08456588 -0.5555422
#>         X198       X199       X200        X201        X202       X203
#> 1  0.6434478  1.1051588  1.0894927  0.07272802 -0.22011421  1.3690946
#> 2  1.6493664  0.4600980 -0.4109516 -1.17399263 -1.25882109 -0.8825570
#> 3  0.4771085 -0.6813210 -0.3253141  0.92358287  1.25239989 -2.0012819
#> 4  0.2957004  0.1999659 -0.4538907  0.92295554  1.61099530  0.0282984
#> 5 -1.4742259  0.8589891  0.2537824 -0.65120503 -0.07388956 -0.8826129
#> 6 -0.4354278 -0.1245197 -0.7032313  0.33494685  2.06018780  0.7830976
#>           X204        X205       X206       X207        X208       X209
#> 1 -0.222325101 -0.00549717  1.9994509  1.5377801  0.93242295  0.4816581
#> 2  2.812385150  1.06787725  0.9701442  1.0609987 -0.12013945  0.1831305
#> 3  0.002162673  0.48814696 -0.1417581  1.4102710 -0.13396148 -1.1731615
#> 4 -0.243564599 -0.47681147 -0.9187395 -0.6728219 -0.03985775  0.9792975
#> 5 -1.020340024  1.04350924  0.4522968  1.4905831 -0.48266761  0.2029030
#> 6 -0.820958766  0.04525610  0.8699441 -0.3001852 -1.10187500 -0.8800532
#>           X210       X211        X212        X213       X214       X215
#> 1  0.099053512 -1.4585700 -1.52547098  0.06629857 -1.3403652 -1.4965711
#> 2 -1.637430725 -0.4067005 -0.64475160 -0.58997586 -0.3583157 -0.6961132
#> 3  0.547621700  0.2699606 -0.70106386 -0.59433253  0.2428388 -0.5991764
#> 4  2.256795282  0.3998170  0.15774777  0.94931892  0.4972175 -1.2685580
#> 5  0.002089642 -0.6323029 -0.48441542 -0.96896705  1.1263264  1.1250144
#> 6 -0.587781663 -1.9301280  0.09584982 -0.01871266 -0.1549052  0.8357363
#>          X216       X217       X218       X219       X220         X221
#> 1 -0.51961483  0.8656851 -0.6987798 0.02835873 -0.4018857 -0.288149778
#> 2 -0.10070671 -1.2030685 -0.6023619 2.30126081  3.1993802 -0.003589331
#> 3 -0.37007354 -0.9678791 -0.2054858 1.59569326 -1.2852053 -0.631084805
#> 4 -0.74891634 -1.0053193  0.3833145 0.37563642  0.6041550  0.164619385
#> 5  1.68721454  1.2844755  0.7833528 1.03239171  0.3912802  0.511975608
#> 6  0.08644119  2.0274721  1.7022917 0.38903606  1.0840728  0.535252455
#>           X222        X223        X224        X225        X226        X227
#> 1 -1.005734385 -0.07421506  1.37507984  1.09112693  1.16100051  1.29642532
#> 2 -0.016800870 -2.36437510  0.04349664 -1.26060413  0.04180809 -0.58090976
#> 3  0.001241544  0.69311875  0.36058277 -0.26683897  1.61242352  0.04131083
#> 4  0.617276112 -0.54348209  1.10323611 -0.90738940 -0.69356885 -0.65729390
#> 5  0.686326480  0.29229431  1.43354605 -0.02545199 -0.09589063  1.35210726
#> 6  0.783527747  1.05296226 -0.33314852 -0.68470215 -0.76159051 -0.50926165
#>         X228       X229       X230        X231       X232        X233
#> 1  1.6481207 -0.1626178  0.2363379 -1.44063654  0.1210269 -1.09854904
#> 2 -0.8626307 -0.2856931  1.1893363 -0.48428380  1.4181101  1.35158981
#> 3 -1.1609242  0.6355475 -0.8367776 -0.09590364  0.3251061 -0.59623249
#> 4 -0.2409797  1.7614099  1.2654241  0.79354377 -0.9172666 -0.06194346
#> 5 -0.6348286 -0.2520910  1.0833218 -0.23903428  1.2241432  0.53416780
#> 6 -0.1058128  0.1623740 -0.7829428 -1.25630330 -0.2660328 -0.56137908
#>          X234        X235        X236        X237        X238        X239
#> 1 -0.70853631  0.90820348 -0.20321665 -0.75204309 -2.13378713 -1.22002843
#> 2  0.02379948  0.16681403 -1.09964074 -0.79993853 -0.06642208 -0.86439545
#> 3 -0.33083863  0.16250378  1.03939932 -1.10464998 -0.91800399 -0.08264266
#> 4  0.61617966 -0.01803839  1.06294890 -0.03861657 -0.16134052  1.07342250
#> 5  1.19052788 -0.23956338  1.51828312  1.73955793  0.91547223  0.81109258
#> 6 -0.26820169  0.04722347 -0.09558757  2.38831727  0.29691707 -1.51489019
#>          X240        X241       X242       X243       X244       X245
#> 1  0.88571971  0.04356703  0.7257754  0.4090373 -0.9272162 -1.3025947
#> 2  0.96886544  2.05001507  0.6463644 -0.2219816  1.9983689  0.1562161
#> 3  1.17052603 -0.22552072 -1.5672544 -0.6707912 -1.0548326 -0.2970339
#> 4  0.26036696 -0.23791236 -0.1999719  0.6671132 -1.0640540 -1.2365794
#> 5 -0.01771912 -0.34451999 -0.1862346 -0.7854628 -0.6558766  0.1918200
#> 6 -0.66224151  0.78103555  1.1560845  0.8959679  1.4194036  1.3200467
#>         X246        X247       X248       X249       X250        X251
#> 1  0.6251739 -1.32745919  2.3796560  0.2238440 -0.5139124 -0.53301889
#> 2 -0.1535438 -0.67761443  0.4191524 -0.3042010  0.4354943 -1.02732045
#> 3 -0.8047782  2.02600341 -0.5453755 -0.8845392 -0.6727571 -0.56145289
#> 4 -0.9714754  0.24616466  0.8432263  0.4650551 -0.4124306 -0.92763747
#> 5 -1.7086903 -0.26758101  0.5396153 -0.1200510 -0.2184201  1.66668032
#> 6  0.8541421 -0.01951382  1.6252566  0.0759471 -0.2163886 -0.07527849
#>         X252        X253       X254        X255        X256        X257
#> 1 -0.8443316 -0.57594247  0.8121782 -0.64357342  1.12456124 -0.38722999
#> 2 -0.2434715 -0.08422807 -1.3941573  0.81496468  0.31942404 -0.56380513
#> 3 -1.5568662 -1.86881995 -2.0224077 -0.42460196 -0.43090979  0.41141886
#> 4  1.2947739 -0.21564880 -1.0268293 -0.54208406  0.22987719  0.68652336
#> 5  1.4589172 -1.27507735 -1.0963908  0.08897653 -1.22529825 -0.97703382
#> 6 -1.8046619 -0.96307446 -0.2100406  1.33064411  0.07401082 -0.03351956
#>         X258        X259       X260        X261       X262        X263
#> 1 -1.4645883  0.43909755  1.2179515  0.66593158 -0.0873880  0.14922384
#> 2 -0.6451787 -0.72552125 -0.3413894  0.35785453  0.1197210 -0.03085362
#> 3 -0.9373448 -0.68971903  1.2189712 -0.14877267  0.8911154 -0.30502856
#> 4  0.1866250 -0.07320368  1.3484312  0.07272707  0.7701622  0.50397542
#> 5 -0.5120317 -2.35815796  0.6071513  0.71359290 -1.0340068 -1.12714467
#> 6  1.8520696  1.24395821  0.1498446  1.43947875  0.1744756 -0.85660030
#>         X264       X265        X266       X267       X268        X269
#> 1 -1.2071671  0.2092158 -0.31487583  0.6067665 -0.3744050  0.02159509
#> 2  0.2245261  0.6849236 -0.71181609 -1.6507598  0.3815349 -0.53372972
#> 3  1.3607866 -0.5607578 -0.07097029 -1.0970468 -0.4359135  0.04443325
#> 4  0.9168014 -1.4315774 -0.65485484  1.0079629  1.1704625 -0.68689855
#> 5  0.2542013  1.0387759  1.14050344 -0.1240224  0.3957994 -0.91635042
#> 6 -0.1857488  0.3631655  0.91463127  0.3970286 -0.6556292 -1.50804715
#>         X270       X271       X272       X273        X274       X275
#> 1  0.5727111 0.05076005  0.1130778 -1.2046165 -1.13523979  0.1036500
#> 2 -1.1233557 0.60024318  0.6830107 -0.9897334 -1.38452913 -0.7094186
#> 3 -0.2350573 2.83643350  1.5509105 -0.1729952 -1.14497103  0.8292530
#> 4 -0.5466028 0.15629098 -0.3810947  0.4014974 -0.25806529  0.5150763
#> 5 -1.4628700 0.35959132 -1.7708680  2.2426665  1.97854399  0.8203712
#> 6  0.2568770 1.41421876 -1.4457216 -2.3294392  0.06410551  1.5157192
#>           X276       X277       X278        X279        X280       X281
#> 1 -0.916819559 -2.1603865 -1.8218718  0.02105143 -0.09379371 -0.4263479
#> 2 -2.374249789  0.2841007  0.3972375 -0.55101697  0.31318876  1.1896640
#> 3  0.005919255 -0.4612264  0.2205832 -1.69461143 -0.75597660  0.3845228
#> 4 -0.551867201  0.1709322 -1.6129117  0.70075418  2.14700499  1.2614162
#> 5  1.384781871 -1.4795718  0.2147929  0.36491627 -0.66559156 -0.2019937
#> 6 -0.232947535  0.9188680 -0.8112144  1.00196315  1.17396501  1.0006645
#>         X282      X283       X284         X285       X286       X287       X288
#> 1  0.6705183 1.0580430  1.4104854 -1.548396310  0.5936496  1.3025629  1.2402591
#> 2 -0.6365169 1.0003241  1.6650201  1.360214300 -0.2975044 -1.8690806 -0.6511994
#> 3  0.1997727 0.8055615 -0.9506582  0.210152257 -0.2165539 -1.4903995 -1.2433374
#> 4  0.8917731 1.2873255  1.5197904 -1.815831451 -1.5493933  0.7565726 -0.4888909
#> 5 -0.5673866 0.2738310 -0.1781208  0.007073544 -0.1442073 -0.7275044 -0.1798687
#> 6  1.7905761 1.9699685  0.1829820  0.063540704  0.8156206 -0.5483775 -0.9403746
#>         X289        X290        X291         X292       X293        X294
#> 1  1.1548428 -0.65242713  0.04195788 -0.006142419  0.5432357  0.44276957
#> 2  0.6256335 -0.10954762  0.04806724 -0.022989222  1.7382488 -0.73119042
#> 3  0.3593957 -0.63578493 -0.32208560  0.473603548  1.2280690  0.72673559
#> 4 -1.3237376 -0.01570483 -1.09709984  1.417186896  1.0177419 -1.82699247
#> 5 -1.0641493 -1.12141598 -0.28797277 -0.044188123 -1.4191642  0.17660793
#> 6  1.2019610  0.62879470 -0.75787457  0.350924354  0.7398702  0.08282233
#>         X295       X296        X297       X298       X299        X300
#> 1  1.5199810  1.1307357  0.88751922 -0.3948801 -1.2669337  0.54679168
#> 2  1.5922415 -0.1659462 -0.03161674  1.6916495  0.6381978  1.75235850
#> 3  0.1399664 -1.6051752 -0.12105972  0.5761076  0.9849913 -1.79113996
#> 4  0.9430456  1.0447721 -0.13913606  0.4701038 -0.5051468 -0.99963399
#> 5 -0.1970609 -0.5259855 -0.90914364 -0.3500582  0.2207893  1.59047367
#> 6  0.7929756 -1.4660249  0.29231230  0.4626924  0.3265621 -0.06211219
#>         X301       X302        X303       X304        X305       X306
#> 1 -0.3186645  0.5125293  0.40282447  0.3743498 -0.73031282 -2.4319337
#> 2 -1.0373912 -0.5961858 -0.08098642 -0.9824434  0.96122371 -0.9991700
#> 3 -1.8544295 -0.7205546 -1.22977687  0.4573398 -0.32082445  0.5770083
#> 4  1.0758107 -1.6340641 -2.90232933 -0.7821798 -0.06839193  0.2548152
#> 5  1.0020133  0.7766502 -0.34351251  0.3467082  0.35114113  0.8749228
#> 6  0.3081632  0.4113899 -0.90724396 -0.1399231 -1.34718727  0.9153954
#>         X307        X308       X309       X310        X311        X312
#> 1 -0.6481848  0.58850839  1.4321988 -0.6948954  0.65952079 -0.59839269
#> 2 -0.9815976 -1.33477435 -2.2008607 -0.7886168  0.97422042  0.21714357
#> 3 -0.7993859 -0.25884814 -0.1721511 -1.5884127  0.26506160  0.54434099
#> 4  0.1941927  0.02852025  1.1500157  0.9892778 -0.87156855 -2.37187521
#> 5 -1.0472404 -0.86252740 -0.3749325  1.9354322  0.12426130 -0.04159006
#> 6  0.7388770 -0.15731184  1.0063803  1.2294763  0.04556713  1.40469299
#>          X313       X314       X315       X316       X317       X318       X319
#> 1 -0.07167262 -0.7294957  0.2229969 -0.3294802  0.1010917  0.7689329 -1.3509191
#> 2 -0.42705927  1.1401319 -0.9920633 -1.8657936 -1.3860138 -0.2971770  0.2851457
#> 3 -0.04468502  0.9545358  0.3279447  0.9783073 -0.6830234  0.5833831 -1.2868811
#> 4 -1.08818210 -0.4866063 -0.3982319  0.2001459  0.2188876  0.4374406  0.4471082
#> 5  2.10397193  1.1039373  1.9218077  0.5025566  1.2623922  0.7638867  0.8436142
#> 6 -0.33332974  0.4180038  0.6499060 -0.8004601 -1.2769727  0.8135279 -0.4465292
#>         X320       X321       X322       X323        X324        X325
#> 1 -2.1435297 -0.8674950 -0.1563538 -0.7673651  0.11485067 -0.24449409
#> 2 -0.5614898 -0.4186052 -2.7058516 -1.1068101 -1.34515784 -0.58544189
#> 3 -2.0048909 -1.2112892  0.2446127  0.2407043 -0.10856624  0.06659243
#> 4  0.1535055 -1.7156959 -0.8983558  0.2927762 -2.38496291 -1.96634895
#> 5  1.2908494  2.0193095  0.9415077 -0.2133673 -0.02790898 -0.05137755
#> 6 -1.1394506  1.1268975  1.1037389 -0.0824529  0.41508544  0.77600932
#>         X326       X327       X328         X329       X330       X331
#> 1  1.5688260  0.6480260 -0.1030040  0.486090810  1.4416621  0.9707897
#> 2  0.2718711 -1.9328794 -2.3688506 -0.980602998  0.5653498 -2.2683718
#> 3 -0.6860884 -0.1081366  1.0176714  1.420104043  1.3303415 -0.5311035
#> 4 -1.5450471 -0.9323082 -0.2745770  1.175548361 -0.6738379 -0.5271932
#> 5  0.4306601 -0.2730898  0.8877102  1.013266611  1.0895713 -0.9785908
#> 6 -1.7262267  0.2066046 -0.8847860 -0.004666254  0.1434009  1.1561521
#>         X332        X333        X334       X335       X336       X337
#> 1 -1.4117182 -0.96901023 -1.59268847 -1.2731807 -0.8294189 -0.7595408
#> 2 -0.3434907 -2.02236956  0.87205861 -0.8294540 -0.4447075 -1.8994018
#> 3  0.2285557  0.01334419  0.63821089  0.4949225  0.9726414  0.9273645
#> 4  1.3759695 -0.36130052  0.46424254  1.1545274 -0.5656811  1.2378365
#> 5  0.5810578  0.26048056 -0.08026512  0.9894853 -1.0201994 -0.8155629
#> 6 -0.2106042 -1.40659920 -1.88190654 -0.5518999  0.5261957  1.8924106
#>         X338        X339         X340       X341       X342       X343
#> 1  1.4522563  1.00448361 -0.125223590  0.1291172 -0.4387767  1.8831752
#> 2  0.9018899  0.23865272 -0.915493850 -0.5231032  0.4722556 -0.4607920
#> 3  0.6761715 -0.07109757 -0.795955953  1.3188722  0.6230444 -1.3199128
#> 4  0.9375274  1.25556263  0.073016917 -0.4200916  0.6240066  0.2211419
#> 5  0.3901034 -0.01977242  0.369542089 -0.2126406  0.8152851  1.0896279
#> 6 -0.4726401 -0.38521174  0.007058437  0.4214292  0.3491038  0.2999846
#>          X344       X345       X346       X347        X348       X349
#> 1  1.05456263  0.7808196 -1.0868561 -0.5803648 -2.19590104 -0.4694756
#> 2  0.96822265  0.7154165  1.5970961 -0.1682938  0.12924530 -0.5150453
#> 3 -2.12612798 -1.0594384 -0.1464269  0.4988982  0.09982783  0.1383777
#> 4  0.03742035  0.9614750  1.4542937  1.2482063 -0.60761650 -1.2278631
#> 5 -1.13106179 -1.9307453 -0.7006742  0.5765273  0.04745856 -0.2891909
#> 6  1.59189896 -0.6305341 -0.1621209 -0.5721630 -1.77522412 -1.1751947
#>         X350       X351         X352       X353       X354        X355
#> 1 -0.5763300 -0.4874169  1.143065375  1.9590808  1.3889908  1.55741335
#> 2 -0.1944479 -0.4116632  0.231580604 -1.7745762 -1.0805764  0.08907743
#> 3 -0.7541471 -1.9945866  0.183332256 -0.1285501  1.5302487 -0.94041850
#> 4 -1.8800284  0.3893717 -0.922572394 -0.3035779 -0.4509116  0.03561057
#> 5  0.4780341  0.7925932 -0.978700088  0.6211215  0.1089897  0.16914234
#> 6 -0.7299619 -0.1856786 -0.001949867  1.1450032 -0.4932022  0.86859115
#>          X356        X357        X358       X359        X360       X361
#> 1  1.34110747  1.67480921 -0.23630529 -0.1855638  0.51881920 -0.1023495
#> 2 -0.01443255 -0.57664694 -0.63696409 -0.3766214  0.16834553 -0.9669529
#> 3 -0.71313839 -0.47183474  1.81057605 -1.4965244 -0.91471983 -1.3881268
#> 4 -0.66100152 -2.07578361  0.09037602 -0.4725833  0.13046107  0.2170781
#> 5  0.13871292  0.04960682  0.10880338 -1.2677487 -0.23089128 -0.1061870
#> 6  0.42700225  0.89927629  0.55149434  1.1133563  0.08692721  0.1309513
#>         X362       X363        X364       X365       X366       X367
#> 1 -0.3338787  0.3460405  1.34251633  0.2073064  0.8109664  0.3768728
#> 2  0.1807535 -1.4904957  0.38669934  0.4250867 -1.2214609 -1.0061373
#> 3 -1.6904821 -0.9278306  0.52360615  0.6164457  1.3999446  0.8825006
#> 4  0.9184483 -1.2152038  0.81807468  0.3070300 -0.7035000 -1.0153974
#> 5  0.1579374 -0.3050388  0.38572717  0.5795383 -0.5382562 -0.3361546
#> 6 -1.2675731 -0.6184011 -0.03429525 -0.0680599 -0.5483209  0.5841530
#>          X368       X369       X370         X371       X372        X373
#> 1 -0.25597101  0.3306716  0.9344536 -0.115591917 -2.6669555 -0.80096667
#> 2  1.82094649 -0.6711687 -0.9800979  0.001697063  0.3436341 -0.01383075
#> 3  1.08736229  0.7572766 -0.8145293  0.278020511  0.7759090  0.83581034
#> 4  0.53782624  0.3990298 -0.7670697 -1.172534241 -0.3356959 -0.45885189
#> 5  0.17416274  2.0418238  0.4126786  1.080664553 -0.7430375 -0.40362513
#> 6  0.07257508  0.5175616 -0.3703792  0.614286006  0.5838504  0.91921964
#>         X374        X375        X376       X377         X378        X379
#> 1  0.7995781 -1.36648620  0.81402612  1.0482061 -0.618392209  0.41819667
#> 2  1.2616667  0.42597934 -0.32754318  0.1949539  1.389701295  1.93005298
#> 3 -0.3797300  1.02086297 -0.09061669 -0.4073795  0.822386433  1.94143995
#> 4  0.4770855 -0.44491702  0.24009946  0.3557975  0.470307526 -0.10680678
#> 5 -0.9075291 -2.03900529 -1.60202974 -1.9010314  0.004556248  1.37486267
#> 6 -1.4888586  0.09356001 -0.32484418  1.6506672  0.553441159  0.00810146
#>         X380        X381        X382        X383        X384        X385
#> 1 -1.5455107  0.09186982  0.08140656 -0.07205307 -0.80730245  0.06980815
#> 2  1.2328231 -0.26373107  1.28504698  3.12777472  0.51570500  0.34803625
#> 3  0.1529144 -0.89175971 -0.20029389  0.02783346  1.57834016  0.32395710
#> 4 -1.2077440 -1.10387413  0.65341744  1.06738473 -0.06681201  0.31022595
#> 5  1.2159295 -0.02985407  0.56533806 -0.56159797  2.25237175 -0.97875555
#> 6  0.1844261 -0.81252740 -0.63136022  0.72025808  0.57601693 -0.60748856
#>         X386       X387        X388        X389        X390        X391
#> 1  0.7094755 -0.7707031  0.09624336  0.18412485 -0.05659223  1.32617906
#> 2  1.3408148  0.2340517  0.41099560  1.35098944  1.72351134  0.05708806
#> 3 -1.8147557 -1.4708410  1.33904183  1.52632190 -0.55038469 -0.28865258
#> 4 -1.7400646 -0.2919194 -0.01007099 -0.03920857 -0.60304175 -2.23221697
#> 5 -0.1627298 -0.6403354  0.43417687 -1.20315444 -1.10628297 -2.07399948
#> 6  1.6004521  1.0317681  0.08320513  0.05413454  1.62705990  2.64606369
#>         X392       X393       X394       X395       X396        X397       X398
#> 1  0.1733875 -0.1663148 -1.3371297 -2.0838650 -1.4897510 -0.15639526  1.3154959
#> 2  0.8360285  1.7467916  0.3614086  1.6246131  1.4101439  0.04169779 -0.4867042
#> 3  0.2594530  0.6088535 -2.4769089 -1.3750982 -1.8875357 -1.24972978 -1.1391266
#> 4 -1.3096091 -1.0131374  0.7560722 -0.8270300 -0.5284078 -1.14069430 -0.1808490
#> 5 -2.2762319 -0.5862119 -1.2195958 -2.3819086 -1.3119944  0.43525981  0.5979399
#> 6  0.8354764  0.1210665 -0.8602682 -0.5171122 -1.0204603  0.17948505 -0.6118340
#>         X399       X400      X401         X402       X403       X404
#> 1 -0.1381733  1.0244739 0.6883925 -0.512051699 -0.6523292  1.4872958
#> 2 -0.6668932  0.1508860 0.5397705  2.389310973  1.2747219 -0.2263345
#> 3 -0.2850113  0.8689091 0.4885157  1.003991756 -1.7290451  0.7946909
#> 4 -0.4638517 -0.2802177 1.4094728  0.266396454  0.3712989  1.1029518
#> 5  1.5611240  0.1508330 0.8832075  0.378458825  0.2556359  0.2548531
#> 6  0.7009908  0.5342693 0.1034455 -0.005180591  0.4303863  1.7530167
#>          X405        X406       X407       X408       X409       X410
#> 1  0.34952391 -1.04719376  0.9899132 -1.4775147 -1.1529343 -0.3015002
#> 2 -1.81846771 -1.49830749 -0.5285816 -0.2283704 -1.5349513 -1.5405864
#> 3  1.71399875  0.06375599 -1.5331440  0.4436200 -0.4803697 -0.6999563
#> 4 -1.19453792  1.04365732 -0.2521512  0.1481070  1.2814428  0.8623362
#> 5  1.29378951  0.71138202  0.4044142 -0.4371119 -1.0608106 -0.4698129
#> 6  0.05888013  1.49010176 -0.2217055  0.2999319  1.2667187  1.4883680
#>         X411        X412       X413        X414       X415        X416
#> 1 -1.8634310 -0.88867292 -0.5619229 -0.75393742 -0.7263326 -0.37269059
#> 2 -0.7939142 -1.34791310 -0.4387678 -0.53369541 -0.9355543  0.27219300
#> 3 -1.1215161  0.23081359 -1.0849872  0.22675672 -0.6506080  1.31969930
#> 4 -0.6029207  0.31821119  0.3786893  0.02081133 -0.8217779  1.50369470
#> 5  0.3570842  0.05820295 -1.2635561 -0.35892742 -0.2226874 -0.02843637
#> 6  2.1720527  0.05131109 -0.4396078 -2.82906961  1.3850803  0.51139980
#>          X417       X418         X419       X420        X421        X422
#> 1  0.41151536  0.4413087 -0.009877738  1.0461649 -1.20619363 -0.28195819
#> 2  1.49741741  0.4898395 -0.239302020  1.4881999 -0.75030920 -1.30201043
#> 3 -1.95154449 -0.7368567 -0.220112542  0.7653788 -0.42013832 -0.05453821
#> 4 -0.03489639 -2.3639476 -1.895658954 -1.8768548 -0.65213874  1.46802239
#> 5  0.77229897  0.8459702 -1.202086426 -0.4418273 -0.09683219  0.19741858
#> 6 -2.29822878 -0.7174376 -1.058370111 -0.9444392 -1.22382017  0.79652963
#>         X423        X424        X425       X426       X427        X428
#> 1  1.2501391  0.36772139  0.95945047  1.4715263 -0.3254791 -0.05670833
#> 2 -2.1203216  0.05230702  0.53600826 -0.1569419  0.2239871  0.21635535
#> 3  0.0760787 -0.10154633 -0.48993433 -0.3066632 -0.6303456 -0.13110421
#> 4 -0.6488934 -0.77036360 -1.65650255 -0.7262504 -0.1598404 -0.12254740
#> 5 -1.1804811  0.85430122 -0.02306447 -0.5546898 -0.8943116 -0.34287846
#> 6  0.5632263 -0.22730845  0.23750401  0.1814062  1.1953567 -1.07367208
#>         X429       X430       X431        X432       X433       X434       X435
#> 1 -0.1220488 -0.1017792  1.0910014  1.59617341 -2.6748405 -0.1423268 -0.2298976
#> 2  2.0743844  2.1978061  0.3172260  2.01963739  1.7772074  0.9833077 -0.8008108
#> 3  0.5642855  0.3305870 -0.2759734 -0.69723909  0.4411140  2.0261539  0.1424940
#> 4  0.9770947  1.6480821  0.0828454  0.68130785 -0.7165759 -0.1413251 -2.3464218
#> 5 -1.1157946  0.9439630  1.8387633  0.59074139  0.2425375  0.4675318 -0.0351038
#> 6 -1.2822151  0.8251960  0.8891633  0.08288579  0.4972371  2.0334474 -1.1633994
#>         X436        X437       X438       X439       X440        X441
#> 1 -0.4463842 -0.96047542 -0.9439687 -0.5003036 -1.3485313  0.37310747
#> 2  0.5148142 -1.44784771 -0.4466190  2.1369953 -0.4668822 -0.14303244
#> 3 -1.4800729  0.01479832  0.2927146  0.9789814  0.4767590 -1.39115022
#> 4 -1.2951849 -0.75632863 -2.4199043  0.3838555  0.4571446 -1.50092746
#> 5 -0.3566463  0.33702485 -1.6989657 -0.2410254  1.6016561  2.86903639
#> 6  0.8175621  0.05033549 -1.1727736  0.1154018  0.9118889 -0.01934529
#>         X442       X443        X444       X445       X446         X447
#> 1 -0.5399470 -0.5181662  0.14507924 -0.4267214  0.4744283 -1.617655598
#> 2 -0.7706916  0.4081368  1.44545464  1.1961276  0.8358159 -0.125949545
#> 3  0.1429548  0.8867409 -0.42189404 -0.8553549  0.4407923 -0.155305923
#> 4 -0.8346855 -0.8492248 -0.13985287 -0.8426949  0.5301579  0.621220962
#> 5  0.2802654 -0.7261521  0.02737762 -0.3732042 -0.6941115  0.009414157
#> 6  0.2007734 -0.7811126 -0.44603351 -0.7237251 -0.1899705 -0.004553905
#>         X448       X449       X450       X451         X452        X453
#> 1 -1.7610202 -1.6977644  0.3274169 -1.0836769 -1.124750175 -1.15136478
#> 2 -0.1659391 -0.9207808 -1.7408852 -0.6603405 -0.003678582 -0.85434364
#> 3 -0.2367704 -0.4985737  0.3213825 -0.9911284  1.738682169 -0.18961079
#> 4  0.7680489  1.6870361  0.7408168 -1.0831712 -0.455322808 -0.34406204
#> 5  0.4638966  0.7883120  0.1980552 -0.2729838  1.142257640 -0.04780054
#> 6 -1.3860732 -1.1204288 -0.1737031 -0.3226924  2.081588822 -0.47445182
#>          X454        X455        X456       X457        X458       X459
#> 1  0.22517979  0.04577029 -1.47526209  0.6051234  0.89947390  1.0562770
#> 2 -0.03384579 -0.45120100  0.64400715  0.8201271 -1.63230719 -0.8602228
#> 3 -1.00037087 -0.24151840  0.27758776 -1.2575565 -0.52141557  1.1314487
#> 4  0.72312099 -0.15920784 -0.68959044 -0.9449918 -1.10325035  0.6295882
#> 5 -0.71127943 -0.76730761  0.02937465  1.6210576  1.34883981 -1.2773951
#> 6 -0.70849569  0.04433550  1.37404853  1.0607840  0.09674681 -0.3678762
#>          X460       X461       X462       X463        X464       X465
#> 1  0.71368067 -0.1252270  0.6859730  1.3581371  0.62549881  0.7850068
#> 2 -1.34072075 -0.9144944  0.9498866 -0.6830506 -0.47255776 -0.4753344
#> 3 -0.18351342 -1.5792372  0.7188021 -0.6546796  1.42091662  1.5831993
#> 4 -1.89491120 -1.2356550 -0.3478137 -1.1892249 -0.35284058  0.1136640
#> 5  0.01505803 -1.0262318  0.2428035 -1.5806947 -0.00251106  1.3081566
#> 6 -0.45184284 -0.5169102  0.3141520 -0.5523362 -1.11424597 -0.8674141
#>          X466       X467       X468        X469       X470       X471
#> 1  0.39648379  1.5791751 -0.2861155 -0.95138987  0.6859826  0.4521264
#> 2  0.07126986  0.3907265  0.6838803  0.08522132  0.8427921  2.0915681
#> 3  0.46428453 -0.7201378 -0.9675126 -1.25773724 -0.3936989 -1.1063283
#> 4  2.32909804  1.5950136  0.4269518  0.01263899 -0.7772995 -0.3200600
#> 5 -0.31292112  1.8252997  0.8720268  0.31625427  1.5592549  1.9142095
#> 6 -0.36440082  0.4799438  1.6288847  1.55760176 -0.5904524  1.1901723
#>          X472       X473       X474       X475        X476       X477
#> 1 -0.07743891  0.9158586 -0.6318370 -0.5656648 -0.26016912 -1.0031590
#> 2 -0.03225137  1.0405215 -0.3364456 -0.6067088 -0.28217823 -0.5528918
#> 3 -0.65624249 -0.3681743 -1.0404402 -0.7573746 -0.52781854 -0.2057193
#> 4 -0.09915201 -0.3235939  0.9915522  1.3419162  1.04385265  1.7391504
#> 5  0.53443957 -0.1980450  2.1315515  1.7609073  0.36729501 -1.0381937
#> 6 -0.44192825 -0.2266534 -1.6258851  0.1957255  0.04511285  0.6991253
#>         X478       X479        X480        X481       X482        X483
#> 1 -0.4223600  1.3318864  0.43283547 -1.11615428 -2.2552660  0.10362794
#> 2 -1.0218669  0.2545116  0.19874760 -0.08087853 -1.1345589 -0.01716615
#> 3 -1.1288848 -1.0604024  1.32382815 -0.69713527  0.6147110 -0.58025406
#> 4  1.8998617  0.8616436 -0.06578089  1.03889953  2.9912248 -0.34904830
#> 5 -0.1149909  0.3706450  0.29271870 -1.49037885 -0.5609999  0.26746704
#> 6  1.5723641 -0.2171089  0.35081588  1.60091400  0.6076551 -0.38936837
#>         X484         X485      X486       X487       X488       X489       X490
#> 1  0.2497306  0.078144387 1.0910119 -0.2661981 -1.2054928 -0.3110851 -0.2297720
#> 2 -0.3750209  0.085457412 1.9458254  1.2819071 -0.4891080 -0.5187225  0.5430969
#> 3 -0.8382870  0.455923551 1.5893382  0.6849318  1.3597193  1.0958429  1.2908295
#> 4  0.5474589 -0.985377115 0.2202032  1.4677285  0.3728935 -0.9217471 -0.1522221
#> 5 -0.3666891 -0.006689443 0.5006465  2.4097054  2.0566073  3.2225850  0.3898649
#> 6 -0.5248417 -0.506873970 0.3407116 -0.7670411  0.2015443  2.0192841  0.5929865
#>         X491       X492        X493       X494       X495       X496       X497
#> 1 -0.5917707 -0.9426303 -0.36314851  0.2952473 -1.4670347 -1.4824212 -1.6651313
#> 2  0.8941004  0.9846885  1.96832536  1.3602567  0.1230812  1.6355481 -0.3738510
#> 3  1.0197547 -0.2654841 -0.55489908 -1.7430083 -1.4308969  0.9392121 -1.3309001
#> 4  0.2039489  1.0370147  0.53882397  0.2384023  1.0094826 -0.5649906  1.4980335
#> 5  0.6393037  0.1428939 -0.53705739  0.2966435  0.1949726 -0.3165069  0.8182568
#> 6 -1.0074843 -0.1281243 -0.00490374  0.3659625  0.3454741 -0.4591514 -1.0473922
#>         X498       X499       X500
#> 1 -0.8432194 -2.0343430 -1.6385027
#> 2  0.8908853 -0.6120472  0.1104062
#> 3  0.5983662  0.1354075 -0.2303969
#> 4  1.3041907  1.9590968  1.3128086
#> 5 -0.1943304 -0.3462852 -0.4256174
#> 6 -0.4463024 -1.6996340  1.2181634