A high dimensional dataset with n equals to 200 and p equals to 500.

A dataset with 200 observations and 500 dimensions is generated from the following process: linear regression model with only first three non-zero coefficients equal to 3, 2, and 1.5 respectively. The covariates are correlated with AR structure (rho=0.3). The error term is normally distributed with zero mean and sd equals to 0.5.

data(HighDim)

Examples

# HighDim dataset is generated from the following process:
n <- 200; p <- 500; sigma <- 0.5
beta <- rep(0, p); nonzero <- c(1, 2, 3); zero <- setdiff(1:p, nonzero)
beta[nonzero] <- c(3, 2, 1.5)
Sigma <- 0.3^(abs(outer(1:p,1:p,"-")))
library(MASS)
X <- mvrnorm(n, rep(0,p), Sigma)
error <- rnorm(n, 0, sigma)

X <- apply(X, 2, scale) * sqrt(n)/sqrt(n-1)
error <- error - mean(error)

Y <- X %*% beta + error
HighDim <- data.frame(Y, X)
head(HighDim)
#>             Y         X1          X2         X3          X4         X5
#> 1  1.89108429 -0.3070360  0.86961241  0.9923989  1.58079878  0.6595641
#> 2 -0.08924071 -0.9902033  0.53645752  1.1095344  0.27630189 -0.2281768
#> 3  1.37990583  0.1307625 -0.00336475  0.6360038 -0.55099697 -1.1728225
#> 4 10.18075509  1.8293298  2.07451748  0.6383554 -0.07772763 -0.9784346
#> 5  0.71250328  0.6597689 -0.59222001 -0.2003166 -0.91248855 -1.2858675
#> 6 -2.46994875 -0.4336750  0.09418981 -0.6372049  0.97581054  0.4781152
#>            X6          X7            X8         X9        X10        X11
#> 1  1.51242614 -0.01179862  0.5841591741  0.7668326  0.8369557 -0.3755716
#> 2  0.90700578  0.13417784  0.0002651955 -0.4027333 -1.4111966 -0.9745750
#> 3 -0.05574752 -1.47031401 -1.1860989676  1.0438029  1.0679717  1.0015504
#> 4  0.84111192  0.97177258  0.2540315450 -0.3637201 -0.4857901 -0.6398150
#> 5  1.14471013  0.03386173  0.7343998284  0.2719695 -0.6250627 -0.9367425
#> 6 -0.42552695  0.01044251  0.9405291418  0.4242204  0.6290035 -1.1141326
#>          X12         X13        X14         X15        X16        X17
#> 1  1.1176385 -0.54813820  1.2162940  0.15536680 -0.2468835  0.3658807
#> 2 -1.2165730 -0.03499565 -2.3112079 -0.72313953  0.3161909 -1.0574513
#> 3  0.9711626  1.23096582  0.9110432  0.71188951  1.2138594 -0.1270607
#> 4 -0.7926303  0.35090563  0.6937871  0.09554522 -0.2717883 -0.5879625
#> 5  0.4972332  1.13832400  1.4988797  0.65416250 -0.6764579 -0.1926140
#> 6 -2.6077494 -0.22364893  1.9272530 -0.94558225 -0.2519016  0.5021375
#>           X18         X19        X20        X21        X22        X23
#> 1  0.04124491  2.34667012  0.1561011  1.6878768  0.2484364  0.7400416
#> 2  0.87138007 -1.83817396 -2.5846502 -2.2274222 -1.1857846 -0.2832620
#> 3 -0.29266026  0.48073783  1.4275694  0.1618043 -0.6328445 -0.5951107
#> 4 -1.33213182 -0.04545135 -0.1130256 -0.7405443  0.2864389  0.4448595
#> 5 -1.38473679 -0.36045487  0.3506360  0.2478425 -0.2802474  1.4612815
#> 6 -1.02195317 -0.74547360 -1.0754526 -1.9835893  1.5759501 -1.1334117
#>          X24         X25        X26         X27         X28        X29
#> 1  0.2132481  0.00567740  0.9144099  1.24376901 -1.68844942  1.3161164
#> 2 -0.8217211  0.77970122  1.0381665  0.94947426  0.05338592  0.1581217
#> 3 -1.3563412 -0.99004706  2.2647363 -0.09649343  0.94663806 -0.4562623
#> 4  0.7066578 -1.62715868  0.3489468  0.81531289  0.50454425  1.4964477
#> 5  0.8353616  0.85304517 -1.7595088 -0.21898608  1.69959125 -0.2101778
#> 6  1.7289515 -0.08889561 -0.4856278 -0.36218991 -0.03421626 -0.1219180
#>           X30          X31         X32         X33        X34        X35
#> 1  2.04695085 -0.140752891 -0.02238511  0.14267863  0.2472522  0.9251447
#> 2  1.12355192 -0.555496173  1.92073384 -0.17706749  0.2064840 -0.7202100
#> 3 -0.40955090  0.518811197  1.94288658  1.45848528  0.3404728 -0.9372861
#> 4 -0.43026319  0.055105458 -0.09242445  0.10840819  0.3891730 -0.0595259
#> 5  0.18188726  0.841438330  1.15202776  0.81990019  0.7965717 -0.2645935
#> 6  0.02376636  0.005634171  0.65046514  0.07864008 -1.2532773  1.5830844
#>           X36        X37        X38        X39          X40        X41
#> 1  2.17361277 -0.2967238 0.71842106 -0.9932422  0.972859601 -0.3024219
#> 2  0.85083040  1.3459611 0.49859452 -0.9248474 -0.723600230 -0.5240830
#> 3 -0.39664233  1.3476896 1.21892028 -0.3321915  0.006646654 -2.1934547
#> 4 -1.15144942  0.9083003 0.01219874  0.5109953  0.328999692  2.2473952
#> 5 -0.02882009  0.7112093 2.84911143 -0.3002890  0.401606739 -1.4417456
#> 6  0.77282579 -0.3432321 0.56365119 -1.5316979  0.798516489 -0.7623861
#>          X42        X43         X44        X45        X46         X47
#> 1 -0.6952499  0.3863003  0.32414847  1.2577063 -0.3993634  0.38018908
#> 2  1.0502885 -0.4682663  1.22574708 -1.1827207 -1.6270145  1.17843585
#> 3 -0.7542356  1.6261151  0.04723493  1.4942059 -0.1653231  0.14224561
#> 4  1.4650774 -1.9154992 -0.93647080  0.2344444  1.9611190 -1.77136926
#> 5  0.8865932  1.1175986 -1.60309933 -1.7050497  0.7287889  0.23340030
#> 6 -0.9129416  0.2320265  1.74879941  1.6027824  0.7951280 -0.06797563
#>           X48        X49         X50        X51         X52          X53
#> 1  0.11030890 -1.4221391 -2.01944624 0.34604219 -0.68547097 -0.394630918
#> 2  0.65025882 -0.4377844  0.25797850 0.60392179 -0.92673830  1.525475419
#> 3  0.76377624  1.0551209 -0.12326764 0.57364528  0.78989384 -0.004843461
#> 4  0.07126116 -1.6573777 -2.59568352 0.22076386  1.62270325  0.251979237
#> 5 -0.40873051 -0.3524663  0.08878341 0.07358692  0.09320564  0.690444572
#> 6  0.14403769 -0.4519287  1.96366790 1.40824699  1.00966171 -0.290931163
#>           X54        X55        X56         X57        X58         X59
#> 1  0.35377658 -0.9205477 -0.8236142  0.08290852 0.91749880  1.01837628
#> 2 -0.36087281 -1.1023845 -0.9484863 -0.18014862 1.07844525 -0.88579712
#> 3 -0.11641399  0.7828191 -0.1493977  0.56920820 1.95058828  0.22891277
#> 4  0.07026118 -0.2704780 -1.2434538 -0.43040993 0.07028263  0.30895904
#> 5  0.94793519 -0.3119700 -0.4830996  0.29092747 1.27344386  0.67806033
#> 6  0.94937207  0.6706208 -0.1874026 -0.95563875 0.52921620  0.05060483
#>          X60       X61        X62        X63         X64           X65
#> 1  2.3107008 3.1270874 -0.5237514 -0.3921895  1.14326423 -1.4601171506
#> 2 -0.2535040 0.8345699  1.4465978  2.7149482 -0.06552772  0.8558156031
#> 3  0.3235121 0.9721547  1.6032665 -0.1647301  0.64168541 -0.7729006137
#> 4  1.8907817 1.6854558  0.7470138 -0.8531706 -0.54283878  0.5122087370
#> 5  0.9875953 0.7994978  0.7722693  0.5499443  0.92293407  1.4670357130
#> 6 -0.9685184 0.0615092 -1.0834997  2.1082359  0.33258258  0.0004970675
#>           X66        X67        X68       X69        X70         X71        X72
#> 1 -0.38755424  0.4563110 -0.5012543  1.454577  1.5753450  0.04546998 -1.3483422
#> 2  1.05815934  1.2280155  0.4701458 -1.377854 -0.2759846 -1.26206335 -0.7863773
#> 3 -0.80575246 -0.8435334 -1.3797479 -1.698572  0.1792602 -0.42844564 -0.5706950
#> 4 -0.01733834 -0.6789301  0.7573100  1.512266 -0.3370283  0.02167612  0.3542796
#> 5  0.73064633 -0.1721299  0.7783340 -0.572700  0.1786613 -0.36290951  0.2605347
#> 6 -1.22389855 -1.6978688  0.1385747 -2.741231 -0.3426589  1.15554138 -0.9128818
#>          X73         X74        X75          X76         X77         X78
#> 1 -0.1090311 -0.04384560 -1.0602841  0.995334081  0.70410965 -0.43790813
#> 2  0.1617076  0.06343711  1.8587756  0.835075227  0.14917153 -0.59858966
#> 3 -1.3038543 -1.50059159  0.3851280  0.058633553 -0.33093590  0.02268399
#> 4 -0.8056300 -0.31936190 -0.8095702  1.052652615  0.42823117  1.10533713
#> 5 -1.1835009 -0.88972797 -2.4061327  0.366175631 -1.38397097  0.13641473
#> 6 -0.6039962  0.32748913  0.7853122 -0.004187656 -0.03254314 -1.28266358
#>          X79        X80        X81        X82       X83        X84        X85
#> 1 -0.8987866 -0.7323661 -0.9966824 -1.2248068  1.106871 -0.7756909 -0.4267993
#> 2  0.1025770  1.6793192 -0.7638994 -0.8298836 -1.115789 -1.2204547 -2.0386092
#> 3 -0.8490860 -0.8546092  0.2941746  0.1549862 -1.442802  0.7458696  1.4167289
#> 4 -0.2788687  0.4623707  1.1031935 -0.3832782  1.287577 -1.4196198 -1.3887516
#> 5 -1.2762465 -1.3085103 -0.8007582 -0.9943848  1.042623  1.2970734  0.1429795
#> 6  0.2669221 -1.9898637  0.7125369 -0.7267149  1.059888 -0.0730565 -1.1610989
#>          X86          X87        X88         X89        X90         X91
#> 1  0.7747923  0.190632591 -1.1940298  0.01280818 -0.2061481 -1.35556615
#> 2 -3.3466422 -1.459213691  0.7763560 -0.26617138 -0.2552885 -1.73033777
#> 3  1.4390864 -0.002089004 -0.3082276 -0.12008389  0.2820215  0.17203938
#> 4 -0.9013441 -0.265277288 -0.5782281 -0.58957398 -0.7431356  0.06195968
#> 5  0.4346295  1.467006864  1.6169182  0.17712242  1.4156702  0.68604119
#> 6 -0.5943855 -0.876982423 -0.4846519  0.01449309  1.2183103  2.22586801
#>          X92        X93        X94        X95        X96        X97        X98
#> 1  0.6396603  0.3418710 -0.2933618 -0.6153963 -0.2225001 -0.8037554  1.0078531
#> 2 -1.3773612  0.8309598 -1.4966101  0.4429719 -1.6550782 -2.0712376 -1.0572472
#> 3  0.9466213  1.2138043 -0.2026935 -1.8824881 -0.4487036 -1.2599584 -0.9639513
#> 4 -0.2104075 -0.3905365 -0.1664314 -1.7662297 -2.2543494  1.2390219  1.4765581
#> 5  0.9777480  0.6871736  0.5332954  0.6724185 -0.6938952 -1.5529026 -0.4222340
#> 6  0.7991332 -0.5245512 -0.3746172 -1.9748705 -0.4538469 -1.6427061 -0.2161602
#>          X99      X100       X101       X102        X103       X104        X105
#> 1  0.8686764 2.4415264 -0.4102982  0.6742904  0.77357448  0.4093876  0.06333886
#> 2 -1.2896236 0.3235894  0.7912535 -0.1917772 -0.70733550 -0.3798602 -1.16245563
#> 3  0.4375375 0.6158077  0.4601815 -0.5650022 -0.39371759 -0.9430726  0.50521342
#> 4  0.7928099 2.2201166  0.5564108 -1.2645771 -0.49904946 -1.0146853  0.69461431
#> 5  1.3917271 0.4880314  1.8334137  0.5341898 -0.29584638 -0.4405108 -1.64703101
#> 6  0.1500177 0.3304506 -1.5563299  0.1957031  0.08177661 -0.1934144 -0.91472058
#>         X106        X107       X108        X109        X110       X111
#> 1  0.8310567 -0.32788282  1.1306671  0.73597835 -0.32841917  1.2248175
#> 2 -0.4010713  0.01521563 -0.3058645 -0.36809529 -0.56610398 -0.8537317
#> 3 -0.6875802 -1.02731530  0.1811960 -0.04973728 -0.93347193  0.1369485
#> 4 -0.4470050 -1.34069079 -0.4076112 -0.44773289 -0.14341603  0.8643638
#> 5  1.5057806  1.13451213  0.3773588  0.63029858  0.08710264 -0.1999290
#> 6  0.1946700 -0.98175939 -1.1501651 -0.24747167 -0.66896454  0.5515711
#>          X112        X113       X114       X115       X116       X117
#> 1  0.80170415  0.25123500 -0.3535492  0.2330940  0.7289574  1.9764250
#> 2 -0.62759934  0.08722526  1.5862122  0.2843811  0.6655758 -0.2464600
#> 3 -0.02660133 -1.03326086  1.7802422  1.2004694  1.0398953  0.7090295
#> 4  0.74499936  0.73996422 -0.6640388  0.7354513  0.6343360  0.1744831
#> 5  0.24045857  1.70766367 -0.4039814 -1.4759453 -0.4780217 -0.0852973
#> 6 -0.61160295  0.37564678  0.2145840  0.6635292 -0.5198558  1.0645330
#>         X118       X119       X120       X121       X122      X123       X124
#> 1  0.4487929  1.1779751 -0.1001302 -0.1306902 -0.2085974 0.5658344  1.1218818
#> 2 -1.4576040 -3.0298125 -1.4625072  0.8379102  0.5662218 0.6692540 -0.4581515
#> 3  1.5360800 -0.3351594 -0.5601482 -1.0267297 -1.3179698 0.4078296  0.5131081
#> 4  0.6307292  0.2371000 -0.4567922  0.0476556 -1.7965523 0.7481460 -0.3043297
#> 5  0.5444850  0.4009112  1.6144178  0.3674574 -0.5153150 0.3583038  1.0372711
#> 6  1.6167932 -0.1639012 -0.9428483  0.0698732  0.1014625 0.1548415  0.9468598
#>          X125        X126        X127       X128       X129       X130
#> 1  0.04596962 -1.41457558  0.03228951  0.3090784 -0.5649114  2.0224306
#> 2  0.23694520 -1.61125093 -0.80290401  1.1115387 -0.8014967 -0.3495551
#> 3 -0.03824027 -0.09717478  0.28581845  0.3052434  1.4947789  1.2765933
#> 4  0.81082959  0.60842333  0.50307381 -0.2435177  1.1033847  0.1032977
#> 5  1.36959825  0.73401215  1.26095775  1.3318071  1.8330829  1.8844792
#> 6 -0.96285841 -2.85299941  1.14713540 -0.2359973 -0.8949845 -0.6232938
#>         X131       X132        X133         X134        X135       X136
#> 1  0.2635642  0.1979599 -0.04130499 -0.471403022  1.13783126 -0.1838361
#> 2 -0.5618143  1.3703067  0.65014101 -0.435164676  0.56088521  0.7497355
#> 3 -0.0742601  0.2291608  0.44549076  1.165466744 -0.04327429 -1.0722731
#> 4 -0.9824794  1.0195192 -0.13182646  0.004396861  0.44626313  0.5444562
#> 5  1.1968441 -0.1573951 -0.32501725 -1.710547336  1.13907510 -0.2677517
#> 6  0.6627668  0.8796993  0.12900759 -0.684198229  0.12044507 -1.1459794
#>           X137       X138       X139        X140        X141        X142
#> 1 0.7580327211  0.6265221  0.1699783  0.68333865 -0.27828877  1.54091579
#> 2 0.2606983677  0.9588217  1.8478306  0.52819706 -0.13294625  0.53145538
#> 3 0.0001976427  0.3367890 -1.0215844 -1.74313604 -0.74619765  0.24668363
#> 4 1.4179084071 -2.0764042 -0.3944387  0.28713573 -0.16079752 -0.43457114
#> 5 0.0897828630  0.9861557  1.4997693 -0.06565385  0.04026499 -0.05592473
#> 6 0.8225178460 -1.2991231  0.3962961  0.78217272  0.14808846  1.37498943
#>         X143        X144       X145       X146       X147       X148
#> 1  1.2633972 -0.02186158  1.2643582  0.7868763 -1.0936392 -0.8141987
#> 2 -0.3122395  0.61962336  0.6955585  2.3297076 -0.3637079 -1.2127417
#> 3 -0.2653453 -1.08818817 -0.5652379 -0.4226617  1.0081540  1.2319678
#> 4 -1.0666815 -0.95059032 -1.1495104 -0.3052898 -1.2607934 -0.1671793
#> 5  1.4009671  0.84677068 -0.8378226 -1.2427411 -0.8928769 -0.4125613
#> 6  0.9542668 -0.40731528 -0.3946770  1.8926794  2.5055262 -0.5603381
#>          X149        X150       X151       X152       X153        X154
#> 1 -0.78054574 -1.71369555  0.6859905 -0.6941039  2.3481201  1.46886723
#> 2 -0.01310674  0.05078282 -0.2554282 -0.1321772 -0.5351309 -2.66139090
#> 3  1.62671528  1.36108994 -0.2560665 -0.1096834 -0.7290936 -0.07742781
#> 4  0.03043721 -0.06411067 -0.1003191  0.6463784  0.0695332 -0.53810863
#> 5  0.74271246  0.99436151  1.0244420  0.7593263 -0.7156758  0.37308024
#> 6  0.22846677 -0.15747493 -1.4170068 -0.1488111 -1.3079176  0.23332374
#>         X155       X156        X157        X158       X159        X160
#> 1  1.7538089  0.4638266 -0.43794065 -0.92365482 -0.5812304 -0.47018670
#> 2 -0.7727986  0.1305592 -0.41512109  0.37030784  0.2433260 -0.23942499
#> 3  2.1045931  0.3020068 -0.15962897  0.33008174  0.4042489 -0.04691772
#> 4  0.4853554 -0.1513719  1.00581134 -0.02702556  1.0463916 -1.24004504
#> 5  1.6842047 -0.6459914 -0.09074592  1.16070044  1.1442375  1.77044610
#> 6  0.6884861  0.4242210  0.50416990 -0.27469552  0.3343381  1.22030995
#>         X161       X162       X163       X164       X165       X166       X167
#> 1 -0.9427347  1.1141748  0.1486585  1.1042139  0.8306034  1.3476351  0.0114767
#> 2 -1.9796898 -2.0719179 -1.3455864  1.4227412  0.6863734 -1.0030337 -0.9959757
#> 3  0.8154628  0.4595369  0.7812819  0.3259677  0.3271599 -1.5857618 -1.1716705
#> 4  1.4361839 -0.7082141 -0.0682755 -1.7712249 -1.5409601 -2.3890782 -0.1133307
#> 5  1.2529251  0.3440159 -0.8127463 -0.6161548 -1.5615750 -2.4932927 -0.9671807
#> 6 -1.8900136 -0.7166256 -1.1732441 -0.4221584  0.9497474 -0.3883874 -0.5628448
#>          X168        X169       X170       X171       X172       X173
#> 1 -0.79097743 -0.74297923  2.3209484  0.8566948  0.8075452  0.9760260
#> 2 -0.12690636  0.07222327  0.7430703 -0.3801731  1.7910577  0.7518308
#> 3 -0.06562288  1.35523380  1.8508683  0.9703827  1.6447735  0.6048961
#> 4 -1.00293542  0.48792451 -1.4070123 -0.4566090  0.4159143 -1.3050608
#> 5 -1.35437284 -1.17663502 -2.0641713 -1.4893944 -0.6081602 -0.4475807
#> 6 -0.73793628  0.20143093  1.1714289  0.1828258  0.8278809  0.6896394
#>          X174        X175       X176       X177       X178       X179
#> 1  1.26511277  0.74850648  0.2616886 -0.3697715  0.3653683  1.3972850
#> 2 -0.60656436 -0.07576646 -1.7414018 -2.2389108 -2.9371741 -1.4598493
#> 3  0.02325087  0.75084038  1.3220264 -1.3009078 -0.3857771 -0.7592478
#> 4 -0.42771440 -0.16724605 -1.2969437 -1.8922909 -1.2154360 -0.5935043
#> 5 -0.68187394 -1.11080024  0.2738241 -0.8462913 -0.5990130 -0.7140300
#> 6  0.79371964  0.45595215 -1.6541290  1.2054974  0.4887409 -1.7733674
#>          X180        X181        X182       X183       X184         X185
#> 1  0.41649323  0.26461393  0.59955333  0.1963364 -0.3077965 -0.005387813
#> 2  0.08421992  0.03444604 -0.01975356  1.2287507  0.9460849  1.071155712
#> 3 -1.50746337 -0.21522004  1.02645177 -0.5418708  0.1615854  0.893951725
#> 4  0.20383712 -2.42468588 -0.05145068 -0.5899048  0.5570090  0.585099018
#> 5 -2.00079033  0.21708691 -1.13089662 -0.7715695 -2.0631615 -0.707971653
#> 6 -0.12707356  0.88921945  1.24007354  0.7709966  0.9784725  0.047105453
#>         X186        X187       X188        X189       X190        X191
#> 1  0.6175907 -0.28717065 -0.7368600 -1.46435212 -0.5912287 -0.58659263
#> 2  0.9622096 -0.06515856  0.9488235  0.97339195 -1.3307020 -0.99049030
#> 3  0.6623345 -0.11866924 -0.4234591 -0.74973493 -1.4469289 -0.05604352
#> 4 -0.3822571  0.48063695 -0.7831346  1.22809635  1.7820951  1.06755251
#> 5  1.1428651 -0.60399942 -0.7700676  0.08647268 -1.8066055 -0.75396070
#> 6  0.5490106  1.13622703  1.2362019  1.60535069  0.4874523  0.27710562
#>          X192       X193       X194       X195       X196       X197
#> 1  0.95772231  1.7574290  1.1159921  0.5501743 -0.4951706  0.1202991
#> 2 -2.51772432  0.6690978  1.8398711  0.3029829 -0.4012666  0.8371652
#> 3 -0.00954492 -1.9959178  0.4565406  0.2209838  1.2638565  0.7470339
#> 4 -0.69110547 -0.1625245  0.0614438  0.7447007  1.1505512  0.3491664
#> 5 -0.30771781  1.5799149  0.2622330  0.5323405 -0.3184577  0.2196803
#> 6 -0.06967077 -1.5404812 -2.4144859 -1.6301382 -1.5785066 -1.0645359
#>          X198       X199       X200       X201        X202       X203
#> 1  0.73708055 -0.4476583  0.1871886 -1.3994412 -1.17498996 -0.5031882
#> 2 -0.73333044 -1.8958699 -1.1864865 -1.2185184  0.55584263 -0.8187153
#> 3 -0.35443065  0.3130586 -0.5834284 -0.3261862  0.83964472  0.2337803
#> 4  0.32464138  0.5566260  0.1353127 -0.2028269 -0.09356902  0.6531007
#> 5 -0.03474109 -0.7215613  0.9766694 -0.5501282 -1.62341106 -1.7163152
#> 6 -0.67256676  0.4547916  0.4050697  0.1476218 -0.99577047 -0.6179759
#>          X204       X205       X206         X207       X208       X209
#> 1 -0.82433888 -0.3726849 -0.3495241 -1.233285005 -1.3251535 -1.7574722
#> 2 -0.04268035  1.4524408  0.5063112 -0.009457366 -0.5102496  1.9015721
#> 3  0.64251194 -0.1104337 -0.7916963 -1.621233147 -0.6756590 -0.7978848
#> 4  0.93090900  0.4494971 -0.6488323 -0.709960997  1.0297530 -0.6995234
#> 5  1.24640063  0.8852768  0.8422103  1.576118326 -0.7986630 -0.1726549
#> 6  0.59523080  0.3503629  0.6802202 -1.659166354 -1.8402263 -1.5973844
#>         X210        X211       X212         X213       X214        X215
#> 1 -1.4240864 -0.96031634 -1.2722411 -0.418584858 -0.9788542  0.03809238
#> 2  2.2839520  0.27973154  2.2691786  1.517666276  0.2805720  0.45058606
#> 3  0.6417541  0.13190773 -0.1711640  0.637589473 -1.6061717  0.10343541
#> 4 -0.4640564  0.72160729  1.8956983 -0.587145035 -1.2409659 -1.41944330
#> 5 -0.7174904 -0.03778152 -0.4182971  2.270748041  0.7168884  0.09095302
#> 6  0.4107438 -0.36342007 -0.4436264  0.007171209  1.1696242 -0.31115074
#>         X216       X217       X218        X219       X220         X221
#> 1 -0.6340018 -1.1432179 -0.3952515 -0.69640870 -1.4893983 -0.827719017
#> 2  0.7182234 -0.5270580  1.4557505  1.10530396  0.8989737 -0.006453045
#> 3 -0.6482751 -0.3273633 -0.1319364 -2.70716775 -1.2487196 -0.118262870
#> 4  1.5624861 -0.6216332 -0.1171562  0.02016233  1.3577980  0.680794213
#> 5  0.9511511 -0.6594102 -2.0296720 -1.52689500 -2.0694279 -0.943740409
#> 6  0.5364433 -0.4404635 -0.4649695  0.07924878 -0.1331090  1.103095506
#>         X222        X223       X224       X225        X226       X227
#> 1 -0.2773911 -0.25692844 -0.7385879  0.6983754  0.44223548  0.8307781
#> 2  0.5954355 -1.29408659 -1.6703511 -0.2961622  0.10891362  0.0255446
#> 3 -0.3992220 -0.62909416  1.0129484 -0.5762548  1.11253487 -1.0214373
#> 4  0.3080727  0.84407766  0.8539970 -1.6668577 -0.17763439  0.2670056
#> 5 -0.5356750  0.99475175  0.1079359  1.3038722  0.02212047  0.5870846
#> 6 -0.5265513 -0.03303848  1.7919034  1.6563863  0.46402765  0.9328896
#>         X228      X229       X230       X231       X232        X233       X234
#> 1 -0.9555524 -1.205082 -1.1778736  1.2902912  0.5930763  0.58947891 -0.4185251
#> 2 -0.1149316 -1.369845 -0.1261978 -0.2205265 -0.7542051 -0.20952921  0.9647016
#> 3 -0.6812422 -1.511931  0.5824585  0.9220910  0.2681167 -0.53521383  1.1122822
#> 4  0.1171932  0.430978 -0.2117386 -0.7977888 -0.2818034  0.29148694  0.1878026
#> 5 -0.1582310 -1.040655 -0.9857187 -1.0424897  0.5458696  0.08669764 -0.7646489
#> 6  0.9009423  1.135829  1.1862009 -1.3466614 -0.5524940 -1.71701461 -0.3755631
#>          X235        X236       X237       X238       X239       X240
#> 1 -0.47548670  0.09207527  1.0321383  0.7109934  0.8255713 -0.4161988
#> 2  0.09708553 -0.10060622  0.2701914 -0.5106745 -0.4990446  0.6077865
#> 3  1.67232247  1.28150472  0.4265391  2.1306169  0.4401761 -1.0177164
#> 4  1.29172327  0.68311842  1.0959355 -0.6455495 -1.4656157 -0.4836792
#> 5 -2.12465577 -1.63544875 -0.2615905  0.3557176 -1.0573382 -1.8392045
#> 6 -1.06937652 -0.42435266  0.1879172 -0.9513527 -0.9327161 -2.8464942
#>           X241       X242       X243       X244          X245         X246
#> 1 -0.446090839 -1.0805282 -0.9092361 -1.6639907  1.5280860215  1.165059297
#> 2 -1.179012597 -0.6486036  1.6575087  0.8533281 -0.2480232532 -0.281017406
#> 3  0.056787739  0.4660708  0.7920569 -0.2016572 -0.2675440236 -0.007079594
#> 4  0.231317028 -2.0428131 -0.8970532 -1.7477804  0.0003774286  0.360669773
#> 5 -1.451782882 -0.6574483  0.8572207  1.4421357 -0.1278732843 -0.467017730
#> 6  0.003980781  0.4876352 -0.7944051 -0.7507294 -0.0005605428 -0.011464237
#>         X247       X248       X249       X250        X251        X252
#> 1  1.5648209  0.4509246  1.4634077 -0.1211142  1.71149033  0.07985383
#> 2 -1.5421054  0.2197690 -0.7308549 -0.7502405  0.15854070 -0.72742036
#> 3  0.2470768 -0.2038317  1.7363132  0.2264742  0.04882068 -0.31916099
#> 4  1.1896208  0.5344679 -0.9737749 -0.5897848  0.83279035  0.48929303
#> 5 -0.1089504 -0.5134665  1.1345005  0.7821529 -0.39952527  0.90086046
#> 6  0.7232190 -0.6388596  1.0967814 -0.8455037 -0.43814614 -0.52235718
#>         X253      X254        X255       X256       X257       X258        X259
#> 1  0.2717248 1.4095470  0.76074253  0.6273689  0.1889082 -0.5752646 -0.49193892
#> 2  0.7822770 2.3707224 -0.04338935  1.4907733  0.1737879 -0.2581393 -0.29207877
#> 3 -0.2185956 0.1134090  0.95550923 -0.4989688 -0.5303747 -0.2860306  0.06942401
#> 4 -0.8625666 0.2757923  0.05287596  1.0300647 -1.0597124  0.2557628  0.13009938
#> 5  2.1441383 1.9934373  0.33568812  0.5082163 -0.5420507 -1.3319390  0.27100639
#> 6 -2.1451070 0.1863851 -1.91067830 -0.5084561  0.8357834  1.8363432  0.26475955
#>          X260       X261        X262        X263       X264       X265
#> 1  1.50426804 -0.6795187 -1.22809611 -0.03718369 -0.3264559 -0.2999336
#> 2 -0.38788461 -1.0422463 -0.64714988  0.98823686  1.7198862 -0.7140983
#> 3  0.18765592  1.2641661 -0.67490579  0.23799108  0.6453406  0.5395840
#> 4  1.27510592  1.5570782 -0.09340547  1.41612892  0.1714416  1.0571474
#> 5  0.90895935 -0.5442268 -1.96172639 -0.12881726  0.2547288  0.8865248
#> 6  0.02335469 -0.8096666  0.22269278  0.02637156  0.9448343  1.0769616
#>           X266       X267        X268       X269        X270       X271
#> 1  0.003081686  0.8200993 -0.12471743 -0.7145741 -1.00701023 0.65232271
#> 2 -0.314102392 -0.4968770 -0.99182042 -0.6622181  0.97953039 1.30947100
#> 3  0.115269591  1.3879201  0.02786066 -0.1391930  0.23631875 1.46251015
#> 4 -0.196603244 -0.5866012  0.94194731  1.1860374  0.19837258 1.89530543
#> 5 -0.540070946  0.9224373  1.53763541  0.6902027 -0.02973225 0.07745501
#> 6  0.442856331 -0.6664793 -0.19190381  0.6986218  1.04464523 0.76614132
#>         X272       X273        X274        X275       X276       X277
#> 1 -1.1173879 -1.0416226 -0.57047668  0.54998581 -0.2560796  0.5696192
#> 2  0.1744615 -0.9819244 -0.29957458  0.49458052 -0.3048442  0.4571912
#> 3  1.5361797  0.6233809  0.59445955 -0.63471049  0.3005541  0.6668729
#> 4 -0.4864288 -0.2101357  0.40095610 -1.44136126 -0.8428885  0.1201993
#> 5 -0.3364820  0.8257944  1.50205173  0.18667915 -1.3018290 -1.2637108
#> 6  0.3078152  0.4670742 -0.08586428  0.07076486 -0.1526438 -0.4835027
#>         X278       X279       X280       X281        X282       X283       X284
#> 1 -0.9857387 -0.4252669 -1.4239262 -0.4888123  1.07816951  1.4366102  1.4250752
#> 2  0.4468870  1.7089666 -0.2626601  0.6753299 -0.09909308 -1.4713333 -1.6957309
#> 3  0.1807285  0.9536175  1.4835165 -0.3624518 -2.51709089 -1.4431543  1.1978188
#> 4 -0.7327982 -0.6696208 -0.7188507 -2.5657035 -0.50018970 -0.9409508 -0.1563982
#> 5 -0.6065518 -0.9335962 -0.3022784  0.3514947 -0.46582125  0.8401051  0.0840493
#> 6  0.3686639 -1.2543946 -1.8207652 -0.6118929 -0.45083372  0.3778318 -0.2428725
#>          X285       X286       X287        X288       X289       X290
#> 1  0.92414675 -0.7698197 -0.3955329  0.79251091 -0.4093298 -0.2348374
#> 2 -0.52040194 -0.9344540 -1.2828504 -0.04452456 -0.5035118 -0.1381464
#> 3  0.06651053  1.3927606  0.4418992  0.39619207 -0.3988125  1.8209536
#> 4  0.07892625  0.4617982 -0.3667014  1.93649958  0.6407955 -1.0563229
#> 5 -0.84091575 -1.7331593 -0.5653215  0.55909267  0.1511628 -0.1535666
#> 6  0.26491273 -0.9541995  0.4982360 -0.89113160  0.8017731  0.5443078
#>           X291        X292       X293       X294       X295       X296
#> 1  0.638096349  0.95437992  1.0147087 -0.2518866 -0.8129511  1.1102229
#> 2 -0.079593603 -0.86266309 -0.1795000  0.4606381 -0.2779045 -0.8451810
#> 3  0.716452704  0.07894708  0.9361825  1.5091185 -0.3925423  0.3360614
#> 4  0.886627926  0.24074452  1.5068617  0.2706549 -0.2784989 -1.8826966
#> 5 -0.007792470 -0.59447771  1.1236162  0.3456584 -0.4967203  1.7157179
#> 6 -0.001143212 -0.05266597 -0.8562922 -0.0273284 -0.8854307  0.1995474
#>         X297        X298       X299        X300       X301       X302
#> 1 -1.1724988  0.32064134  0.1655166 -0.72301625 -0.7914319 -1.7125551
#> 2 -0.3083750 -0.90615966 -0.8848734  0.25998294 -1.2957401 -1.2030056
#> 3 -1.5242791 -1.17606005 -0.2201740 -0.34610036 -1.8233864  0.2754101
#> 4  0.7295904  0.86698162 -0.9255281 -0.21271207  0.1577303  1.7496093
#> 5 -0.6905727 -0.04485839 -0.2241721  0.09547892  0.2240036  0.2125692
#> 6  0.2915201 -0.58708321  0.8835390  0.11115831  0.7594624  0.5163813
#>          X303       X304       X305        X306       X307      X308       X309
#> 1 -0.84889060 -1.1056671  1.2893619  1.15003959  1.2758503 0.1722025 -1.9478631
#> 2 -1.09269370 -0.2102537 -0.5863680 -1.31524129 -0.7535874 0.4608043  0.6633460
#> 3  0.85881717  0.9960133  0.7910329  1.08015352  0.4415003 0.1419186  0.4626925
#> 4 -0.56082049 -0.5313922  0.4893983 -0.01119888  0.4134633 0.2628562 -0.8041215
#> 5  0.50186532  0.3506956  0.1886179  2.16271288  1.1397400 0.5029286  0.0741424
#> 6 -0.07733878  0.9661318  0.9633727  2.05352853  0.4076159 0.7326001 -1.2405169
#>         X310        X311       X312       X313         X314       X315
#> 1  0.4294622 -0.81132345 -1.3341147  0.1758190 -1.152785630 -1.4323026
#> 2  0.7397383  0.08981475 -0.6954377 -0.1379107  0.436968765  1.6189598
#> 3  1.7145412  2.01192847  1.6483425  0.3575146  0.003947486 -1.9399169
#> 4 -1.1868011 -1.30924887  0.1289110 -0.4734917  0.292836450  0.3719458
#> 5 -0.3941531  0.87887528 -0.2638047 -1.1299987 -0.177624340  0.5799660
#> 6  0.3634921  0.14836774  0.3046551  0.8504902  1.136742933  0.0173906
#>         X316        X317       X318       X319       X320        X321
#> 1 -0.6179543  0.07408107  0.3463980  2.0516207 -0.4380006 -0.85611960
#> 2  1.4360692  0.53274207  0.4333953  2.2956491  1.7609930  0.16436440
#> 3 -1.1665860  1.31520194  0.8759255 -1.3533704 -1.0336063 -0.71863681
#> 4 -0.7507503 -0.07156901 -1.7086663 -0.7578445 -1.1665401 -0.08723571
#> 5  0.9151231 -0.49758703 -1.4512523 -0.4320736 -0.1753365 -0.86362930
#> 6 -0.1436614  2.02140393 -0.3800417  0.2725962  0.3986918 -0.22941041
#>           X322        X323       X324        X325        X326       X327
#> 1  0.003669011 -0.29885484  0.0411156 -0.82127929 -0.19113281  0.2404461
#> 2  0.568034091 -0.08648794 -1.4714509 -2.84895046  0.36287327 -0.6339246
#> 3  0.748720920  1.19346500  1.0176908  1.01936711  0.07561448  0.0503839
#> 4 -0.473268771  0.57685611 -0.4027838 -0.06967238 -0.40981735 -0.4520294
#> 5 -0.395364534 -0.53357832 -1.0826945  0.01435656 -0.37647480  1.2664844
#> 6  0.810295974 -0.68384424  0.5075929  0.27801193  0.27418380 -0.2664559
#>         X328       X329       X330       X331        X332       X333       X334
#> 1  0.3750903 -1.4682586 -0.9403019  0.5791866  0.02529286  0.4390753 -2.0671503
#> 2  1.1184538  0.3580460 -0.4730398 -0.7694188 -0.48811227  0.1984883  1.4185576
#> 3 -0.5144728  0.5132462  0.2004251 -0.4229409 -0.76883909 -0.6634712  0.5418215
#> 4 -1.6997713  0.2223922 -0.2817547 -0.0683566  1.20135052 -0.2455727  0.8952226
#> 5 -0.1430625 -0.4320772  0.9195157 -0.1818547 -0.30775217 -1.0255751  0.3446567
#> 6 -0.5285199  1.4209136  2.5668882  1.4835136  0.48338203 -0.4864299 -0.2333702
#>         X335       X336       X337        X338       X339       X340
#> 1 -3.3299155 -0.3153476  0.9631795  1.63628310 -0.7037023  0.7343677
#> 2  0.2958380  0.8014809  0.9187234 -0.46850974  0.4816581 -1.6386941
#> 3 -1.1373595 -0.7448117  0.4949331 -0.18587551 -0.2058760 -0.6625784
#> 4 -0.7504140 -0.3928512 -0.2011249 -0.79619357 -2.0165907 -0.3509699
#> 5 -0.4505505 -0.8250626  0.4573457  0.07458883 -1.5174707  1.0661574
#> 6 -0.1146043 -0.2965659 -0.6023609  0.90159571 -0.1974580 -1.2575308
#>          X341       X342       X343       X344       X345       X346       X347
#> 1  0.59951032  0.5486772 -0.3713459 0.13789311 -0.5976858 -0.6032075  0.3789227
#> 2  0.01954649  1.9669821  0.3069960 0.18156429  0.5860366 -1.3732597  1.4336661
#> 3 -1.27638658 -0.1418469  1.3433848 1.80422925  1.3662370  0.7624197 -1.5120900
#> 4  0.47051024 -1.9938850 -1.2586535 1.29827593  1.4192739  1.6642974  1.6564255
#> 5 -0.40945121 -1.2011843 -1.4998317 0.39879432 -0.2419726  0.3673951  1.0514631
#> 6 -2.11429714 -1.3965165 -1.5369046 0.04895686  1.5822977  0.6314102 -0.6320098
#>         X348       X349       X350       X351       X352       X353
#> 1  0.5459117  1.4336203  1.0479862  0.1002950  0.8189928  0.8036642
#> 2  2.1958556  1.5644663  0.2091406 -0.7466460 -0.1062157  0.7385571
#> 3 -0.7030840 -0.6935721 -1.0542085 -1.0322914  0.1145898 -0.1290555
#> 4  0.5774167  1.3318005 -0.9604822  1.0464297  1.1191892  0.6822484
#> 5  0.3687354  0.8845233  1.1156784  0.5203774 -1.6124098 -0.1468312
#> 6 -2.2642765  0.1356856 -1.4979380 -0.7000794  0.7803482 -0.9857597
#>           X354       X355        X356       X357       X358        X359
#> 1  0.650172438  1.5441668  0.27716044 -0.6456186 -0.6451585 -0.01026877
#> 2  0.354384908  1.7145354 -0.07328055  0.7180643  0.3911936  0.95933018
#> 3  0.465510525 -0.3086968  0.39882468  1.5616868 -0.5301381  0.03501465
#> 4 -0.309916134  0.2444494  0.65076426 -0.5507225  1.6113915  1.48780991
#> 5 -1.890429950 -0.8973505  0.87362306  2.4748979  0.3130001  0.13960346
#> 6 -0.005460547  0.8926568  0.63188963  0.8991836  0.3492417  0.52001476
#>         X360       X361        X362       X363        X364       X365
#> 1  1.3154782 -1.1824371  1.23111993  0.2004018 -0.47308490  0.1583219
#> 2 -0.8977736  0.7777124 -0.24457617  0.4766562 -0.16970951 -1.2237888
#> 3  0.4658521  0.9685372  0.09195548  0.1710152  0.76050065  0.7471953
#> 4 -1.5296992 -1.4393639 -0.01481136 -0.4944536  0.15373421 -1.1354727
#> 5  0.5780527  0.1090853 -0.67269761 -0.2723852  0.07735436 -2.1043327
#> 6  2.0665792  0.6667466 -0.60066080  0.2742564 -0.44381421  1.2348583
#>         X366       X367       X368        X369        X370       X371
#> 1  0.3813031  0.6816790 -1.4418117 -0.81719146  0.83627354 -1.0487378
#> 2 -0.8790156 -1.7516818 -1.0463180  0.07874839 -2.35716138 -0.2531421
#> 3  0.2182264  0.3367305  0.2690043 -1.69206238 -1.58422041 -2.2465535
#> 4  0.1483050  0.3096659  1.0499963  0.17157581 -0.08082745  1.2432391
#> 5  0.1637860 -1.3448083  0.1332821  1.90383720  0.49305368  0.9270356
#> 6  0.2829927  0.9795570  0.9765076  1.59345655 -0.47325383  0.5883221
#>         X372       X373       X374       X375       X376       X377       X378
#> 1 -0.3970176  0.8343542 -0.9735985 -0.6493845 -0.3408729  1.3473207  1.6580412
#> 2  0.8623596 -0.5616262 -0.6632541 -0.2009445  0.5338768  1.3222688 -0.1428757
#> 3  0.4735766 -0.4400860 -0.8907809 -0.6722705  0.9677167  1.3674468  2.0308195
#> 4 -0.1545187  0.7137804  0.4405268 -1.3776398  0.5543288 -0.2251293 -0.4419080
#> 5  0.1802474  0.2446507 -0.8668774 -0.9298952 -0.4557426 -1.5024824  0.1617444
#> 6  0.9403137  0.8178667 -0.1847622  0.3354105  0.7523572  0.7481406  0.6937696
#>          X379        X380       X381       X382       X383       X384
#> 1  0.87798147  1.35065601  2.2747095  0.6836663  0.5042835 -0.9360393
#> 2 -0.55501165 -2.54397753 -0.1865400 -0.6306176 -1.2932957 -0.1022629
#> 3 -0.72027949 -0.01843769  0.6008408  1.3866068  1.0889796  0.1728094
#> 4 -0.07171085 -0.50744478  0.8285505  1.3071143 -1.4758800 -2.3576326
#> 5  1.23506121  0.75496394  0.4546464 -0.9073781  0.5040339 -0.4498264
#> 6  0.09507966 -0.08002738  0.1982719 -0.2083075 -0.3597844  2.2852991
#>         X385        X386        X387         X388        X389       X390
#> 1  0.2692607  1.12861733  0.32037895 -0.955370780 -0.42108521  0.3394762
#> 2  0.7748792 -2.30161712 -0.64341061 -0.338436342  0.43943055  0.8324055
#> 3 -0.3127288  0.03461603  0.02400878 -1.276754244  0.03635204 -0.1790587
#> 4 -2.6793571  0.38737193 -0.44696083  0.135103110  1.93183448  0.4721125
#> 5  1.1112379  1.69428712  0.21940568  0.396114914  1.23639748  0.4980284
#> 6  0.8017268 -0.49827822 -1.91278317  0.009956543  0.42402200  1.4345269
#>         X391       X392        X393        X394       X395       X396
#> 1  0.4523608 -0.3689498 -1.07203407  0.88853021  1.6764541  0.1426245
#> 2  0.8339089 -0.3168080 -0.88275631 -0.86868398 -0.0572841 -1.1655846
#> 3 -0.6197104  1.4363638 -0.07970779 -1.43804396  0.1692509  0.6612290
#> 4  1.0183184  0.0642016  1.18903561  0.06433664 -0.1403028  1.9256830
#> 5  0.2522367 -1.3574796 -0.67105286 -0.41930204  1.5291838  0.5369772
#> 6  2.1755599  1.4267112 -0.53169757 -1.19361828 -1.2116080 -1.3630243
#>          X397       X398       X399       X400        X401       X402
#> 1 -1.76606661 -1.0800043 -1.0618196  0.1083012  0.60159907  1.3001096
#> 2 -0.09468655 -0.2725320  1.5164964  1.6372888  0.35268086  0.9753610
#> 3  2.30649282  1.5114811  1.0757069 -0.7708041 -0.21146204 -0.2709782
#> 4  1.38756904  0.3373725  1.0284504 -0.1247479  0.05684237 -0.2349871
#> 5  0.61400047  0.5214531 -0.8645610 -1.4676599  1.13636305  0.5988430
#> 6 -1.15553397 -0.9889395  0.4326292 -0.8025603 -0.12227067  0.8100503
#>         X403        X404       X405       X406       X407        X408
#> 1  0.4337998  0.44181449  0.2791466  1.3161357  0.9945521 -1.60365658
#> 2  2.1938811 -0.43781335 -0.9359081  0.2304120  0.4028320  1.03694558
#> 3 -0.2467972  0.60131578 -0.2781792  0.3529103 -0.6100361 -0.20239671
#> 4  0.5173270 -0.04931615 -1.0356605  0.2531028  0.4651685 -0.85698650
#> 5 -0.4886954  0.28692789 -0.8482138 -1.2426445 -0.6263805  0.38793169
#> 6 -0.4584368  1.50581161  1.8655790 -0.1533048 -1.3995608 -0.01570867
#>          X409        X410        X411       X412       X413        X414
#> 1  0.48874921  0.56922093  2.88968193  0.9714101  0.4829644  0.02986614
#> 2  0.37239512  0.42689174  2.03801147  0.8337508 -0.1479745 -0.47249728
#> 3 -0.55673105 -0.20394087 -0.52476131  0.5396100 -0.8882600 -0.66274592
#> 4 -0.01295494 -0.17609604  0.05926883 -1.1654793 -1.0834149  1.54749655
#> 5 -0.16220419  0.09178461 -0.03574402  0.2428263 -0.0207459 -0.39143245
#> 6  0.39549058  1.05650749  0.79918176  1.9366040  1.3316131  0.10797974
#>         X415       X416       X417       X418       X419       X420       X421
#> 1 -0.4435294 -0.3507339 -1.2685397  0.7864395  0.3038254  1.4852754  1.6414172
#> 2  0.3636669  0.4835306 -1.3896243 -0.5233066 -0.1678615 -0.1613832 -0.2224204
#> 3  0.2394846 -0.7209744 -0.9293627 -0.2211532 -0.9022764 -0.7791730 -0.5522315
#> 4  0.4946996 -0.3414536 -1.8553304  1.2689673  0.1235262  0.3236443  0.5935021
#> 5 -1.5604383  0.4818520  0.2577484  0.5164234  0.3450566 -1.0179056 -0.6234578
#> 6  1.9484873  0.3605839  0.8641656  0.9289839 -0.3775872 -1.5137358 -1.7596241
#>         X422       X423       X424       X425        X426       X427
#> 1  0.8779762 -0.2042316 -0.7007522 -0.8465612  0.57267536 -0.5655257
#> 2  0.7547961  1.4325522  2.0982905  1.2094193 -0.05575363 -0.1948244
#> 3  1.6515923  1.1262206 -1.9607370 -0.2844329 -1.17881159 -0.9024416
#> 4  0.7134937 -0.1895489  0.2549576 -0.8812455  1.78770587  1.7570699
#> 5  2.6978193  0.3968773  0.8588648  1.1834865  0.44421659  1.9544028
#> 6 -0.1964024 -0.6758096 -2.3267311 -1.9068811 -0.74080905 -0.8249204
#>           X428       X429        X430       X431       X432       X433
#> 1 -1.098171024 -0.6617423 -0.44786757  0.9410642 -1.7445858 -0.6198548
#> 2  0.023174802 -0.2325852 -0.06337624 -1.6256906 -0.1192036 -0.7697216
#> 3  0.007796970  0.5770254  0.26977656  0.6898650 -0.3751476 -0.3281541
#> 4  0.715547937  0.6563119 -0.11989427  0.9828706  0.6074702 -0.2123378
#> 5 -0.256187680  0.8837284 -0.04125246  1.6607508  1.4780132  1.5696989
#> 6 -0.002428455  0.4379796  1.77695870  0.5300222 -0.2679364 -0.9572409
#>          X434       X435       X436       X437       X438       X439
#> 1 -1.50696311 -1.4801794 -0.6650920 -0.7996698 -0.6114844  0.1098349
#> 2 -0.43905065  0.2354773 -0.6420135 -0.9634598 -1.1780926 -1.8768639
#> 3  0.03897772 -1.4480435 -0.4785543 -1.8944486  0.9167825  1.4140121
#> 4  0.14122593  0.8904970  0.3721029  0.6051823  0.3219351 -0.6711732
#> 5 -1.02206297  0.4071490 -0.1882257 -0.7789361  0.9163526 -0.2109300
#> 6  0.05202875 -0.8064581 -0.2830392  1.1796410 -0.4685660  0.9595473
#>          X440        X441       X442        X443       X444        X445
#> 1  1.08919910 -1.59064206 -0.8558018  0.01043147  0.1921331  0.05525157
#> 2 -0.02342658 -1.46804135 -1.5751698 -0.37210448  0.8442296  0.88480530
#> 3 -0.21536198 -0.08684034  2.6099569  2.52576776  0.9905818  0.29382822
#> 4 -1.35572003  1.61956379 -0.9753212 -2.42324437 -1.4035838  0.18716283
#> 5  0.25110432 -0.40514471 -0.1916472  0.53584040 -0.3522198 -1.51101884
#> 6  0.30326045 -0.41332101  0.3505748 -1.73341139  0.2763448  0.21595282
#>         X446       X447        X448       X449       X450       X451
#> 1  1.4649891  1.5546400  1.14575970 -1.3126430  0.4664015  2.0136820
#> 2 -0.3575544 -1.3078710  0.44824260 -0.7281077  0.5462387 -0.6925865
#> 3  0.7346585  0.7754523 -0.26087874 -0.3798093 -0.7445731  0.4630689
#> 4 -0.1517274 -0.2036697 -0.75016626 -0.5386333 -1.2000319  0.7607299
#> 5 -0.4926257 -0.5481448 -2.19154194  0.8851557  1.3049053 -0.7970735
#> 6  0.6533369  1.3151126 -0.04077988 -0.4980739 -0.3266844  1.4025355
#>          X452       X453        X454       X455        X456       X457
#> 1  2.04347223  1.3469190 0.938394817 -1.1067303 -0.77051028 -0.9411762
#> 2 -0.59826228 -0.2459937 0.346604006  0.6589137  0.50699486 -0.4535064
#> 3 -0.03793485 -0.5844638 0.377004465  0.4372706 -0.03376785 -0.1304945
#> 4  0.43919634  0.4039223 0.301469088  0.6613826 -0.11062016  0.8277654
#> 5 -1.38412873 -0.5804096 1.187479015 -0.1074107  0.66020177 -0.7097091
#> 6 -0.99134941 -0.3476950 0.002907146  1.0857472 -0.40255888  0.6510277
#>         X458       X459       X460       X461         X462        X463
#> 1 -0.3419244 -1.2678488  0.5005845 -0.1169687  1.051249523  0.14765294
#> 2  0.3153275 -0.1055651  0.1155962  1.6979560  0.697217451 -0.18146185
#> 3  0.9914828  1.7075449  0.7428200 -0.2050241  0.163909438 -0.54424913
#> 4 -0.6476980 -0.2137932  0.8821430  1.2206484  0.003218964 -0.07991735
#> 5  0.1034929 -0.7401752 -0.5435101  0.6425253  0.824385633  0.48794575
#> 6 -0.5769825  0.3116790  0.3179102 -0.2021597 -0.415549002 -0.39543027
#>         X464       X465        X466       X467       X468       X469       X470
#> 1  0.8016550  0.3140042 -0.96234652 -2.3630587 -1.3917849  0.6172025  0.8633972
#> 2 -0.5160644  0.6229312  1.35239178  0.1132807 -0.6042615  0.2588579  0.3895141
#> 3  0.8776152  1.5649585  1.05800729 -0.7121260  0.1303582  1.2654335  1.5074861
#> 4  0.1480206  0.2439816  0.08254722  1.2150290  0.5840778 -0.1134858  1.1149434
#> 5  1.2153132 -0.9225858  0.51880801  0.1171082  0.8276566 -1.2900139 -0.7102652
#> 6 -0.1472049  1.1003717 -1.27829610 -0.9351825 -0.2009047 -0.8341980 -0.6675604
#>          X471       X472       X473        X474       X475       X476
#> 1  0.86475827  0.4147083  0.6824631  0.80031102  0.8152677  0.1882739
#> 2 -1.26820396 -0.1494749 -1.6517013  0.44774073  0.5687410 -0.6699747
#> 3 -0.79592694  0.9662241  0.1774535 -0.01107221 -0.3102342 -0.1491607
#> 4  0.29253614  1.2083093 -0.8425884  0.37490666 -0.8748780 -1.1418366
#> 5 -0.19431234  0.3434303 -1.2486058  1.14123469  0.4219296 -1.4170530
#> 6  0.06162091 -0.9792139 -2.0299093 -0.77116716 -0.2581275  0.2841326
#>          X477       X478        X479        X480       X481        X482
#> 1  1.41461106  0.7197726  0.92389425 -0.13226460 -0.9722145  0.04916306
#> 2  0.91867257  0.5468401  0.01912301  0.47938245  2.2578412  0.13385294
#> 3  0.01384019 -0.2199067  0.73209287 -0.91969345 -0.7179875 -0.43790873
#> 4  0.31625576  1.3899980  0.53932534 -0.43044283 -0.4331077 -0.11401388
#> 5 -0.04095310 -0.7615276 -0.46092891  0.87183681  0.1653446 -0.49929972
#> 6  0.18265805  0.4564953 -0.27752388  0.04505884 -0.4374218 -1.56704674
#>          X483       X484       X485        X486       X487        X488
#> 1 -0.71601051 -1.3222170 -0.9541601  0.78966580 -0.2923563  0.03401479
#> 2 -1.01680526 -0.9282225 -1.0660104 -0.09142612  0.1985304  2.45302814
#> 3 -1.39660562 -0.6599578  0.5959916 -0.06342731  1.2201527  0.40818927
#> 4 -1.45724324  0.1783025 -0.1713405  0.21379103 -1.0545969 -0.31128749
#> 5 -0.64647711 -0.7270700  0.1674159 -0.11117956  1.3618884  0.79847404
#> 6  0.07911676 -0.3866911  0.2679059  0.12913141 -1.4285903 -0.76998627
#>         X489       X490       X491        X492       X493       X494       X495
#> 1 -1.4637377 -0.2007863  0.1831389  0.04646011  1.3393101  1.8768860 -1.0766492
#> 2  1.1521711 -0.3229562 -0.1715151  0.84022734  2.3291745  1.1370344  0.2472020
#> 3 -0.9363256 -2.5579186 -0.5083217 -1.42629028 -1.1354588 -0.1305529 -1.9648540
#> 4  0.8785165  1.9732152 -0.0966934  1.33892187 -0.9011467 -0.7206268 -0.2707169
#> 5  0.3621652 -2.1408266 -2.0652674 -1.37429057 -1.8440738 -0.4491974 -1.4225536
#> 6 -0.5570056  0.7649388  0.5934545  1.72591264  1.3376593  2.0751997 -2.1989572
#>         X496        X497         X498       X499        X500
#> 1 -0.3330593 -0.04195416 -0.734061297 -1.5431765 -1.17832002
#> 2 -0.9319084 -0.88320479 -0.247050083  0.2909866  0.32179805
#> 3 -1.4669536  0.30178918  0.629556067  0.8273646  0.66573902
#> 4 -0.7234366  0.43784434 -0.025229322 -1.6643781 -0.04430914
#> 5  0.6355531 -0.17067388 -0.007289339  2.9353461  1.13973140
#> 6 -3.3443856 -2.46268140 -0.093005237  0.3575185  1.19402499