A high dimensional dataset with n equals to 200 and p equals to 500.

A dataset with 200 observations and 500 dimensions is generated from the following process: linear regression model with only first three non-zero coefficients equal to 3, 2, and 1.5 respectively. The covariates are correlated with AR structure (rho=0.3). The error term is normally distributed with zero mean and sd equals to 0.5.

data(HighDim)

Examples

# HighDim dataset is generated from the following process:
n <- 200; p <- 500; sigma <- 0.5
beta <- rep(0, p); nonzero <- c(1, 2, 3); zero <- setdiff(1:p, nonzero)
beta[nonzero] <- c(3, 2, 1.5)
Sigma <- 0.3^(abs(outer(1:p,1:p,"-")))
library(MASS)
X <- mvrnorm(n, rep(0,p), Sigma)
error <- rnorm(n, 0, sigma)

X <- apply(X, 2, scale) * sqrt(n)/sqrt(n-1)
error <- error - mean(error)

Y <- X %*% beta + error
HighDim <- data.frame(Y, X)
head(HighDim)
#>            Y          X1         X2           X3         X4           X5
#> 1 -4.9273248 -0.49648759 -1.1591004 -0.870994528 -0.8912255 -0.583646997
#> 2 -1.2610797 -0.36116851 -1.2675696  1.459497537  0.3129178  0.045758758
#> 3  5.0252230  0.73962421  1.2498021 -0.003536993 -0.4089725  0.022661389
#> 4 -2.7734547  0.21671622 -1.1781131 -1.179565318 -0.3885327 -0.002958705
#> 5  0.5691676 -0.05575727  0.1006341  0.583144432  0.6437334 -0.363883583
#> 6 -1.9566591 -0.22514722 -0.5237663 -0.519533259  0.4634841  0.811691763
#>           X6          X7         X8         X9        X10        X11
#> 1 -1.4398373  0.31171853  0.8939405 -0.2310409 -0.6252843 -0.2102258
#> 2  1.0440058 -0.52585391 -1.2503146 -0.1076673  0.1925203 -0.9501790
#> 3  0.4758118  0.81944286 -0.3692389 -1.8819081  0.1554386  0.1009003
#> 4  1.6449230 -1.03280112 -0.7303120  0.5749514  0.1826491  0.6972098
#> 5 -0.4356742 -1.05185004 -0.7409355 -0.6764412  0.5497792 -0.1110020
#> 6  0.8937378 -0.01561612  1.1439277  0.8492386 -0.1251891  0.1779967
#>            X12        X13        X14          X15        X16        X17
#> 1  1.329175899 -0.3040962 -0.1813565  0.136443641 -0.8116860 -0.3869187
#> 2 -0.355228596  0.9134739  1.2497592  1.882074712  2.4654404 -0.4634860
#> 3 -0.006870159 -0.6386774  0.2737667  0.008671303  2.1296087  0.1708272
#> 4 -0.081156752  0.3517628  1.5521297  0.123937208 -0.8078996  0.5056432
#> 5 -1.070434706 -0.3496775 -0.2954134 -1.208941095  0.9662667 -0.4707201
#> 6  0.354479083 -0.6562447 -1.3864081  1.094383791 -0.7707564  1.3242252
#>          X18        X19        X20         X21         X22        X23
#> 1 -0.5921660  0.6911438 -0.9346142 -1.59095477 -1.27600496  1.0646855
#> 2  1.1669643  1.0203620  1.4296178  0.03862881 -0.90813392 -0.6070295
#> 3 -0.0381209  0.2072237  0.8366009  1.48069143  0.08731648 -1.7106616
#> 4 -1.1984159 -0.1077866  1.1424656 -0.15734852 -0.42183262  1.3449080
#> 5 -0.5444763 -1.0524784 -1.9232822  1.03393410 -0.45610803 -0.1146577
#> 6  1.2104153  0.9564866 -0.6126311 -0.54166928 -0.84525994 -0.6795648
#>            X24        X25        X26        X27         X28        X29
#> 1  0.003082204  1.9285962 -0.5102627 -0.4280794  0.65296695 -0.2619373
#> 2 -1.632509540  0.1366045  0.4801924  0.5698933 -0.60674129 -1.3795448
#> 3 -1.653401673 -0.2741195  2.2685732  0.6663376 -0.20736891 -0.7774179
#> 4  0.143197570 -1.3765840 -0.5430574  0.2914722 -0.69082579 -0.2128864
#> 5 -0.567658698 -0.4603565  0.9729878 -0.2796950 -0.27134744 -0.8156245
#> 6  0.368304499 -0.4754349 -2.4998933 -0.5220479  0.05764434 -0.7791423
#>           X30         X31        X32        X33        X34        X35
#> 1 -1.12380494 -0.32213214 -0.2847857  1.0183655  0.2484291 -0.1039039
#> 2  0.04507557 -1.31480225  2.0213229  0.4749670 -0.6283638  0.9812198
#> 3 -0.64446616 -1.23352156 -0.4750161  0.2386523 -0.3924472 -0.1537918
#> 4 -0.72225434 -0.06198977 -0.4869046 -0.6973763 -0.1056404  0.9274824
#> 5 -0.18391725  2.32134466  1.4236692 -0.9412451  1.2474581 -0.5047275
#> 6 -0.65529623  1.03669675  1.8229605  1.4779153 -0.8861377 -0.7394729
#>           X36         X37        X38        X39        X40        X41
#> 1  0.78241240 -0.08723399 -1.2621521 -1.4137301 -1.6707348 -0.9455960
#> 2  1.04936236 -1.21030008 -0.6699273 -0.3318908 -0.7853797 -0.5453972
#> 3  0.07452355  2.20007378 -0.4939446 -2.4062668 -1.7409998 -0.1653565
#> 4 -1.25577348  0.14883005  0.2374338 -0.7505668 -0.5088937 -0.4969846
#> 5 -0.36072708  0.19571735  0.4296940 -0.1203030 -0.4499516  0.6883464
#> 6 -1.34895697 -0.47290047  0.9787689  0.2341223 -0.5438877 -0.0751496
#>          X42         X43        X44         X45        X46        X47
#> 1 -0.5564008 -0.35950654  0.4630894  0.04630753 -1.1725005 -0.3017665
#> 2 -1.2134763  0.38340754  1.2102387 -0.01175446  1.1880954  1.1096764
#> 3  0.9640938 -0.02887584 -0.5621974  0.27255293  1.0461598 -0.6696450
#> 4 -1.6663574 -0.28073087 -0.1891644 -0.10285160 -0.6852228 -0.7971429
#> 5  0.2968590 -2.04396889 -0.2020791  1.48079169 -0.6434604  0.2841301
#> 6  0.8775990  0.43398457  1.0504330  1.53574370  1.9937961  1.1718557
#>          X48         X49         X50        X51         X52        X53
#> 1  0.3169005 -0.05937712  0.01726557  2.0692924  0.82478209  1.3698077
#> 2  0.4163126  1.70599470  1.67484387  0.8906340 -0.03657716 -0.2029179
#> 3 -1.2673351 -1.26682150 -0.57424230 -0.7061492 -0.82740805 -1.7935417
#> 4 -0.5874554 -0.67053713 -0.31631230  0.2889590 -0.99720685 -0.7864377
#> 5  0.3992232 -0.25042006 -0.22199939  0.3033887  0.49041338  0.3797652
#> 6 -0.8746885 -0.45125962 -0.57404274 -0.2098010 -1.96428899 -1.1809669
#>          X54        X55         X56          X57         X58        X59
#> 1 -0.6215574 -1.1941590  0.15668526 -0.008274775 -0.07393536 -0.1579171
#> 2  0.5411796  0.8228635  0.94135383  1.166542636 -1.24189652 -1.1065777
#> 3 -0.5822465 -0.4294952 -1.42135937  0.183693594  0.52856793 -0.7532188
#> 4 -0.3245136 -0.3647888 -0.99930580 -0.256343630  0.81610454  0.4621033
#> 5 -0.1578448  1.1075763  0.07436446 -0.950176676  0.70010321 -0.1400980
#> 6 -1.6492282  0.7704777  0.69435397 -2.276659649 -1.75385158 -0.6228121
#>          X60        X61        X62        X63        X64        X65        X66
#> 1 -0.3066131  0.4770152  0.6302320  0.3500566  0.7461710  0.9347110  0.4864870
#> 2  0.2608677  1.7335383  0.7200766 -0.3191901 -1.6453476 -1.6268723 -1.1272550
#> 3 -0.4981837  0.5387920 -0.9723045 -0.1428871 -1.6180520 -0.6823191  0.2728577
#> 4  0.1780658  0.1508320 -0.1946723 -0.8372448 -0.4674779  0.1148807  0.6287614
#> 5  1.5550172  2.7354532  3.1149868  0.9767710 -1.3062907 -0.6374545 -1.2939277
#> 6  0.0540609 -0.3170400  0.9464045  0.6668438 -0.7972283  0.1083628  0.7379949
#>          X67        X68        X69         X70        X71        X72        X73
#> 1  0.6425458 -0.4464285 -0.8142732 -0.16631425 -0.6182362  0.8824326 -1.0697768
#> 2 -2.1452088  0.5596092  1.8481922  2.54380939 -0.0966148  0.8591654  0.9821257
#> 3  0.2784739  0.8931323  1.6613485  1.61832353  0.9699362  0.5433249 -0.2744027
#> 4 -2.4233190  0.6052827 -0.6994357  0.03356792 -0.5919741 -1.2551736 -0.9098053
#> 5 -0.6718538 -1.4817966  0.2135353 -0.24954118 -0.5434415 -0.7259506 -0.6794089
#> 6  0.8148224 -0.5823466  1.3321719  0.37289205  0.7045341  0.3080297 -0.4796853
#>          X74        X75        X76        X77           X78        X79
#> 1 -1.4963171  0.7712160  2.3239231 -0.2590156  0.1200812462  0.1661376
#> 2  0.4824306  0.3740681  1.4309075  0.1403800 -0.6234929576  1.4988323
#> 3 -0.8426662 -1.2247375  0.1231684 -0.7757135 -0.0007012895  1.6838071
#> 4 -0.4591479 -1.2897341 -1.1850009 -0.9045393 -2.2770554282 -0.2813189
#> 5  1.3679609 -1.0786707  1.1932443 -0.8955444 -1.0687339185  0.4136876
#> 6  0.7412685  0.6808058 -0.1953919  0.8646249 -0.8678531527 -0.4926206
#>           X80        X81        X82        X83        X84        X85        X86
#> 1  0.09431312 2.74879242  1.6053838  0.7736926  0.5500681  0.1614981  0.9195872
#> 2  1.56633574 0.04724317  0.7791985 -0.2941125 -0.3890119 -0.3621824 -0.9217389
#> 3  0.89600229 1.33578484 -0.5087224 -0.9914317 -0.2684780 -0.1344933  2.0947077
#> 4 -0.28549649 0.31214772  0.3051674 -0.8287112 -1.2255758 -0.9582038  0.3501407
#> 5  0.48371335 0.33945339  1.3175813 -1.0806078 -0.5121008  0.4409943  1.4190383
#> 6  1.71789308 1.01022065  0.3821983  0.4450647  0.3594967 -0.5401892  2.2694045
#>          X87           X88        X89        X90         X91        X92
#> 1 -0.4949303  7.730586e-05  1.3471183 -0.5872528 -1.15738162  1.0706338
#> 2  0.8699408 -1.548376e+00  0.2128917 -1.4234226 -0.02498938  0.3448398
#> 3  0.1797124 -8.794240e-01  1.8625143  0.9997122  1.10101600 -1.4000174
#> 4 -0.3522095  5.177157e-01 -0.1375804 -0.3134841 -1.67626195 -1.4858785
#> 5 -0.6087617  8.358448e-01 -0.4734458 -1.0353737 -1.52912086  1.1413796
#> 6  1.7504181 -2.717530e-01 -0.1994842 -1.2404751 -0.01626724 -0.5790771
#>           X93        X94        X95         X96        X97        X98
#> 1  0.76458388 1.24169690 -0.1779951 -0.37706275 -0.9038943 -1.3426272
#> 2  0.48346253 1.41192147 -0.6228428  0.01042669 -1.6430996 -1.7225313
#> 3  0.21730469 1.66518177  0.6761559  0.38704971 -0.2440572  1.1349362
#> 4 -0.33217041 0.08892574  1.6273799  1.03128358 -0.8508194 -0.1421802
#> 5  0.72295402 0.82160777 -0.5815807 -1.04894548 -1.0725273 -0.9959160
#> 6 -0.07072461 0.29642059  0.6442044  1.22185762  0.9831620  0.2117524
#>          X99         X100       X101       X102       X103       X104      X105
#> 1 -1.7743005 -2.125174683 -1.1594064 -0.9875204 -0.6859197  0.8427068 1.0678358
#> 2 -1.9534216  0.328380953  1.4423795  0.6134761  1.5131130  0.3963897 1.1107564
#> 3  0.3118511 -0.054133100  0.5433136  0.8733282  0.1220558  0.5064504 0.8214291
#> 4 -0.4299334 -1.256921837  0.5016600  1.2320924  1.5091691  2.8207978 1.2605382
#> 5 -2.1852362  0.921799755  0.0311615 -0.5741090 -0.7640833 -0.5806215 1.4695649
#> 6  0.8897174 -0.009611792 -0.5680940  1.2063704 -0.7592012  2.0858136 0.3486389
#>          X106       X107        X108       X109        X110       X111
#> 1  0.87100657  1.7350068  0.15694638 -0.7428152 -1.41984867  0.6341924
#> 2 -0.33182339 -0.1702302 -0.09876249 -0.7938108  0.01305360 -0.5492613
#> 3  0.03942009 -1.6124811  0.34401460 -1.1870476 -0.86428790 -1.3301326
#> 4  1.98818242  0.1867846  0.52926796  0.3838324 -1.11887912 -0.2569091
#> 5  0.90989842 -0.7751210  0.46532331  0.3080411  0.10492361  0.7950255
#> 6 -1.62124637 -0.4574410  1.20852522  1.9099887 -0.05971252  0.8985425
#>         X112       X113        X114       X115       X116       X117       X118
#> 1  0.2210882 -0.5528232 -0.04339463  0.2007938  1.0455908  0.4131344  0.9367544
#> 2 -1.0447656 -0.1364327 -2.80651036  0.5158493  0.7061772  0.5812050  0.2335870
#> 3  0.3778894 -0.1212367 -0.18455588  0.8561621 -0.6928575 -1.8360838  0.3616210
#> 4 -0.3472715 -0.2882350 -0.54640931  1.2019899 -0.2752422 -1.5061752 -1.0555613
#> 5 -0.6983910 -1.3991756 -0.50671449 -2.4953709 -0.8192371 -0.6818661 -3.2192040
#> 6  0.3672422  0.9209373 -0.49331931 -1.2547270 -0.3349151 -1.1888534  0.3072596
#>          X119       X120       X121       X122        X123       X124
#> 1  0.07171699  0.8970112 -0.8978625 -1.7462091  0.58632319  1.6629532
#> 2 -0.93638511  1.4434578 -0.5492204  0.2306793  0.07251624  0.1115802
#> 3 -0.59411141 -0.9643927 -1.3264685 -1.5726429 -2.00871127 -0.3763420
#> 4 -1.16623521 -0.5422641 -0.5872903 -0.3654744  1.83470031  0.4262813
#> 5 -2.29153011 -1.2680558 -0.3722150 -0.1045667  0.17579482  0.5167327
#> 6  0.80049578 -0.0769994 -0.4739097  0.3747454 -0.58940219 -0.6429317
#>         X125       X126       X127       X128       X129       X130        X131
#> 1 -0.4151444 -0.5150503 -1.1622176 -0.4850252  0.1349663  0.2844644 -1.02491867
#> 2  1.1920753  1.0801408 -1.3849975 -0.4819368  0.3177237 -0.3520701  0.10206505
#> 3  1.3957394  0.4623026  0.2203761 -1.9527338 -0.8904578  0.8085615 -0.96546196
#> 4  2.1383530 -0.3700620 -2.6878757  0.5130652 -0.1333342  0.3173922  0.03279776
#> 5 -1.7222664 -0.8918018 -0.2712474 -0.9660878  0.2806935 -0.6676538 -1.72622393
#> 6  0.7551668  0.2180316 -0.9475919 -2.4546919 -0.9586756 -1.3482553  0.49423471
#>         X132       X133       X134        X135       X136        X137
#> 1 -0.3227568  0.3098919  1.1120260 -0.53289146 -0.5671472 -0.32764132
#> 2  0.1548129 -1.5727002 -0.9522641 -1.14207394 -0.9050179 -0.49570072
#> 3  1.0642469  0.2449023 -0.9531774  1.03728172  1.7983316  1.40468463
#> 4 -0.5085839  0.9741116  0.2329341  0.03484529 -0.2831916  0.73603352
#> 5 -1.8989073 -0.5871001 -0.2588232  0.01634346 -0.8421339  0.07114411
#> 6  0.2777448  0.6169768  0.7554989  0.39652815 -1.6351336 -0.78830954
#>         X138       X139       X140       X141        X142       X143      X144
#> 1 -1.0804329 -0.4142851  0.6882319 -0.8387884 -0.13716971  0.8690964 0.3262188
#> 2  0.8392872  0.1583544  0.0756185 -2.3515142 -2.08194677 -0.9497215 0.7034796
#> 3  1.9083090  2.1555843 -0.5637997 -0.4358883 -1.25564244 -1.5808785 0.4865763
#> 4  2.1700281  2.1953726 -1.9025652 -0.6516647  1.91343293 -0.5211674 0.4984544
#> 5  0.9768940 -0.2047704 -0.2489122 -0.3978189  0.02808872  1.0553315 0.4049841
#> 6 -0.4026694 -1.1938825  0.7725758  1.6450310  1.08463121  0.6291419 2.1885154
#>          X145       X146          X147        X148       X149       X150
#> 1  0.35184576  0.7012056 -0.0009909228  0.19014987  1.0513381 -1.0375492
#> 2 -0.07312375  1.2405259 -1.7332100239 -0.27850092 -0.8978032  0.5659009
#> 3  0.77992682  0.3375340  1.2320077924 -0.33621746  0.4852854  1.5643719
#> 4  0.71171496  0.6467647 -0.2595551951 -1.57635162 -2.3280934  0.1588430
#> 5  1.87189608  1.4330349 -0.6441664420 -0.01294793 -1.0104064 -0.2223527
#> 6 -1.06331135 -0.7970938 -1.2870205343 -0.53844061 -0.5136115  0.2510737
#>          X151        X152       X153       X154       X155       X156
#> 1 -0.14051661 -1.02600635 -0.2989478 -0.3005382 -1.1843376 -1.2263920
#> 2  0.78402748  0.01836247  1.0115295  0.3119908  0.4822336 -0.4934275
#> 3  1.82330670  0.97866229  0.1585152 -0.2536805 -0.9186793  0.4924438
#> 4  0.84091996 -1.23564225 -0.7058659 -2.1628626  0.1004068 -1.1860122
#> 5 -0.24480074  0.44953973 -0.1860199  0.1182236 -1.9032818 -0.6587135
#> 6 -0.01651935  0.51828088  0.1912833  0.2118959 -0.6450140 -0.4362218
#>          X157         X158       X159       X160       X161       X162
#> 1 -1.61126177 -1.475677070 -0.2272143 -1.0880530  0.8653866 -0.3482734
#> 2  0.55646665  0.102821542 -0.7065599 -0.1404028 -1.7310148 -1.5535410
#> 3  1.93698581  1.226896285  0.7106710 -1.1643333 -1.1664873 -1.2598300
#> 4 -0.67401940 -0.009855193  0.5924932  0.2625351  0.6459550  0.1336541
#> 5  0.93197951  0.545204804 -1.1812253 -1.7776705 -0.1368919  0.9766531
#> 6 -0.06069121  0.749805820 -0.4170785 -0.3883202 -0.3050844 -0.9936232
#>          X163         X164       X165       X166       X167       X168
#> 1  1.41831149 -0.565161823 -1.8559914  1.2013906 -0.5195191 -0.9020094
#> 2  3.41692490  2.220487250  1.3138291  0.4303510  1.7128350  0.9410181
#> 3  0.85439224 -1.169260371  0.6194979 -0.1650444 -0.5103291 -0.4717351
#> 4  0.80572101 -2.004720901 -0.5229508 -1.0042187  0.5946832  0.4489574
#> 5 -1.00004956  0.002231791 -1.8844740  0.1503315 -1.1732818 -0.5448116
#> 6 -0.05338011  0.340275220 -1.3972249 -1.2726563 -2.1213228 -1.0337179
#>         X169        X170        X171        X172        X173       X174
#> 1 -0.5426382  0.22586151 -0.29523194  0.01262964  0.21395064  2.4716377
#> 2 -0.2094410  0.25936062 -1.79403857 -1.38325864 -1.89883156 -0.5339121
#> 3  0.5481280 -0.04522522  0.04905495  0.94683778 -0.91468888 -1.1647324
#> 4  1.0203570  1.56376435 -1.08611926  1.57310422 -0.02815447  0.2334637
#> 5 -1.1778316 -0.72746992 -1.19542942 -0.85311726  0.08251165  0.1058618
#> 6  0.7684614 -1.05327106  1.33050205  0.55915880  0.50745885  1.8763995
#>         X175        X176       X177        X178       X179        X180
#> 1  1.8356082 -0.08323486 -0.3634940 -0.64191754 -1.4413995 -2.61512005
#> 2  0.5671420 -0.64314035 -1.5791212 -0.67684596  0.5076676  2.34916147
#> 3  0.3783227  0.18083448 -1.2960309  0.04817141 -0.9587185 -0.02364759
#> 4  0.4958413  1.06453679  0.6275103  0.79446602  0.9073389  1.66561523
#> 5 -0.8738423 -0.01219826 -0.3738307  1.89062962  2.1874426  0.26982553
#> 6  1.8100507 -0.57651142  0.1652460  1.02817351  0.1794698 -1.22804361
#>         X181       X182       X183        X184       X185       X186       X187
#> 1  0.4352698  2.6095045 -0.8372008 -0.15149336 -1.4543032 -2.4447310 -0.1067977
#> 2 -1.4069799 -0.6374038 -0.6904436 -0.84277477  0.2237251  0.3553767  0.2732787
#> 3 -0.4418813 -0.6671420  1.2346781 -0.07803611  0.3213119  0.5184638  0.8381045
#> 4 -0.2660560 -0.1741280  0.2199450 -0.28865051  0.1815547  1.7673558 -0.4771178
#> 5 -1.3936481  1.7197411  0.1246965 -1.09915139  1.2212904  0.1330408 -0.1880631
#> 6 -0.3921859  0.8530683  0.6164930 -0.30760174  0.4831584 -0.9090413  1.3995609
#>         X188         X189       X190       X191        X192       X193
#> 1  0.5091969  1.056378635  0.5339734  1.2520857  0.44426398  2.2776033
#> 2 -0.7241337 -1.086033733 -0.5050559 -0.1549469 -1.17940421  0.2501638
#> 3  1.4564332 -0.005814349 -1.6274730 -1.9214161 -1.10514631 -1.7975190
#> 4  0.2937080  0.431954497 -0.4145344  1.0622399  0.06163781 -0.2593276
#> 5  0.2624940  0.864394548 -0.4577136 -1.2612362  0.63905996  2.8407894
#> 6  1.4381762  1.225311574  0.5094783 -1.6240214  0.38337736 -2.2030780
#>         X194        X195       X196       X197        X198        X199
#> 1 -0.5609339  0.86152139  0.7547062  0.5349041 -0.31705468  1.55266031
#> 2  1.1319823  0.01551378  0.8963109  0.7017396  0.55978569  0.39954391
#> 3 -0.2374337 -0.50553547  0.5198748  1.6343084  2.43951057  2.61981162
#> 4 -0.5760806 -1.83542392 -0.5254673  0.1409647 -0.08713788 -0.44967037
#> 5 -0.0801304  0.34400490 -0.2666633 -0.9278147 -0.45106657  0.50964275
#> 6 -2.3114059 -0.93624625 -0.1752779 -0.9054463 -1.49802461 -0.02070402
#>         X200       X201        X202      X203       X204       X205       X206
#> 1 -0.2603372 -1.0556679 -0.08389173 0.2066671  0.9473283  0.4534112 -0.5180714
#> 2  0.1773504  0.1236748  2.76332641 0.5078841 -1.4086909  1.3573166  0.9634645
#> 3 -0.1037677  1.3509597  0.73885965 0.2578502 -0.3700185 -0.6793541 -0.4734766
#> 4 -0.1130070  0.5007018 -0.42126706 2.6510592  1.4848892 -0.3404573  1.8073061
#> 5 -1.0144762 -1.0996940  1.93780445 0.2313786  0.5875465  1.0918579  0.2278973
#> 6  0.8487598  0.8928889  0.21263082 0.3931321  1.3974484  1.7737415  2.1823322
#>         X207       X208        X209       X210       X211       X212       X213
#> 1  0.5204887  0.5473469 -0.18139170 -0.1764614  0.9302595 -0.2421856  0.2921473
#> 2  0.2687328 -0.1182517 -1.47566579 -0.1938265  0.9425538  0.8000938 -0.4689394
#> 3  0.9835279  1.0955067  0.29645972  0.8119736 -0.2696041  0.3962050 -1.0831374
#> 4  1.8933097 -0.4260567 -0.03273133  1.1504336 -0.5972821  0.1258906  0.8630214
#> 5 -0.9435005  1.1468915 -1.66849064  0.4295248  0.8311693  1.1254772  0.2562097
#> 6  2.0846702  0.3015547  0.81684118  1.0045451 -0.8408250 -0.4860529  0.2694896
#>          X214         X215       X216       X217       X218       X219
#> 1  1.43660341 -0.469122178 -1.4283960 -1.7277940 -1.2085012  0.1397868
#> 2  1.93066795  0.689066625 -0.1918024  0.3292692  0.7762449 -0.4340287
#> 3  0.55013452  0.089497040 -0.1206421  0.2503283 -1.7894180 -0.4253034
#> 4 -0.84480003 -0.391474025 -0.1135343  0.3677299 -0.8769060 -0.2328947
#> 5  0.49119596  0.005107677 -0.2599375 -2.0392888 -1.5106481 -0.1872898
#> 6  0.02395157  0.096657311  1.6739946  0.6057611 -1.4186911  0.7789866
#>         X220      X221       X222       X223       X224        X225        X226
#> 1 -1.6800499 0.9149331  0.6073957 -1.0035478  1.1697768 -2.36171111 -0.75597652
#> 2 -0.1661750 1.3570668 -1.0862059 -0.7780847  0.6186510  1.26742800 -0.61529530
#> 3  0.6097635 0.7660990  1.2130619 -0.2307803  0.2771424  1.28176030 -0.50880692
#> 4 -1.1921539 1.3836597  0.3953304 -0.6880851 -1.8844241 -0.28333136  2.18808359
#> 5 -0.8096839 1.3182372 -1.1122187  0.5267791  0.4304798  0.02171862 -0.02268314
#> 6  2.3034464 0.9687388  1.9956143  1.1938149  0.4726268  0.30764820 -1.08208336
#>         X227       X228        X229        X230       X231       X232
#> 1  0.7389265  0.8303521  1.00882848 -0.17126451 -0.1893644  0.6309058
#> 2  0.1839954 -1.0100247  0.30444354 -0.68591353  2.2053239 -0.2727257
#> 3 -1.9257363 -0.4214521 -0.24654282 -0.06837505 -1.1802059 -0.5338109
#> 4  2.3858832 -0.4521942 -1.37819497 -0.96106930 -0.3953473 -0.4391334
#> 5  0.7206735 -0.5381337 -0.04342397  0.15639529 -1.5429918 -0.6847918
#> 6 -0.4316196 -0.2658457 -0.57035129  2.41885580  0.2059356  1.2383982
#>         X233       X234       X235       X236        X237       X238
#> 1 -2.2642405 -0.7272746 -1.2545500 -0.3298849 -0.74537344 -1.0286922
#> 2  1.2597350  0.7181977  0.2202849 -1.2871513 -1.52228961 -0.2473848
#> 3 -0.2153468 -0.1868743  0.5526952  1.3160333  0.07896817  0.4917521
#> 4  1.4279732 -0.3696765  0.4900377 -0.9586010 -0.32445385  2.6190973
#> 5  0.8704468 -0.6002690  1.0578494  2.0552731  0.43696480 -0.1145005
#> 6 -1.2095690 -0.4065337 -0.4640055 -0.8194820 -1.71557283 -0.3710854
#>          X239        X240      X241       X242        X243        X244
#> 1 -0.83607471  0.22742478  1.054044 -0.7994717 -0.07587895 -0.14250935
#> 2 -0.03004157 -0.45515896 -1.083328 -0.7684864 -0.91762224 -1.05615057
#> 3 -0.33448363 -0.32822565 -1.330762  0.4640866 -1.71181250 -0.06292688
#> 4  1.13846847 -0.04540536 -1.362356  0.2287054  0.19758490  0.76918105
#> 5 -0.56397972 -0.70002150 -1.652079 -0.7952006 -0.75625783  0.05086555
#> 6 -1.58362685 -1.35357079  1.736421 -0.2869807 -0.74061891  1.62528575
#>          X245       X246       X247        X248       X249       X250
#> 1  1.45583155  1.2737868  1.8896406  0.68880538  0.1633289  0.1405859
#> 2 -0.22445094  0.1362006 -0.3683612 -1.75699865 -0.8180771 -0.5139959
#> 3 -0.07555078 -2.0994047 -1.8615169 -1.26226997  0.4219700  0.3878744
#> 4 -1.78027904 -3.7991287 -1.3073235  0.06234095  1.9900436  0.5588631
#> 5  0.14801825  1.0327783  0.1862932  1.04191186  0.8244050 -0.1412361
#> 6  0.25582359  0.9845562  0.9831637  0.34875474  1.1996009  1.5368601
#>         X251       X252       X253        X254       X255         X256
#> 1  0.6907770 -1.1112215  0.3152322  0.28158729  1.6000169  1.518715957
#> 2  1.3747584 -0.2523278 -0.0775611  0.35248279  1.3491632  0.001727146
#> 3  1.0742242  0.8477129 -1.2138424  0.49461652  1.1294454  0.149102866
#> 4 -0.6300055  0.9630991  0.4537786 -0.06433873  0.3160568  0.404020673
#> 5 -0.4427932 -1.3733294  0.3013744 -2.97902794 -1.0379587 -0.193713429
#> 6  1.1154540  0.6622706 -1.1856162  0.41876567  0.7924355 -0.156556602
#>         X257       X258       X259       X260       X261       X262
#> 1  1.9226793  0.1091837  0.9935103  0.2293983  0.5170656 -0.2244141
#> 2  1.0128417 -0.3247510 -0.1932752  0.2863186  0.4181255  0.6291936
#> 3 -0.8030516  1.2610183  2.5772479  0.7999408  0.4526977  0.6933156
#> 4  0.3470706  0.9415133 -1.0193835  0.7925846  0.2197835 -0.2801188
#> 5 -0.4371024 -1.1003353 -0.7169660  0.3619419 -1.6567778 -0.8752837
#> 6 -0.1092806 -0.1566224  0.4317975 -0.9652380 -1.1298551  1.3959691
#>           X263        X264       X265         X266        X267        X268
#> 1 -0.220917835 -0.18280614 -0.7064248 -0.936596062 -1.36267255 -0.36726478
#> 2  1.006878263 -1.59306420  0.5232729  0.215322148 -0.05924850  1.43093053
#> 3 -1.064056135  0.13266683 -0.1141935 -0.156610807 -0.06719159 -0.38654736
#> 4  0.888721480 -0.80306015  0.3390674  0.138775544 -1.15901524 -0.04847865
#> 5  0.001349282  0.01739708  1.0553412  0.001053502 -1.04705389 -0.49647513
#> 6  0.085910923 -0.15338014 -0.2386803 -0.500255315 -0.19707180  1.22845641
#>         X269        X270       X271       X272       X273        X274
#> 1 -0.1642231 -2.13811483 -0.1671347 -0.1101149  1.1391257  0.68852828
#> 2  1.5559650  1.49819682  0.8572722 -0.1224998 -0.5675517  0.25764275
#> 3  2.4981690 -0.95821197 -0.3578560  0.5264881  0.2916500 -0.08305891
#> 4 -1.5263523  0.01243244  1.1793733 -0.5264911  0.1074347 -0.70907066
#> 5 -1.0989955  0.33994138 -0.8027905  0.6120201  0.1586356  0.80881430
#> 6  1.1618051 -1.47046970  0.6535739 -1.1025188  0.2171155 -0.10787053
#>         X275       X276       X277       X278       X279       X280        X281
#> 1  0.9914189  0.9103862 -0.2525983 -2.1222846 -0.2242281  0.3203450 -0.12669809
#> 2  1.4030812 -0.1891085 -0.6270551  0.8791106  0.7531686  0.6498088  2.02191778
#> 3 -0.1245320  0.5053531 -0.3639890  1.0921673  1.2098685 -1.0942826 -1.60303911
#> 4 -0.4287224  0.6293995 -0.2412167  0.1608945  0.2332015 -1.4372388 -0.07077322
#> 5 -1.1723585 -1.7784327 -1.2474636 -0.8256378  0.4000876  1.8160133 -0.82138680
#> 6  0.5461884 -0.9401190 -0.6432463 -1.1385615  0.2401216  0.8842608  0.87075020
#>          X282       X283         X284       X285        X286       X287
#> 1 -1.04131000 -0.1981118 -0.054214252 -0.3331026 -0.47205626 -0.4328084
#> 2 -0.02709759 -0.6471805 -0.559342906  0.2596435  0.03958992  0.9283761
#> 3 -0.07251416 -0.8446014 -0.701358486  0.0193632  1.00979110  1.2230405
#> 4  0.36858855 -1.2975753 -1.268597518 -0.5879059  0.92487298  0.7103701
#> 5  1.69491208 -0.9613468  0.695595721  0.6853154  0.68134212 -1.3129719
#> 6  0.83973146  1.1409498 -0.005481283  0.2465569  1.40441140 -0.4541943
#>         X288        X289       X290         X291        X292        X293
#> 1 -2.5031520 -0.93224448  0.9420625  0.007234049 -1.72808078 -0.10279665
#> 2 -0.3873901 -1.21897870 -0.6219036 -0.018855962 -1.42122671 -0.46848280
#> 3  1.0232227 -0.02912036  1.7131478  0.734940240  1.54963610  0.02682117
#> 4 -0.1317899 -0.74855197 -1.4903964 -1.139701314 -0.56772743 -0.21595085
#> 5 -0.7120566 -1.05214865 -0.5447286  0.535589049 -0.57618191  0.66620675
#> 6  0.2990882  1.53529982 -1.1147446 -1.982636171  0.03533303 -0.11441014
#>          X294       X295       X296       X297       X298       X299       X300
#> 1 -0.79427674 -0.4227457  0.3260102  0.0334264  1.4213107  1.2388054  0.4224890
#> 2 -1.24006503 -0.9048228  0.1412420 -0.8994048 -1.9653314 -0.4612442  0.9315197
#> 3  0.07436272  0.4805329 -1.3716016 -1.3811461 -1.0991509 -0.5952033  0.5204821
#> 4 -0.25591536  0.9559126  0.8385460 -0.8157053 -0.8796163 -0.1378388 -0.4861781
#> 5  0.44920807  1.7891175 -0.1539999 -1.6617070 -1.1280985 -1.4385429  0.5957134
#> 6 -1.20118269  0.1994871  1.1385208  1.5149310  0.6125911  0.6002847 -1.6663067
#>          X301       X302        X303       X304        X305        X306
#> 1  0.73807416 -1.5408054 -0.75939937  0.7289775  0.09899581 -0.49299130
#> 2 -0.82173284 -0.4830365 -1.19008835  0.4954371  0.24303418 -0.10892728
#> 3  0.90943776  0.7238141 -0.34672709 -0.5762556  0.08840714  1.29543546
#> 4  2.07729596  0.3769818 -0.31356729  0.3900935 -0.26308382 -0.98864915
#> 5 -0.05918501  1.3583124 -0.22254565 -0.0924414 -0.04033241 -0.09607346
#> 6  1.09470275  1.0903523  0.07318992  1.1903769  0.96241923  1.17564130
#>         X307       X308       X309        X310        X311       X312
#> 1  0.3998853 -0.1533692  0.5334667  0.66497202  0.20564395  1.8972151
#> 2  0.2148964  1.5435182  1.3571568  0.61620522 -0.56663854  0.4932072
#> 3 -0.2382629  1.3313068 -0.2234453 -0.03244999 -0.02336688  1.4359813
#> 4 -2.6712087  0.2646890  0.7112295  0.21991024  0.23985248 -0.5088323
#> 5 -0.7121376 -0.0247485 -0.6705912 -0.12855848  0.08593787  1.1132996
#> 6 -2.1192943  1.1745438 -1.4581754  0.67995094  0.12017299 -1.1419248
#>          X313        X314       X315       X316       X317       X318
#> 1  0.96962268  1.34887376 -0.1316500 -0.7127396  1.2515056 -0.9183146
#> 2  0.29278649  0.08448806 -0.2053572 -0.3836008 -0.9944413  0.2719946
#> 3 -0.43152061 -0.85781305 -0.1631233  0.9445314 -0.0336078  0.6282150
#> 4  0.41527548 -0.83872362  0.8067989 -1.2515233 -0.0704080 -0.5197193
#> 5 -0.08046065  0.34453012 -1.8062872 -0.4268316 -1.3159651 -0.3083836
#> 6 -0.59529712  0.75382861  0.6988149  0.6354083 -2.2024668 -1.4831723
#>         X319       X320       X321       X322        X323       X324       X325
#> 1  0.4836262  0.1286205 -1.3323041 -1.0368176 -1.87838793 -0.3765226 -0.4840705
#> 2 -0.9071815 -1.0711496 -0.4316719 -0.3377574  1.17826805 -0.1310282 -0.4573560
#> 3 -1.2657238  1.0169635  0.3920737  0.2027212 -0.91732787 -1.1548004 -0.5882307
#> 4 -1.1484573 -1.1172293 -1.3178177 -0.8615209 -0.06890278  0.7648338 -0.3867542
#> 5 -0.1547887 -1.1829227 -0.2076403 -1.8366309 -0.06329331 -0.3580686 -0.6552092
#> 6 -0.3162390 -1.0321084  1.4448597 -0.9484522 -0.95005261 -2.6242779 -1.2174664
#>         X326       X327        X328        X329       X330        X331
#> 1 -0.2852462 -0.8863642  1.21946229  1.42198112 -0.3679827  0.20319693
#> 2 -0.5648901  1.0064841 -0.28074878  1.91959005  1.0561132 -0.72250016
#> 3  0.9406686 -0.1732483 -0.07244581  0.07873971  1.1285927 -1.56463787
#> 4 -1.2427725  0.1396443 -0.59169779 -1.47521773 -1.0063356  0.07683419
#> 5 -0.5218863 -0.2291406 -0.85484343 -0.51072743  2.5854474  1.55529354
#> 6 -1.8792118 -1.7042073 -0.37176945  0.53996121 -2.1471322 -0.95585764
#>         X332       X333        X334       X335       X336       X337       X338
#> 1  1.4927127  0.7346322 -1.16009398 -0.9811197 -1.8801210 -0.7014136  0.6136988
#> 2 -0.8051884 -0.8939276  1.38964221 -0.2533573  1.8230643  1.2502144  0.8332725
#> 3 -1.0594954 -0.7778221  0.02795818  0.7829370 -0.7006499  0.1121002 -0.3787973
#> 4 -1.6300134 -0.8418264 -0.40705142 -0.1930363  0.2071796  1.5764331  0.9317780
#> 5  1.2205266 -0.9794721 -0.74905969 -0.8073735 -1.7185979  0.5195702 -1.1202195
#> 6  1.1892963  0.5106741  2.19702589  0.8293806  0.5547543  0.1375359  0.4979791
#>          X339       X340       X341       X342       X343        X344
#> 1  0.22074569  0.1440936 -0.2819831 -0.7682808  0.5176950 -0.69587835
#> 2 -1.10318374 -1.2064924 -1.9848238 -1.2661607 -0.5263165 -1.12073163
#> 3 -0.94690305 -1.3451901 -0.3244192 -0.2460230 -0.9012191 -0.05293026
#> 4  0.89727295 -0.4961933 -0.2093773 -1.0265111  1.5909458 -0.48227907
#> 5 -0.08269496  0.3156809  0.7166713 -0.2274105  0.3420406 -0.09851846
#> 6 -0.40712230  1.0192221  0.3484802 -0.3739863 -1.6260131 -0.25355179
#>          X345        X346        X347       X348        X349        X350
#> 1  0.74378292 -0.00778906  0.37018741  0.4696143  0.30739241 -0.66182067
#> 2 -0.12876668 -0.26949047 -0.06003956  0.1550650  0.49194196  0.04433926
#> 3 -1.51508099 -1.41484024 -2.71103320 -2.7727974  0.02121904  0.18819388
#> 4 -0.57268747 -2.13678462 -1.07202006 -1.4597767  0.21891412  0.28420321
#> 5  0.08073814 -0.77488223 -1.63356811 -0.1362861 -1.52590522  1.08416866
#> 6  0.67757345 -0.60901371 -0.98180899 -0.6423267 -0.59951384 -1.23072890
#>          X351       X352       X353        X354        X355       X356
#> 1 -0.42384038 -0.0893775 -0.3728124 -0.33471728 -0.06085773 -0.4520427
#> 2  0.01272781  0.6073053 -1.3377014  0.91221747  0.87363051  0.2898855
#> 3 -0.12524013 -0.8315336 -0.9169878  0.38276600  1.80583357  2.2722978
#> 4  1.38448321  0.2113168 -0.8358794 -0.06750933  0.35678141 -0.6298398
#> 5  0.80876275  0.5529675  1.5016319  0.53034856  0.94450228 -0.4415377
#> 6  0.32110440  0.3565386 -0.1792988 -1.37065559 -0.94706105  1.2681108
#>         X357       X358       X359       X360         X361       X362
#> 1  0.1544908  0.2870484  1.4060650 -1.8874398  1.691841059 -0.6056704
#> 2  1.2013976  0.6850185  0.6230445  1.3903205 -0.005206828 -0.5542808
#> 3 -0.2593907 -1.3334651 -1.5561174  0.6368556 -0.088329835  0.6189264
#> 4 -0.5357414 -0.3286796 -0.2800655  1.2846681  1.196994057  0.2757965
#> 5 -0.6483874 -1.0047563 -0.8972163 -1.6852433  0.034075656  0.8096095
#> 6  1.8003648 -0.1963477  0.5822600  1.2536745  0.340269850 -0.6168585
#>          X363        X364        X365       X366        X367       X368
#> 1  0.16981905 -0.02503565 -0.44520395 -0.8709752 -0.94477062 -1.7042245
#> 2  0.60652532  1.79320354 -0.69870121  0.5002763  0.07336375  0.1425468
#> 3 -1.74724503 -0.46835205 -1.35482751  0.3616228 -0.05988744  2.8878279
#> 4  0.03264433  0.42781633 -1.02861907  0.3944775  0.69073116 -0.4287456
#> 5  0.34571170 -0.52738363  0.03812239  2.0549998 -1.23127169  1.0739033
#> 6  0.17769281  0.97539069  2.15563916  1.8836819 -0.71036240 -0.4219242
#>         X369        X370       X371       X372        X373        X374
#> 1 -1.6624616 -1.69941268 -0.4234963 -1.3655320 0.507383935  0.48514903
#> 2  1.1626602  0.42982286  0.9344436 -1.0290987 0.648960150 -0.07843269
#> 3  0.5386783  0.09408483 -1.8996917  0.3971274 0.002431945  0.56177466
#> 4  0.1929108  0.08569146  0.8622399  1.1032000 0.295281718 -0.30094257
#> 5  0.8453223 -0.56821855  1.1728264  0.9308318 0.280671848 -0.61964031
#> 6 -0.5284270 -1.34475128  1.4371058  0.7053492 1.471622909  0.41169776
#>         X375       X376       X377        X378       X379       X380       X381
#> 1  2.2402359 -1.0071930 -0.8058665 -0.64455725 -0.2405722 -1.0175689  0.2794655
#> 2  0.3908510  0.4493128  0.7042946 -0.33011100  1.2973914  0.2549233 -0.8558681
#> 3 -2.1444120 -1.1943676  0.7812122 -1.57908526 -1.5371154 -2.1821005 -0.5763152
#> 4 -1.6236692 -0.8039232  1.3609824  1.57554838  0.2826067 -1.9018167 -0.8300635
#> 5  0.0282042  0.2748350 -0.5426683 -1.62299249  0.4854969  0.1840973  0.4982805
#> 6  1.7730007 -0.5584407  0.9345020  0.05606397 -0.6836943  0.1000876 -0.5799818
#>            X382        X383       X384        X385       X386       X387
#> 1  1.6642858624  0.99547100 -0.8905477 -0.25387116  0.7900944  1.1746904
#> 2  1.6341810196  0.82391875 -0.6304818  0.05213627 -1.0328345 -0.3571488
#> 3  0.2242893185  2.34348879  0.3576196 -0.79223142  0.1911190  0.3897028
#> 4  0.0001381177  0.29323099  0.1890377 -0.21835569  0.9527288  1.0520852
#> 5 -0.8814347058 -0.06837577  1.6986324  0.99003486 -0.1999754  0.5035739
#> 6  0.2390366467  0.76006725 -0.5378960 -0.07739084 -1.8846741 -0.1308146
#>          X388       X389      X390       X391        X392       X393       X394
#> 1 -0.62034889  0.5415939 1.2554280  0.2573681  1.53664426 -0.1637086  1.4628441
#> 2  1.95794537  0.3537739 1.3309808 -0.4185341 -0.08515516 -0.4662796 -0.3299013
#> 3  0.44502762  0.2878566 0.0769314 -0.4452951  0.06658935 -1.4663741 -0.4469066
#> 4 -0.43332678  0.2149218 1.2056645  0.5669552  1.62119075  1.0385379  0.3711790
#> 5 -0.33108988 -0.6107567 0.2536178  0.5216622  1.00665091  0.1335569 -0.8502991
#> 6  0.09487121  0.6168592 0.6521736 -2.3202009 -1.20506113  0.5154417 -1.2174913
#>         X395       X396       X397       X398        X399       X400       X401
#> 1  0.7188557 -0.5839087 -0.4479836  1.8586637 -1.16912622 -0.2682781  1.3360279
#> 2 -1.2884454  0.8058254  0.2676545 -0.2629249 -0.53790810 -1.8900067  0.1190327
#> 3  0.2370925  0.3646161 -1.1434094 -0.2778450  1.46669854  0.2526368 -1.1890509
#> 4  1.8837793  0.5937286  1.1979271  0.7112275  0.03689197 -0.3653209 -0.2195610
#> 5 -0.5056225 -1.4589588  0.1469597  1.4112243  1.54693172 -1.0763066 -1.2668376
#> 6 -0.3571367  1.9137553  1.3663412  2.1508469  1.28542090  1.8990990  2.2957391
#>         X402       X403       X404       X405       X406       X407        X408
#> 1  0.4438644 -1.0370433 -0.4575925  0.8372551 -0.8747374 -0.3120033  0.68108448
#> 2  0.5853838  0.6592607 -0.4383077 -1.2516469 -1.3809094 -0.2829238 -0.04707154
#> 3  0.1542860 -0.2705740 -0.8286408 -1.5455491 -1.4815540 -0.8115706  0.64480793
#> 4  1.0088372  0.4122904  0.0415562 -0.6032457  0.1834953  1.8202291 -1.88753971
#> 5 -1.4445876 -0.5147829 -0.2770125 -0.2167939  0.1346864  0.5532775 -2.25159250
#> 6 -0.1588564  1.7251975  1.4189592  1.3466712  1.2072040  0.9873123 -0.55865389
#>         X409       X410        X411       X412       X413        X414
#> 1  2.1029593  0.9342774  0.76396862  0.2402926  1.0141736 -0.45158148
#> 2  0.9986970 -0.4649932 -0.04824158  0.5245429  1.0154739  1.54588832
#> 3 -0.4009145 -0.8305641 -1.24849019 -1.4821528 -0.2712026  0.06344529
#> 4 -0.6353100  0.1713520  0.80604414  1.9538344  1.3291714 -1.14237537
#> 5  0.9863276  0.3110521  1.57123131  1.2930545 -0.1065211 -0.92685903
#> 6 -1.1176878  0.5511123  0.38759753  1.5801489  0.5168251 -0.44351197
#>         X415        X416      X417        X418       X419       X420       X421
#> 1 -0.9216024  1.45187321  1.430950 -0.31959422 -0.3091379  1.5075976  0.6074717
#> 2  0.6022946  0.23330131  1.509417  1.05143037  1.0094956 -0.4933456  0.7312203
#> 3  0.8958306  0.08607498  1.738033 -0.23683949 -0.4156383 -0.2687178 -0.2804526
#> 4 -0.9263586  0.75681352 -1.351180 -2.30617954  0.4016877  0.4358291 -0.2825942
#> 5  1.2902676  0.82851170  1.081248 -0.14494089 -0.9719775 -0.4783751 -0.3835414
#> 6  0.3018968 -0.57906410  1.172126  0.08568142 -0.3882803 -0.0406491  0.3944693
#>         X422       X423        X424        X425        X426       X427
#> 1  2.0243796  0.6252548  0.07187638 -1.90646767 -1.07397629  0.6960625
#> 2  0.4340146  0.6488935  0.69616074 -1.67017354 -0.36530347  0.2957811
#> 3  0.3256678 -1.1847622 -0.25208101  0.26649994  0.42240997 -1.2726325
#> 4 -1.2556657 -0.3721642 -1.20952840 -0.02583255  1.10363146  0.6407071
#> 5 -0.7774338 -1.3058417  1.40938452 -0.03816203  0.80655906  1.5033466
#> 6 -1.4546561 -2.0232912 -0.61283645 -0.46218920  0.02891704 -1.9904791
#>         X428       X429       X430         X431       X432       X433
#> 1 -1.0680920 -0.8049452 -1.4536737 -0.790063345  0.4241007 -0.5375194
#> 2 -1.8109583 -0.8222823  0.1685291 -0.001597909 -0.6949724 -1.2247576
#> 3 -0.3697302 -0.7289512 -0.4212292  2.052725326  1.2468159  0.2545101
#> 4 -1.3839196 -0.7399672 -0.5620286 -1.101730913 -0.5664943  2.2420468
#> 5 -0.2230168  0.6422187  0.1741887 -0.463872256  0.7062606  0.1255902
#> 6 -1.2102249 -2.5973832 -0.6913074 -1.558689707 -1.4324299 -2.1446420
#>          X434        X435       X436       X437       X438       X439
#> 1  1.03534239 -0.79302548 -0.3820252  0.8331335 -0.1778023  0.9307396
#> 2  0.08906259  0.26986119  0.1221914 -0.6844559  0.9040689 -0.5449462
#> 3  0.59293591  1.10471947 -0.1035279  0.4100970 -0.7549154 -0.5979343
#> 4 -0.48945219  0.04036964 -0.4583300 -1.0281243 -0.6729122 -1.1045986
#> 5 -0.93917520  0.80509525  0.5816991  1.3777927  1.2836368  1.2382184
#> 6 -1.27667002  0.29436009 -0.1327547 -0.3624462  0.4013184  2.1199344
#>          X440       X441         X442       X443        X444        X445
#> 1 -0.07135176 -0.3059145 -1.547181785  0.2597266 -0.16289972 -0.91083319
#> 2 -1.10353078 -0.9828413 -0.004637419  0.5548494 -0.98433038  1.07668594
#> 3  0.59984484 -0.5103642 -0.372464089  0.9703255 -0.65703342 -0.26776186
#> 4 -2.02170999 -1.5601147 -0.041888459  0.6407474  1.65121400  0.92193181
#> 5  0.73500138 -0.4977233 -2.686245238 -1.4898515 -0.05312586  0.08946329
#> 6  0.14933678  0.6387183 -0.496467333 -0.4318717 -0.91467601 -0.24943904
#>         X446       X447       X448        X449       X450        X451
#> 1 -0.6797412 -0.8908345 -0.8231254  0.08165473  0.8566355 -0.62469136
#> 2  0.3236696  0.6253509  0.8541184  1.68844937  1.1083861  2.30487744
#> 3  0.8087031  0.0584135  0.3585717 -0.26046171 -0.7683210 -2.05479311
#> 4  1.0767922  0.3086716 -0.1302756 -2.10637506 -0.7035185 -0.71929809
#> 5  0.3792318 -0.6013868 -1.1726398 -1.18367103  0.7310724  0.09949616
#> 6  1.0947670 -0.9775397  0.3095591 -0.54313047  0.3414435  1.31701183
#>          X452        X453       X454        X455       X456       X457
#> 1  0.37375544  0.07469484  1.9253291  0.14006022  0.4558704  0.3651022
#> 2  1.14794089 -1.16320204 -1.2787489 -0.22337285 -0.3088596 -0.1150128
#> 3 -0.76463330 -0.63159161 -1.4996529 -1.21378063 -0.3545558  0.1846070
#> 4 -0.56795375 -0.15189917  0.7159510  0.09414505  1.4403065 -0.1548911
#> 5  0.07649966 -0.76644799  1.1290552 -0.66596971 -1.1112166  0.2258028
#> 6  0.43182549  0.92896446  0.1648817 -0.64596299  0.2793447  1.2357514
#>          X458        X459       X460        X461       X462       X463
#> 1  0.05267859 -0.74067563 -0.2543763  0.49794385  0.7745271 -1.6165600
#> 2 -0.21319639 -0.79236014  0.4610235  0.71005033  0.3810799 -0.4157105
#> 3 -0.25577878 -0.50500173  0.8734852 -0.52252646 -0.6711419 -0.3183986
#> 4 -1.27695247 -0.12569338 -0.2356562 -0.07386808  0.2868140 -0.2897283
#> 5  0.29450807 -0.02693413  0.6380819 -1.49903577 -0.7351780  0.7832162
#> 6 -0.84945576 -1.08881007  1.1249471  0.40831080 -0.3643766  1.1344963
#>          X464       X465       X466       X467       X468       X469
#> 1 -0.95680929 -0.2784592 -0.3804674  1.2360706  1.2787512  1.2276261
#> 2 -0.77406120 -0.5679496  0.5202125  1.7608049  2.6979720  0.1779255
#> 3 -0.08308884  0.3701328  2.2134662 -0.7515920  0.7683876 -1.2408200
#> 4 -0.37192736 -2.6261936 -1.3954119  0.2755166 -0.9989574  0.8243674
#> 5  0.90264442  0.4747268 -0.1299556  0.1281311 -0.1950927 -1.2766036
#> 6  1.08745384  0.8445519  0.1800763  0.1465037 -0.3826061  0.8853504
#>          X470       X471       X472       X473       X474       X475       X476
#> 1 -0.23032028 -0.1975855 -0.8559229 -0.9103440 -1.2570734 -0.6394487 -0.3624045
#> 2  0.64333647  1.0280425  0.4580840  0.8563058  0.9483450 -0.8001237 -0.2097978
#> 3  0.77139009  1.4822913  1.4615931 -0.6109150 -0.5242181 -0.1983477  0.7457740
#> 4 -1.23252796  1.1258460  1.1471325  1.5537443  0.6623620 -0.9466158 -0.0841964
#> 5 -0.06367505  0.2419178  1.4833470  0.7339679  0.6141022  0.3440765 -0.3791752
#> 6 -0.65746330 -0.6473458  0.8389909  0.4851680  0.6548957  0.2151860  1.4087223
#>          X477       X478         X479        X480         X481       X482
#> 1 -0.47877905  2.4857624 -0.004845227 -0.55571904  0.007051657  0.2730539
#> 2 -1.25056627 -0.2484357  0.330041003  0.22598143 -0.069802325  0.9348659
#> 3 -2.41092774 -1.3600341 -0.258245228 -0.31760067 -0.362743570  0.0242058
#> 4 -0.83073469 -0.5045065  0.457015394 -0.44964462  0.749447919 -0.3208759
#> 5 -0.04734646  1.0326300 -0.338390987  0.39480587  1.390138673  2.0253522
#> 6  0.76980639  0.1385568  0.808239398  0.09297068  0.239663302  1.6184659
#>          X483       X484       X485        X486       X487        X488
#> 1 -0.20659328  0.1313118 -0.2072735 -1.09554666 -0.5735090 -0.72822901
#> 2 -0.58885317  0.1124052  1.0593820  0.73800896 -0.9522971  0.02222864
#> 3  1.24141059  1.8012018  0.5943529  0.71584651 -0.9040710 -0.68965570
#> 4 -1.08876520  0.3924733  0.4004019  0.66414909  0.2428327  0.05212577
#> 5  0.52265092  0.3420198  1.1342267  0.03791072  1.2754936 -0.53476717
#> 6 -0.04933869 -0.4219958 -0.2649525 -0.11656328 -1.7690137 -0.84267634
#>         X489       X490       X491       X492       X493        X494       X495
#> 1  0.3469606 -0.7174714 -0.3314545 -2.3402682  0.1892302 -0.90567999  0.7075705
#> 2 -1.1254536 -0.3156169  0.1876416 -0.8994184  0.1952924 -0.04945303 -2.6989048
#> 3 -0.2771303 -0.7286070  0.7287732  0.2940881  1.7526985  1.03523352  0.3997525
#> 4  1.2350850 -1.0475017 -2.0362408 -0.3374081 -0.2617359 -2.22604235 -0.1169866
#> 5 -1.1729876 -0.6564676  0.5681478  1.6333526  1.3626138 -0.57513352  0.4325790
#> 6 -1.5256677 -1.9438887  0.9278489  0.8044309 -0.2243260  0.68904412  1.5493952
#>         X496       X497       X498       X499       X500
#> 1 -0.3489099 -1.0432585 -1.0635463 -1.9301481 -3.0078907
#> 2  0.6273140  1.9355363  0.9736631  1.0486007 -0.2938534
#> 3  0.9358627 -0.3283781  1.1811321 -0.3764565 -0.1663633
#> 4 -1.3388864 -1.3810972  0.1129085  1.3869744  0.3833798
#> 5 -1.5360412 -0.3997460 -0.9389794  0.1592148  1.4573464
#> 6  2.0556492  0.7832976 -0.7149012 -1.0246483 -1.5576497