A high dimensional dataset with n equals to 200 and p equals to 500.

A dataset with 200 observations and 500 dimensions is generated from the following process: linear regression model with only first three non-zero coefficients equal to 3, 2, and 1.5 respectively. The covariates are correlated with AR structure (rho=0.3). The error term is normally distributed with zero mean and sd equals to 0.5.

Usage

data(HighDim)

Examples

# HighDim dataset is generated from the following process:
n <- 200; p <- 500; sigma <- 0.5
beta <- rep(0, p); nonzero <- c(1, 2, 3); zero <- setdiff(1:p, nonzero)
beta[nonzero] <- c(3, 2, 1.5)
Sigma <- 0.3^(abs(outer(1:p,1:p,"-")))
library(MASS)
X <- mvrnorm(n, rep(0,p), Sigma)
error <- rnorm(n, 0, sigma)

X <- apply(X, 2, scale) * sqrt(n)/sqrt(n-1)
error <- error - mean(error)

Y <- X %*% beta + error
HighDim <- data.frame(Y, X)
head(HighDim)
#>           Y         X1         X2           X3         X4         X5        X6
#> 1  4.376884  0.6386625  0.7598428  0.008698697 -0.5647407 -1.0988383 0.4020907
#> 2 -1.937395  0.2303227 -0.3160325 -1.244542896  0.7257375  0.2601853 0.2432956
#> 3  4.598954  0.8846548  1.4036599 -1.000387900 -1.1805405  0.2525938 0.4274621
#> 4  2.425983 -1.1868544  1.5981317  2.198167015 -0.9012518 -0.2345649 1.2038079
#> 5 -2.835150 -0.3034995 -0.7141646 -0.349819146  2.3529434  2.9516077 0.7585057
#> 6 -1.306194 -0.5470160 -0.2422698  0.519329244  0.3485572 -1.1882432 1.7169486
#>           X7         X8         X9         X10        X11        X12        X13
#> 1  1.7658695  0.1833270  1.1182544  0.81780271 -0.5640617 -0.5484711  1.6305228
#> 2 -0.8986104 -0.5613678  0.2624679 -1.27226715 -0.4576510  0.2204798  0.2922984
#> 3  0.5768412 -0.5038509 -0.1605494  1.15311371  0.4176065  2.9842547  2.5187979
#> 4  1.2889591  0.6075300  0.7006633 -0.66112948 -2.4144092 -0.5419897 -0.6490864
#> 5  0.6838910 -1.1036420 -0.2761015 -0.09458969  1.0170246  0.9395104 -0.1099176
#> 6 -0.6977547  1.7669388  0.1140248  0.60614322 -1.0363524  0.1870216 -0.6753479
#>          X14        X15        X16        X17        X18        X19         X20
#> 1  1.3692870  0.1600934 -0.3747611  0.1003025 -0.2991597 -0.4412207  2.30728400
#> 2  0.4286124  0.7671220  0.3240434  0.4669384  0.2362224 -0.8921022  1.37119255
#> 3  1.0256804 -0.7529653  1.0257455  1.1102159 -0.4931071 -0.3980987 -0.07026782
#> 4  0.9499126  1.0194589  2.3053234  1.3203805  2.1048351  0.9667856 -1.58968950
#> 5 -0.2487455  0.2472858 -0.2713448 -0.8720972 -0.3733266  0.8665565  2.05019462
#> 6  0.3468014  0.9587650  1.2535917  2.3702683  0.3470507 -1.1220527  0.00011920
#>           X21        X22         X23         X24        X25        X26
#> 1  1.43347391 -0.3855762 -0.53622339  1.11385562 -0.9944273 -1.4242293
#> 2 -0.03198674  1.2271644  1.54104767 -0.90751978  0.4555100  1.4718311
#> 3  1.32462804 -1.0512417  0.06308676  0.97523242  0.2392488  0.8667079
#> 4  0.51708827  1.0732733  0.82731055 -0.06034375  0.6733257  0.3930891
#> 5  0.51812267  0.1320002 -0.75932165 -0.83614510 -0.3193516  1.4473175
#> 6 -0.88760141  0.2216573  0.41094801  1.21498895 -1.7426341  0.7422544
#>           X27        X28        X29         X30        X31         X32
#> 1  0.18415588 -0.3780573 -1.1450481 -0.00729652 -0.6662430  0.79224769
#> 2  0.77895958 -0.5907555 -0.8448803 -0.54352032  0.1222618 -0.01146404
#> 3 -0.75593552 -0.6501123  0.4147131  0.28014282  0.5860742 -0.38996850
#> 4 -0.15963448 -0.4135124  0.5063679  0.83294791  0.7662505 -0.05084181
#> 5 -0.36924894 -0.5640191 -1.3716358 -0.15453297 -0.2380782 -1.56089798
#> 6 -0.09503365 -0.1643379  0.6037312  0.22722621 -0.3836994 -2.62160111
#>          X33         X34        X35        X36         X37        X38
#> 1 -0.3797479  0.02557384  0.7407251  0.3176647  1.50866418  1.5754077
#> 2  1.1126153  0.84013965  0.7420368  1.5924700  0.55367100  2.2758240
#> 3 -0.6271085  1.20426159 -0.2328736 -1.1333227 -0.86438495 -0.6888285
#> 4  0.5727153 -1.01922163 -1.1919369  0.8940662 -0.17274456  0.6942602
#> 5  0.8308494  0.57981731  0.1022077  0.7143287 -0.04297019  0.6918063
#> 6 -0.6840233 -2.90806088 -1.8653121  0.4388052  1.26461870  0.4079428
#>          X39        X40         X41         X42        X43        X44
#> 1 0.08679435 -0.8879805 -2.17070594  0.86572809  0.8664708 -0.1828948
#> 2 0.86778701  0.9665135  1.51248557  0.87238182  0.4808157 -1.6645986
#> 3 0.31579643  1.5114986  0.78825652 -0.48092899 -1.4529626 -0.5866480
#> 4 0.39866872  0.1411677 -0.09906144 -0.26483409 -1.6635342 -1.1372479
#> 5 1.37874159  1.1265727  1.41276294 -0.09608206 -0.7030908 -1.6028221
#> 6 0.96431532  0.4824854 -1.19653888 -1.58526017 -0.1489691  0.1270472
#>          X45         X46        X47         X48        X49        X50
#> 1 -0.1281340 -0.03144019  0.5413813 -0.25861318 -1.0636559 -2.0523316
#> 2  0.5401956  0.49142274  2.0983567  0.06275197  0.3793115 -0.2758983
#> 3  0.2452463 -0.25322342 -1.7616858 -2.00214875 -0.9586681 -1.2548083
#> 4 -1.0149666  1.60280096  2.0461535 -1.22473728 -0.5309369 -1.3333588
#> 5 -2.2195851  0.92922434 -1.2100659 -0.10293223 -2.1895127 -0.7217744
#> 6  0.1788411 -0.34359348 -0.6835591  1.12375040  2.5900425 -0.2402536
#>           X51        X52         X53         X54         X55         X56
#> 1  0.21744266 -0.6217535 -0.05283824 -0.23493340 -0.05290526  0.97586059
#> 2 -1.01915784 -0.6347558  1.68704902  0.71701657  0.22245684  0.96260828
#> 3 -0.03534937 -2.3135478  0.85749045  1.96512895  0.43612170  0.02681082
#> 4 -1.41246484 -0.7198080 -0.20221097 -0.79979306 -1.86859822 -0.85633086
#> 5 -0.45064437 -1.3083412  0.69060116  1.12088347  1.55647114 -0.01490555
#> 6  0.24180824 -0.8375396  0.07658838  0.02544841 -0.24982465 -0.37499137
#>           X57        X58        X59         X60        X61         X62
#> 1 -0.33946128 -0.4613056 -0.9994831  0.06266007  0.2336199 -1.20638713
#> 2  1.95863501  0.1279531 -0.2332412  0.13079549 -1.4067138  0.64484780
#> 3 -0.03467852  0.5899378  1.0372503  0.19413311 -1.6498586 -0.48488336
#> 4 -1.30236704 -1.5861704 -0.3735165 -1.41344848  1.0760500  0.19429724
#> 5  0.89946126  0.2418655 -0.6427325  0.23894674  0.2330954  1.16467222
#> 6 -0.96862739 -2.2949695 -0.5286291 -1.28826925 -0.7171102 -0.02998356
#>           X63          X64        X65         X66         X67         X68
#> 1  0.34016276  0.009827454 -0.3786835 -2.27743553 -0.20508125  0.03786959
#> 2 -0.77699658 -1.377078071 -0.6054126 -0.34054205  0.17148494  1.64404724
#> 3  0.33075662  1.187235632 -0.2271363 -1.05284418 -1.91829196  0.25447556
#> 4 -0.20404308  0.942384273  0.6955766  0.04268964  0.06222142 -1.72343052
#> 5  0.20808031  0.761390184  1.1087524 -0.66709618  0.33049880  0.09780942
#> 6  0.03997176  2.112254717 -0.1947440  0.26583893  0.58566153 -1.58798163
#>           X69         X70         X71        X72         X73        X74
#> 1  0.03556541  2.04608233  0.36926887  0.6413418  0.03932231 -0.4784107
#> 2  1.03236558 -0.64282490  0.09582809 -1.1592103  1.34161685  0.7972159
#> 3 -0.61165660 -1.33818120 -2.06412109 -0.8896165 -0.95584566  0.1820526
#> 4 -1.90396249 -1.24447389  0.06947160  0.2995510  2.78503622  0.1528461
#> 5 -0.65758156 -0.03076139 -1.37587782 -0.3332909 -0.54456576 -1.0526384
#> 6  0.60470633 -0.39371034  0.18176879 -0.8665721  0.73835495 -2.1929701
#>          X75         X76         X77        X78        X79         X80
#> 1  0.2214755 -0.73904874 -0.09062732  0.6033907 -0.2124096 -0.55257702
#> 2  0.7497090  0.23600330  0.55071369  0.2746224  0.5580319  1.09606026
#> 3  0.1993776  2.09323160  1.00895540 -1.2814139 -0.2759879  2.57083095
#> 4  1.3427363  0.08253422 -0.15008855 -0.1756831  0.9671825 -0.06497785
#> 5 -0.3951481 -2.25345002 -2.21869649 -1.6360681 -0.3934517 -1.06495352
#> 6 -0.6394331  0.64361640 -1.16096520 -0.6169242 -0.7113541  0.11495258
#>          X81        X82         X83         X84           X85        X86
#> 1 -1.2251491  1.5724740  0.64920708  0.02808344  1.0715469896  0.5542670
#> 2  0.8928384 -0.4968474 -0.45475485 -1.17371183 -0.0216999427  0.3960756
#> 3  0.4793558 -0.4034824  0.03653837  1.09415814 -0.0009210677  0.7981032
#> 4 -1.7402796 -1.0211243 -1.61485073 -2.72226374 -0.8184498263 -1.1171825
#> 5 -0.9476411  1.0074199  0.78209047 -0.22276698  0.5412765651 -0.5518855
#> 6  0.6201453 -0.4253418  0.13727492 -0.02925905 -0.5479732940 -0.1885825
#>           X87        X88        X89         X90        X91         X92
#> 1 -1.10661518  0.1776363 -0.5991924  0.96969376  0.8314684 -0.56660969
#> 2 -0.26431741  0.6448915  1.5330181  0.39476443  0.5176831  1.07411092
#> 3  0.06121652  1.9216696  0.7617940  0.51540641 -0.2174463 -1.14997072
#> 4 -0.36179766 -1.2320278 -0.6134159 -0.06141963 -0.3511629 -1.00001904
#> 5  1.43811391  1.8094979  0.5356657  0.93648134  0.2934536 -0.02167531
#> 6 -0.64455159  2.1592601  1.6853803  0.17423211 -0.1440215  1.39348908
#>          X93        X94        X95         X96        X97        X98        X99
#> 1 -2.0153658 -0.9890460  0.6232306 -0.17516977  0.3749144  0.7134140  1.4652203
#> 2  0.8023578  1.0845958  1.4053598  1.38056120  1.7528692 -0.7463929  0.4072029
#> 3  0.7303737 -0.5212767 -0.7551647  0.04645662  0.9379995  1.8884434  0.1061960
#> 4 -1.1980113 -0.1116050  0.5470846  0.71774970  1.3945520 -0.6810394 -0.9681979
#> 5 -1.7458102 -0.7915956  0.4787216 -1.85546268  0.6373143  1.0420822  0.4544047
#> 6  0.3527339  1.9459036  0.5436023  0.12045739 -0.9319517 -1.2620625 -1.9538078
#>         X100       X101       X102        X103        X104       X105
#> 1  1.3819228 -0.4218951 -0.5979192 -0.37410054 -0.37768201  0.3114408
#> 2  0.2364045  0.3644002 -1.4947606 -1.16273997  1.21280451  0.5048562
#> 3  0.9082823 -0.7788715 -0.6530495 -1.74146786  0.54314586  0.5164928
#> 4 -0.7274132 -1.3074833 -0.3482967 -1.18860970 -0.64119192  0.7826111
#> 5  0.7555863  1.3680737  1.8034376 -1.59297560  0.02691505 -0.6069384
#> 6 -1.8396587 -0.1610747  0.9470916  0.01204541 -0.42972341  0.1859040
#>         X106       X107        X108       X109        X110       X111
#> 1 -0.6838202  0.7836370 -0.09797856  0.7126275 -1.07646414 -0.3679881
#> 2 -0.7462328  0.4102529 -1.42911360  1.5729188  0.32753321 -0.1684194
#> 3 -1.3038986 -0.7601153  0.91398416  0.3204298 -1.75751194 -1.3391811
#> 4 -1.6091701 -0.9279536 -0.31717728  0.2686502  0.01159489  0.4655929
#> 5 -0.3455163  1.3856154 -1.38522816 -1.0269033 -1.41023021  1.5318749
#> 6  1.7255861  0.5414859 -0.55374783  0.1538162 -0.41932901 -1.4295535
#>         X112       X113       X114        X115        X116       X117
#> 1  0.8595499 -0.4628458 -2.3285571 -0.64584173 -2.21203131 -0.3308934
#> 2 -0.9081414  1.4026351  0.3299353  1.19374664  0.74050547  0.2188928
#> 3  1.0501084  0.3656611  1.3765197  1.83312715 -1.76468743 -1.1424465
#> 4 -0.5067823 -0.6290373 -1.1701759  0.40653770  0.63880256  0.1434828
#> 5  0.3177168  1.1694837 -0.2946756  0.14152047 -0.60287685 -2.2268045
#> 6 -0.2520265  1.3579594  0.0639291 -0.03953441 -0.02563127 -0.1823060
#>          X118       X119       X120        X121        X122        X123
#> 1 -0.61670514 -0.6379752 -1.2680194 -0.31347319  0.05136527  0.69128344
#> 2  0.09561246 -0.7365990 -0.5212512  0.39077995 -0.36884634 -1.78188973
#> 3 -0.01244217 -1.0352112 -0.1241169  0.71441019  2.09466878  0.34548943
#> 4  1.36294802 -0.2662948  1.8738072  0.10887812 -0.61229643 -0.08256753
#> 5 -0.65709329 -2.2055801 -1.1291733  0.03543947  0.99413781  0.62967031
#> 6  1.39838935  0.5835346  0.6982081 -1.01213355  0.05255125 -1.54861900
#>         X124       X125       X126        X127      X128        X129
#> 1  0.3074973 -0.3688987 -0.1819807 -0.08690946  0.327585  0.08784169
#> 2 -0.1276865  0.1470808  0.9138061  1.15498871 -1.122866  0.88332754
#> 3 -0.8535957 -0.7467980  0.6811712 -0.86193338 -1.090147 -0.73795594
#> 4  0.8616455  1.6587326  0.7268385  1.09653892  1.366659  1.00280375
#> 5  1.1731885  1.3924784  1.7664005  1.53785181  1.074907  0.70741622
#> 6 -0.2323255  1.3191659  1.7161023  1.07997543 -1.066415 -2.30533832
#>          X130        X131        X132        X133       X134        X135
#> 1 -0.32769224 -1.83005692  0.38466949  0.55212493 -0.4372538  0.37093200
#> 2  1.15298270  0.39482597  1.48467987  1.65681263 -0.2743136 -0.89700751
#> 3 -2.05191430 -0.82325022  0.01012360  0.01549631 -0.8040514  0.18712759
#> 4 -0.35953562 -0.89783767  0.05506087 -0.76316145  0.2315137  0.32469326
#> 5 -0.22790595  0.70495817 -0.02869734  0.57933681  0.2275783  0.06038678
#> 6 -0.09771137 -0.02960706  0.27535715 -0.63957118 -0.6315650  0.19529657
#>         X136       X137        X138       X139       X140       X141       X142
#> 1 -0.7750890 -2.2732615  0.01800631 -0.7141928  0.3530581 -1.1959290 -0.8637029
#> 2 -0.2453936  0.8986410 -0.92510379 -0.9707477 -0.5681699 -0.4108097  1.4019924
#> 3  0.9329811  0.2127724 -0.43491267  0.8049850 -0.2100094  0.7428116 -0.4956601
#> 4 -1.9257131  0.3531044 -1.68589338  0.3084628  1.1741941 -0.9414627 -0.4464728
#> 5 -1.2937635 -0.6669076 -1.16121788 -0.6410292 -1.0210223 -0.2468947 -0.6172130
#> 6  1.6501406  2.1258250  1.15288391  1.2928435  0.7593428 -0.9576274  1.1883059
#>         X143        X144       X145        X146        X147        X148
#> 1 -0.3355757 -0.37141323  1.2415481  0.60031434 -1.84730872  0.01286814
#> 2 -0.8035714  0.08827369 -0.1547225  0.04874311 -0.43929824  0.36238071
#> 3 -0.8279799  1.25336116  1.1369707 -1.05414764  0.82585280 -0.99587121
#> 4 -0.6920435 -0.26591641 -0.8602420  1.35299935  1.76717911  1.31718456
#> 5 -1.4242431 -0.73335064 -1.6722735  1.00092277  0.02293785 -0.99651056
#> 6  0.5977661  0.31177912 -1.6959404  0.60369424 -0.26768770  0.15289390
#>         X149       X150       X151       X152       X153       X154        X155
#> 1  2.0082475  1.4441257  0.7519835  0.4677201  0.1084877  0.2346856 -0.31809486
#> 2 -0.6273086  1.1391390  1.8252300 -0.2694851 -0.4295405  0.7408823  0.22849955
#> 3  0.6108359  1.7949200  2.1982857 -0.3662772  1.5429706  0.7416384  0.04466985
#> 4 -1.1836359 -1.5722555 -1.1322895 -0.5972148 -0.7209431 -2.1833485 -1.81487006
#> 5  0.2845328  0.8767722  2.0249483  1.3370513  0.8440917  0.4838952  2.15593909
#> 6  0.1282312 -0.8393641  0.3984677  0.7968131 -1.5985458 -0.9439376 -1.06976868
#>          X156       X157        X158        X159        X160       X161
#> 1  1.22854449  1.0252087  0.06219307  0.77410648  0.71883979  0.6766454
#> 2  0.97379865  0.5941593 -0.41954206 -0.08032774  1.30965510 -0.6008765
#> 3  0.05893983 -0.5980857 -0.23551045  1.46930942 -1.71713085  0.1878560
#> 4 -1.16241163 -0.1581981 -0.70256176 -0.30451148 -0.21129175  1.8715205
#> 5 -0.88208846  1.7227738 -0.78949503 -0.54135939  0.03472291 -0.4668589
#> 6 -2.55353399 -1.8692834 -0.17493556 -0.45437910 -0.57390083 -1.5859041
#>         X162       X163       X164       X165      X166        X167      X168
#> 1  1.8989706  0.8216773  0.8720078  1.6206826 0.2359837  1.32370614  1.002641
#> 2 -0.3466839 -0.1841563 -0.4539701  0.5597253 1.6473817 -1.80470795 -2.027769
#> 3 -1.6865094 -1.1032856 -0.9335771 -0.6831025 0.6378453  0.70590969  0.433797
#> 4  1.3633977  0.2767019 -0.4143984  0.4445977 0.1623493  0.85622550  1.265418
#> 5 -0.6912873 -0.3721618 -0.1913864  0.4829342 0.2139305  0.00423923  1.159485
#> 6  0.4746418 -0.5875103  0.6905454  0.2355466 0.5347573 -1.16983047  1.312629
#>         X169        X170       X171        X172        X173         X174
#> 1 -1.1497996 -0.49538339  0.3837356 -0.51816551  0.03647802 -0.250130107
#> 2 -1.0310994  0.09614489 -0.2045727  0.83854503 -0.17106003 -0.109927287
#> 3  1.2934854  0.72962296 -0.3335906 -1.34634887 -2.44368961  0.119814517
#> 4  1.1430550 -0.05051958 -0.3385373  0.07348757 -0.36116015 -1.033599176
#> 5  0.1967947  1.31143002  2.5019355 -1.24367361 -0.97919670  0.005494916
#> 6 -1.1568776 -0.58511090 -0.9217806  0.82947075  1.99023120  2.019149841
#>          X175          X176       X177       X178       X179        X180
#> 1 -0.02382820 -0.1296079664 -0.1381700  0.4836398  0.1992857 -0.08710126
#> 2 -0.04000786 -0.3965044128  1.3510810  0.3455097  0.7795777  0.10761031
#> 3 -0.20314561  0.1055183289  0.2773396  1.2159993  1.6554104  1.15509040
#> 4 -0.06743365 -0.1262499699 -0.4317792 -0.9183815  0.1830273  0.20914873
#> 5 -0.10329231  1.9782279727  1.5044200  0.2293705  0.4289782 -0.88346700
#> 6 -0.49141977 -0.0004045424  2.4143696 -1.0691161 -0.9958349  0.21747043
#>          X181        X182       X183        X184        X185         X186
#> 1  0.20812479  0.25760942  1.3977147  1.18872530 -0.48948006 -0.322412324
#> 2 -0.86842757 -0.83873764  0.3642024 -0.54379932 -0.07028458 -0.524743632
#> 3  0.66931624  0.24351657 -0.9274616 -0.05874264 -0.60730744  0.006670508
#> 4 -0.05234172  2.37469388  2.5151921 -0.07258738  0.44509633 -0.791346392
#> 5 -0.06664604 -0.04992044  0.3481919 -0.86397506  0.60960168  0.244151406
#> 6  1.42808590 -2.72914770 -0.7284917 -0.51945959  0.95411536 -0.159021638
#>         X187       X188       X189        X190       X191        X192
#> 1 -1.3850217 -0.2538247 -0.3570757  1.29997301 -0.7815002  0.27454632
#> 2 -1.2121857 -0.1914237  1.0955403 -0.92332728 -0.1925039 -0.38213764
#> 3 -0.3761925 -0.2386589 -1.3360722 -0.21674010 -1.2293685 -0.88651522
#> 4  0.0964307  0.0593653  0.3743896  0.79619328  0.3172683  2.28458407
#> 5  0.9791676  0.6512649  0.7385207 -0.34746437  0.9588115 -0.66004436
#> 6 -1.4513079 -1.4026086 -0.3024329  0.08095013  1.0577539  0.04437807
#>          X193         X194       X195        X196        X197      X198
#> 1 -0.06433742  0.826253095 -0.4110578  0.06543048  1.23321067 1.3997537
#> 2  1.08302204 -0.002583513 -0.4735164  0.64633337 -1.60358107 1.0448025
#> 3 -0.40894659 -1.341901168 -0.1444604 -0.69985260 -0.76372031 0.1107892
#> 4  1.52064746 -0.368358711 -0.3652663  1.25997927 -0.08769981 0.7159865
#> 5  0.24877827  0.271035847 -0.2643735 -0.95632655 -0.91628025 0.1872811
#> 6  0.25645059 -0.894141256 -0.1249214  1.14780282  1.44465475 1.5400169
#>          X199       X200       X201        X202          X203       X204
#> 1 -0.07254690  0.5313700 -0.0311028 -1.02640438  2.1727133174  1.7780823
#> 2  1.95122657  1.4014320  0.9157782 -0.13881082  0.2286706589 -0.5144662
#> 3  0.08388528 -0.5086967 -0.4739643 -0.30837863  0.0006245873 -0.7824864
#> 4 -1.65608668 -1.7284629 -1.9676917  1.02027156 -1.0031792984 -2.0618604
#> 5  1.78711896  0.1666374 -0.5299509 -0.04881914 -0.6821460034  0.8320918
#> 6  1.89242703 -0.8823541  0.1523175 -0.34706343  1.7113608101  1.0426088
#>         X205       X206         X207       X208       X209       X210
#> 1  0.8857929 -0.8576945  0.755127338  1.1847094 -0.4645767 -1.1820384
#> 2 -0.7125260 -0.9894950 -0.993584991  0.4757612 -0.5991213  0.7962107
#> 3 -0.4625267  0.2754559  0.868423263  1.3837277 -0.7658181 -0.6387462
#> 4  1.0804150  0.4889191 -0.304634437 -0.5475990 -0.1572808  1.5415147
#> 5  0.4602265 -1.3883938  0.004406618  1.0576079  0.2466672 -0.9214728
#> 6  0.5012547 -1.2864788 -0.709979842  0.4526824  0.8887735  0.9119433
#>         X211       X212       X213       X214       X215       X216        X217
#> 1  0.9024612  0.6462818  1.0881669  1.3697377  1.5509187  0.5002264 -1.72766146
#> 2 -0.5004055  0.1541108  2.3996225  0.7519902 -1.4189410  0.6337327 -0.05550528
#> 3 -0.6342220 -1.8113027  1.5547261  0.2407787  0.7401423 -1.0646334 -0.51489112
#> 4 -1.4661864 -0.3844375 -2.4400638 -0.5475770 -1.8784328  0.3572076 -0.19068161
#> 5 -1.4193939 -1.8516265  0.3748003 -0.6630076 -0.4713856  0.4488995  0.53205792
#> 6  1.6115803 -0.1448669 -0.6866857  1.2859085 -0.6256114 -1.4970699 -2.16527539
#>         X218       X219        X220       X221       X222       X223       X224
#> 1 -0.9852041 -1.1399595  0.31599438  0.6424570 -0.3711925 -0.5341722 -3.1255441
#> 2  0.8854625  2.2472428  0.06590645  0.7446481 -2.3388393 -0.2179203  0.3481963
#> 3  2.3427761  0.4465334  0.71613883  0.4713749 -0.9385072 -0.7481020  1.1033882
#> 4 -0.7447087 -0.7697879 -1.25288201 -0.5622900 -0.2833656 -0.4441165  0.8351532
#> 5 -1.4904828 -1.1195851 -1.14799383 -1.0717482  0.3712876  1.8371988 -0.9177442
#> 6 -0.3885134  0.2396599  0.68303157  1.9229840  1.2514026  1.2036695  1.0206194
#>         X225        X226        X227       X228        X229       X230
#> 1 -0.5217956 -0.39788056 -1.28817228 -1.1763384  0.63271587 -1.2503659
#> 2  0.4905529 -0.07361549 -0.03489369  1.0796731  0.80465235  0.5713717
#> 3  0.3019550  0.83035321 -1.35567486 -1.1998423  0.98938031  1.6109558
#> 4 -0.3882066 -0.48648741 -1.52450352 -1.4329195 -0.17836954 -1.1938947
#> 5  0.3874346  0.31722366  0.50013733 -0.2362259 -0.49075628  0.9766334
#> 6 -0.4907636 -0.29911272 -0.70678382  0.4654365  0.09655395 -0.3050777
#>          X231        X232       X233        X234        X235       X236
#> 1 -0.74483223  0.32274948 -0.2658112  0.75198196  1.25913283  1.8363430
#> 2  0.98323397 -0.01007798  1.5988838  0.01027659 -0.28090534  1.4171040
#> 3  1.15639030  1.21199431  0.2726681  0.47484882 -0.03455924  1.1044989
#> 4  0.31103832  0.40145378 -0.7041406  0.76000643  1.11164249 -1.3185637
#> 5 -0.02107968  0.87801675  1.0447126  2.27399881 -0.50113940 -0.8359250
#> 6  0.18278513 -0.69318115  0.8771156 -0.21674141  1.97922043  0.2946002
#>         X237       X238        X239       X240       X241       X242       X243
#> 1 -1.2257643 -0.4437356  1.22149013  0.1300116 -0.3509821  1.2470437  0.1467304
#> 2  0.8945856  1.2144703  0.06041172 -0.6697088  0.4874442  1.1866295  0.6476502
#> 3 -0.8693070  0.2556676 -0.15096920 -0.8787096 -1.0913822  0.5187994  1.5382286
#> 4 -0.3726429 -0.1866151  0.33405190 -1.1447340 -0.4899869  0.7749991 -0.4542659
#> 5 -1.2090475 -1.0453242  0.36710119 -0.1273890 -0.6313337 -0.9350339  0.1830294
#> 6 -0.5088812  0.1168100 -2.53962425 -1.1863721 -2.0095731 -1.1699982  0.7029302
#>         X244        X245       X246        X247       X248       X249
#> 1  0.1835167 -1.38288210 -2.5325048 -0.94611051 -0.8461256  0.6958179
#> 2  0.3417367  1.50710997 -0.5163891 -0.45501675 -2.3248083 -1.2914536
#> 3  0.7965040 -0.05409208 -0.4584467  0.29492328 -0.8613347 -0.6959195
#> 4 -1.3377432 -0.74339459 -1.0884549  0.25489688  1.4238613 -1.8465710
#> 5 -0.6030336  0.59670948  0.5720888  0.08735000  0.1047316 -0.6045985
#> 6  1.3013876  0.22011696 -1.5241773 -0.06984076 -0.6576390  0.4801087
#>         X250       X251       X252        X253       X254        X255
#> 1  0.9109164  0.9293796 -0.3160620 -0.71804430  0.6191531  1.68987841
#> 2  0.6888584 -1.1946772 -0.2789699 -0.38287682  1.0774492  1.91576997
#> 3 -0.8941712  0.8588128 -0.3618273  0.02747764  0.0373096 -0.76201562
#> 4 -0.6448773 -0.2213588  0.2235703  2.30757243  1.0202771 -0.30331560
#> 5 -0.8476468  0.4700096 -1.4701710 -1.38462453 -0.1968507 -2.04705258
#> 6  0.8702351  0.2889126  0.2046640 -1.82092240 -0.9023134 -0.05405798
#>         X256       X257       X258       X259       X260        X261       X262
#> 1 -0.4835611  0.2451781  1.3329868  0.6703154  0.1480744  0.94114128 -0.7594220
#> 2 -0.9571768  0.5353525 -0.6203164 -0.2644426  0.6274602  1.57260622  1.4154896
#> 3 -2.8978442 -2.0330163 -0.3692648 -0.6778353 -0.5759169  1.58833903  0.1085596
#> 4  0.6899373  0.5357275  1.9203771  0.8716876  0.0186613  0.07657764  0.1826169
#> 5 -0.5444262  0.2077547  0.8786257 -1.0452126 -0.4548868 -1.35774182 -1.2388146
#> 6 -0.8751468 -0.6174080  0.6401566  0.4680309  0.8070612  0.95948775  0.1082436
#>         X263       X264        X265        X266        X267       X268
#> 1 -1.1825613 -0.5654893  0.59417684 -0.08713781 -0.24699033  1.0963445
#> 2 -0.5096426 -0.8189065 -1.64294150  0.94837696  0.09578430  1.3787298
#> 3  0.8242907  1.1033175  0.16876329 -0.29496626 -0.76474237  0.1805309
#> 4 -0.5537874  1.0864108 -0.01002871 -0.86723429  0.03654176  0.1427225
#> 5  0.5117001  0.3380091  0.36318814 -1.11308235  0.02587918 -0.1488502
#> 6 -0.1152584  1.1125831  0.68917207  0.44040642  0.02023757 -0.8448325
#>         X269       X270         X271       X272        X273       X274
#> 1  0.4013224  0.4776501 -0.711981293 -0.4925804 -1.20997952 -0.7734149
#> 2 -0.1554355  0.2093945  0.919872961  1.3946096 -0.05635374 -0.7421148
#> 3 -0.3670770  0.3772945 -0.412442259 -1.3178728  0.20367395 -0.5591997
#> 4  1.8987460  2.2959974  1.427688514 -1.4304187 -0.48121063 -0.5593500
#> 5  0.5736978  0.5219353 -0.306611929 -0.1458992 -0.64434690 -0.4474372
#> 6  0.9455028 -0.7606374  0.006591003  1.4268247  0.74286278  1.4351805
#>          X275       X276        X277       X278       X279       X280
#> 1 -1.53968010 -1.8530706 -1.19994376 -1.0318157  0.5311858  0.3200103
#> 2 -0.41628559  0.7957915  1.33957822  0.3715581  1.0304628  0.7688439
#> 3 -0.04693699 -0.9876556  1.09150650  0.7415097 -1.6738707 -1.1241906
#> 4  1.01758485 -0.6525439  0.07966143  0.5421083 -0.7442237 -0.3441896
#> 5 -0.74189278  0.4206835 -0.71772704 -0.1096947  1.7972367  0.4950582
#> 6 -1.35466744  0.7355275 -0.34184271  0.8304400 -0.1828936 -0.4228098
#>          X281        X282       X283       X284       X285       X286
#> 1 -0.58992676 -0.75595100 -0.4235282 -0.4718512  0.6913621 -0.1542750
#> 2  0.55268857  1.08112267  0.7670678  0.8767088  1.6553191  0.1976164
#> 3  0.02898442  0.91763357 -0.2312635  0.2545630 -0.5002775 -0.5438145
#> 4  1.41881182 -1.36197084  0.8525310 -0.8045310  1.0779352 -0.5725344
#> 5 -0.15871743  0.06878613  0.4354195  1.2133546 -0.6985014  1.0014390
#> 6 -0.09984678 -0.27487582  1.5578227  0.8643817  0.1060493  0.6332643
#>         X287        X288        X289        X290        X291       X292
#> 1 -0.5859855 -1.64171713 -2.65684188 -1.10985364  0.47015781 -1.1382535
#> 2 -0.2780619 -2.38687475 -0.86698217  0.57856978  1.82084201 -1.2589578
#> 3 -0.1261617 -0.12303900  0.03491643  0.69536421  0.20415404  0.7029078
#> 4 -1.5334949  0.04413138 -0.06922679  1.17808433  0.09470086  0.2727650
#> 5  0.5608235 -0.61673939 -0.57938623  1.21873620 -0.26589971  0.3208658
#> 6 -0.6429411 -0.22994531 -0.30660701 -0.03815788  0.54200211  0.1882744
#>         X293       X294       X295       X296       X297       X298
#> 1 -0.7456679 -2.6834735 -2.2264824  0.2591657 -1.3468736 -1.0932361
#> 2  0.3041080  1.6243994  1.6820561 -0.2591598  1.3659803  0.8920152
#> 3  0.9164255 -0.7477072  0.3834148  0.1269107 -1.8858212 -0.6584762
#> 4 -0.4631173 -1.5090846  0.5204872  1.4737662 -1.2733382  1.2614707
#> 5 -0.8195618 -0.6807291  1.5177331  1.6312311  1.1651017 -0.6505553
#> 6 -1.0597432  0.4202163  1.5728619  0.1926764 -0.8917468 -0.9387069
#>            X299       X300        X301      X302         X303      X304
#> 1 -0.0937116347 -0.4674909 -1.27929222 1.7267067  2.667648412 2.1214011
#> 2 -0.1436381730  0.2551077 -0.01496003 0.3158678  3.255544063 2.8275698
#> 3  0.0007150745 -2.4703388 -2.63712707 0.7870620 -0.008587158 1.3404626
#> 4  1.6756082930  1.6486052  1.58225528 1.0083631  2.429874993 1.6382989
#> 5 -1.2745810169 -0.6289049 -1.67338441 0.8424228 -0.403442437 0.1585114
#> 6  0.9578267275  0.4027333  1.25480990 0.4341530  1.165917668 0.8474975
#>          X305        X306       X307       X308       X309       X310
#> 1  0.09101085 -0.42625865  0.2345759 -0.8738026 -1.7471883 -1.9377980
#> 2  0.72439917  1.22585061  0.3185538  2.6546742  0.9722184  0.2585024
#> 3  0.71299053  0.74676407  0.6801622  0.3831786 -2.1023398 -0.1621177
#> 4  1.61679089  3.02859758  0.6862236 -0.4170136 -1.0950259  0.2497323
#> 5  0.63861918  0.06416161 -0.4181771  2.0243011  0.4707096 -0.9440489
#> 6 -0.11094266 -0.42241195 -0.4647136  1.2080772  0.7803734 -0.6162235
#>         X311       X312       X313        X314        X315        X316
#> 1 -0.9261938  0.4417176 -0.2948959  0.60087087  1.49674383  0.41932598
#> 2  0.5825660  0.6630357  0.5119250 -1.14808556 -1.41543993  0.06947842
#> 3  0.3108003 -0.5524318  1.1045660 -0.55620276 -0.15288752 -0.50366633
#> 4  0.2797128 -1.0745344  0.4014144 -0.06029024 -0.85469886 -1.38988998
#> 5 -0.1167444  0.7479251 -0.6804145 -0.37815302  0.09155101 -1.02674467
#> 6  1.3883256  0.2173088  0.4106079  0.89669441  1.07766679 -0.13233968
#>         X317       X318       X319        X320        X321       X322
#> 1 -0.2799848 -1.1641202 -0.4217896 -1.40999448 -1.03385981  0.2130562
#> 2  1.3572707 -0.2655103 -1.7042554 -1.55484395  0.12599347  0.8871657
#> 3 -0.8770187 -1.0923015 -0.2819128 -0.76777152  0.40494045  0.1797544
#> 4 -2.7238544 -0.8566270 -0.1663554  0.02781928  0.01387305  0.9584860
#> 5 -0.7765215  1.1188310  0.1879746  0.01080180 -1.81741248 -0.6082996
#> 6 -2.4061124  0.2071628 -1.3818427  0.64798088  2.28185931  1.4312780
#>         X323       X324       X325        X326        X327       X328
#> 1  0.5368267  0.2879569 -0.8265096  0.19677218  0.81921267  0.9963526
#> 2 -0.8707561  0.6437056  0.8717350  0.02534016  1.03590252  0.6539785
#> 3 -1.3768329 -0.2902336  0.4460281  1.48511340  0.42219914  0.7960787
#> 4  1.0443393  1.1016013 -0.3539938 -0.12879190  0.01507308 -0.6265697
#> 5  0.5639682  0.5515174  0.8916800  0.87188187 -0.98812188 -0.3385796
#> 6  0.7850373 -0.3666184 -0.7804183  0.87496657  1.00368420  1.0266590
#>          X329        X330       X331       X332       X333        X334
#> 1 -0.09184258  0.29183812  0.0582008  1.5988160  0.6388953  2.10109399
#> 2 -0.20241127 -0.58387320 -0.9945344 -1.3435339 -0.2197506 -0.09329585
#> 3  0.75160938  0.08163491 -0.7522106  0.4269362 -1.3049084 -0.31772454
#> 4  1.23382020  1.22566449 -0.5758219 -0.6019720 -1.1509049 -1.11257540
#> 5  0.81362999  0.64497147  1.5899213  2.1000141  0.7503279  0.55794851
#> 6  0.87063717  0.42781635 -0.5243669  1.1218359  0.1700400 -2.64652579
#>         X335       X336       X337        X338       X339       X340       X341
#> 1  1.7001634 -0.9260879 -0.8641142  0.61836638 -0.2986905  0.4587552 -0.5504838
#> 2  2.2989769  0.9424343  0.3281748 -1.20999625  0.8850113  1.3958991  0.4963700
#> 3  0.4133526  1.4721054  0.9578255  0.22701428 -0.4914032  0.5625207  2.2651754
#> 4  2.6855264  1.0395694 -1.0428929 -0.09897806 -1.3220359 -1.6077496 -1.0980014
#> 5 -0.1011717 -0.8549150  0.4569555  1.14336034 -0.7716480  1.0553538 -0.4189176
#> 6 -0.7427837 -0.7696233 -0.1565048  1.82959994  1.0342614  0.7112563 -1.0591585
#>          X342       X343        X344        X345        X346       X347
#> 1 -0.37054745 -1.7712862  1.69963613  1.66470955 -0.14605541 -0.1858096
#> 2  0.73541474  0.1710412  0.05983014  0.08537823 -0.31490193 -1.3783870
#> 3  1.46899724  1.6597461 -0.59437509 -0.60748863 -0.38171775 -0.5703641
#> 4 -3.52333808 -2.1530102 -0.85819456  1.22038597 -0.48029320 -0.4005023
#> 5 -0.04793013 -0.1241411 -1.31059286 -1.74104030  1.44185897 -0.2890845
#> 6 -0.98401609  1.1159123 -0.85255259 -0.22470312  0.07855491  0.5986326
#>         X348       X349       X350        X351       X352       X353       X354
#> 1  1.1883808 -1.2009120  0.2958302 -0.39814150 -0.4892421 -1.6175563 -0.1174005
#> 2  0.1945730  1.2989436 -0.1871895  1.67087432 -0.5777402  0.2748654 -0.9994416
#> 3  0.6160829  0.5688427  0.6360451 -0.28851196  0.1287653 -2.2502448  1.0628638
#> 4  0.5597078 -1.0288962  1.1750935  0.59763225 -1.2924345  0.6454919  0.4508952
#> 5 -0.9574182  0.2620773  0.4766460  0.51984546 -2.0594532 -1.0434820 -1.5361491
#> 6  0.1813995 -1.1276641  1.2358707  0.06664891  1.1451990  1.1534070 -0.1220524
#>          X355       X356         X357       X358       X359        X360
#> 1 -0.05938776 -0.9932655 -1.119305283 -0.8325642 -0.9653303  0.03096563
#> 2 -1.53378545  1.1307243  0.460380179  0.5596310 -1.3267627 -1.38824744
#> 3  1.07188013  1.0667986  0.340608773 -1.6184050 -0.1316789 -0.77676493
#> 4  0.08530319  0.2486472  0.003549818  1.0114447  1.0187133  1.41633354
#> 5 -0.12156823 -0.5153584  1.349614181  0.3711660  0.2794616  0.44925933
#> 6  0.15282381 -0.8924679 -1.547458323 -1.1674164 -0.8442083 -0.16659141
#>         X361       X362       X363        X364       X365        X366
#> 1 -0.7988796  1.6245081  1.5778345 -0.78699191 -0.6888955 -1.23570885
#> 2  0.7386564 -0.7954800  1.3369749 -0.06794514 -1.4162563 -0.45855501
#> 3  1.5073489 -1.1847106 -0.7231118 -0.45014299 -0.6885106  0.04002034
#> 4  0.6439574  1.7303051  1.1094474 -0.26230565  2.9780184  0.52115231
#> 5 -1.2533708 -1.3212105 -0.4207985  0.93155689  0.2664559  0.27225167
#> 6 -0.2332240 -0.3990724  0.6930507  1.78448366 -1.0346753 -0.87553465
#>          X367       X368       X369       X370       X371        X372
#> 1  0.37131046 -0.2857877  1.2983654  0.8102755  0.7750976 -2.21571806
#> 2 -1.20088074  0.6134055  0.6265172 -1.2707821 -0.4587694  0.03159852
#> 3 -0.01592845 -0.3112262 -1.3072055  0.2122786 -1.2937428 -0.85580352
#> 4  0.45522200 -0.1976168  0.7627225  1.0875237  0.3111115 -0.69410607
#> 5 -1.05541466 -1.5974869 -1.7838649 -1.8515567 -1.3996448 -1.47663636
#> 6  0.53362699  1.8573468  2.3026379  0.8099008  2.1478568  0.76991575
#>         X373       X374       X375       X376       X377        X378       X379
#> 1  0.2965938 -0.1525208  1.2244350 -0.8210852 -1.3377307  0.60661485  1.6838078
#> 2  0.2385862  0.2331387 -1.2352592 -0.3166369  1.3030780 -0.76805461 -0.2174262
#> 3  1.6714284 -0.2468034 -1.3930311 -1.1259507  1.3254899  1.23888589  1.4115573
#> 4  0.1173108 -0.3513494  2.2985839  1.5167273 -0.6425482 -0.56416851 -0.8958599
#> 5 -1.0379808 -0.7265838  1.8615874 -1.5833221  1.2770543  0.83214836  0.8222189
#> 6 -0.8831487 -0.6897725  0.2585461  0.6447046  0.5095223  0.01595631 -1.7185264
#>          X380        X381        X382       X383         X384       X385
#> 1  0.08729119 -1.29985699 -0.60017037 -1.6693919  0.461155745  0.1212282
#> 2  0.34556564 -1.19491895 -1.84910067 -1.1585582 -0.401712782  2.1266294
#> 3 -0.32138846  0.05265236 -0.06435567 -0.2521725  1.810312789 -0.1573187
#> 4  1.52064810  0.89564015  1.56984663  0.8895171 -0.002674652 -0.3659427
#> 5 -0.71321567  0.52681713 -0.48649516  0.3006066  0.952690110  0.1207966
#> 6 -0.74606089 -0.93266493  0.26777097 -1.1481447 -0.805798314 -0.2964119
#>         X386       X387       X388       X389       X390        X391       X392
#> 1  0.2904451 -0.1194425  0.4111594 -1.7505500 -0.9075924 -0.60600294  0.3354597
#> 2  1.3771350 -0.3249652  0.5376267  1.1980511  1.1655260  1.12609860  0.5721209
#> 3 -0.5958792 -0.4268356  0.9788791 -0.7959936  0.8414174  0.15787921  1.5760428
#> 4 -0.8103593 -0.3289114  0.3234834  0.2442041  0.3583194 -0.40126802  2.9184765
#> 5  0.5360922 -1.9597337 -2.0773122 -1.3786416 -0.6406621  0.05832189 -0.8135288
#> 6 -1.8039269  1.1316186 -0.9483639 -1.6228629 -2.9425203 -1.28971406 -1.1289040
#>          X393       X394        X395       X396        X397        X398
#> 1 -0.07754883 -1.4244247 -0.08860402  0.6660142 -0.03192624 -0.91382006
#> 2  1.38516036  0.8269086 -0.37965353 -1.7312589 -0.10633885 -0.55455373
#> 3 -0.41361477 -0.2685401 -0.02382564  0.6201564  0.30714394 -0.14108730
#> 4  2.00816827  0.9746265  1.25359993  0.3714397 -0.79399603 -0.34868071
#> 5 -0.49678582  0.1726818 -0.07032695  0.7286305 -0.02971122  0.06267331
#> 6 -0.86672993 -1.1477474  0.32452424  2.2107597  0.82287696 -0.02381182
#>         X399       X400       X401       X402       X403       X404        X405
#> 1  0.1109930  0.2739676  1.6188365  0.2619822 -1.6324965 -0.1104779  0.15522556
#> 2 -1.4149472 -1.9802954 -1.1539674 -1.0057697  0.4889209 -0.3957438  0.32321392
#> 3 -0.8378580 -2.1654791  0.9814171  1.2994456  0.2200329  0.2920700  0.09048108
#> 4 -1.4250939 -0.5797861  0.6838300 -0.5440092 -2.5084090  0.3639032 -0.29795187
#> 5  0.4317461 -0.7822367  0.5141681 -1.2881039 -0.6034393  0.7244707  0.50908563
#> 6  0.1965983 -1.4660631 -0.8942419  0.3346099 -0.4283100 -2.2868915 -1.69581942
#>          X406       X407        X408       X409       X410         X411
#> 1  1.94897196  1.4789168 -0.83784676  0.8482711  2.0990473  0.472103059
#> 2  0.29346980 -0.4592686 -0.01840548 -0.5526227 -2.4715188 -0.003048029
#> 3  0.04814511 -0.9597411 -0.06188623  1.9008417  0.0863923  0.593512380
#> 4 -0.11074456 -0.8922030 -0.72314902 -0.1530357 -1.4790623 -1.477600108
#> 5 -0.95474930  0.1378297  0.25405287  0.1180248  0.8626163 -0.155020600
#> 6  0.44562165 -1.2149862 -1.10249709 -0.6373694 -0.6673538 -0.253235770
#>         X412        X413       X414       X415       X416       X417       X418
#> 1  1.2853821  0.59073292 -0.6127083  0.3641950 -0.4551047 -0.5076732 -0.3967667
#> 2  1.9773853  1.88067837  0.7123144  0.2686516  0.5064998 -0.1838720  0.3169461
#> 3 -0.4545815  0.34223822 -0.4435516  1.6249693 -0.2861434  0.5234756 -1.6068175
#> 4 -0.3828926 -0.33194393 -0.8196534 -0.4770486 -0.2120361 -2.3312784 -0.8949515
#> 5  1.0281520 -0.07705256 -1.7752333  0.2927618  1.4543578  0.3800127 -1.2500179
#> 6  2.2486017  1.34731792 -0.6613715 -0.6517804  0.1683774 -1.6315170 -0.7538588
#>         X419       X420        X421        X422       X423       X424
#> 1 -1.4148113 -0.3185488  0.02421454  0.06551171 0.52191232  0.8974322
#> 2 -0.5549401  0.8407395 -0.26371165  1.18348485 0.04903432  0.2466480
#> 3  1.1845209  2.1846356  0.63180472  1.34284003 1.05303010  1.7319953
#> 4 -0.6418496 -1.1087110  0.45113001  0.94706543 1.70157053  0.1905189
#> 5 -0.5809070  1.3391342  0.16184113 -1.04957466 1.35513444 -1.0056428
#> 6  0.1581742 -0.2376590  0.16135797 -0.04736039 0.61499952  1.2158221
#>         X425       X426       X427       X428        X429       X430       X431
#> 1  1.0163426  0.3770453  1.9044320  0.1612020  0.04473769  1.1721631 -1.2618587
#> 2  0.7186814  0.3126385 -2.4788558 -0.6634863 -0.39637433  0.5335338  0.4589368
#> 3 -0.8571149 -0.4822759 -1.3031958 -1.3098659 -0.55196087  0.3439561  1.1281973
#> 4  0.2409685  0.5470237  0.3941247 -0.2956130 -0.44950354  1.8120722 -0.1351709
#> 5 -0.1958521 -0.5410285  0.7322148 -0.2115575 -1.04695794  0.8452455  1.1710350
#> 6  0.4810885  1.0071650 -1.1876098 -1.2973146 -1.13942094 -0.7446137  0.6590433
#>         X432       X433        X434          X435       X436       X437
#> 1  0.1408057  0.8911167  1.89599196 -0.0009351073  0.3066826  1.2825799
#> 2  0.1971752  0.1266117  0.01127606  0.6546604739 -0.9192120 -1.3579030
#> 3  0.1513279  1.4316758  0.41480105 -1.6745988021 -0.6524607  0.1214681
#> 4  0.6323494 -1.7221924 -1.96173623 -1.2393341286 -2.1241489 -1.5867126
#> 5  1.5947082  2.5644223  1.91934666 -0.3377076556  0.8384399 -0.3947282
#> 6 -0.7205398 -0.1085427 -0.99522488 -1.0322948295 -0.8647180  0.6787250
#>          X438        X439       X440       X441       X442       X443
#> 1 -0.12383585 -0.83941205  1.9652890  0.8114734  0.7230973 -0.3613384
#> 2  0.35998519  0.64802974 -0.7016312  0.1130515 -0.7752798 -1.6736113
#> 3  0.56866297  0.70520458  0.6612711  1.2411092  0.3375172  0.0389090
#> 4 -1.03305179 -0.08885239  0.4480516  2.0374489  0.8036212  0.9552622
#> 5 -0.85669278 -0.02390560  0.7470066 -1.6479928 -1.2220029 -0.4586110
#> 6  0.03678541  0.21372953  1.5157040  0.1757768 -0.1387441 -1.6390941
#>         X444       X445       X446       X447       X448        X449
#> 1  0.2867806 -0.2493531  0.3712097  0.5166497 -0.9288262 -1.20500247
#> 2 -0.1743196 -0.3056875 -0.4407120  0.8265232 -0.7664746  1.22001162
#> 3 -0.6897965 -0.2615747 -1.3144022 -0.2267625 -2.9455876  0.96382171
#> 4 -0.3131362 -0.4535501 -0.4639069 -1.4305171 -0.2067842  0.22811152
#> 5 -1.3555066 -1.9667250 -1.6245791 -0.6473584 -0.3250907 -0.35625404
#> 6 -1.5016399 -0.7548021 -0.2327583  0.4962956  0.6087932  0.06812354
#>          X450         X451       X452        X453        X454       X455
#> 1  0.03659401 -0.008248559 -0.8638316 -0.83288663 -1.81947021 -1.9477641
#> 2  0.28438057 -0.252116946 -0.1433490  0.76924741  1.67265299  0.6997692
#> 3 -1.16407312  1.786058426  1.9535890 -0.08182438  0.73346285 -0.2884042
#> 4 -0.76697656 -0.543225382 -0.6040109 -2.27941947  0.09686078 -1.0850403
#> 5 -1.32933048  0.812766956  0.2424772  1.75323692 -0.93720519  0.2112463
#> 6 -1.20527151 -0.762921970  1.7492353  1.39986027 -0.16195666 -0.7318675
#>         X456       X457       X458       X459       X460       X461       X462
#> 1 -0.9512165  0.1455232 -0.1100247  0.1376823 -1.7448402 -0.5570352  0.6682515
#> 2  0.1745326 -0.3591138 -0.6105915 -1.0382030  0.4007868  0.5547551 -1.6659971
#> 3 -0.2451345  0.1360092 -1.1170890 -1.3492692 -0.9112736  0.9246756  0.9767302
#> 4  0.4984745  0.9811046  1.6503280 -0.5674485 -2.0174985 -0.4697883  1.4335174
#> 5 -0.1578377 -2.7992330 -0.8363768 -0.6065037 -0.6921967  1.2913003  0.1008094
#> 6 -0.6068526 -0.2399822  0.3462625 -0.1723190 -2.5227707 -1.7815466 -0.7138301
#>         X463       X464       X465       X466         X467        X468
#> 1  1.3167000  0.7883336  0.1168518 -0.9859769 -0.284185440  0.90455476
#> 2  1.9942090  0.6420137 -0.1293085 -0.2025632  0.172937504 -0.97933899
#> 3  0.7078309  0.3416816 -0.8735687  1.2933905 -0.350932096  1.06138967
#> 4 -0.4651415  0.7035921 -0.2216062 -0.2688878  0.007237282  0.57265645
#> 5  1.2124229 -1.2299297 -0.9433116 -1.3875027  0.550311037 -1.60479869
#> 6 -0.2384445 -0.6639138  0.7548360  0.8107179  0.286399528  0.06175271
#>         X469        X470       X471       X472       X473       X474       X475
#> 1 -0.2544324 -0.34379892  1.5376698  1.9040723  1.0052588  0.3651854 -2.0512827
#> 2  0.6409403 -1.71696427  0.0166513 -0.3460744  0.6929146 -0.6253452  0.2711422
#> 3 -2.6716133 -1.86646409 -0.3850740  0.1526578 -0.3766575 -0.5204961 -1.2407695
#> 4 -0.3647292  0.09895341 -0.1629314 -0.8538290 -0.3146342  0.2757974 -0.5026825
#> 5 -2.6633039 -0.85070715  0.1379637  0.9542730  0.1049527 -0.2899301  0.6656496
#> 6  1.3300783  0.35819740  0.1700401 -0.9562654  0.5164147  1.1172012  0.9962769
#>         X476       X477       X478       X479        X480       X481
#> 1 -1.4990237 -1.1572305 -0.7715807 -0.4213657  0.30703458 -0.1528425
#> 2  0.7984405 -0.2211935  0.5362630  1.1082296  0.41514849 -1.4635133
#> 3  0.3048481  1.1232879  0.1211546 -1.3204662 -0.83939050 -0.2479119
#> 4 -1.8944517 -1.0473675 -0.8129602  0.2238537 -0.07871452  2.7861773
#> 5 -0.7613014  1.2303107  1.3247804  1.4793284 -0.15940606  0.1147122
#> 6  1.5422290  0.2799092 -0.2234897 -1.3555108  0.09556488  1.6008540
#>          X482       X483        X484       X485       X486        X487
#> 1 -0.55228436 -1.8463194  0.66238285  1.0954122  0.1406328 -0.70673396
#> 2 -1.62046788 -0.3544886 -0.41572359 -1.1607726  0.4415227  0.40139678
#> 3  0.14438248 -0.1788045 -0.50109144  0.6406934  1.6157543  0.56034692
#> 4  1.86640177  0.4772138  0.58113389  0.7536208 -0.9063684 -0.05355484
#> 5  0.02601585 -0.1773441  0.07727136  2.2378360  0.6327297 -0.04272353
#> 6  0.39043281 -0.2728488  0.05937776  1.2106093  1.2223833  0.17330295
#>         X488       X489        X490       X491        X492        X493
#> 1 -0.1703188  1.5462463 -0.05884645 -1.1090196  0.48186967  0.03514956
#> 2 -0.2045443 -1.7772914 -0.37362929  0.1526945 -0.19068916  1.18731694
#> 3  0.7690254 -0.3623323 -0.06281004 -0.5089515 -1.12569224  0.39120239
#> 4  1.5857391 -0.3301127  0.54727034 -0.1112612  1.13740856  0.95526109
#> 5  2.5093012  2.2403090  0.84729915 -1.1608047 -0.05587286 -1.48909349
#> 6 -2.9650123 -0.5345334  0.06312235  0.6921504  0.87270888  0.07931799
#>         X494       X495       X496        X497       X498       X499       X500
#> 1 -0.1360599  0.5916489  0.5597106 -1.55399910 -0.9204243  0.8115009 -0.4467533
#> 2  0.7516570  0.2010400 -0.2849151 -1.84166776 -0.7443139 -0.6923109 -1.4686679
#> 3 -0.9961510 -0.2737043 -0.2473623 -0.33877464 -1.4937167  0.1480711 -0.9316575
#> 4  0.8972594  1.2840144 -0.4288300 -1.22995654 -1.1565237 -1.1564881 -1.7458184
#> 5  0.5897580 -0.4099132  0.2631772  0.01478195  0.4535785  0.6543322  1.2211632
#> 6  0.7446063 -0.5900595  0.8627196  1.17586719  0.4452388  0.5607436 -0.1751433