A high dimensional dataset with n equals to 200 and p equals to 500.

A dataset with 200 observations and 500 dimensions is generated from the following process: linear regression model with only first three non-zero coefficients equal to 3, 2, and 1.5 respectively. The covariates are correlated with AR structure (rho=0.3). The error term is normally distributed with zero mean and sd equals to 0.5.

data(HighDim)

Examples

# HighDim dataset is generated from the following process:
n <- 200; p <- 500; sigma <- 0.5
beta <- rep(0, p); nonzero <- c(1, 2, 3); zero <- setdiff(1:p, nonzero)
beta[nonzero] <- c(3, 2, 1.5)
Sigma <- 0.3^(abs(outer(1:p,1:p,"-")))
library(MASS)
X <- mvrnorm(n, rep(0,p), Sigma)
error <- rnorm(n, 0, sigma)

X <- apply(X, 2, scale) * sqrt(n)/sqrt(n-1)
error <- error - mean(error)

Y <- X %*% beta + error
HighDim <- data.frame(Y, X)
head(HighDim)
#>           Y         X1          X2          X3         X4         X5
#> 1 -8.484475 -1.8022130 -1.89992802  0.58407990 -0.2221759  1.5070074
#> 2  3.573816  1.5952714 -0.23653862 -0.48289374 -0.4173879 -0.3889912
#> 3 -1.812384 -1.1711864  0.09532456  0.67888668  0.1673915  0.2425634
#> 4 -3.272272 -1.8594119  0.19964452  1.33902484  1.1499743  0.1884645
#> 5  4.284667  0.7528476  0.66021472  0.58420708  0.6763003  0.9270759
#> 6 -2.682273  0.1039846 -1.80509817 -0.09690419  0.1668564 -0.7801318
#>            X6         X7         X8          X9        X10        X11
#> 1  0.78285801 -1.5008686 -0.4784529 -0.03073041  0.9020240  0.9463057
#> 2 -1.20638323 -1.1090745 -1.4194414 -1.42596193 -1.6734443  0.7852283
#> 3 -0.49277990 -0.2478749 -1.7537732 -0.13585237  1.3948473  0.2136423
#> 4  0.68016979 -1.1922197  0.7466195  0.68124582  0.8972375  1.7742660
#> 5 -0.08160925  0.6853427 -0.4936589 -1.87592121 -1.1713217 -0.6684019
#> 6  1.03258889  0.4529595  1.1240056  0.69976851  1.1597014  1.0377703
#>          X12         X13        X14         X15        X16         X17
#> 1  0.4871772  0.26433802  0.5239505  0.43693310 -0.3304200 -0.01132258
#> 2  2.9245195 -0.07497869 -1.2005728 -0.71158212 -0.8168727 -0.54468134
#> 3  0.5102400 -0.05948139  0.8059274  1.55291925 -0.1092364 -0.77533914
#> 4 -0.6980370 -0.15528744  0.7898448 -0.03448883  1.1997302  1.03958338
#> 5  0.2592505  0.90474552  2.2462259 -0.55770797  0.6642254 -1.51312737
#> 6  0.4732077  0.12371713 -0.1973646  0.13274514  0.7207371  0.73563332
#>           X18        X19        X20        X21          X22        X23
#> 1  0.72774860 -2.1776320  1.8649301 -1.4473332 -0.477849709  0.4763638
#> 2 -0.39093178  0.5367058  2.1015105 -1.5673611 -2.033880608 -2.0184726
#> 3 -0.53968972  2.2422404  0.3795690  0.3786309  1.637273274  1.5923413
#> 4  0.07274656 -1.0277247 -0.5570155  0.7945556  0.739854052  1.1607449
#> 5  0.61015959 -0.1987227 -2.2211040 -1.6283868 -0.003673259  0.5976032
#> 6 -0.32879576 -0.3971687 -0.4377958 -0.9207241 -0.476485851 -0.7552187
#>          X24         X25        X26         X27         X28         X29
#> 1  2.5482051  0.68690056  0.3230196 -0.43452956 -1.14396190 -1.28243986
#> 2  0.2482020 -0.01156484 -0.1980769 -0.40128248  0.02735899 -0.01111342
#> 3 -0.4723248  0.13040961 -1.5557156 -0.01379643 -0.16745306 -0.51531248
#> 4 -1.0002712 -0.13348411  1.0377832  0.39493461  0.02490606  1.25036798
#> 5 -0.0614386  0.51917714  0.1819684  0.76205172 -1.33330527  0.43193315
#> 6 -2.3917480 -0.15606578 -0.1832203 -0.94434924  1.11822654 -0.74251443
#>          X30        X31        X32         X33         X34        X35
#> 1 -2.2334250 -0.8992518  1.2842694  1.30437998 -0.43190241  1.2183791
#> 2  1.1074238 -0.3966662  0.9942345  0.63535193 -0.20493722  0.2149853
#> 3 -1.1033906  1.2565476 -0.7436041 -0.77018746 -0.44051856 -0.0597575
#> 4  1.7230259  0.8973942  1.6698544 -1.53137466  0.05835287  1.1900902
#> 5 -0.7591482  0.2436343 -0.1752570 -0.06677108 -0.34579787 -1.3010668
#> 6 -1.6614160  0.1656242  0.4118590  0.93172573  0.21549470 -0.8637779
#>          X36        X37         X38        X39        X40        X41
#> 1 -0.1740705  1.4310703  0.99205023 -0.2062483  1.2086362  0.4624019
#> 2  0.1947873 -0.7463721 -0.58684579  0.3443027  0.3965955 -0.3610342
#> 3 -0.1488413 -1.0322971 -2.37996216 -0.6082388 -1.8265516  0.2572237
#> 4 -0.3486376  0.5848005  0.08664613 -0.2332889 -1.3821810 -0.7044208
#> 5 -1.0483334  0.2205739 -0.20969194  0.6996671  0.5182328  0.5908201
#> 6 -0.1263657  1.2601514  1.76507568  0.5310928  0.1517165  0.6838437
#>           X42        X43        X44        X45        X46        X47
#> 1  0.74510339  0.1660602  1.2090322  0.4917110 -0.3547475  0.3728337
#> 2 -0.27806297 -0.1513634  0.4157289  0.5413404  0.2336381 -0.7092447
#> 3  0.02108632 -0.6201309 -0.6774347  0.1600726  0.3501042  0.9805517
#> 4 -1.31308882 -0.8953523  0.8941760 -0.1277918  0.7423933  0.1335630
#> 5 -0.60003126 -0.3227476 -0.1705652 -2.8133485 -1.2341722 -1.0396804
#> 6  2.18622302  1.0919823 -1.4044138 -1.5397372 -1.8137934  0.6226977
#>           X48        X49        X50          X51        X52        X53
#> 1 -0.22207905 -0.3886569  0.3946802  0.417196103  1.3550363 -0.2132473
#> 2 -0.03204564  0.9746357  0.6146223 -1.115370675  0.7210927  0.8226461
#> 3  1.32006093 -0.1318285  0.8693133 -0.880849295  0.7702195  1.2169065
#> 4 -0.62132389  3.5092499  1.5042709  1.133191732  0.7061738 -0.6572408
#> 5 -0.27662349 -1.9191824 -1.7596774  0.002343115 -2.0717083 -0.5978526
#> 6 -0.05174005 -0.2207417  0.6999428  2.090216627  0.1294970  0.1272667
#>           X54         X55        X56        X57        X58        X59
#> 1 -0.44486013  0.95573471  1.4441384  2.8964646  1.6702300 0.26948088
#> 2 -0.02888514  0.66351279  0.6604263  0.5738339 -0.8848456 0.01306246
#> 3 -0.20248581 -1.07826172 -0.3124325 -1.7571525 -1.8897978 1.30126139
#> 4 -0.75562261 -1.51561819 -0.3590235  0.3445610  0.9287208 0.28421606
#> 5 -0.83185120 -0.04781233 -1.0398650 -0.1838082  1.0179570 0.08335943
#> 6 -1.78190237 -0.59993938  0.1328578 -1.1261134  0.4929374 1.73967702
#>          X60         X61        X62          X63        X64         X65
#> 1  0.4680378 -0.39582382  0.9477744  1.375324353  0.9462485 -0.20859147
#> 2 -1.2112366 -0.04426084  0.2073974 -0.525281419 -0.7880081 -0.64035179
#> 3 -0.1315502  0.01753059  0.5094526  0.855730542  0.5446938  1.03697639
#> 4  0.6837661 -1.18298979 -2.1189467 -1.147186511  0.2932354  1.23660207
#> 5  0.4773811  0.12824127 -0.9434702 -0.446392399  0.8017964 -0.05392271
#> 6  0.8221498 -0.14191949  0.4237377 -0.008435274  1.8291122  0.56078248
#>            X66        X67        X68        X69        X70         X71
#> 1  0.450797657  1.2603798  0.4734832 -0.6868446  0.9036872  1.19797785
#> 2 -0.481140467 -0.3991205  0.6870242  0.8936234 -0.6091724 -0.13584368
#> 3  0.625225748 -0.5177528 -0.6734098  0.1276112  1.4504076 -0.32699498
#> 4 -1.222555823 -1.0225824 -0.7260535  1.1797411  0.2639829  0.01382335
#> 5  1.141762832  0.2920340  0.4973113  0.1450884  1.0359699 -2.20559710
#> 6  0.009885487  0.3681532  0.1616008  1.9119825  0.3012305 -0.71784016
#>           X72         X73        X74        X75         X76        X77
#> 1  0.16165930 -0.18911626 -0.3084659  0.1162579  0.01367225 -0.5290439
#> 2 -0.02021653  0.52613807  0.8029848 -0.5759745 -0.05162179 -1.7939926
#> 3 -0.27113213 -0.51523785 -0.6022743 -1.0288094  1.01428139  1.1425540
#> 4 -0.01743250  1.58219591  0.9650202  0.6970022  0.48828393  1.2068935
#> 5 -1.32039515  0.07588806 -0.1908271  0.5896536  1.29608232 -0.6802810
#> 6 -1.95105216 -0.59819301  0.1181629  0.1192893  0.44612967  1.6228303
#>          X78        X79        X80          X81        X82        X83
#> 1  0.7035354 -0.1689112 -1.3059147 -1.368030484 -2.3353577 -0.9677502
#> 2 -0.3778561  0.6130017  0.8320682  0.826754459 -0.7074700 -0.4363315
#> 3  1.2309409  0.2261657  0.5112170  0.284309595  1.3291254  0.8460504
#> 4  0.1590663  1.5696527 -0.6808683 -0.597993517 -0.4387426 -1.0850749
#> 5 -0.4857493 -1.5026227  1.5878232 -1.328112682 -1.4804399 -0.6372391
#> 6  0.3738338  0.5245790 -0.2794591  0.001774452 -0.5444610 -0.1327783
#>           X84         X85        X86         X87          X88        X89
#> 1  1.47353526  0.25387688 -0.3960525 -0.01975553 -0.455196496 -1.1096169
#> 2  0.03275118 -0.31716737  0.6456691 -1.02871243 -0.732556264 -0.1456110
#> 3 -0.84387708  1.12434412  0.3824573 -0.94544856  0.134741330  0.3870678
#> 4 -0.82301279 -0.79890859 -0.6880204 -1.15626476 -1.519432990 -0.8792539
#> 5 -1.09293714 -0.24040556 -0.7821436 -1.18421722 -0.366289611  1.3601379
#> 6 -0.42917867 -0.01518489 -1.8026347 -0.07161072 -0.007603626  0.3527757
#>          X90        X91        X92        X93          X94         X95
#> 1  0.3809905  1.7807559  0.8571143  0.5086036 -0.009537112 -0.61714359
#> 2  1.8226579 -0.3210184 -1.2093151 -1.1157037 -0.167821348  0.27452654
#> 3 -1.5900620  0.5202222 -1.7631547 -1.4363155 -2.132433762 -0.03809051
#> 4 -0.6090202 -0.1144151 -0.1614406 -0.7642317 -0.309401678 -1.91382684
#> 5  0.6892234  0.9342345 -0.4573933 -2.0241073 -1.212954009 -1.83331882
#> 6  0.9896738 -1.1596130 -1.3469452 -0.5671084  0.613365178  0.19087591
#>          X96        X97        X98         X99       X100          X101
#> 1  1.5288327  0.3865903 -1.2694554 -1.90162249 -0.6546406  1.8014662376
#> 2 -0.6698911  1.2099743  1.0938747 -1.38166618 -0.7938543  0.0002432089
#> 3 -1.6872487 -1.5587876 -0.2348199  0.57757577 -1.2621216 -0.7954823487
#> 4  0.2907657  0.7099451  0.7093340 -0.05340326 -2.8085030  0.1647867769
#> 5 -1.2736908  0.3826391 -0.0824775 -0.97341963 -0.7414768 -0.2217017484
#> 6 -0.3223477  0.2021087  0.4940407  0.14452152 -0.1050100  0.5515429142
#>          X102       X103        X104        X105       X106       X107
#> 1  1.74157870  0.3888315  0.01145676 -0.06725049  0.1359246  0.7415682
#> 2  0.77979086 -0.8226714 -0.45360424 -0.65074628 -0.8565000 -1.0655674
#> 3  0.01710274 -1.5875482 -1.30389843 -0.14800759 -1.5631568 -0.6071564
#> 4 -0.28447070  0.5198222 -0.52751548  1.15554017  0.3046720  0.8450471
#> 5 -0.70048965  0.4254566 -2.02812242 -0.62987245  0.3186573  1.1898841
#> 6 -1.01864995 -0.9321753 -0.94486641  0.90233656  0.4314837 -0.6198256
#>         X108       X109       X110       X111       X112        X113       X114
#> 1  0.8674191 -0.2530369 -0.4974867  1.2489201  0.5598045  0.70144537  0.2403362
#> 2  0.5944305 -0.5928722  0.4074082  0.2067320 -0.0630550  1.63121029  1.2073312
#> 3  0.6529822  2.4068777 -0.5535363  0.8663476 -0.1631730 -0.03671643 -0.1475930
#> 4 -0.3471347 -0.1186669 -1.2860372 -0.7222105 -0.4275952 -0.80729944  0.4067572
#> 5  1.7035439  0.5333573 -0.5072816  1.2803876  1.1888537  2.11540649  0.1130474
#> 6  1.1356391  0.2907990 -0.0383611 -0.1872566  0.1045454  0.59761696 -0.6106274
#>         X115       X116       X117       X118       X119       X120       X121
#> 1  1.5378741  1.5024159  0.3938754 -0.1302426  1.3320626  0.7818471  0.5037448
#> 2  0.5273268  0.3177221  0.4548350 -0.6415157 -0.2473151  0.9771072  0.5667648
#> 3  0.7354000 -0.5485907 -0.1232797 -0.8093266 -1.5541442 -1.3077412 -0.8376027
#> 4 -1.3146719  0.6827405 -1.0019917  0.3165367 -1.0994403 -0.6109515  0.3128668
#> 5 -0.2200146 -0.1479140 -0.2867227 -0.5734948 -1.2660597 -1.0752046 -1.3319021
#> 6 -0.9914051 -1.5023354 -1.3595751  0.9082136 -0.3569342  0.9187130  0.3220096
#>          X122       X123       X124        X125        X126        X127
#> 1  0.03252245  0.9846677  1.8314257  0.43638136 -1.21721090  1.66849845
#> 2 -0.79355955  1.5745537  0.5595622 -0.49107172 -0.08246327  0.02166739
#> 3 -0.53469924 -0.7706543 -0.9352515 -0.21866753  1.13424756 -1.42409434
#> 4 -0.21587293 -0.7191980  0.8689487  2.18784425  0.38766707  0.43996597
#> 5 -0.35680447  1.2748088  0.6888717 -0.03650452  0.95500073  1.72497058
#> 6 -0.04087406  0.6446782  1.6901064 -0.26454174  0.48986836 -1.25920939
#>          X128       X129        X130       X131        X132        X133
#> 1  0.71807355  0.4787075  1.22865141 -0.6298387  0.70901049 -0.03951205
#> 2 -0.05809742 -1.2812415 -0.30934043 -2.1843244 -0.75481982 -0.74408300
#> 3 -0.10669949  0.2022426 -0.21613751 -1.4631428 -1.21705079 -1.45683615
#> 4 -0.57999143 -0.5873778  0.08807702  0.3312312  0.05884195  2.00351465
#> 5  1.10042945  0.1849149 -0.73356891 -0.8785816  0.10657805  1.75209951
#> 6 -0.72065902 -1.4143853 -2.43540951 -1.0803758 -1.77768505 -0.57078263
#>         X134       X135       X136        X137       X138       X139       X140
#> 1 -0.3926865 -1.4055994 -1.3263113 -0.54721239 -1.7011835 -0.5257256  0.3466403
#> 2 -0.9131010 -0.5446325  1.6989402 -0.75265821 -1.1795379 -0.6111775  0.5239828
#> 3 -0.1427343 -0.7050635  0.9843103 -0.29049927 -0.5985827  0.7244025 -0.2818341
#> 4  0.6370460  0.2941539  0.1722936 -0.04244506  1.6577191  0.7177380  0.7475459
#> 5  0.8288534  1.4992806 -0.0408500 -0.39555578  0.5753429 -1.4249198 -0.5294933
#> 6 -0.4671612 -0.7231986 -1.2161190 -1.68379251 -0.7879817 -2.5377212 -0.8603225
#>           X141       X142       X143       X144        X145         X146
#> 1  0.047323282 -1.6825604 -3.0445668 -0.3741477 -0.64969731  1.508218740
#> 2  1.512704491  0.4351022  2.0224755 -1.0712364  0.01375536 -0.007988851
#> 3 -0.008221891  0.8466386  0.9521036  0.9342296  0.06066434  0.346950453
#> 4  0.169379762  0.7662638  0.9669426  1.8630608  0.18264713  1.095919001
#> 5 -0.995445117 -0.3922948  0.5750742 -0.6167780  0.38195728 -1.023619139
#> 6 -0.352085323 -0.4828243 -1.2529156 -1.2638673 -0.02407032 -0.418032190
#>          X147       X148         X149       X150        X151       X152
#> 1  0.29438155 -0.3326722 -0.006699392  1.0612465  1.22959174  0.6123100
#> 2 -0.16045918  1.0999599 -0.062549599  0.8208810  1.43202038  0.1384056
#> 3  0.62674798  0.6415454  0.307684633 -0.5000518 -0.04710709  1.3191120
#> 4  0.17118865  1.7436234  0.634978715 -1.1569511 -0.92904330  1.0155253
#> 5 -0.09413391  0.5867017 -0.346594646 -0.8685398 -1.92748117 -1.0734402
#> 6 -1.41760756 -0.6025782 -1.893647614 -0.9639848 -0.52445368 -0.4472273
#>          X153       X154       X155      X156       X157       X158        X159
#> 1  1.24330865  0.2719778 -0.2709411  0.360567 -0.5151004 -1.2475373 -0.19699487
#> 2  0.36554911 -0.1943124 -1.5580813 -1.102618 -0.6177318  0.6225137  0.04595675
#> 3  0.13740460  0.7167118 -0.3627833 -1.336445 -2.3982349 -1.5680092 -0.39613376
#> 4 -0.96186016  1.0873507  0.9089836  1.032644  1.0084529  0.2971383  1.44658237
#> 5 -1.99189300 -1.5060423 -1.8732912 -1.156554 -0.1719352 -0.3510610  0.09554054
#> 6  0.04241034  0.7272137  2.6339107  2.437412  0.7722665  0.6871216  0.15262288
#>          X160       X161        X162        X163       X164       X165
#> 1  0.47210054  0.8323402  0.92494317  0.35638986 -0.2466222  0.6336512
#> 2 -0.02510657  0.7185119  0.07812016 -1.03748839  0.4317022 -0.2282787
#> 3 -0.23100421 -0.5448099  0.27378895 -0.37275031  0.4534456  0.4104271
#> 4  1.79573487  0.2202127 -0.75819526 -0.92923653  0.3918328 -2.8826396
#> 5  0.02250830  0.6131037  1.40377489  1.18995278 -0.8886683 -0.5454264
#> 6  0.88741675  1.2169687  0.14791753  0.03143531 -1.0203255 -0.9227712
#>         X166        X167        X168       X169         X170       X171
#> 1  1.5879617 -0.07029924 -0.24871737 -1.8583344 -1.220773532 -1.9355656
#> 2 -0.1768365 -0.44481613  0.08079981  2.0636656  0.005756653 -0.4647740
#> 3  1.7453140  1.81637587  0.59857051  2.1781796 -1.784753020  0.1506637
#> 4 -2.1316836 -0.74305507 -0.18504334 -0.3924606  0.978717533  0.6797782
#> 5 -0.5585893  0.69007879  0.55780667  0.9125855 -0.613534270 -0.2976833
#> 6  2.4022376  1.97066579  1.92313104 -0.7643436  2.436846145 -0.3023334
#>         X172       X173       X174       X175        X176        X177
#> 1 -0.8408170 -0.3515469  0.5188226  0.5316278  0.20022081 -0.98611448
#> 2 -0.2978934  0.2989701 -1.0432815 -1.0379431  0.31760348  0.00468597
#> 3  0.6200160  0.1681401 -0.3180777 -2.6310531 -1.64315187 -0.39793192
#> 4  0.6908883  0.6097147 -0.5737393  1.7933687  0.04717313  1.03966507
#> 5  0.2519215 -0.4712644 -0.6572198  0.9222434  0.40401144  0.65005628
#> 6 -0.7248941 -0.1741899 -1.9738868  0.8375537 -1.35835060 -0.28660198
#>         X178        X179       X180        X181        X182       X183
#> 1  1.3543321  0.48896471  0.4606687 -0.78639930  0.89755096 -0.6478063
#> 2  0.1325001  0.56808440  0.2233511  0.44256933 -0.64730298 -0.5388188
#> 3 -0.4330972  0.10541033  1.1730965 -0.04576125  0.69499382 -0.8540639
#> 4  0.2216966  0.03422356  2.9277436  0.61842493 -0.24895897  1.9318507
#> 5  0.3847420  0.11713480 -1.0854922  0.47228402 -0.58294083 -1.4462676
#> 6 -0.6429158 -0.90886656 -1.1454371 -0.73422781 -0.09014939 -1.4890181
#>          X184        X185       X186       X187       X188        X189
#> 1 -0.01575805  1.14921182  1.7718049  2.1052353  1.2928668 -0.38960980
#> 2 -1.11051649  1.13282047  0.7916346 -0.4340557 -1.1813132  0.74199722
#> 3 -0.44981747  0.89946668 -0.6104599 -1.3547756 -2.2927525 -1.12912885
#> 4  2.60245603  0.61373348  2.2135754  0.8184908  2.0143510  1.16954990
#> 5 -1.78579409 -0.69262378 -2.0159229  0.7226477 -0.1272586 -0.41086651
#> 6 -0.75064449  0.05579808 -0.5122820 -0.4894188 -0.2469818  0.03086097
#>         X190        X191       X192       X193        X194       X195
#> 1 -0.6722392 -0.42682760 -1.6693261  0.5512923 -0.96858904 -2.5260110
#> 2  0.6981245 -0.11926926 -0.5470896  1.1880951  1.89726668  0.5272888
#> 3  0.8898445  0.97833267  0.9156795  1.8912925  1.09673348  1.2214038
#> 4 -0.5514885 -0.00429340 -1.6825296 -1.0246224 -0.06161121  0.8008646
#> 5  1.4338937 -0.93533808 -2.5044935 -1.9836305 -0.60259560 -1.6733550
#> 6  0.7996753  0.01945875  0.8223133  0.6284080 -0.06169054 -1.5958250
#>         X196       X197       X198       X199       X200       X201       X202
#> 1 -1.5215154  1.1246300  0.1808497 -0.6442758 -1.0825513 -1.7219347 -1.3308255
#> 2  0.3857189 -0.5783362  0.4373293  0.7723157  1.5352946  1.5648309 -0.2373250
#> 3 -1.1738280 -1.0009374 -0.4622664  0.7057006  0.0649161  1.7341315  0.7772003
#> 4 -0.1572591 -0.6898325  1.0021664  0.5942218 -0.1129260 -0.3988485 -0.6469693
#> 5 -1.2068658  1.3927545  1.0925773  0.4508065 -0.1145712 -0.5597539  0.3320223
#> 6 -2.5804808 -0.6595074 -2.1874103  0.5342704 -1.0312194  1.1443792  1.2506507
#>          X203        X204       X205       X206        X207       X208
#> 1  0.48332447  0.05127684 -0.4841956 -0.5689872 -0.16595267 -1.7462147
#> 2  0.76847780 -1.46907715 -0.5806838  0.7078476  0.57996842 -0.1310401
#> 3  0.77774375  1.18414477  0.5869629 -0.1945027  0.07217842 -0.6577388
#> 4 -0.45148356  1.48710278  0.5125617 -0.1712478  0.69641315  0.1331536
#> 5  0.58772039 -0.08530634  0.9798274 -0.5404856 -0.12720035 -1.0846076
#> 6 -0.01515552  1.35053120  0.7306218  1.2369920  1.76967049  1.0485084
#>         X209        X210       X211        X212       X213       X214
#> 1 -0.6530878  0.10027573  0.4871096 -0.81191836 -1.7887271 -0.8505041
#> 2  0.7552123 -0.14307464 -1.3141528 -0.01606217 -1.0979318 -0.2129394
#> 3  0.2256557 -0.06474843 -0.3209520 -0.82493283  0.5787218  0.1629158
#> 4 -0.1909367 -1.51127605 -0.3977751 -0.33651493 -0.6631766 -0.6846946
#> 5 -0.7411135 -0.74611705  0.1196911 -0.22630222 -1.8692315 -0.6982456
#> 6  1.4695512 -0.66804617 -0.8296384 -0.90793894 -0.3813945 -0.4692601
#>         X215       X216        X217       X218       X219        X220
#> 1 -0.4024036 -0.9529985 -0.68764385  0.6310217  0.9678931 -0.35781133
#> 2  1.7051336 -1.4051670 -0.52385420 -2.5758698 -0.7502606 -0.01066651
#> 3  1.0751754  0.2620003 -1.40812554  0.5363823 -1.0666512  0.48637562
#> 4  0.5316808 -0.2902031  0.07032731  0.5322299 -0.9346864 -2.73556140
#> 5  0.3219325  1.1914490  0.62803668  0.7918478  0.9985400  0.59181726
#> 6 -0.2721988 -0.3122287 -1.15223627 -1.1431375 -0.1881071 -1.60089184
#>         X221       X222       X223        X224       X225       X226       X227
#> 1 -0.5252966 -0.8920875 -0.6739487 -1.37846685  0.1964016 -1.0337947 -0.4356861
#> 2 -1.0343193  0.5333717  1.5746757 -0.10376121 -0.4684912 -0.6954315 -0.7555973
#> 3  0.7339555  1.0807091  0.1110812  0.18316362  0.6775188 -0.6776435  0.1365760
#> 4 -0.6047603 -0.2008567  0.4418887  0.70636827  0.3694158 -0.3070604 -0.6834776
#> 5  0.8234054 -0.7611362 -0.3360429  1.12554361 -0.4828969  1.4647788  0.4103611
#> 6 -0.9060314  0.1174107  1.3649325 -0.07992089  1.0815490  0.1103057  0.2704949
#>         X228       X229      X230       X231       X232       X233       X234
#> 1 -0.4084910  0.6313125 0.5720827 -0.1811162  0.3528165 -1.2962167  0.2638451
#> 2 -0.3421924 -0.9084688 0.8712518  0.2728275 -1.1180080  0.0310052 -0.1066670
#> 3  0.2818418 -0.1411662 1.2238770  0.1929312  1.9674570  1.8638365  2.1356584
#> 4  0.7060953 -0.3308644 0.6881719  0.2595924  1.0220887  1.3211285  0.8834512
#> 5  0.1198824 -0.1475995 0.5211200 -0.9449161  0.8597935 -1.5363711 -1.3836604
#> 6  0.2331384 -0.5327750 0.9447922 -1.0680948 -0.6052140 -0.2809180 -1.1683377
#>         X235        X236        X237        X238       X239        X240
#> 1 -0.5124082  0.76542255 -1.23655880 -0.91495223 -1.0020019 -0.45662151
#> 2 -0.5769163  0.06111834  0.34823691  0.49688159 -0.8684756 -1.88352964
#> 3  1.1406811 -0.83756726  2.11137976  0.03713355 -1.2811613 -1.18837424
#> 4  0.4899861  1.74388204 -0.44916789 -1.84894255 -0.7275959 -0.12977537
#> 5  0.2321392  1.08563125 -0.92082002  1.20662763  2.5594508 -0.04843518
#> 6 -1.5223899  0.49860955 -0.03834813  0.01120968  0.6720651  1.31461694
#>          X241       X242       X243       X244        X245        X246
#> 1  0.56241797  0.8178191 -0.3677116  0.2004289  0.86844531  0.47755896
#> 2 -1.04792085  0.7628407 -0.5722552 -1.2204452  0.63244863  1.40772065
#> 3 -0.33365303 -0.9665818 -1.0673729  0.4905556  0.07074257  1.19276070
#> 4  0.09671555  0.2722669 -0.2781690  1.1695784 -0.35036907  0.32894080
#> 5 -2.13698081 -1.2290068 -0.9813867 -0.9377670 -0.68091572 -0.09504932
#> 6  0.98982352  0.2059287  0.4386526 -0.6404534 -0.88903550 -0.87954792
#>         X247       X248       X249        X250        X251        X252
#> 1  0.4306534  1.4306814  0.4068084  0.11920626  0.98082334  0.02459445
#> 2  0.7326562 -0.7609871 -0.6421389 -0.07177741 -0.76369015  0.26404444
#> 3  0.2663700  0.7085227  0.5343969  0.79734686  0.03877371  1.51733495
#> 4  0.5850845  0.7723215  0.2637436  1.46730680  1.61342132 -1.29024312
#> 5  0.6729280  0.2725553 -1.1672467 -0.31460924  1.11760416  0.87436668
#> 6 -0.6852423 -1.5390684 -0.3681297  1.05434966  0.17229048 -0.40368266
#>           X253       X254       X255       X256       X257       X258
#> 1  1.505847633 -0.4119071 -0.3886938  1.1412615  1.6635610 -1.2859720
#> 2 -0.005861514 -1.6151806 -0.3412225  0.3247206  0.4462802 -0.6144046
#> 3  0.614032918 -0.4501480 -1.1351793 -2.1757520 -1.1474468  0.3267941
#> 4 -0.135465095  0.5853792  2.1958889  1.4565290  1.2892512  0.8917540
#> 5  1.281161034 -0.5310862  1.5997330  0.3681267  0.8881270 -0.8549759
#> 6 -0.076694739 -0.6119429 -1.7013313 -2.0870771 -0.9658028  0.1751127
#>         X259        X260       X261        X262       X263        X264
#> 1 -0.2004182 -0.09051953  1.2587964  0.97323347  0.1465528 -0.82130911
#> 2  0.2343085  1.17789948  0.9727397  0.05081215 -1.0547188 -0.73893146
#> 3 -0.9607448  0.13897612  1.3216599  0.35781487  0.2986993  0.58895626
#> 4  1.1361683 -0.33946448 -0.2918426 -0.38113647 -0.7707142 -0.09940366
#> 5 -0.5961164  0.91473104  1.1116190 -0.72350864  1.1689807  0.35065933
#> 6  0.3628299  1.05976779  1.2857606  1.26079073  0.2330717  0.01332804
#>         X265        X266       X267        X268       X269       X270
#> 1 -0.5797225  0.54391824  1.4635724 -0.01850569  0.6031381 -0.6451686
#> 2 -1.5994540 -1.67161895 -0.2565147  1.07655042  0.8746610 -0.1146951
#> 3  1.0652332 -0.06603288  0.7768979  0.39335984 -0.0496617  0.7508463
#> 4  0.2790682 -0.79570578 -0.2517336  0.21151633 -1.2501789 -0.7831055
#> 5  0.6277450  0.07041137 -1.0997242 -1.24576457 -0.5001021 -1.5238833
#> 6 -1.3800279  0.70073434  0.9537074 -0.84027666  1.2814770  0.2281292
#>         X271       X272       X273        X274       X275        X276
#> 1 -1.2502416  0.3890002 -0.2487148  1.01946956 -0.6869645  0.81723507
#> 2 -0.2872695  0.9366525 -0.1292581 -0.07790283 -0.7871297 -0.87521638
#> 3  0.7593226 -1.0385663  0.6843812 -0.07924954  0.5600838 -0.13654894
#> 4  2.2958841 -0.2686728 -1.3192008  0.11865896  0.6383388  0.88884224
#> 5 -1.6584735 -1.8437841  0.6411452  1.50028525 -1.1715332  0.46704656
#> 6  1.6793113  0.5187344 -1.3073417 -0.56566225  1.2093212  0.08126737
#>         X277       X278       X279       X280       X281        X282       X283
#> 1  0.1483566 -0.9932276 -1.0101393 -0.5245736 -0.0150743 -0.32185364 -0.9925534
#> 2  0.3648025  0.2766257  0.2077078 -1.2175178  0.7861869  0.96630079  0.9008164
#> 3 -0.2348380 -1.2563646 -0.5916919 -1.0424910 -0.1927302  1.20518102  1.4232125
#> 4 -0.1323124  0.2660712 -0.1518268 -1.2621868 -1.6564369 -0.62654482 -0.6649715
#> 5  1.9347464 -0.1455003 -0.4302014 -1.3084123  0.1463434 -0.02645212  0.3673587
#> 6  0.3793412  0.3258904  0.1994006 -1.8474887  0.9532617  0.88338780 -0.4698831
#>         X284       X285       X286       X287        X288       X289       X290
#> 1  1.3671525  1.0504361 -0.7763253 -0.6476424  0.19025622 -1.7485847 -1.3026513
#> 2 -1.6801229 -1.0229303  0.1699124  0.2937637  0.98134556  1.8879267  1.1530423
#> 3 -0.6544842  1.1890101  0.5180357  0.2942223 -0.87064558  0.2027100  1.4415885
#> 4 -2.1348202 -0.8178115 -1.3493423  0.1438308  0.84211656 -0.3501345 -0.2818443
#> 5 -0.9581136 -0.2541505 -1.0230022 -0.8747080 -0.02546761 -0.6909295 -0.5332985
#> 6 -0.5970737  0.6439136  0.7056915 -0.8153693  0.40592168 -0.6886560 -1.0626147
#>         X291       X292         X293       X294       X295       X296
#> 1 -0.4139934 -1.4591730 -1.773394041 -1.6814927 -0.3886297  0.4127235
#> 2  0.4315173  1.0539303  0.014179998  0.7376974  0.2385385  0.4988189
#> 3  0.7174765  0.6697436  1.054413359 -0.8207024 -0.5371510  0.2306203
#> 4  0.5651815  0.6524804  0.008223591 -1.0306916 -0.9211161 -0.2334470
#> 5  0.9195600 -0.7501810 -0.028781926 -1.3385760 -0.5141586 -0.2028373
#> 6  1.0762706 -0.8442151  0.524916405  2.3179624 -0.3493414 -0.1787013
#>         X297       X298       X299        X300       X301      X302       X303
#> 1  2.2394765  2.4128035  1.0003013 -0.07069926  0.0862794 0.1709873 -1.0745265
#> 2  0.2639195 -0.7805141 -1.7843857 -0.37504133 -0.6810461 1.6277029 -0.4776386
#> 3 -0.2993604  0.6793124  0.9600137  0.17613338  0.2331979 1.8988802  1.3260805
#> 4  1.2744227 -0.2795474  1.1220920 -0.81489452  0.4357687 0.2350634  0.9855607
#> 5 -0.0556065 -1.6448809 -0.7922121 -1.39589754 -0.6193909 0.8639871  1.8331075
#> 6  0.6172994  0.5235972 -1.5092293  1.33524342  1.9472467 1.8778780  0.8922388
#>         X304       X305        X306       X307       X308       X309       X310
#> 1 -2.4001909  1.3532203 -0.06685359  1.8481196 -0.8876926  0.2438688 -1.4370564
#> 2  0.4891489  1.3499366  0.58047555  0.9916877  0.4548793 -1.3039541 -1.8184609
#> 3  0.6203648 -0.8739892  0.62412790  0.1553433  0.0640409  0.6439174  0.3161857
#> 4  1.9057447  0.8071003  0.68597951  0.3022193 -1.5771934 -1.1404909 -0.4284042
#> 5 -1.3987707  0.9962494  1.43451530  0.7591915 -0.8043939 -0.3053129  1.0578417
#> 6 -1.6314202 -0.6655878 -0.52612867 -0.8147951  1.5170516  0.8914945  1.1345652
#>          X311       X312       X313       X314       X315      X316       X317
#> 1  0.20619731 -1.1261968 -1.2007321 -0.6236774  1.8212496 0.2983085 -1.0339047
#> 2 -0.99026197  0.5933287  0.4541778 -0.2627222  0.2097885 1.2488425 -0.2617574
#> 3 -1.51365959  0.2967347  1.9164799  2.4375257  1.5592260 0.1961319 -0.7005184
#> 4 -0.74191707 -0.4487540  0.1741488  0.7668100  0.4765623 0.7144563  0.1103807
#> 5  0.03924926 -1.2928171 -0.9231552 -0.6350574  2.1274845 0.9781414 -0.1794456
#> 6 -0.50909828 -2.3469787 -1.0660272 -0.4952506 -0.7311710 0.2414958 -0.7553107
#>         X318        X319       X320       X321       X322        X323
#> 1 -0.7792126  0.89285090  0.9966118  0.9732618  1.3585784  0.13081849
#> 2  0.6859778  0.12674208 -0.3053500 -1.4798398 -0.7760782 -1.16129283
#> 3  0.8839433  0.74618606  0.7538914  1.6798121  1.3384547  0.08985116
#> 4  1.4214261 -0.04605265  0.5023481  0.9652342  0.9329178 -0.51794881
#> 5 -0.9166483 -0.56849334  1.3879126  1.1423884 -1.0334192 -1.15603438
#> 6 -1.3114072 -0.54748223 -0.1934956  0.3968767 -0.7355002 -0.76384904
#>          X324       X325        X326       X327        X328       X329
#> 1 -1.56292973 -0.8142939 -1.04753812 -0.3165400 -1.14687076  0.3487017
#> 2  0.25441517 -1.3482761  0.24523892  1.0829231  0.11719740 -2.6680134
#> 3  1.21467215  0.7273647  1.22044381  1.4612253  2.06494410 -0.5017005
#> 4 -0.47144478 -1.7594242  0.86824228 -0.4827212 -0.54622983  0.9804295
#> 5 -0.68375528  0.6939642 -0.08499531 -0.7357448 -0.86372091  0.4884331
#> 6  0.02581194 -1.4980637 -0.92405067 -0.5397668  0.03343625  0.3161559
#>         X330       X331        X332        X333       X334       X335
#> 1  0.8650242 -1.1475395  1.84125115  0.39854063 -0.8006925 -0.5964519
#> 2 -0.2694860 -0.1999315  0.12000198  2.65369083  1.4218347  0.2388425
#> 3 -1.4763158  0.7988894 -0.02660519  0.73809303  0.1188962  0.3342356
#> 4  1.6203113 -1.3588477 -0.35715059 -2.02336061 -1.7405228 -0.4476264
#> 5  0.6190545  0.1683677 -2.04208171  1.36859987  0.9867984 -1.2324240
#> 6  0.9064866  0.4980364 -2.17850126  0.08822446 -0.7523708 -0.5591798
#>          X336       X337         X338       X339       X340       X341
#> 1 -0.70505686  0.2063160 -1.220776247 -1.5540304 -0.5608681  0.3385609
#> 2  1.49214083 -0.7791919 -0.450056862  0.8737058 -0.2539137 -1.2940969
#> 3  0.43683168  0.1192129  0.644816208 -1.0518942 -1.7798113 -0.2087187
#> 4  0.32384201  0.9170479 -0.001826077 -0.6998404 -0.6392083 -1.2543906
#> 5  0.58649251 -0.5510689  0.179770128 -0.6333900  0.4509074  0.7204069
#> 6 -0.08831286 -1.3144028 -0.659304333  0.2601504  1.7414244  1.6270677
#>          X342       X343        X344       X345        X346        X347
#> 1 -1.13041774 -1.6797603  0.02879311  2.0630651  0.58235399 -0.96975875
#> 2 -1.01546613  0.1572026  0.11011032  0.5488672 -0.08079158  1.13265889
#> 3  0.14698273 -1.6162467 -1.37900560 -0.2437123 -1.91459023 -0.06137284
#> 4 -1.28850761  0.1359140 -0.13142998 -1.6608790 -1.04721563  0.03500916
#> 5  0.47694194  0.4842211 -0.39285858 -0.9539056  0.42990556  1.20567383
#> 6 -0.07439282  0.7538894 -0.32798021 -0.8508714 -0.67929644 -0.00264641
#>          X348       X349        X350         X351       X352       X353
#> 1  0.21894484  1.7114043  0.82520148  0.375846085 -1.5394524 -2.1810932
#> 2 -0.04777503 -0.5598588 -0.82509567  0.073885957  1.7527354  1.5911619
#> 3  0.28707845  1.8192800 -0.68525274 -1.479634453  0.5017139  0.4084576
#> 4 -1.93001046 -0.5220508 -0.04565879 -1.478178632 -0.9264078 -1.2511057
#> 5  0.09419394  0.3531373  0.34536452  0.319825750 -0.6201761 -1.0453573
#> 6  0.37744055 -1.0930303  0.53233936 -0.001694919  0.6859137 -0.7743943
#>         X354       X355       X356       X357       X358       X359       X360
#> 1  0.6924866  0.3195531  0.4588724  0.3469328  1.2238017 0.03068049  0.5138508
#> 2 -0.1128641  0.1214365  1.2784713 -0.6424013 -1.0167996 0.16083487 -0.9580727
#> 3  0.3365580  0.3778106  0.9871045 -1.1502281  1.0573141 0.16255654  0.4809294
#> 4  0.1281182 -0.9546251 -0.7919148  0.6248593  0.2092835 0.49043890  0.2098421
#> 5  1.1843634  0.7489029  0.8961476  0.5941286 -1.4696811 1.13198589  0.3222725
#> 6  0.3627138 -1.7537139 -0.6894451 -2.3849814 -2.4085242 0.64929277  0.8247543
#>         X361         X362       X363       X364       X365       X366
#> 1  0.4754741  0.002246158  0.2333825  0.5501937 -0.1862159 -1.2432342
#> 2 -0.8685924 -0.205698114  1.0707685  0.1442765  0.2548413  0.5452771
#> 3  0.7738749  0.736203304 -0.5536637 -0.5177434 -0.5384890  0.4441541
#> 4  0.8287797 -0.030372690  0.8700372  0.7579933  1.2997259  1.9811776
#> 5 -1.5125764  0.108838113  1.8326072 -0.6068886 -0.4994693 -1.1321109
#> 6  0.2613573  0.157676734  1.0894092 -1.0867931 -1.5006972 -0.9022589
#>         X367       X368       X369         X370      X371        X372
#> 1 -1.8687452  1.0374024  0.6614926  1.279314482  2.386035 -1.70751922
#> 2 -1.4652620 -1.3847637  0.2290302 -0.795160962 -2.918227 -2.45605062
#> 3 -0.2262367 -1.1553771  0.4075617 -0.139227702 -1.223991 -0.17780311
#> 4  0.8094787 -0.3117751  0.4250468  0.944335386  0.293490  0.26988030
#> 5  0.3970753 -1.1559420 -0.8192150 -0.004679073 -1.188442  0.06465259
#> 6  0.2478495  0.4858959 -0.9484756  0.837703676  1.143852 -0.71226075
#>         X373       X374       X375       X376        X377       X378      X379
#> 1 -1.7452610 -0.7762823 -0.9929676  0.3571689 -0.12073912 -0.2067959 0.6062019
#> 2 -0.8943374  0.4347683  0.6192716 -1.2955581 -1.53017488 -0.3520490 0.4325271
#> 3 -0.6484248 -1.6255938 -0.3571092 -1.8591681 -0.15205514  0.2747124 2.3578834
#> 4 -0.5881448  0.9929588 -0.1777360 -0.1796383 -0.06233213 -0.7453199 0.8873460
#> 5  1.4383596 -0.2858159  0.4184730  0.6152210 -0.68388536  1.3508492 0.2504864
#> 6 -1.2543053 -0.6396095 -1.0552424 -0.3863612 -0.36875283 -0.6531546 0.7923930
#>          X380        X381       X382       X383       X384       X385
#> 1  2.09270960 -1.87126758  0.5040130 -0.7751231  1.3186915 -0.1612692
#> 2  1.69880011  2.96854146  2.2226161  1.0931289  1.9590177 -0.6292649
#> 3 -1.14300487 -1.42697899 -0.2679652  1.4687203 -0.7635450 -0.8183806
#> 4  0.69296005 -0.07125623  0.3385029  0.6041983  0.1715389 -0.7345405
#> 5 -0.01884694 -0.35846665 -0.5239495 -0.8689126 -1.3813523 -0.1556920
#> 6 -1.13203753  0.64558915  1.2343025 -2.0489441 -0.7070063 -0.5507818
#>         X386       X387        X388       X389        X390        X391
#> 1 -2.0206842 -0.3521093 -0.05619020  0.1162463 -0.39149194 -1.02343499
#> 2  0.3224419  0.3107369  1.97964194  0.1825441 -0.33405003 -0.19278824
#> 3 -1.7049282 -0.4113596 -1.04976684 -1.7553739 -0.97388170  1.15628316
#> 4 -1.3688895  0.7773489  0.06207889 -0.1061614  0.47464690  0.36010971
#> 5 -0.3533023  0.7349150  0.34033508  0.6673433 -0.07143846  0.01437702
#> 6  1.8660104  1.0698276  1.71020396  1.0152173  0.07748929 -2.25697860
#>          X392       X393        X394        X395        X396       X397
#> 1 -1.29339050  0.9167827  1.21302503 -0.43469675  0.18505728  2.2711405
#> 2  0.02976703  1.2471976 -0.30674687 -0.02197698  0.04533768  2.5042941
#> 3 -1.17356564  0.3026014 -0.31431314  0.39685473  1.08180652 -0.2569867
#> 4  0.14630144  0.1811425 -0.07967136 -0.31589967  1.45463874  0.9613834
#> 5 -0.62317374 -1.1234247  0.62034355  1.11759970  0.66051263  1.0836407
#> 6 -1.55082776  0.5579635  0.79524372  0.83149067 -0.12814779  0.5217370
#>          X398       X399        X400       X401        X402       X403
#> 1  1.68209519  1.3969022  1.55153735 -0.6467643  1.13562818 -0.3905001
#> 2 -0.09351403 -0.2482559 -0.23200368  0.7490902 -1.50609132 -0.3217893
#> 3  0.26013166  1.2132877 -0.02803224  0.2496502  0.06400344 -0.5386091
#> 4  0.50803552 -1.0966109 -2.37168119 -1.3383475 -0.78137189 -0.9282860
#> 5 -0.75629637 -0.9828031 -0.25052583  0.4113495 -0.11746542 -1.5508286
#> 6  0.54879741 -1.7765728 -3.17558425  0.2417248 -0.83693304  0.5496170
#>          X404       X405         X406       X407        X408       X409
#> 1 -0.37446803  0.9464406  0.538479050 -0.9314642 -1.66993970 -1.0928606
#> 2  1.35987573  0.2720858 -0.629494684 -2.0258194  0.02781622 -0.5918606
#> 3  0.30675740  2.1687819  0.875306021  0.3593720  0.91757541  0.3921276
#> 4 -0.46991080 -0.0732420  0.929867963  1.7690975  1.95816647  1.4767668
#> 5 -1.84029978 -2.4809677 -0.002675477  0.1907699  0.80029571  0.6573713
#> 6 -0.07002037  1.6491870  0.861565528  2.5522688  0.61639088 -1.5722557
#>          X410          X411        X412        X413       X414       X415
#> 1 -1.48602988 -1.3748843505 -0.42679787  0.54218854 -0.4777460  0.3149553
#> 2  0.73108421 -0.4882228820  2.33089840 -1.02008052 -0.5802723 -0.7670349
#> 3  0.01651012 -1.3475157075 -1.24394755 -0.69527024  0.3719306  0.6532357
#> 4  0.09286640  0.6645928954 -1.06064241 -1.08439742 -0.2792008 -0.8680589
#> 5 -1.29424707 -1.4154581745 -1.49210117  0.02561544  1.3601794  1.6411452
#> 6 -1.20905179  0.0003562964 -0.07534041  0.13957004 -0.6070608 -1.5736125
#>           X416        X417       X418        X419        X420       X421
#> 1  1.261246509 -0.46604204 -0.3363670  0.72048885  1.44650533  2.2731808
#> 2  0.300292890  0.55399086  0.5790783 -0.65130981 -0.17822281  0.6507280
#> 3  1.647282605  0.71468665  0.5611630  0.97011558  0.37332699 -0.1167409
#> 4 -0.008655146  0.68546558  0.8284472 -0.08609129  0.26915881 -0.5342740
#> 5  0.705546860  2.69847464 -0.6710378  0.13446715 -0.09395148 -0.9758739
#> 6 -1.505565604 -0.06785549  0.4976620  0.26124363 -0.87717900  0.2552227
#>          X422        X423      X424       X425        X426       X427
#> 1  1.28314514  1.20432805 -1.678094 -0.6176831 -1.64752483  0.5798350
#> 2 -0.66944729 -0.07038105  1.686979  1.0984267  1.66531392  1.4351861
#> 3 -0.54072726 -0.08387652 -0.785023 -1.4553932  1.06374249  2.1123462
#> 4 -1.35859049 -0.58651557  0.726055 -0.9566625  0.08000686 -0.5123323
#> 5  0.07734442 -0.26937237 -1.117496 -0.5191840 -0.10711377  0.2392184
#> 6 -0.67742176 -1.54890287 -1.554192 -0.4905098  0.68358918 -0.8290597
#>         X428       X429       X430         X431       X432          X433
#> 1  0.1838632 -0.3203266 -0.8296824 -0.496240592  1.0032481 -2.977021e-01
#> 2  0.3146717  1.3713927  0.9680705 -0.400339960  1.1803538 -1.794458e-02
#> 3 -0.4890219 -0.4895408 -1.0262967  0.415681207  1.0919665  5.677206e-01
#> 4  1.0829910  1.2018520  1.1600968  1.179211399  0.7433498  2.312589e+00
#> 5  0.2703041 -0.1339784  1.1641457  0.745784987 -1.0323765  1.363803e+00
#> 6 -1.2246770 -1.1406586 -0.7946364  0.008639273  1.4646740  1.490387e-05
#>          X434        X435       X436       X437        X438       X439
#> 1  1.16913478  0.62838912 -1.5026434 -1.8387410  0.06045205  0.5292438
#> 2 -0.53654842  0.19905900  2.3517869  0.6786272 -0.18485964 -0.5591239
#> 3  0.35817827  2.22487428  0.3691447 -0.7391004 -0.28691125  0.8984423
#> 4  0.78542606  0.63732112 -1.8547330 -1.9119964 -2.31788959 -0.2780494
#> 5 -0.51394514  0.55726924  0.7060316  1.1654599 -0.72927112 -0.1608369
#> 6 -0.07199491 -0.04192567 -0.4463179 -0.8101330 -0.56312012  0.7125466
#>         X440       X441       X442       X443       X444        X445
#> 1 -0.9933020 -1.6653634 -1.5380180 -1.4915684 -1.4619423 -0.68482066
#> 2 -0.2559343  1.1568777  0.8414511 -0.5214226  0.3508162 -0.06489156
#> 3  2.4020834  1.9278589  1.3718076 -0.2280151  0.8037062  0.08254275
#> 4 -1.1842069  0.5526659 -1.6374774 -2.2026776 -1.1270286  0.71246447
#> 5  1.6168877 -0.9117401 -0.4490875  0.2510133  0.8031252  0.02641153
#> 6 -0.3031379 -0.8844618  0.5582073 -0.0802503 -0.9200613 -0.01262863
#>          X446        X447       X448        X449       X450       X451
#> 1 -1.07089129 -0.28531816 -0.6929992  0.08794049  0.3551507  0.2480373
#> 2  0.01080244  0.36351470  1.0794362 -0.12360076 -1.1252411  0.3738847
#> 3 -1.05772924  0.06422146 -0.1944891  2.11792452  0.2950578 -0.6598641
#> 4  0.37026931  0.88194126  0.1630310 -0.69963155 -0.3517269 -1.3186698
#> 5  0.56565850  1.12725106 -0.5449672 -0.23770148 -0.6448337 -0.5817640
#> 6  0.29748590 -0.87567894 -0.7780197  0.38593849 -0.6439692 -1.3843612
#>          X452       X453       X454       X455        X456        X457
#> 1  1.29945327 -1.0186822  1.1268120 -0.8224834 -1.51883947 -0.00996089
#> 2 -0.33063223 -0.7420880  2.1413820  1.0038282 -0.30636496  1.54627235
#> 3 -0.55417101 -0.2478346 -0.9768908 -0.7928340  0.09248785  1.13135843
#> 4  0.03753799 -0.5027156 -0.4050427  0.4654366 -0.13210343  0.63241010
#> 5  0.20996983 -0.6371895  0.4489622 -1.9731022 -2.06538543  2.01960739
#> 6  1.02998099  0.6954117  1.4831816  1.7345095  2.06924208  0.09791695
#>         X458        X459       X460       X461       X462       X463       X464
#> 1 1.30705984  0.07714976 -1.2991910 -2.3673134 -2.1837248 -1.3362169 -0.6512716
#> 2 0.42812563  0.96615407  0.2381634  0.5489071 -0.6023685 -0.6203108 -1.1580645
#> 3 0.03783505  0.94652377 -0.3972943  0.4001452  2.3993840  1.7841176  2.0473115
#> 4 0.69552030  2.16704324  0.1475155  1.1535659 -0.9399811 -0.1261846 -0.6304986
#> 5 0.53701643 -0.43591852  0.7688647  1.5736766  1.4233322  1.8834591  0.4562778
#> 6 0.76090685 -0.92399168 -1.7940967 -0.9921775 -0.8888798 -1.0798351 -1.3729699
#>         X465       X466        X467       X468       X469       X470       X471
#> 1 -0.1010232 -1.0131998 -1.19034648  0.5100430  1.2359313  0.9294600  1.4067879
#> 2 -0.2936412 -1.6622360  1.42046111  0.3728430  0.3510162 -1.7689658 -0.2904463
#> 3 -0.3739902 -1.3126062 -0.03454805 -1.3287674 -1.1697267 -1.4555102 -0.9134736
#> 4  0.5556055  1.3692547  0.41232030 -0.9983033 -0.2435460  0.4900234 -0.7141694
#> 5  0.1281720  0.6686265 -1.90675367 -0.8131866 -1.4930920 -0.9587903  0.7195040
#> 6  1.0360323  1.4250785  0.11234955  1.5162737 -1.0732246 -1.8362691 -2.2239304
#>          X472       X473         X474         X475        X476       X477
#> 1  0.01887018 -1.2165620 -0.445132861 -0.923606759  0.67926384  0.6135814
#> 2  0.29630554 -0.6076395 -1.560715346 -0.959444319  0.36011092 -1.2190831
#> 3  0.90145370  0.8784988  0.001823713  0.562548663  0.80416941  0.3590277
#> 4  1.13718294  1.3660460  1.402401500 -1.049707521  1.07094448  0.2817889
#> 5  1.49371552  2.1288695  1.363150799  0.762718115  0.72806855  0.0870734
#> 6 -1.28945112 -0.1224240 -0.816365190 -0.004108797 -0.04896413 -0.1098261
#>          X478        X479       X480       X481       X482       X483
#> 1 -0.51811578 -1.22756694 -1.3751162  1.4681679  1.6129420  0.4383245
#> 2 -3.05918151 -0.83938368 -0.9603778 -0.2747921  0.5645936 -0.8397516
#> 3 -0.28249989 -0.25955841 -0.9272949 -2.5893646 -1.7983847 -0.7630834
#> 4 -0.54843373  0.46795327 -0.5720260 -0.2222767  0.6639503  0.4040933
#> 5  0.47366022  0.07411268  0.4950384  0.2865692  0.5090050 -0.3103592
#> 6 -0.02413904 -1.00451034  0.3993354  0.8553566  1.3726093 -0.4928974
#>          X484        X485       X486        X487       X488       X489
#> 1 -0.09615354  0.81831840  1.2376082  0.99293791 -0.1754057 -0.9084751
#> 2 -0.15585917  0.19132961 -0.6616413  0.30040880  1.0950023  1.0216514
#> 3 -1.51984779 -0.62845606 -0.3994574 -1.43993983  1.0481165  0.5200617
#> 4  0.74670341  0.10909135  1.4819178 -0.80997050 -0.5201977  1.1091577
#> 5 -0.40885693 -0.04721803  0.1599485 -0.48806686 -1.8379618 -0.1780670
#> 6  0.52076713  1.27824612  1.8048754 -0.09204548  0.6830070  0.9955078
#>         X490       X491       X492       X493       X494        X495
#> 1 -0.8248246 -1.1757598 -0.5126921  0.3165394 -1.2591368 -0.51118602
#> 2 -1.0259128 -1.1322837  1.9136717  1.5107356  0.9261247 -0.11685926
#> 3  0.6666079 -0.5749369 -0.3591456 -0.9930355 -0.8149010 -0.19462072
#> 4  0.7172945 -1.2142071 -0.9276829 -0.3647419  0.6780960 -0.98104162
#> 5  0.1069541  1.3648527  2.1632753  1.1397526 -0.9455900  1.49682938
#> 6 -1.0255570 -0.6795388  2.0517235  0.7866867  0.9936949 -0.05412441
#>          X496       X497       X498       X499       X500
#> 1 -0.89405788  0.7845457  0.1058604  0.7975191  0.9372415
#> 2  0.06264754  0.6666539  0.4089920  0.4682468  0.1089008
#> 3 -0.30505732  0.1981965 -1.2405608 -0.5977843 -1.0914567
#> 4  0.22739886  1.3942333  0.4765515  0.3354033 -1.0327549
#> 5  0.36202547  1.4964203 -1.6001454  0.2044743  1.4555652
#> 6  0.85599410 -1.6043856 -2.6246257 -0.8065283 -0.2854864