A high dimensional dataset with n equals to 200 and p equals to 500.

A dataset with 200 observations and 500 dimensions is generated from the following process: linear regression model with only first three non-zero coefficients equal to 3, 2, and 1.5 respectively. The covariates are correlated with AR structure (rho=0.3). The error term is normally distributed with zero mean and sd equals to 0.5.

data(HighDim)

Examples

# HighDim dataset is generated from the following process:
n <- 200; p <- 500; sigma <- 0.5
beta <- rep(0, p); nonzero <- c(1, 2, 3); zero <- setdiff(1:p, nonzero)
beta[nonzero] <- c(3, 2, 1.5)
Sigma <- 0.3^(abs(outer(1:p,1:p,"-")))
library(MASS)
X <- mvrnorm(n, rep(0,p), Sigma)
error <- rnorm(n, 0, sigma)

X <- apply(X, 2, scale) * sqrt(n)/sqrt(n-1)
error <- error - mean(error)

Y <- X %*% beta + error
HighDim <- data.frame(Y, X)
head(HighDim)

#>            Y         X1         X2         X3         X4         X5          X6
#> 1 -7.4192877  0.0863133 -1.6682149 -2.4581793 -1.3992108  0.4848767  0.09811417
#> 2 -6.8238845 -1.0987566 -0.7250576 -1.4241543 -1.5847569 -1.3581466 -2.12442419
#> 3 -3.2694255 -0.3927008  0.1385922 -1.7148533  0.4354128 -0.8640058 -1.09989444
#> 4 -8.2596589 -0.1885928 -1.9488392 -2.5213063 -0.4411994  0.2586261  1.23635219
#> 5  0.1363329 -0.2048344 -0.4176927  0.8146770  2.3550950 -0.6023713 -0.59703258
#> 6 -1.8472711 -0.6566556 -0.1497862  0.1879229  0.6278829  0.5006136  2.04351731
#>             X7         X8         X9         X10         X11        X12
#> 1  0.894739597  1.7048723 -0.3812686 -0.54167271 -0.29373981  0.4673539
#> 2 -0.960616641 -1.9882099 -1.1655335 -1.25495818 -0.32677601 -1.8788898
#> 3 -0.626743252 -0.7722067 -1.0355069 -0.09733266 -0.02641062 -0.1411804
#> 4 -0.001223792 -1.1425224 -0.2857225  0.12678314 -0.09671955 -1.3367465
#> 5 -0.413528858  0.3175860  0.9299019 -1.22075334 -0.71608346  0.4077383
#> 6  0.061531301 -0.7070893  0.1754162 -0.07056827  0.26210868 -0.5849007
#>          X13        X14         X15        X16          X17        X18
#> 1 -1.1537577  1.6916594  2.23926261  0.3901963  0.007724614 0.09153433
#> 2 -0.4044667  0.5197467  0.94681278  0.2565733  1.939428609 1.12256119
#> 3  0.2482573  1.3225770  0.03654341  0.4616450 -0.164043935 0.10515155
#> 4 -0.1024193  1.7386812  1.81059455  0.1783461  0.847258074 1.29721092
#> 5 -0.2071572 -1.5710672 -1.86950658  0.7703391 -0.311238163 1.14314859
#> 6  0.3819103  1.0667663 -1.67165597 -0.1646461  0.619606925 0.52063731
#>          X19         X20         X21         X22         X23        X24
#> 1 0.45197787 -0.09497406 -0.09066815  0.86200281  0.33519144 -0.0417935
#> 2 0.75105088 -0.46201420  0.67339464  0.09218585  1.34353890 -0.3583386
#> 3 0.08720415  1.64381978  0.01893093  0.19588962  2.25169831  0.9927850
#> 4 0.41159080  0.93378199  1.29934050  0.83552336 -0.13417625 -1.4265676
#> 5 1.24373714  1.39889097  0.27674452 -1.55043156  0.47917223 -0.4454968
#> 6 0.50697647  2.10065909  0.77193654  0.88439488 -0.07031015 -0.6709955
#>          X25        X26          X27          X28         X29        X30
#> 1  1.0053395  0.8559862  0.824354264  0.695746613  1.02131075  1.2838495
#> 2  0.2490671 -0.4939814  0.008282046  0.146798677 -0.74019436 -0.4403265
#> 3  0.8791750  0.0599409 -1.316620574 -0.006192229 -0.29422220  1.2912811
#> 4 -0.8891836  0.7842697  1.729810205  0.199364830 -0.08642575  0.7823949
#> 5  0.5616753 -0.1896454  1.425591882  1.463234967  0.43194674 -1.3087875
#> 6  0.7560114 -0.5641884  1.400654114 -0.655474050 -0.55102939 -1.0268196
#>           X31         X32        X33        X34         X35        X36
#> 1 -0.74998998  0.52409027  0.6196002  1.0789872 -0.89384707 -1.2808035
#> 2 -0.21207859 -0.74777542 -1.0925308  0.3581958 -2.59753282 -1.5158064
#> 3 -0.47661482  1.05106266  0.9852890 -0.4914321  2.79117548  1.7703718
#> 4  0.09448202  1.04534359 -0.4486098 -0.3731394  0.05944598  0.1893283
#> 5 -1.11085899  1.39040358  0.3850753  0.8515244  1.62041932  0.2453646
#> 6  0.18915369  0.05948124 -0.8691234 -0.5299421 -1.12454541 -0.9043471
#>           X37         X38        X39         X40         X41         X42
#> 1 -0.76479892 -0.14091610 -0.6427983 -0.48952653 -0.77868285 -0.60035019
#> 2 -2.17481951  0.53734562  0.8780186  0.05218139 -0.43617689 -0.32648908
#> 3  1.17936872  0.31457159 -0.9318813 -2.92397781 -0.07573153 -0.02935757
#> 4  0.24011441 -0.27437253  0.5495846 -0.34035877  0.17032329  0.22497772
#> 5 -0.68529849  0.03701933 -0.1277075  0.69288831  0.25235430  1.34305060
#> 6 -0.07397658 -1.06785477  2.3762861  0.36201330  0.87011913 -1.06316605
#>            X43        X44        X45        X46        X47        X48
#> 1  0.525608880 -0.4132931  0.0775867 -1.4834023  1.0050608  1.7486029
#> 2 -0.005063632  0.2221750  0.3584473  0.1782370 -1.1098349 -1.2604861
#> 3 -0.977167295 -1.8086370  0.9701936  0.5705638 -0.7833326  1.1555604
#> 4 -1.613005593 -0.1726590  1.1252327 -1.4517712  0.3034873  0.4914330
#> 5  0.186129828 -0.9556878  0.9232345  1.1837787  0.3566797 -1.1927331
#> 6  0.074986873 -0.2662723 -0.6454281  1.0371828  0.7843891 -0.4269854
#>          X49        X50        X51        X52         X53        X54        X55
#> 1  0.3986590  1.0404220 -0.5090334 -0.6137870  0.68254290  0.8057285  1.6034682
#> 2 -0.3251228 -1.3527459 -0.5469487  0.7429457 -0.19012021 -1.2681607 -0.5226127
#> 3 -0.0434279  0.5695731 -1.0183414  0.2925825  0.85265570 -1.1651621  0.6953967
#> 4 -0.5608560 -1.8301380 -1.6247730 -1.2548088  0.04405838  0.2806365 -1.0428896
#> 5 -0.6336825 -0.2122067 -1.0217443 -0.4500265  0.10515648  0.9749978  0.2481684
#> 6  1.3154600 -0.7885738 -0.1996135  0.2687673 -1.05081585  1.0975978  1.1692226
#>          X56         X57        X58        X59        X60        X61        X62
#> 1  1.1388560  0.52534126  0.8073616 -0.2171774  0.1699467  1.2874744  1.0564189
#> 2 -1.1862560  0.02739554  0.1773070  0.8053536 -1.5366835 -0.2143169 -1.5279292
#> 3  1.3161719 -0.28054652  0.2233597 -0.6135499  1.0676061 -0.5401349 -2.0023776
#> 4 -0.5711471 -1.37271757 -0.2832928  0.7853216 -2.0594895  0.4567300  0.4902770
#> 5  0.7916488  0.48363403  0.5348962 -0.6562884 -1.9229293 -0.7143591  1.3492666
#> 6  0.1218965  0.31653975 -0.1930719 -1.9424273 -0.5885549  1.7644486  0.5078149
#>          X63         X64         X65        X66        X67         X68
#> 1  0.4646357  0.26750147 -0.60565130 -1.5891568 -1.5624011  0.73589797
#> 2 -0.7357409 -0.00361001 -0.87573571  0.3341086  1.0847916  1.42101002
#> 3 -0.5850324  1.55786914 -0.08505970 -1.0932850 -0.3464121 -0.38983822
#> 4 -1.0029295 -0.19054103 -0.02016796  1.0145106  0.7870210  1.76690963
#> 5  0.1331473  0.30199729 -1.21039019  0.7878133  0.1984179  0.56861459
#> 6 -1.5233868 -0.84987908  0.50170628  0.2653979  0.1945886 -0.04829023
#>          X69        X70        X71        X72        X73        X74        X75
#> 1  0.8309449  1.9783849  1.5505889  1.4872937 -0.4550441 -0.1647442 -0.1638710
#> 2 -0.2576336 -2.6065057  0.5222685  0.3520017 -0.7904856 -0.8184488 -0.2343579
#> 3 -0.9106852 -0.1041861 -1.0135676 -0.4502460  0.6854112 -1.8779166 -0.2614689
#> 4  1.6731033  0.6214964 -0.9625132 -0.7039227 -1.8824175 -1.1175242 -0.6044139
#> 5  0.1400703  1.3069576  0.2437936 -0.8209301  0.5053668  1.7693909 -0.2284906
#> 6 -1.2457427 -0.4988039  0.1562255 -0.2465310  0.1581898  0.7515600  1.9015066
#>          X76       X77         X78        X79        X80         X81        X82
#> 1  0.3310857 -1.722739 -0.46498700 -0.1724646  0.2297013 -0.07740975 -0.6441468
#> 2 -2.1367245 -1.625934 -1.53483914 -1.6034701 -0.4144630  0.57727291  0.5842728
#> 3  0.1335889 -1.142001 -0.87402046 -1.5581167 -0.4655357 -0.32412249  1.9391854
#> 4 -1.6496600 -1.679289  0.08405058  1.2124201 -0.2474226  1.04548904  0.4302691
#> 5  0.2092789 -2.175937 -1.69616824 -0.9397166 -0.3351982  1.47598233  1.2098001
#> 6  1.0129356 -1.443364 -0.53960001 -2.1906950 -1.1280249 -0.93151139 -1.2280110
#>           X83        X84        X85         X86        X87        X88
#> 1 -1.29928509 -3.0648334 -0.6813363  0.32589730 -0.4806588 -0.1465259
#> 2  0.74230682  0.2799684  0.7257448  0.29920819  0.5162454  1.6916943
#> 3 -0.03377261  0.3112656  0.9207632  1.62825987  1.1302060  1.1340010
#> 4  1.97591530  1.8530399 -2.1437831 -0.38852343 -0.6159186  0.5152997
#> 5  1.43130098  0.5613014  1.2973897  0.02879234 -1.9400775 -0.6408424
#> 6  0.93614018  2.2526643  1.8848620  0.62918014 -0.1271799 -0.2540553
#>          X89        X90        X91         X92        X93         X94
#> 1 -0.3981970 -0.9748223 -1.7374961 -1.51162967 -0.1331828 -0.15498589
#> 2  0.3823159  1.2787146  0.5815489  1.31074437  2.3470358  0.68882399
#> 3  1.0458283 -0.6898783 -1.4796275 -0.06194143  0.0658251  0.54668716
#> 4  0.8258080  0.7356478  0.5083930 -0.02943333 -0.4431335 -0.00840658
#> 5  1.3353873  1.5055445  0.6658304 -0.72700351  0.4376099 -0.56986455
#> 6 -2.0163936 -1.9701467 -0.4724200  0.52263909 -0.4667615 -0.81802193
#>          X95          X96        X97        X98         X99       X100
#> 1  1.0205667  1.182578765  0.9692685 -0.2509226  0.53386088 -0.2743824
#> 2  0.9480390  0.114065671  0.4417699  0.7513991  0.02920487 -1.9551589
#> 3  1.1733272 -0.044259281 -0.3751073  0.8322965  0.75021766  0.6646437
#> 4 -0.4340610  1.380854177  1.0894564 -0.6997614  2.02142566  1.6496641
#> 5 -1.2547010  0.606495605 -0.6690775  0.1841269 -0.19704350  0.5150799
#> 6  0.2749854  0.008322667  1.0533372  1.2151955 -0.63511239  0.6854757
#>         X101        X102       X103       X104       X105        X106
#> 1  1.3043152  1.66719765  1.7927616  0.3341564  0.9651138 -0.46911092
#> 2  0.5625566 -0.03719485 -0.4736125  0.4191185 -1.6544370  0.04125054
#> 3  1.0809530  0.65508505 -0.9851073  1.4356846  0.4912063 -2.87529810
#> 4 -1.6286299 -1.54348980 -1.4302709 -1.5123570  0.3778741 -0.59732810
#> 5  0.5626754 -1.13413152 -0.6368043 -1.6190818 -0.4484732 -0.64998839
#> 6 -0.4144002  1.44675418  0.5501496  0.1886222 -0.7936947  1.30303061
#>           X107       X108        X109       X110       X111       X112
#> 1  0.575362356  1.4347299 -0.37718180 -0.3185801 -0.6936739 -0.2619430
#> 2  0.090980733  0.1690479 -0.33716361 -1.2365761 -0.8184045 -0.4053778
#> 3 -1.645106695 -1.5990527  0.96286055 -1.4079528 -1.2604873 -0.6603737
#> 4  0.029267856 -0.3654326  1.49694728  0.8974475  0.4240709  1.5633942
#> 5 -0.009754471  0.5349848  0.54964897 -0.1813398  2.0919497  0.2583283
#> 6  0.002484703 -1.6362571 -0.09907612  1.7250229  0.4896366 -0.4484729
#>         X113       X114       X115       X116        X117        X118
#> 1 -0.8632458 -1.2678315 -0.8945801  0.1593412 -0.13547257  1.01534016
#> 2  1.7295317 -0.2952637  0.9227206  0.4044431  0.56356753 -0.50655166
#> 3 -0.2095007  1.0493215  0.7317118  0.5734394  1.28244431 -0.51093677
#> 4  1.6610926  1.4395578 -0.4614187  0.6514022 -0.24780983  0.56970910
#> 5  0.4076597  1.0954580 -1.8104283 -0.7102204 -0.43796389 -0.47357968
#> 6  1.8361362  0.1694150  1.4436657 -0.7468673 -0.01033799 -0.03119562
#>         X119       X120        X121        X122         X123      X124
#> 1  1.6538802  0.5704348  0.63221848 -0.21297435  0.100303284 1.0109466
#> 2  0.7719030 -0.3099151 -0.51652391 -0.57022819  0.106359903 0.1854815
#> 3 -0.6303322  0.7539878  0.37338683  0.05040826  0.703999347 0.5925632
#> 4 -0.2971519  0.3061660 -0.92574892 -1.55452799 -0.917010540 1.5144350
#> 5 -2.0726145 -1.6643729 -0.03602387  0.96959573  0.866783231 1.0062442
#> 6 -0.2911930 -1.3819041  0.91413519 -0.11518249 -0.009601252 0.8717430
#>         X125        X126         X127       X128        X129       X130
#> 1  1.8236682 -0.42917612  0.692946375  0.9545833 -1.26259719 -0.4150164
#> 2  0.2539376 -0.59665228 -1.045981570  0.5327420  0.50770417  0.5066916
#> 3 -0.4782635 -0.90388883 -0.585924965 -1.8204225 -0.04433963  0.3620451
#> 4 -0.5341415  0.17750960  0.002227367 -0.6439823  0.91950873  0.2854091
#> 5 -0.5036558  0.08009913  1.075803761  1.1661918 -0.12852839 -0.3700623
#> 6  0.6069912 -0.45002999 -0.361073298  0.1245817 -1.50978818 -1.8923508
#>         X131        X132       X133        X134        X135        X136
#> 1 -0.3148570 -0.09744342  0.1524850 -0.29490904  0.41546092  1.81970201
#> 2  1.7160157 -1.12881964 -1.7939446 -1.88692348 -0.18271490  0.02617961
#> 3  1.7162458  1.15039773  0.8421764  0.29675750  1.57778706  0.60073817
#> 4  0.3564466 -0.83562934 -0.1068135  0.02652326 -0.10221108 -1.19081065
#> 5 -0.5920821 -0.94775649  0.3523086  1.08032759  0.82942716  0.55507160
#> 6 -0.9204986 -0.24794680  0.1899083  0.84019548 -0.04738725  0.37557581
#>          X137        X138       X139       X140       X141       X142
#> 1  0.05576211 -0.05898426  2.2415406  1.1320943  2.0590538  0.8629434
#> 2 -0.21077521 -0.98970874 -0.9761880 -0.2222881 -1.0164320 -1.6249809
#> 3  1.16846043 -0.07680672 -0.8721614 -0.8366857 -2.0476436 -1.4448673
#> 4 -0.08897954 -0.25916004  1.1196033 -0.2099422 -0.5995701 -0.4754445
#> 5  1.27645892  0.70056477 -1.4754112 -0.9600210  0.3122309  1.0409402
#> 6 -0.18705055 -2.12479885 -1.1009456 -0.6694591 -0.6548489 -0.3101703
#>         X143       X144       X145        X146       X147       X148       X149
#> 1 -0.2677058  0.4244238  0.9102769  1.73543670  1.1780508 -0.0265151 -0.1112813
#> 2 -0.3872955  0.3670144  0.1328578  1.28561778 -1.1124699  0.2167214  0.9295177
#> 3 -0.3343480  0.2319883  1.0533371  1.52031406  0.8527566 -0.1797337 -1.1566792
#> 4 -1.0800016  1.3996874 -0.1584964  0.05045487 -1.5068454 -0.1844026  0.3443952
#> 5  0.5503022  0.9239654  0.1208996  0.60752707 -0.4306277 -1.4603178  0.1243335
#> 6  0.7860221 -0.9759368 -0.1685733 -0.73088645 -1.0055447 -1.3174856 -1.4977160
#>         X150       X151       X152       X153        X154       X155       X156
#> 1  1.3384924  1.9336148  0.5885317 -0.7721812  0.89186474  2.0170143  1.2944397
#> 2 -0.9995716 -0.5523338 -0.1141568 -0.6143679  0.01372035 -0.9188334  0.7947000
#> 3 -1.3387655 -1.1044285  0.8078906  1.8416221  0.25726239  1.7728134  0.1095708
#> 4  0.1991794 -0.2225943 -1.0521246 -1.7665682  1.88223224 -1.8859283  0.3670041
#> 5 -0.2820307  1.8035120  0.1808058 -1.8871388 -2.13506590 -0.4216202  0.7209906
#> 6 -1.9647046  0.1334967 -0.6856272  1.2269438  0.04530219 -1.8833741 -0.1050976
#>         X157        X158        X159       X160       X161       X162
#> 1 -0.1275631  1.54294249  0.50745383  0.6880611  0.5137167 -1.1293496
#> 2 -0.1415264 -0.86577388  0.09672465  0.4716734  0.2985378 -0.2261466
#> 3  1.9180858  1.07027972 -1.71483420 -0.5723285  2.1551026  0.7230081
#> 4  0.2306871  0.58051668 -0.71523562  0.3508378  1.3245645 -0.4101892
#> 5  1.5680687  1.07362359  0.49082637 -0.5915040  0.5402850  0.8045326
#> 6 -0.5600143  0.07748908  1.32496906  1.2558395 -1.1124445  0.1541634
#>         X163       X164       X165       X166       X167       X168       X169
#> 1 -0.6141693  0.3616457  0.9747146 -1.0624597 -0.8625007  0.2034896  0.8077815
#> 2 -0.4163661 -0.1429503 -0.4289183 -1.3408719 -1.1915478 -0.2311664 -1.5322128
#> 3  0.7773293  0.9677715  1.5427140 -0.0262071  0.1448452 -0.2767146 -0.4054665
#> 4  1.2384771  0.8722754  0.2378787 -0.1565486 -0.8695492 -1.0035101  0.4800888
#> 5  2.8089901  1.9201705 -0.9395079  1.1255131 -1.1267563 -0.9207563  0.3116325
#> 6  0.7562868  1.7570962  0.8200663 -1.0189878 -0.2843881 -0.3151174 -0.7795558
#>         X170        X171       X172       X173      X174       X175       X176
#> 1 -0.8154905 -2.09528452 -1.1854954  0.1530233 1.1184866  0.5557792  0.5310673
#> 2 -0.4877308  0.07982924  0.4339566  0.8781066 1.4865615  1.0327686 -0.1496489
#> 3 -0.1019161  1.24249887 -0.3026686 -1.4714432 0.9380371  1.3078809  0.7222437
#> 4  1.2090481  0.22005764  1.3883805  0.5390915 1.0976152  0.7431737  0.8721957
#> 5  0.3335338  0.27619062  0.6059145 -0.3291898 1.6507513 -0.1724559 -0.3228749
#> 6 -1.3835053 -0.49703982 -1.1938339  1.8243101 1.1155773 -0.8623341  0.2293990
#>         X177       X178       X179       X180       X181        X182       X183
#> 1 -0.2200556 -0.7689655  1.1034686  0.2874198  0.5421739  1.50471287  1.5518979
#> 2 -1.3555498  0.0330733 -0.8952344 -0.6699159  1.2702600  0.02121954  0.6339259
#> 3  1.3497907  0.5833136 -0.8896760 -1.1250216 -2.0563469 -1.62756179  0.2773097
#> 4 -1.1278645  0.8325716 -0.8116010  0.2244590 -0.8795758  0.85736112  1.8671500
#> 5  0.4129995  0.4366801 -0.1140778  2.1664650  1.1232028 -1.18140476  0.7692849
#> 6  0.9869097 -0.2371834 -1.5759782 -0.9764493 -0.4567902 -0.51583004 -0.1522577
#>          X184       X185       X186       X187       X188        X189
#> 1  0.96817878  0.2553316  0.4598947  0.5233585 -0.2652502  0.69214814
#> 2 -0.78506500  0.4478831 -0.1040749  1.2030426 -0.1539781  0.05787226
#> 3 -0.78207746  1.1083147  0.8642326  0.5504105  0.9647437 -0.22424441
#> 4  1.55986490 -0.3040008  0.6684721  2.0777821 -0.7444708  2.82070662
#> 5  0.04551867  0.5048402 -1.8548192 -0.1084203  0.5984813 -0.24231832
#> 6  0.72300700 -0.4071837 -0.8827858 -1.0226582 -1.2080867  0.04485525
#>         X190       X191        X192       X193       X194        X195
#> 1  0.4777947 -0.1654527  0.05243336 -0.7370134 -1.8798878  0.48021145
#> 2  0.5654217 -1.0177150 -1.40568575  0.2593722 -0.1367416 -1.17122677
#> 3  0.0711770  1.1044724 -0.09165511  0.4796800  0.2645815  0.32757305
#> 4  0.7513436 -0.7086087  0.55136507  0.8604388  0.8489800  3.16292764
#> 5 -0.2833712  1.3392494  0.04871784 -0.4637578 -0.4994733 -1.12228075
#> 6 -1.1855265 -0.5999479 -0.14745685 -1.2556513 -1.6982345  0.05572583
#>         X196        X197       X198       X199       X200        X201
#> 1  1.1945075  1.01718052  2.0922656 -0.3861936  0.5899896 -0.06542833
#> 2 -1.5852298 -1.66791202 -1.4724811  0.8869030  0.3948993 -0.99809663
#> 3  1.0007362  2.01151893  0.9417381  1.7663031  0.4156684 -0.84320549
#> 4  0.5546283  0.04126106  1.9564840  0.6582992  1.5311616  1.14579706
#> 5 -0.7143901 -2.13314583  1.5380785  0.9610552 -1.1423079 -2.34598542
#> 6 -1.3499508 -1.34640504 -0.2063524  0.9519872  0.2676631  0.34254275
#>         X202       X203       X204       X205      X206        X207        X208
#> 1 -2.3012605 -1.0241017 -0.2490300  0.6578273 1.0185198 -0.14670631 -0.05026207
#> 2  0.2245940 -1.6625275 -0.4802622 -0.4668896 0.7784017  0.46250662  0.10007279
#> 3 -0.7902559 -0.4324670  1.1741865  0.3555161 0.4642072  0.56713072 -0.13386144
#> 4  2.2381450 -0.2482026 -0.1740271 -0.7595278 0.3217822 -0.54394372  0.45314229
#> 5 -1.3312787  0.4032464 -2.3342542  0.7585454 2.5518194 -0.11434123  1.08172748
#> 6  0.3427965  1.3392343  1.1984754  1.4700594 0.3691843  0.01638634 -0.33805401
#>          X209       X210       X211       X212       X213        X214
#> 1 -0.06810901 -2.3899821 -2.1781673 -0.7978544  1.6206035 -0.55813941
#> 2  0.19897843 -0.4867686 -1.4285730 -1.2294596  0.1927870 -0.76792177
#> 3  1.37502426 -0.1309839  1.6485164  0.5598114 -0.4445532 -0.52411829
#> 4  0.88401561  0.3649463 -1.7932844 -1.0753600 -1.2281903 -0.03698137
#> 5 -1.42479904  0.9349967  0.4995404 -0.1543204 -0.9520953  0.87627128
#> 6 -0.48971541  0.9687597 -1.2825295  0.1365010 -1.1151069 -0.57019527
#>         X215       X216       X217        X218       X219       X220       X221
#> 1 -1.1021205 -1.0188920  0.4967689  0.08503305  1.5219621  0.4367171 -0.5605241
#> 2  0.7486865  0.8228408  0.9457875  0.91941279 -0.3252947 -1.3364989  0.8020185
#> 3  1.2209152  0.7576967  0.1142242 -0.71798263 -0.8930215 -0.8032028 -0.4299041
#> 4 -1.2033492 -0.5991553  0.6430813 -1.19753314  1.6157286 -0.3196535 -0.3208083
#> 5  1.3484715  1.1867174  0.5508580  1.26051521  0.8946683  1.1025741 -0.1290918
#> 6 -0.1129112 -0.8116194 -1.4229120 -0.88417899  0.4797792 -0.6969135  0.7463554
#>         X222       X223       X224      X225        X226       X227       X228
#> 1 -0.8942416  0.8112439  1.0014639 0.3727325 -0.92710795  0.0834208 -0.9877032
#> 2  0.5507678  1.6178376  0.2056790 0.2267574 -1.47424020  0.8146455  1.4472143
#> 3 -0.8317682 -0.3207402 -0.2260128 0.3518497  1.79297984  2.0505482  1.4283475
#> 4  0.9515958 -0.1397022 -0.3351098 0.1750065  0.16228747 -1.2318460 -1.8329575
#> 5 -0.5397342 -0.6962639  1.6834472 1.8179094  1.00254433  2.0973422  1.2050696
#> 6  2.9931985  1.6090524  0.8398178 0.5095449 -0.01320195  0.9050559 -0.7034152
#>         X229       X230        X231       X232       X233        X234
#> 1 -0.2324506 -0.1995957  0.47832533  0.2410977  1.5007377 -0.11355146
#> 2  0.9364704  0.7920709  0.93683620  1.1666145  0.8016817  0.90200445
#> 3  1.5577524 -0.7035383  0.96044498  0.3930014 -0.3103687 -0.55988492
#> 4 -2.0221351 -1.7056323  0.02750982 -0.1762688  1.5993756 -0.03345915
#> 5  1.2152892 -0.2549729  0.01980215  0.4964283  0.1743443  0.69969168
#> 6 -0.3091516 -0.3642332 -0.11247161 -0.3132277  0.3332395  0.99868757
#>          X235        X236       X237       X238       X239        X240
#> 1 -0.34485179  1.00236086  0.6846479 -0.3529947  2.1294090 -0.01663769
#> 2  0.37703331 -0.04570056  0.8798812  0.2248849  0.7874205  0.53367233
#> 3 -0.02657617  1.13565987 -1.4075515 -1.0161956 -1.3027838 -1.79248928
#> 4 -0.24485865  0.02310635  1.1150549 -1.3830530  0.5501386  0.42819648
#> 5 -0.30880147  0.13156337  1.6433521 -2.0969477 -0.2639016  0.80435584
#> 6  0.49774400  0.12413323  0.8521553  0.1773522 -0.8503353 -0.19158577
#>          X241        X242        X243        X244       X245       X246
#> 1  0.05951393 -0.99223267 -1.87207921 -1.05609288  0.6046173  2.1155422
#> 2 -0.77695957 -0.00802499  0.06009426  1.15286115  1.1192359  0.8031830
#> 3 -0.18161219 -0.71968427  0.10746227  0.53394088  1.2519993  0.6131768
#> 4  0.44492266  0.95441848  0.52798531 -1.02151092  0.6579512  0.4466746
#> 5 -0.82661091 -1.32884655 -0.36904637 -0.03828021 -2.4836215 -0.4916570
#> 6 -1.30944635 -1.93897182  0.67844975  2.46529054  0.3430260 -1.0678392
#>          X247       X248        X249         X250        X251       X252
#> 1 -0.39915285 -1.7814755  1.90971073  0.342700119  0.58509578 -1.0038127
#> 2 -0.17954498 -0.3396155 -1.44645149  0.001146494  0.02861973 -1.0485175
#> 3  0.91909774  1.6471075  0.78366564  0.504145839  2.43662627  1.3408120
#> 4  0.18782515  0.8916280  0.20210491 -1.414355091 -0.34144663 -2.3456058
#> 5  0.06723691  0.3855918  0.81566276 -0.356608711 -0.46803562 -0.5246081
#> 6  1.23751473 -0.5388054  0.03089295  0.029683116 -0.59231930 -0.7394053
#>          X253        X254        X255       X256       X257        X258
#> 1  0.20140644 -0.79132942 -0.51273016 -0.3480082 -0.7798031 -0.18029381
#> 2  0.41396006  1.05691605  1.48673049 -1.3183256 -0.9217860 -1.59546069
#> 3  1.89370426  0.01829565 -1.04164346  0.2036680  1.3659038  1.65925599
#> 4  0.09045135  0.64186375 -1.87852426 -0.9483667 -0.9500955 -0.02265148
#> 5 -0.16448819  2.04789947  0.66859826  0.8657906  0.7634150  0.13041929
#> 6 -0.20662264 -0.78010613 -0.02095803  1.8628226  0.7076891  0.97080282
#>          X259       X260       X261       X262       X263       X264       X265
#> 1 -0.54434586 -0.2277832 -0.7614339 -1.0931232 -1.7830672 -0.3780731 -0.9572555
#> 2  0.04491098 -0.1553736 -0.9588380  1.5172845  1.8667150  0.8387648 -1.0016470
#> 3  0.57875253  0.9261393  1.0659976 -0.2929955 -0.5887045 -3.1647813  0.8877454
#> 4 -0.62662679  0.1906956  0.1530495 -1.5198868  0.7992136  0.3795584 -0.4048797
#> 5  0.61254912 -0.7182011 -0.9256970 -0.4815147  0.7422171  0.2349171 -1.1424110
#> 6  1.01876888 -1.6733025 -1.1736727  0.3437975  1.9739197  0.5935338  0.1005529
#>         X266       X267       X268       X269       X270       X271       X272
#> 1 -0.5040849  0.1903658 -0.1381013  0.4318116  1.1577052 -0.7395716 -1.0771791
#> 2  0.2327552  0.0680296 -0.6818480 -0.3726826 -0.7881988  0.7918241  1.0815561
#> 3  0.5979739  0.5245709 -0.9453411 -0.3627740 -0.2067324 -1.0066062  1.9079960
#> 4 -0.2055255  0.9215021  0.5437024 -0.4164546 -0.6370993 -1.4640848  0.9043364
#> 5 -0.9364374 -0.6695439 -1.1929899 -0.8959081 -1.0744267 -0.6224456 -0.6488894
#> 6 -0.6608408 -0.5534190 -0.5602253  0.7647800  0.8024386 -0.9704759  0.2165677
#>         X273       X274        X275         X276       X277       X278
#> 1 -1.3173045 -0.9272605 -0.09731150  0.146076942 -0.5752089 -1.2571728
#> 2  0.7820460  0.3113476  1.28507516  0.500125000 -0.2770892 -0.1679347
#> 3  0.7656183 -0.3760519  0.22860541  0.001500835 -0.8309842 -0.8536513
#> 4  0.9661058 -1.3404404  0.02238318  0.124002822  0.1124016 -2.1899703
#> 5 -1.1611073 -1.1983918  0.55108009 -0.606748944  0.8402613  0.4957195
#> 6 -0.5854943  1.0700723 -0.49102295  0.232802859  1.5770673  0.1449754
#>          X279       X280        X281       X282       X283       X284
#> 1 -0.51036245  1.0753427 -0.07757487 -0.6085538  0.2034364 -1.6001185
#> 2 -1.36338062  0.8708637  1.38618231  1.6852528  1.6697627  1.3537868
#> 3 -1.18834766 -0.1155560 -0.78313059  1.9755782  0.2807869 -0.1912048
#> 4  0.08224831  0.6423262 -0.87453467  2.0224959  0.5599049 -1.1846629
#> 5  0.29849393 -0.5851986  0.21154684  0.3389410 -0.8545625  0.2602671
#> 6  0.46392084 -1.8312786 -1.02103854 -0.2437121  0.6839644  1.5185310
#>         X285       X286       X287        X288       X289       X290
#> 1  0.3161837 -1.6797103 -1.0121822  0.29236738  0.8856511  0.3178216
#> 2 -0.2001975  0.9914034  0.4060765 -0.01253678 -0.1380407  1.3345514
#> 3  0.8682214  0.8653578 -0.2340567  0.06200271  0.9157733  0.5930171
#> 4 -1.2336116 -1.5360320 -0.8385940 -0.46715978 -0.3699772  0.4216411
#> 5  1.1935586  1.2167424 -0.5049264 -1.45146902 -0.9301796  0.5037670
#> 6 -1.0598412  1.1700011  0.8266508  0.73628428 -1.2723818 -1.4817043
#>          X291       X292       X293        X294         X295       X296
#> 1  0.04362083  0.1831029  0.0802087  0.04432861  0.360901368  0.6247432
#> 2 -0.94782433 -1.9904508 -0.5993576 -0.80708932 -0.746998922 -1.2306937
#> 3 -0.61928281 -1.0304856  0.2217164  0.39149438  0.757842461  1.3437035
#> 4 -1.51529064 -0.3511452  0.2632775 -1.42626884 -0.403101436 -0.5825843
#> 5 -0.47859561 -1.2106990  0.1507438 -0.90920967  1.353303071  0.9231823
#> 6  0.64197849  0.3738942  0.9012810 -0.26203938 -0.005097034  0.9651165
#>         X297       X298        X299      X300       X301        X302       X303
#> 1  1.1881810  0.5148322  1.08625901 1.0772051 -0.1427242  0.49232242 -0.5936932
#> 2  0.3826669 -0.9616128 -1.66459953 0.0428736  0.1562007  0.14472293  0.4338370
#> 3  1.0675708 -0.2268723 -0.01932557 0.5968146 -0.1073146  0.08398384  2.3498550
#> 4 -2.2715070 -1.2057702  0.14750525 0.8975233 -1.2849250 -0.52515825 -0.4046299
#> 5 -1.1532978  1.5778301  2.34471465 1.1583981  0.4229306 -0.44144545 -1.4735024
#> 6  0.5598455 -0.1365392  0.95132562 1.3775289  0.7687297  1.59495798  0.4965797
#>         X304       X305       X306       X307        X308       X309       X310
#> 1  0.4295898  1.3063853  0.1348970 -0.5580981  2.08399847  0.7682799  0.1840406
#> 2  0.6559704 -0.5010253  0.2554280 -0.8026258  0.02053718  1.1143489 -0.6718423
#> 3  1.5461533  1.6053318  0.5867738 -1.1268310  1.65619618  0.9756360 -0.4285903
#> 4 -0.1493853 -0.4692984 -0.3614860  0.9390311 -0.07619321  0.1959169  0.6666260
#> 5 -2.0203132 -0.7497743  0.4522944  0.6391389  2.18018569  0.5972637 -0.5822820
#> 6 -0.1048863 -1.3483226 -2.4779222 -0.9929741  0.08012798 -1.4920996  0.3563346
#>          X311        X312        X313        X314        X315       X316
#> 1  0.32171918  0.03095059 -0.48693042 -0.04674427  0.52812784 0.02424184
#> 2 -1.80564436 -0.01940267 -0.92147532 -0.81927625 -0.50195630 1.70515261
#> 3  1.06435413  0.38365331 -0.90299338 -0.68792899  2.62680745 2.11828461
#> 4 -1.43565384 -2.45688493 -0.07925139 -0.16695449  0.52929341 0.50918224
#> 5 -1.23748209  2.89895288  0.88255062 -0.60312956  0.34534071 0.37046781
#> 6  0.04517111 -0.09395145 -1.09960738  0.45904168 -0.09192851 0.77723685
#>         X317        X318       X319       X320       X321       X322       X323
#> 1 -0.6370863  0.45544240  0.1358334  1.1411333  1.2012763  0.2698755 -0.6405473
#> 2  0.4156088 -0.38597228 -0.3822251  1.8223860  1.8746393  0.7140081  0.4534456
#> 3  2.0138764  1.81028773 -0.8932085  0.4165519  1.5445864 -0.2497590 -1.1957433
#> 4 -1.7326599  1.35354393  0.4224247 -0.1345444  1.0276972  1.1816390  0.7988554
#> 5  1.5704539  0.04912458  0.5277579  0.9872010 -0.5761727 -1.2008491 -0.2773522
#> 6  1.9843877  2.07366094  0.5608131 -0.8098960 -1.4816670 -1.6535724 -1.7383932
#>          X324       X325       X326       X327         X328       X329
#> 1 -0.54466725 -1.8412044 -1.3486787 -1.0761329  0.223503138  0.4904924
#> 2 -0.76452990 -0.3916670 -1.5927704 -0.3594265 -0.311770915  1.1115105
#> 3 -0.74790681  2.5431093  1.1018413  0.2520711  1.373970216  0.5240499
#> 4 -0.09213893  0.6277757  0.6439898 -0.1195488 -0.005387775 -0.2493966
#> 5 -0.27646044 -1.5030589  0.1764869  1.6703402 -0.099867373 -0.5802928
#> 6 -1.28588299 -1.8472447 -0.3587010 -0.2351798  1.224145127  0.2598958
#>          X330        X331       X332       X333       X334       X335
#> 1 -0.55617027 -0.01767448  1.3579014 -0.2670798 -0.5034452  1.7903445
#> 2  1.21640647 -0.48602845  0.1618378  0.4458523 -1.5441072  1.8551190
#> 3  1.47205671 -1.84931869 -0.2452233  0.5888564 -1.5081567 -1.0305873
#> 4  0.42715911 -0.25803621 -0.6339555 -1.4484755  0.7943067  1.5249074
#> 5 -0.49299444 -0.14580863 -1.0007726 -2.1986324 -2.6125048 -2.4391248
#> 6  0.02886297 -1.48053120 -1.3199850  0.7462396 -1.2174605 -0.1718881
#>         X336        X337       X338       X339        X340       X341
#> 1 -0.2434590  0.07215424  0.3997464 -0.1733166 -0.73295271 -0.3664276
#> 2  0.1434811 -2.20470142 -1.2014748 -1.0766158 -0.08653885  0.7060292
#> 3  0.3190561  0.64815344 -0.5865272 -1.2474641 -0.34111311  0.6055087
#> 4  0.2070692 -0.05670554  0.4126934 -1.5474979 -0.65477889 -1.7752603
#> 5 -2.9133838 -1.75056362 -0.7635323  0.3044199  1.33587375 -1.0083123
#> 6  0.6754543  0.45703092 -1.7868242 -0.2771972 -0.84404352 -1.6671020
#>         X342        X343       X344       X345        X346        X347
#> 1 -0.9397993  0.34567417  0.3517766  1.3031575  0.47145498  0.68155858
#> 2  0.3902390  0.07677674 -1.2067626 -1.4975715 -1.00704062  0.04787833
#> 3 -0.1462516  0.56265815  0.1772857 -0.7083038  0.14159373  0.41767667
#> 4 -1.0688858 -0.62314936  0.1458572  0.3029542  0.51488468  0.87226414
#> 5 -1.2257626  0.59948685  0.4761481  0.6918834  0.07009015 -0.24883225
#> 6 -0.4975222 -0.73106764  0.7139655  1.0317050 -1.06529835  0.79252659
#>         X348       X349        X350       X351        X352         X353
#> 1  1.4208783 -1.2560402 -1.52525283 -0.5049922 -1.27759298  0.943561137
#> 2 -1.1219432 -1.5899249 -0.64705646  0.5971767  1.74761670  1.616924318
#> 3 -0.7807234  0.3468515  0.02274581 -1.5504162 -0.12036292  1.533440327
#> 4  0.1533167  0.3723836 -0.70296891 -0.5258388 -0.08342956 -0.687518936
#> 5 -0.7094721 -0.3807166 -0.06374206 -0.3897530  0.11103396 -0.924032292
#> 6  1.0766600  0.3720963  0.72336321 -0.7387667 -0.51149042 -0.001188722
#>          X354       X355       X356       X357       X358       X359
#> 1 -0.08815217  0.6065650  1.3860657  0.8266370  0.9301774  0.8561470
#> 2 -0.17293500 -0.2175302  0.7776208  1.1167211 -0.2200449 -2.1272365
#> 3  0.14248031 -0.7785216 -1.4401920 -0.7158149 -0.1199620 -0.4288614
#> 4 -1.07489256 -1.0222717  1.1153507  0.7136017  0.1949292  0.8479488
#> 5  1.36740733  0.3441632  0.7114565 -1.9116345  1.6824433  0.8687227
#> 6  2.32727596  0.3341854 -0.3122445 -0.1030077  0.5104585  0.4874629
#>          X360       X361       X362       X363        X364       X365
#> 1  0.05834588  1.5509251 -0.7560967  0.4673550  2.61412825  0.2038575
#> 2 -0.44259521 -0.7814190  1.0103149 -0.6006311  0.44710951  0.8650985
#> 3 -0.57187893 -1.0773597 -0.5299226  0.3989243  1.83618207 -1.1293618
#> 4  0.36522901  0.3297251  0.1546519 -0.4909950  2.18192625  1.3318328
#> 5 -1.32399678 -1.8959987 -0.7140764  0.8076604 -0.14990996 -1.1875076
#> 6  0.31174235  0.9591861  0.9013398 -0.6362619 -0.02912429  0.2507958
#>         X366         X367       X368       X369        X370       X371
#> 1 -0.1036455 -0.247563425  1.5523937  0.2250695 -0.41291240 -0.9613940
#> 2  0.4296791 -1.709388177 -2.0134423 -0.7830313 -0.01309843 -0.7441009
#> 3  0.2394745 -1.084338174  0.3681101  1.9647549 -0.74841280 -0.6496713
#> 4  0.3317087 -0.048818034 -0.2416077 -0.1716671 -0.02199478 -0.5194629
#> 5 -1.3961675  1.387390518  0.7960944  0.4015540 -0.33673986  0.1676573
#> 6  0.1865441 -0.006536482  1.4188373 -0.1176737  1.58086533  1.4746260
#>         X372       X373       X374       X375        X376         X377
#> 1 -0.9009678  0.9089285  0.1461180 -0.5125975 -0.08526663  0.806404567
#> 2  0.3081507  0.1538657  2.0008053  0.3911408 -0.31064633  0.004059439
#> 3 -0.4594469 -1.7898064 -0.8583423 -1.1598936  0.49447832  0.153640203
#> 4 -1.3036901  0.6694361 -0.1678802  1.2428244  0.89258168 -0.721010472
#> 5 -0.4869534  0.3468513 -0.1707914  1.7119799  0.19809284 -0.879019550
#> 6 -1.5020207 -2.8643495 -0.6661209 -0.9726166  0.24437158  1.522051905
#>          X378       X379        X380       X381        X382        X383
#> 1  0.70540049 -0.3100864  0.29564108  0.3292704  1.05397563  0.21198693
#> 2 -0.08128063  0.6486651  0.26531278 -0.6040443  1.20072242 -0.16160209
#> 3 -1.78118189 -1.7155364 -0.49659242 -1.3935232 -0.05806353  1.93802355
#> 4 -0.59325331 -0.4322154  0.01709799  0.1490195 -1.36067340 -1.80375178
#> 5 -1.92845051 -1.3892994 -1.15409363  1.0365522 -0.39309592 -0.03981149
#> 6  0.76389945  0.9877221 -1.34316840 -0.5038039 -0.44067212 -0.60267390
#>         X384        X385       X386       X387        X388       X389
#> 1  0.2885807 2.292592306  2.0454160  1.1477212 -0.89201833 -0.4254459
#> 2 -0.4874252 0.102266114  0.5100438 -0.1747499  0.07452333 -1.0028437
#> 3  1.7021870 0.689870856  1.2918612  1.5071716 -0.15112434  1.2785683
#> 4 -0.5034332 0.007714935 -1.1647040  0.4603397  1.04989021  0.9873543
#> 5 -0.4365634 0.512966968 -0.7720252  0.8501499  0.08831664 -1.2222154
#> 6 -0.4607608 0.613746711  0.8856864 -1.6318715 -0.61828246 -0.7743106
#>         X390        X391       X392       X393       X394       X395       X396
#> 1 -0.2267062  0.83568153 -0.1408593 -0.3144858 -1.4028472 -0.3953604 -0.3211407
#> 2  0.3306439  0.03036692  0.5514044  1.1544644  2.0937910 -1.3941880  1.0335702
#> 3 -0.9587969 -0.51066898 -0.8795202  0.3257161  1.5395577  0.8803889  0.4661187
#> 4  1.7523782  0.58207110  0.2505027 -0.4594638 -0.8637215  1.2543271  0.9602434
#> 5 -0.2062278 -0.54339265  1.7518054  0.3187762  0.1362727 -0.7810039 -3.0967336
#> 6 -0.7039318 -0.88663587 -1.9357585 -0.6323879 -2.3878376 -0.1814913  1.6032802
#>          X397        X398       X399        X400       X401        X402
#> 1  0.96088207 -0.17210768 -1.9672602 -2.03105694  0.3495653 -0.40181526
#> 2  1.62767719  0.46144872 -0.4689856  0.53173994 -0.3527464 -1.07926605
#> 3 -0.03930607  0.08528437  1.2943143  0.57528299  0.4464424  1.86092908
#> 4  0.37588850 -0.53302001  0.3413636  1.21290402 -0.9238451 -0.50632376
#> 5 -0.98873896 -0.64895948  0.9360095 -0.06988719 -1.9560215  0.08676929
#> 6 -0.56493911  0.02323700  0.8587113 -1.78971553  0.6646349 -0.02818202
#>          X403        X404       X405       X406       X407       X408
#> 1  0.77992827  0.61067934 -0.1299045  1.0156999  1.8465145  0.3815288
#> 2 -0.03247955  1.86627491  0.5883028 -0.7617310  0.3627270  1.1942539
#> 3 -0.41387966  0.80146294 -0.3933623  1.0638502 -1.0029464 -0.8640938
#> 4  0.14556001 -0.36755009 -0.1697010 -0.7283094  0.2662991  0.8787287
#> 5  1.48148829 -0.12141692 -1.6126121  1.4445373  0.6947786 -0.3110622
#> 6  0.15460756  0.03643573  0.2896182 -0.1620117 -0.3680696 -0.1631415
#>         X409        X410          X411        X412       X413        X414
#> 1  0.5725803  0.02325933  1.926033e-02 -0.25782013  0.3952034 -0.09239163
#> 2  0.7569754 -1.90313297  1.062903e+00  1.65135804  1.3526148  0.60244091
#> 3 -1.9046198 -0.37666282 -1.741723e+00 -0.07461008 -0.9704680 -1.68498183
#> 4  0.6776671  1.19184183  8.869709e-01  1.98746346  0.9225153  0.89174058
#> 5 -1.3752700  0.56878446 -4.623649e-01  1.12192407  2.1070779 -0.97876755
#> 6 -1.2363706  0.60656059  7.344809e-05  0.54197412 -0.2382802 -0.20392496
#>          X415       X416       X417       X418       X419      X420       X421
#> 1 -1.30761208 -0.5929457  2.5533116  1.3375571 -0.1045676 1.2667767 -1.0304130
#> 2  0.05974125  1.2944083  2.1102564  1.2820842  1.0631376 2.1521840  0.6100080
#> 3  1.22230093  1.0529040  0.4181118  1.0972416 -1.3884183 0.9846551  0.5238640
#> 4  0.18240335 -0.7988480  0.0701177  0.2828281  1.0250817 0.8240508 -0.4899637
#> 5 -0.63758065 -1.9744833 -0.7901694 -1.2632582 -0.8011683 0.2050357 -2.1323520
#> 6 -1.91432396 -1.6242165 -0.1640008 -1.0430321  0.3724886 0.3434573  0.8318407
#>         X422        X423       X424       X425       X426        X427
#> 1 -1.2724093  0.84169886 -0.5136727  1.7259246  0.9294782  0.08667645
#> 2  1.1470496  1.46453715  1.3998575 -0.1349692  0.1208178 -0.73385844
#> 3 -0.4334854 -0.48867975 -0.6123004  0.7768034  2.1095964  0.40282101
#> 4 -0.2682653  0.01042247  1.1122121 -1.5671498  0.8383686 -0.90978122
#> 5 -1.3129503 -0.90954427 -1.8889801 -0.8938426 -0.1157487 -0.20852929
#> 6  1.0687011 -0.13612038 -0.8225158  0.7361925  1.0933677 -1.02132528
#>           X428       X429        X430         X431        X432        X433
#> 1 -0.882606103 -0.4293345  0.26238762  0.045448212  1.18018831  0.65598902
#> 2 -0.122785132 -0.4096617 -0.08227436 -0.831738372 -0.79911688  0.42656530
#> 3  0.527082327 -0.4941084  0.12522577  0.089998105 -0.01535173  0.87597944
#> 4  0.000882219 -0.1373805 -0.37990134 -0.002783095 -0.06518607 -0.08241044
#> 5 -0.293396802  1.8942274  0.51934386  0.269396733 -2.23423863  0.59841397
#> 6  0.455570341 -0.1611116 -1.04532874 -0.548239268 -1.21982869  0.75430511
#>         X434       X435       X436        X437         X438       X439
#> 1  1.2505318  1.1273035 -0.5014827 -0.02446968 -1.244204570  0.7903417
#> 2  0.8347726  0.3789016  0.3670436 -0.39825860 -0.564057304  0.5523776
#> 3 -0.9527188  1.0779853  1.0444368  0.17531342 -0.006130196 -0.1358752
#> 4  1.4183914  2.0335560 -0.2386358 -1.80659777 -0.189813687 -0.2771387
#> 5 -0.5985921  0.4444116 -1.4379832 -1.45131562 -1.750302667 -1.2605993
#> 6 -0.1027559 -0.1085342  0.1212120 -0.55232473  0.109725544  0.9587046
#>         X440        X441       X442        X443       X444        X445
#> 1 -0.8603592 -1.07268023 -0.4959748 -0.62196114 -0.2440178  0.44209057
#> 2 -0.9053220  0.48154799 -0.4856543 -0.57826169 -2.4318513 -1.84186583
#> 3 -1.0211562 -0.09016919  1.1151027  0.47061657  0.6501901  0.33867479
#> 4  1.3719227  0.95078256 -0.2379489  0.36358425  2.3573605  0.06076678
#> 5 -0.1521010 -0.07506430  1.5549442 -0.02368316  0.5112293  0.10622430
#> 6  0.5055859 -0.46988762  2.7456206  0.84088104 -0.9692509 -0.79966911
#>         X446        X447        X448       X449        X450       X451
#> 1  0.5649729  0.43325906  0.09251954  1.0377405 -0.17395456  0.2538948
#> 2 -1.3486744 -1.31926097 -0.61117084 -0.4528584 -1.38894964 -0.4851683
#> 3 -0.4524588  0.16956239 -0.81528243  0.8996520 -0.46449773  1.3627719
#> 4  1.0271951  1.17070943  0.07301557 -0.7390901 -0.07200701  0.8128322
#> 5 -0.2199009  2.72451813 -0.03401541  0.4023416 -0.10941846 -1.9048095
#> 6 -2.3652290  0.03533066 -1.51354470 -1.1245504 -0.91228401 -0.1435152
#>          X452       X453        X454       X455       X456       X457
#> 1 -0.04370048 -0.7592709  0.68784034  0.5164657  1.1707672 -0.5997654
#> 2 -0.44707456  2.0642759  0.57845263  1.8810636 -0.1423135 -0.6373360
#> 3  0.98286514  0.2831167  0.42250929 -1.3062123  0.1594623  0.0459510
#> 4  0.24709315 -0.2635882 -0.12775815  0.6946751 -1.2066018 -1.3510965
#> 5  0.43151545  0.5717415 -0.01575958 -0.3235685 -0.1491476 -0.8679129
#> 6  0.21898097 -1.3831036  0.04227582 -0.7898475 -0.7165878 -1.0358398
#>         X458        X459       X460       X461       X462        X463
#> 1  0.3736380 -0.01824175  0.6717653  1.0773189 -0.7788576 -0.01689863
#> 2 -1.4019278  0.04161693 -0.1581025 -0.5197076 -0.2024937  0.58485623
#> 3  0.7035704  2.04947872  0.8384396  0.5970750 -1.6630185 -0.97826979
#> 4 -0.4797384  0.30786719 -1.1131410 -1.2033725 -0.6321478 -0.38272392
#> 5 -1.5108468 -0.40446052 -0.2060151 -0.4047457 -0.9956792 -1.43416498
#> 6 -0.3881769  0.50413990  0.2061426  0.4448223  0.5336419 -0.63357205
#>          X464        X465       X466       X467        X468       X469
#> 1  0.78991234  0.77836807  0.1860482  0.6919415 -1.18059771 -1.9779049
#> 2  0.58168231 -0.33437992  0.1649139  0.4027749  1.42049725  1.7926456
#> 3 -0.07374198 -0.04772207  0.3352442 -0.2851297  0.04496052  0.7179199
#> 4  0.67356825  0.81771897  0.2220470  0.4341290  1.28490214 -0.3885876
#> 5  0.08451088 -1.84169663 -1.0205690 -0.9229441 -1.83760262 -1.1867567
#> 6 -0.47929700  0.03361102  1.0209351  1.2200380  2.05828949 -2.4458513
#>         X470       X471       X472       X473       X474       X475       X476
#> 1 -0.7915542  0.6125274 -0.4013667 -0.2048909  0.4565955 -1.1059532  0.4615426
#> 2  0.6458546  0.8987679 -1.0485490 -0.5224165 -0.1865990  0.4114497 -0.3104777
#> 3  1.3971304  1.4132249  0.3151329  0.9835260  0.6562657 -0.2120604 -0.4048082
#> 4 -0.4081205 -0.4956285 -0.4632395 -1.2438305  0.9081367  1.3977360  0.2929505
#> 5  0.7535145 -0.5312722 -0.3416822  0.7652843 -0.6800976 -1.7727981 -0.1134908
#> 6 -0.5360689 -1.5128770  0.8449661 -0.8347146  0.1880181 -1.3264846 -0.1294257
#>         X477       X478         X479        X480       X481         X482
#> 1  0.5910120 -0.4940498  0.002785365 -0.02666764  0.8270371  0.165791046
#> 2  0.5781499 -1.1197848 -0.950176003 -1.20688811 -1.7532187 -1.305479742
#> 3  0.4810013 -1.6396316 -0.122433649 -0.71051478 -2.6239487 -0.661403211
#> 4  0.9989063 -0.6270967  0.173962736 -1.41809257 -0.6872429  0.263557790
#> 5 -2.0431501 -0.9694412 -0.998524484  1.18930972  0.7901700  1.740895845
#> 6 -0.2301079 -0.1621700  0.944308907  0.76841242 -0.5975304 -0.001662638
#>         X483        X484          X485        X486        X487        X488
#> 1 -0.7870721 -0.66243079 -0.4779052285  0.04384065  0.61567915 -1.03913206
#> 2 -1.8727884 -1.59578760  0.7368222878  0.86058228  1.10552746  0.88887632
#> 3 -1.3771872  0.29158010 -1.1769093301  0.05977455  1.80859325  0.63418368
#> 4 -1.0714646  0.06824475  0.0007612326  1.93153292 -0.01028083 -0.42336419
#> 5  1.7638005 -1.30874896 -0.3343533246 -0.93598073 -0.24468409 -0.06422972
#> 6 -0.1977117 -0.10382886 -0.0473396292  0.65544324  0.69741904  0.86500921
#>          X489        X490        X491       X492        X493        X494
#> 1 -0.43175885  0.41210017 -0.98012178 -0.7909953 -0.42089234 -0.80848933
#> 2 -1.81913265 -0.52374557 -1.46740676 -1.9785274 -0.84415978  1.05917988
#> 3  0.08099351 -0.02695561 -1.44959864 -0.1958963  0.37634200  1.00242982
#> 4 -0.21239898 -1.05606085 -0.37092054 -0.5107478 -0.73234656  0.36755110
#> 5  0.05654165 -0.08530003  0.01673045 -0.5247859  0.03478482  0.01893690
#> 6  0.59030152 -1.30024524 -0.25075594  0.9161684  1.62839516 -0.08056638
#>         X495        X496       X497        X498        X499      X500
#> 1  0.3877065  1.12575031 -0.7987559  1.18303736  1.05295521  1.238527
#> 2 -1.2148116 -0.91257105 -2.1248513 -0.52190398 -2.08475585 -2.005173
#> 3 -1.1123699 -0.01437611  0.7999073  0.90626533 -1.02394657 -1.116762
#> 4  0.6322546  1.04613154  1.8491579  1.23872254  0.05100059 -0.503154
#> 5  0.2400078 -0.85865768 -2.9424184 -0.35175460 -0.43408456  2.173090
#> 6  0.7473840  1.06106205 -1.1043799  0.09476932  0.26859681  1.228495